+++ docs/article/paper.bib
@@ -76,18 +76,6 @@
   doi={10.1109/10.126609},
 }
 
-@article{riener1996biomechanical,
-  title={Biomechanical model of the human knee evaluated by neuromuscular stimulation},
-  author={Riener, Robert and Quintern, Jochen and Schmidt, G{\"u}nther},
-  journal={Journal of biomechanics},
-  volume={29},
-  number={9},
-  pages={1157--1167},
-  year={1996},
-  publisher={Elsevier},
-  doi={10.1016/0021-9290(96)00012-7},
-}
-
 @article{ding2003mathematical,
   title={Mathematical models for fatigue minimization during functional electrical stimulation},
   author={Ding, Jun and Wexler, Anthony S and Binder-Macleod, Stuart A},
@@ -100,63 +88,6 @@
   doi={10.1016/S1050-6411(03)00102-0},
 }
 
-@article{ding2007mathematical,
-  title={Mathematical model that predicts the force--intensity and force--frequency relationships after spinal cord injuries},
-  author={Ding, Jun and Chou, Li-Wei and Kesar, Trisha M and Lee, Samuel CK and Johnston, Therese E and Wexler, Anthony S and Binder-Macleod, Stuart A},
-  journal={Muscle \& Nerve: Official Journal of the American Association of Electrodiagnostic Medicine},
-  volume={36},
-  number={2},
-  pages={214--222},
-  year={2007},
-  publisher={Wiley Online Library},
-    doi={10.1002/mus.20806},
-}
-
-@article{marion2009predicting,
-  title={Predicting the effect of muscle length on fatigue during electrical stimulation},
-  author={Marion, M Susan and Wexler, Anthony S and Hull, Maury L and Binder-Macleod, Stuart A},
-  journal={Muscle \& Nerve: Official Journal of the American Association of Electrodiagnostic Medicine},
-  volume={40},
-  number={4},
-  pages={573--581},
-  year={2009},
-  publisher={Wiley Online Library},
-  doi={10.1002/mus.21459},
-}
-
-@article{marion2013predicting,
-  title={Predicting non-isometric fatigue induced by electrical stimulation pulse trains as a function of pulse duration},
-  author={Marion, M Susan and Wexler, Anthony S and Hull, Maury L},
-  journal={Journal of neuroengineering and rehabilitation},
-  volume={10},
-  pages={1--16},
-  year={2013},
-  publisher={Springer},
-  doi={10.1186/1743-0003-10-13},
-}
-
-@article{hmed2018approach,
-  title={An approach to a muscle force model with force-pulse amplitude relationship of human quadriceps muscles},
-  author={Hmed, Abdennacer Ben and Bakir, Toufik and Garnier, Yoann M and Sakly, Anis and Lepers, Romuald and Binczak, Stephane},
-  journal={Computers in Biology and Medicine},
-  volume={101},
-  pages={218--228},
-  year={2018},
-  publisher={Elsevier},
-  doi={10.1016/j.compbiomed.2018.08.026},
-}
-
-@article{michaud2021biorbd,
-  title={biorbd: A c++, python and matlab library to analyze and simulate the human body biomechanics},
-  author={Michaud, Benjamin and Begon, Micka{\"e}l},
-  journal={Journal of open source software},
-  volume={6},
-  number={57},
-  pages={2562},
-  year={2021},
-  doi={10.21105/joss.02562},
-}
-
 @article{michaud2022bioptim,
   title={Bioptim, a python framework for musculoskeletal optimal control in biomechanics},
[…]
