### 2025-11-20 d366928
Updating to new bioptim version
