+++ cocofest/models/veltink1992/veltink1992_and_riener1998.py
@@ -2,15 +2,7 @@ from typing import Callable
 from casadi import MX, vertcat
 import numpy as np
 
-from bioptim import (
-    ConfigureProblem,
-    DynamicsEvaluation,
-    NonLinearProgram,
-    OptimalControlProgram,
-)
-
 from cocofest.models.veltink1992.veltink1992 import VeltinkModelPulseIntensity
-from cocofest.models.state_configure import StateConfigure
 
 
 class VeltinkRienerModelPulseIntensityWithFatigue(VeltinkModelPulseIntensity):
@@ -118,100 +110,34 @@ class VeltinkRienerModelPulseIntensityWithFatigue(VeltinkModelPulseIntensity):
 
     def system_dynamics(
         self,
-        a: MX,
-        mu: MX,
-        I: MX,
+        time: MX,
+        states: MX,
+        controls: MX,
+        numerical_timeseries: MX,
     ) -> MX:
         """
         The system dynamics including fatigue effects.
 
         Parameters
         ----------
-        a: MX
-            Muscle activation state (unitless)
-        mu: MX
-            Fatigue state (unitless)
-        I: MX
-            Stimulation current amplitude (mA)
+        time: MX
+            The system's current node time
+        states: MX
+            The state of the system a, mu
+        controls: MX
+            The controls of the system, I
+        numerical_timeseries: MX
+            The numerical timeseries of the system
 
         Returns
         -------
         The value of the derivative of each state
         """
+        a = states[0]
+        mu = states[1]
+        I = controls[0]
         u = self.normalize_current(I)
         a_dot = self.get_muscle_activation(a=a, u=u)
         mu_dot = self.get_mu_dot(a=a, mu=mu)
 
         return vertcat(a_dot, mu_dot)
-
-    @staticmethod
-    def dynamics(
-        time: MX,
-        states: MX,
-        controls: MX,
-        parameters: MX,
-        algebraic_states: MX,
-        numerical_timeseries: MX,
-        nlp: NonLinearProgram,
-        fes_model=None,
-    ) -> DynamicsEvaluation:
-        """
-        Parameters
-        ----------
-        time: MX
-            The system's current node time
-        states: MX
-            The state of the system (muscle activation, fatigue)
-        controls: MX
-            The controls of the system (stimulation current)
-        parameters: MX
-            The parameters acting on the system
-        algebraic_states: MX
-            The stochastic variables of the system
-        numerical_timeseries: MX
-            The numerical timeseries of the system
-        nlp: NonLinearProgram
-            A reference to the phase
-        fes_model: VeltinkRienerModelPulseIntensityWithFatigue
-            The current phase fes model
-
-        Returns
-        -------
-        The derivative of the states
-        """
-        model = fes_model if fes_model else nlp.model
-        dxdt_fun = model.system_dynamics
-
-        return DynamicsEvaluation(
-            dxdt=dxdt_fun(
-                a=states[0],
-                mu=states[1],
-                I=controls[0],
-            ),
-            defects=None,
-        )
-
-    def declare_model_variables(
-        self,
-        ocp: OptimalControlProgram,
-        nlp: NonLinearProgram,
-        numerical_data_timeseries: dict[str, np.ndarray] = None,
-        contact_type: tuple = (),
-    ):
-        """
-        Tell the program which variables are states and controls.
-
-        Parameters
-        ----------
-        ocp: OptimalControlProgram
-            A reference to the ocp
-        nlp: NonLinearProgram
-            A reference to the phase
-        numerical_data_timeseries: dict[str, np.ndarray]
-            A list of values to pass to the dynamics at each node
-        contact_type: tuple
[…]
