+++ cocofest/models/dynamical_model.py
@@ -4,15 +4,17 @@ import numpy as np
 from casadi import vertcat, MX, SX, Function
 from bioptim import (
     BiorbdModel,
-    ExternalForceSetTimeSeries,
-    OptimalControlProgram,
-    NonLinearProgram,
-    ConfigureProblem,
-    DynamicsFunctions,
+    ConfigureVariables,
     DynamicsEvaluation,
-    ParameterList,
-    ContactType,
+    DynamicsFunctions,
+    ExternalForceSetTimeSeries,
+    FcnEnum,
+    NonLinearProgram,
     OdeSolver,
+    OptimalControlProgram,
+    ParameterList,
+    StateDynamics,
+    States
 )
 
 from ..models.fes_model import FesModel
@@ -26,8 +28,7 @@ from .hill_coefficients import (
     muscle_passive_force_coefficient,
 )
 
-
-class FesMskModel(BiorbdModel):
+class FesMskModel(StateDynamics, BiorbdModel):
     def __init__(
         self,
         name: str = None,
@@ -41,6 +42,7 @@ class FesMskModel(BiorbdModel):
         activate_residual_torque: bool = False,
         parameters: ParameterList = None,
         external_force_set: ExternalForceSetTimeSeries = None,
+        with_contact=False,
     ):
         """
         The custom model that will be used in the optimal control program for the FES-MSK models
@@ -64,8 +66,8 @@ class FesMskModel(BiorbdModel):
         parameters: ParameterList
             The parameters that will be used in the model
         """
-        super().__init__(biorbd_path, parameters=parameters, external_force_set=external_force_set)
-        self.bio_model = BiorbdModel(biorbd_path, parameters=parameters, external_force_set=external_force_set)
+        super().__init__(bio_model=biorbd_path, parameters=parameters, external_force_set=external_force_set)
+
         self._name = name
         self.biorbd_path = biorbd_path
 
@@ -84,14 +86,19 @@ class FesMskModel(BiorbdModel):
             )
             self.muscles_dynamics_model[i].all_stim = self.muscles_dynamics_model[i].previous_stim["time"] + stim_time
 
-        self.bio_stim_model = [self.bio_model] + self.muscles_dynamics_model
-
+        self.bio_stim_model = [self] + self.muscles_dynamics_model
         self.activate_force_length_relationship = activate_force_length_relationship
         self.activate_force_velocity_relationship = activate_force_velocity_relationship
         self.activate_passive_force_relationship = activate_passive_force_relationship
         self.activate_residual_torque = activate_residual_torque
         self.parameters_list = parameters
         self.external_forces_set = external_force_set
+        self.muscles_model = muscles_model
+        self.with_contact = with_contact
+
+        self.state_configuration = [CustomStates.my_muscles_states, States.Q, States.QDOT]
+        self.control_configuration = [CustomStates.my_controls]
+
 
     # ---- Absolutely needed methods ---- #
     def serialize(self) -> tuple[Callable, dict]:
@@ -99,7 +106,7 @@ class FesMskModel(BiorbdModel):
             FesMskModel,
             {
                 "name": self._name,
-                "biorbd_path": self.bio_model.path,
+                "biorbd_path": self.path,
                 "muscles_model": self.muscles_dynamics_model,
                 "stim_time": self.muscles_dynamics_model[0].stim_time,
                 "previous_stim": self.muscles_dynamics_model[0].previous_stim,
@@ -113,10 +120,6 @@ class FesMskModel(BiorbdModel):
         )
 
     # ---- Needed for the example ---- #
-    @property
-    def name_dof(self) -> tuple[str]:
-        return self.bio_model.name_dof
-
     def muscle_name_dof(self, index: int = 0) -> list[str]:
[…]
