+++ examples/getting_started/optimization/pulse_width_optimization.py
@@ -13,9 +13,8 @@ def prepare_ocp(model, final_time, pw_max):
     # --- Set dynamics --- #
     n_shooting = model.get_n_shooting(final_time=final_time)
     numerical_data_time_series, stim_idx_at_node_list = model.get_numerical_data_time_series(n_shooting, final_time)
-    dynamics = OcpFes.declare_dynamics(
-        model, numerical_data_time_series, ode_solver=OdeSolver.RK4(n_integration_steps=10)
-    )
+    dynamics_options = OcpFes.declare_dynamics_options(numerical_time_series=numerical_data_time_series,
+                                                       ode_solver=OdeSolver.RK4(n_integration_steps=10))
 
     # --- Set initial guesses and bounds for states and controls --- #
     x_bounds = OcpFes.set_x_bounds(model)
@@ -32,7 +31,7 @@ def prepare_ocp(model, final_time, pw_max):
 
     return OptimalControlProgram(
         bio_model=[model],
-        dynamics=dynamics,
+        dynamics=dynamics_options,
         n_shooting=n_shooting,
         phase_time=final_time,
         objective_functions=objective_functions,
