### 2025-11-21 9326889
nmpc, collocation and contact point update
