+++ tests/shard2/test_run_examples.py
@@ -74,15 +74,15 @@ MULTIBODY_CYCLING_EXAMPLE_MODULES = [
 cycling_biorbd_model_path = biomodel_folder + "/Wu/Modified_Wu_Shoulder_Model_Cycling.bioMod"
 cycling_initial_guess_biorbd_model_path = biomodel_folder + "/Wu/Modified_Wu_Shoulder_Model_Cycling_for_IK.bioMod"
 
-# TODO: Update it
-# @pytest.mark.parametrize("module_name", MULTIBODY_CYCLING_EXAMPLE_MODULES)
-# def test_cycling_multibody_examples(module_name):
-#     ocp_module = importlib.import_module(module_name)
-#     ocp_module.main(
-#         plot=False,
-#         model_path=cycling_biorbd_model_path,
-#         initial_guess_model_path=cycling_initial_guess_biorbd_model_path,
-#     )
+
+@pytest.mark.parametrize("module_name", MULTIBODY_CYCLING_EXAMPLE_MODULES)
+def test_cycling_multibody_examples(module_name):
+    ocp_module = importlib.import_module(module_name)
+    ocp_module.main(
+        plot=False,
+        model_path=cycling_biorbd_model_path,
+        initial_guess_model_path=cycling_initial_guess_biorbd_model_path,
+    )
 
 
 MULTIBODY_FLEXION_EXAMPLE_MODULES = [
