+++ cocofest/models/marion2013/marion2013_modified_with_fatigue.py
@@ -58,8 +58,8 @@ class Marion2013ModelPulseWidthFrequencyWithFatigue(Marion2013ModelPulseWidthFre
         self.beta_a = BETA_A_DEFAULT
 
     @property
-    def name_dof(self, with_muscle_name: bool = False) -> list[str]:
-        muscle_name = "_" + self.muscle_name if self.muscle_name and with_muscle_name else ""
+    def name_dofs(self, with_muscle_name: bool = False) -> list[str]:
+        muscle_name = ("_" + self.muscle_name if self.muscle_name is not None else "")
         return [
             "Cn" + muscle_name,
             "F" + muscle_name,
