+++ cocofest/models/marion2013/marion2013_modified.py
@@ -68,8 +68,8 @@ class Marion2013ModelPulseWidthFrequency(Marion2009ModelPulseWidthFrequency):
         self.FM = FM_DEFAULT
 
     @property
-    def name_dof(self, with_muscle_name: bool = False) -> list[str]:
-        muscle_name = "_" + self.muscle_name if self.muscle_name and with_muscle_name else ""
+    def name_dofs(self, with_muscle_name: bool = False) -> list[str]:
+        muscle_name = ("_" + self.muscle_name if self.muscle_name is not None else "")
         return [
             "Cn" + muscle_name,
             "F" + muscle_name,
