+++ cocofest/models/ding2003/ding2003_with_fatigue.py
@@ -94,7 +94,7 @@ class DingModelFrequencyWithFatigue(DingModelFrequency):
     # ---- Needed for the example ---- #
     @property
     def name_dofs(self) -> list[str]:
-        muscle_name = ("_" + self.muscle_name if self.muscle_name is not None else "")
+        muscle_name = "_" + self.muscle_name if self.muscle_name is not None else ""
         return [
             "Cn" + muscle_name,
             "F" + muscle_name,
