+++ cocofest/models/veltink1992/veltink1992.py
@@ -42,7 +42,6 @@ class VeltinkModelPulseIntensity(StateDynamics):
 
         self.contact_types = ()
 
-
     @property
     def name(self):
         return self._name
@@ -66,7 +65,7 @@ class VeltinkModelPulseIntensity(StateDynamics):
 
     @property
     def name_dofs(self, with_muscle_name: bool = False) -> list[str]:
-        muscle_name = ("_" + self.muscle_name if self.muscle_name is not None else "")
+        muscle_name = "_" + self.muscle_name if self.muscle_name is not None else ""
         return ["a" + muscle_name]  # Only muscle activation state
 
     @property
