### 2025-11-24 f4efa64
Bioptim installation
