+++ examples/fes_multibody/cycling/cost_functions.py
@@ -1,773 +1,0 @@
-from casadi import MX, vertcat, sum1, mmax, fabs, sign
-from bioptim import PenaltyController
-from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
-
-
-class CustomCostFunctions:
-    def __init__(self):
-        self.dict_functions = {
-            "minimize_average_activation": {
-                "function": self.minimize_average_activation,
-                "index": 1,
-                "latex": r"\phi_1 = \frac{1}{n_m}\sum_{j=1}^{n_m} a^{j}, \quad a^{j}=\frac{f^{j}-f^{j}_{\min}}{f^{j}_{\max}-f^{j}_{\min}}",
-                "description": "Minimize the average muscle activation",
-                "power": "1",
-                "state": "pw",
-            },
-            "minimize_root_mean_square_activation": {
-                "function": self.minimize_root_mean_square_activation,
-                "index": 2,
-                "latex": r"\phi_2 = \left(\frac{1}{n_m}\sum_{j=1}^{n_m} (a^{j})^{2}\right)^{\tfrac{1}{2}}, \quad a^{j}=\frac{f^{j}-f^{j}_{\min}}{f^{j}_{\max}-f^{j}_{\min}}",
-                "description": "Minimize the root mean square of muscle activation",
-                "power": "2",
-                "state": "pw",
-            },
-            "minimize_cubic_average_activation": {
-                "function": self.minimize_cubic_average_activation,
-                "index": 3,
-                "latex": r"\phi_3 = \left(\frac{1}{n_m}\sum_{j=1}^{n_m} (a^{j})^{3}\right)^{\tfrac{1}{3}}, \quad a^{j}=\frac{f^{j}-f^{j}_{\min}}{f^{j}_{\max}-f^{j}_{\min}}",
-                "description": "Minimize the cubic average of muscle activation",
-                "power": "3",
-                "state": "pw",
-            },
-            "minimize_peak_activation": {
-                "function": self.minimize_peak_activation,
-                "index": 4,
-                "latex": r"\phi_4 = \max_{j=1,\ldots,n_m} \; a^{j}, \quad a^{j}=\frac{f^{j}-f^{j}_{\min}}{f^{j}_{\max}-f^{j}_{\min}}",
-                "description": "Minimize the peak of muscle activation",
-                "power": "max",
-                "state": "pw",
-            },
-            "minimize_average_force": {
-                "function": self.minimize_average_force,
-                "index": 5,
-                "latex": r"\phi_5 = \frac{1}{n_m}\sum_{j=1}^{n_m} f^{j}",
-                "description": "Minimize the average muscle force",
-                "power": "1",
-                "state": "f",
-            },
-            "minimize_root_mean_square_force": {
-                "function": self.minimize_root_mean_square_force,
-                "index": 6,
-                "latex": r"\phi_6 = \left(\frac{1}{n_m}\sum_{j=1}^{n_m} (f^{j})^{2}\right)^{\tfrac{1}{2}}",
-                "description": "Minimize the root mean square of muscle force",
-                "power": "2",
-                "state": "f",
-            },
-            "minimize_cubic_average_force": {
-                "function": self.minimize_cubic_average_force,
-                "index": 7,
-                "latex": r"\phi_7 = \left(\frac{1}{n_m}\sum_{j=1}^{n_m} (f^{j})^{3}\right)^{\tfrac{1}{3}}",
-                "description": "Minimize the cubic average of muscle force",
-                "power": "3",
-                "state": "f",
-            },
-            "minimize_peak_force": {
-                "function": self.minimize_peak_force,
-                "index": 8,
-                "latex": r"\phi_8 = \max_{j=1,\ldots,n_m} \; f^{j}",
[…]
