+++ cocofest/models/ding2007/ding2007_with_fatigue.py
@@ -126,7 +126,7 @@ class DingModelPulseWidthFrequencyWithFatigue(DingModelPulseWidthFrequency):
         The system dynamics is the function that describes the models.
 
         Parameters
-        # ----------
+        ----------
         time: MX
             The system's current node time
         states: MX
