### 2025-11-26 12d9a6c
New cost fun
