+++ examples/fes_multibody/cycling/cost_functions.py
@@ -780,7 +780,7 @@ class CustomCostFunctions:
         eps = 1e-8
         muscle_fatigue_decay = vertcat(
             *[
-                ((A_t[x] - A_t_plus_one[x]) / (A_t[x] - A_end[x])) ** 2
+                ((A_t[x] - A_t_plus_one[x]) / (1 + ((A_t[x] - A_end[x])/A_end[x]))) ** 2
                 for x in range(len(muscle_name_list))
             ]
         )
