### 2025-12-10 d3747e4
scout: Update README equations rendering
- make the rendering of equations better.
