### 2025-12-10 b119d5a
Revise installation instructions in README
Updated bioptim version in installation instructions and removed setup.py installation steps.
