### 2025-12-15 8c137a0
less convergence iter to optim
