+++ examples/fes_multibody/cycling/cost_functions.py
@@ -8,7 +8,7 @@ class CustomCostFunctions:
             "minimize_average_activation": {
                 "function": self.minimize_average_activation,
                 "index": 1,
-                "latex": r"\phi_1 = \frac{1}{n_m}\sum_{j=1}^{n_m} a^{j}, \quad a^{j}=\frac{f^{j}-f^{j}_{\min}}{f^{j}_{\max}-f^{j}_{\min}}",
+                "latex": r"\phi_1 = \frac{1}{M}\sum_{m=1}^{M} a^{m}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
                 "description": "Minimize the average muscle activation",
                 "power": "1",
                 "state": "pw",
@@ -16,7 +16,7 @@ class CustomCostFunctions:
             "minimize_root_mean_square_activation": {
                 "function": self.minimize_root_mean_square_activation,
                 "index": 2,
-                "latex": r"\phi_2 = \left(\frac{1}{n_m}\sum_{j=1}^{n_m} (a^{j})^{2}\right)^{\tfrac{1}{2}}, \quad a^{j}=\frac{f^{j}-f^{j}_{\min}}{f^{j}_{\max}-f^{j}_{\min}}",
+                "latex": r"\phi_2 = \left(\frac{1}{M}\sum_{m=1}^{M} (a^{m})^{2}\right)^{\tfrac{1}{2}}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
                 "description": "Minimize the root mean square of muscle activation",
                 "power": "2",
                 "state": "pw",
@@ -24,7 +24,7 @@ class CustomCostFunctions:
             "minimize_cubic_average_activation": {
                 "function": self.minimize_cubic_average_activation,
                 "index": 3,
-                "latex": r"\phi_3 = \left(\frac{1}{n_m}\sum_{j=1}^{n_m} (a^{j})^{3}\right)^{\tfrac{1}{3}}, \quad a^{j}=\frac{f^{j}-f^{j}_{\min}}{f^{j}_{\max}-f^{j}_{\min}}",
+                "latex": r"\phi_3 = \left(\frac{1}{M}\sum_{m=1}^{M} (a^{m})^{3}\right)^{\tfrac{1}{3}}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
                 "description": "Minimize the cubic average of muscle activation",
                 "power": "3",
                 "state": "pw",
@@ -32,7 +32,7 @@ class CustomCostFunctions:
             "minimize_peak_activation": {
                 "function": self.minimize_peak_activation,
                 "index": 4,
-                "latex": r"\phi_4 = \max_{j=1,\ldots,n_m} \; a^{j}, \quad a^{j}=\frac{f^{j}-f^{j}_{\min}}{f^{j}_{\max}-f^{j}_{\min}}",
+                "latex": r"\phi_4 = \max_{m=1,\ldots,M} \; a^{m}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
                 "description": "Minimize the peak of muscle activation",
                 "power": "max",
                 "state": "pw",
@@ -40,7 +40,7 @@ class CustomCostFunctions:
             "minimize_average_force": {
                 "function": self.minimize_average_force,
                 "index": 5,
-                "latex": r"\phi_5 = \frac{1}{n_m}\sum_{j=1}^{n_m} f^{j}",
+                "latex": r"\phi_5 = \frac{1}{M}\sum_{m=1}^{M} f^{m}",
[…]
