### 2026-01-05 #8989 @Kev1CO
Dear @Kevin-Mattheus-Moerman and @rkurchin,
Thank you again for your time and for all the work you do for JOSS.

We have been informed that the BCM track is currently short on editors/reviewers and that our submission has been waitlisted as a result. We understand that this can slow things down, and we will be patient.

When you have a moment, could you share a rough estimate of when our paper might be picked up for the review process (or when we should check back)? Even a broad timeframe would be really helpful to plan the upcoming year.

Many thanks, and wishing you a happy new year!
