### 2026-01-09 0fc888e
Optimized motor torque for standard FES parameters
