### 2026-01-14 03d40b1
Set solver
