### 2026-01-16 89c40fd
correcting bound shift
