### 2026-01-16 397080b
New cost funs
