+++ examples/fes_multibody/cycling/cost_functions.py
@@ -1,4 +1,4 @@
-from casadi import MX, vertcat, sum1, mmax, fabs, sign
+from casadi import MX, vertcat, sum1, mmax, fabs, sign, tanh
 from bioptim import PenaltyController
 from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
 
@@ -142,38 +142,48 @@ class CustomCostFunctions:
                 "state": "pow",
             },
 
-
-            "minimize_root_mean_square_fatigue_with_weights": {
-                "function": self.minimize_root_mean_square_fatigue_with_weights,
-                "index": 21,
-                "latex": r"\phi_{21} = \left(\frac{1}{\sum_{m=1}^{M} w_m}\sum_{m=1}^{M} w_m (\mathcal{F}^{m})^{2}\right)^{\tfrac{1}{2}}",
-                "description": "Minimize the root mean square of muscle fatigue with weights",
-                "power": "2",
-                "state": "fatw",
-            },
+            # --- Custom cost functions --- #
             "minimize_root_mean_square_scalable_fatigue_decay": {
                 "function": self.minimize_root_mean_square_scalable_fatigue_decay,
-                "index": 22,
-                "latex": r"\phi_{22} = \left(\frac{1}{M}\sum_{t=1}^{M}\left(\frac{A_{t+1} - A_t}{A_t - A_{\text{end}}}\right)^{2}\right)^{\tfrac{1}{2}}",
+                "index": 20,
+                "latex": r"\phi_{20} = \left(\frac{1}{M}\sum_{t=1}^{M}\left(\frac{A_{t+1} - A_t}{A_t - A_{\text{end}}}\right)^{2}\right)^{\tfrac{1}{2}}",
                 "description": "Minimize the root mean square of scalable muscle fatigue decay",
                 "power": "2",
                 "state": "fats",
             },
             "minimize_root_mean_square_fatigue_decay": {
                 "function": self.minimize_root_mean_square_fatigue_decay,
-                "index": 23,
-                "latex": r"\phi_{23} = \left(\frac{1}{M}\sum_{t=1}^{M}\left(\frac{A_{t+1} - A_t}{-A_{\text{rest}} / alpha_a}\right)^{2}\right)^{\tfrac{1}{2}}",
+                "index": 21,
+                "latex": r"\phi_{21} = \left(\frac{1}{M}\sum_{t=1}^{M}\left(\frac{A_{t+1} - A_t}{-A_{\text{rest}} / alpha_a}\right)^{2}\right)^{\tfrac{1}{2}}",
                 "description": "Minimize the root mean square of muscle fatigue decay",
                 "power": "2",
                 "state": "fatd",
             },
             "minimize_peak_fatigue_decay": {
                 "function": self.minimize_peak_fatigue_decay,
-                "index": 24,
-                "latex": r"\phi_{24} = \max_{m=1,\ldots,M} \left( \frac{A_{t+1}^{(m)} - A_t^{(m)}}{-A_{\text{rest}}^{(m)} / \alpha_a} \right)",
+                "index": 22,
+                "latex": r"\phi_{22} = \max_{m=1,\ldots,M} \left( \frac{A_{t+1}^{(m)} - A_t^{(m)}}{-A_{\text{rest}}^{(m)} / \alpha_a} \right)",
                 "description": "Minimize the peak muscle fatigue decay",
                 "power": "max",
                 "state": "fatd",
+            },
+
+            "minimize_root_mean_square_tanh_fatigue_decay": {
+                "function": self.minimize_root_mean_square_tanh_fatigue_decay,
+                "index": 23,
[…]
