+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -1094,25 +1094,31 @@ if __name__ == "__main__":
         n_cycles_simultaneous=[2, 3, 4, 5],
         resistive_torque=-0.20,  # (N.m)
         cost_fun_dict={"optimized_function": [
-            ["minimize_average_activation"],
-            ["minimize_root_mean_square_activation"],
-            ["minimize_cubic_average_activation"],
-            ["minimize_peak_activation"],
-            ["minimize_average_force"],
-            ["minimize_root_mean_square_force"],
-            ["minimize_cubic_average_force"],
-            ["minimize_peak_force"],
-            ["minimize_average_muscle_stress"],
-            ["minimize_root_mean_square_muscle_stress"],
-            ["minimize_cubic_average_muscle_stress"],
-            ["minimize_peak_muscle_stress"],
-            ["minimize_root_mean_square_muscle_power"],
-            ["minimize_average_fatigue"],
-            ["minimize_root_mean_square_fatigue"],
-            ["minimize_cubic_average_fatigue"],
-            ["minimize_peak_fatigue"],
-            ["minimize_root_mean_square_fatigue_with_weights"],
+            # ["minimize_average_activation"],
+            # ["minimize_root_mean_square_activation"],
+            # ["minimize_cubic_average_activation"],
+            # ["minimize_peak_activation"],
+            # ["minimize_average_force"],
+            # ["minimize_root_mean_square_force"],
+            # ["minimize_cubic_average_force"],
+            # ["minimize_peak_force"],
+            # ["minimize_average_muscle_stress"],
+            # ["minimize_root_mean_square_muscle_stress"],
+            # ["minimize_cubic_average_muscle_stress"],
+            # ["minimize_peak_muscle_stress"],
+            # ["minimize_root_mean_square_muscle_power"],
+            # ["minimize_average_fatigue"],
+            # ["minimize_root_mean_square_fatigue"],
+            # ["minimize_cubic_average_fatigue"],
+            # ["minimize_peak_fatigue"],
+            # ["minimize_root_mean_square_fatigue_with_weights"],
+            # ["minimize_root_mean_square_scalable_fatigue_decay"],
+
             ["minimize_root_mean_square_scalable_fatigue_decay"],
+            ["minimize_root_mean_square_fatigue_decay"],
+            ["minimize_peak_fatigue_decay"],
+            ["minimize_root_mean_square_tanh_fatigue_decay"],
+            ["minimize_root_mean_square_weighted_tanh_fatigue_decay"],
             ],
 
             "weight": [
