### 2026-01-19 f009862
fixed qdot
