### 2026-01-19 eff250f
fixing cost functions
