+++ examples/fes_multibody/cycling/cost_functions.py
@@ -697,7 +697,7 @@ class CustomCostFunctions:
         eps = 1e-8
         muscle_fatigue_decay = vertcat(
             *[
-                ((A_t[x] - A_t_plus_one[x]) / -(A_rest[x] / alpha_a[x])) ** 2
+                ((A_t[x] - A_t_plus_one[x]) / (1/(-(A_rest[x] / alpha_a[x])))) ** 2
                 for x in range(len(muscle_name_list))
             ]
         )
@@ -726,7 +726,7 @@ class CustomCostFunctions:
 
         muscle_fatigue_decay = vertcat(
             *[
-                ((A_t[x] - A_t_plus_one[x]) / -(A_rest[x] / alpha_a[x]))
+                ((A_t[x] - A_t_plus_one[x]) / (1/(-(A_rest[x] / alpha_a[x]))))
                 for x in range(len(muscle_name_list))
             ]
         )
@@ -755,7 +755,7 @@ class CustomCostFunctions:
         eps = 1e-8
         muscle_fatigue_decay = vertcat(
             *[
-                (1 + tanh(0.1 * (A_t[x] - A_t_plus_one[x]))) ** 2
+                (1 + tanh(10 * (A_t[x] - A_t_plus_one[x]))) ** 2
                 for x in range(len(muscle_name_list))
             ]
         )
@@ -781,14 +781,11 @@ class CustomCostFunctions:
         alpha_a = vertcat(*[controller.model.muscles_dynamics_model[x].alpha_a for x in range(len(muscle_name_list))])
         A_t = vertcat(*[controller.states["A_" + muscle_name_list[x]].cx for x in range(len(muscle_name_list))])
         A_t_plus_one = CustomCostFunctions.compute_A_t_plus_one(controller)
-        weights = vertcat(*[(A_rest[x] / (-alpha_a[x])) for x in range(len(muscle_name_list))])
-        l1 = sum1(fabs(weights))
-        weights_normed = weights / l1
 
         eps = 1e-8
         muscle_fatigue_decay = vertcat(
             *[
-                (((1 + tanh(0.1 * (A_t[x] - A_t_plus_one[x]))) ** 2) / weights_normed[x])
+                ((1 + tanh(10 * (A_t[x] - A_t_plus_one[x]))) / (1/(-(A_rest[x] / alpha_a[x])))) ** 2
                 for x in range(len(muscle_name_list))
             ]
         )
