### 2026-01-19 6d46965
working on step computation
