### 2026-01-19 e77ff92
fatigability weight
