### 2026-01-20 ce0e546
fixing minmax error
