### 2026-01-20 cd13780
removing factor
