### 2026-02-03 de23b6e
No log_unifor for bayesian weights
