### 2026-02-13 4b1669f
Barrier function
