+++ examples/fes_multibody/cycling/cost_functions.py
@@ -1,4 +1,4 @@
-from casadi import MX, vertcat, sum1, mmax, fabs, sign, tanh, if_else
+from casadi import MX, vertcat, sum1, mmax, fabs, sign, tanh, if_else, log, exp, DM, dot
 from bioptim import PenaltyController
 from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
 
@@ -34,7 +34,7 @@ class CustomCostFunctions:
                 "index": 4,
                 "latex": r"\phi_4 = \max_{m=1,\ldots,M} \; a^{m}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
                 "description": "Minimize the peak of muscle activation",
-                "power": "max",
+                "power": r"\infty",
                 "state": "pw",
             },
             "minimize_average_force": {
@@ -43,7 +43,7 @@ class CustomCostFunctions:
                 "latex": r"\phi_5 = \frac{1}{M}\sum_{m=1}^{M} f^{m}",
                 "description": "Minimize the average muscle force",
                 "power": "1",
-                "state": "f",
+                "state": r"F^{m}",
             },
             "minimize_root_mean_square_force": {
                 "function": self.minimize_root_mean_square_force,
@@ -51,7 +51,7 @@ class CustomCostFunctions:
                 "latex": r"\phi_6 = \left(\frac{1}{M}\sum_{m=1}^{M} (f^{m})^{2}\right)^{\tfrac{1}{2}}",
                 "description": "Minimize the root mean square of muscle force",
                 "power": "2",
-                "state": "f",
+                "state": r"F^{m}",
             },
             "minimize_cubic_average_force": {
                 "function": self.minimize_cubic_average_force,
@@ -59,15 +59,15 @@ class CustomCostFunctions:
                 "latex": r"\phi_7 = \left(\frac{1}{M}\sum_{m=1}^{M} (f^{m})^{3}\right)^{\tfrac{1}{3}}",
                 "description": "Minimize the cubic average of muscle force",
                 "power": "3",
-                "state": "f",
+                "state": r"F^{m}",
             },
             "minimize_peak_force": {
                 "function": self.minimize_peak_force,
                 "index": 8,
                 "latex": r"\phi_8 = \max_{m=1,\ldots,M} \; f^{m}",
                 "description": "Minimize the peak muscle force",
-                "power": "max",
-                "state": "f",
+                "power": r"\infty",
+                "state": r"F^{m}",
             },
             "minimize_average_muscle_stress": {
                 "function": self.minimize_average_muscle_stress,
@@ -75,7 +75,7 @@ class CustomCostFunctions:
                 "latex": r"\phi_9 = \frac{1}{M}\sum_{m=1}^{M} \frac{f^{m}}{S^{m}}",
                 "description":"Minimize the average muscle stress",
                 "power": "1",
-                "state": "str",
+                "state": r"\sigma",
             },
             "minimize_root_mean_square_muscle_stress": {
                 "function": self.minimize_root_mean_square_muscle_stress,
@@ -83,7 +83,7 @@ class CustomCostFunctions:
                 "latex": r"\phi_{10} = \left(\frac{1}{M}\sum_{m=1}^{M} \left(\frac{f^{m}}{S^{m}}\right)^{2}\right)^{\tfrac{1}{2}}",
                 "description":"Minimize the root mean square of muscle stress",
                 "power": "2",
-                "state": "str",
+                "state": r"\sigma",
             },
[…]
