### 2026-02-13 3ad7b39
Resuming beyan optim from saved data + multiple window final calculation
