+++ examples/fes_multibody/cycling/cycling_bayesian_mhe.py
@@ -1,12 +1,12 @@
 """
 This example will perform a bayesian optimization on an optimal control program moving time horizon for a hand cycling
- motion driven by FES to find the best weight to increase number of cycling before failure.
+motion driven by FES to find the best weight to increase number of cycling before failure.
 """
 
 from pathlib import Path
 from sys import platform
-
 import pickle
+
 import numpy as np
 from skopt import gp_minimize
 from skopt.space import Real
@@ -31,21 +31,13 @@ import cycling_pulse_width_mhe as base
 def minimize_root_mean_square_fatigue(controller: PenaltyController, muscle_weights: list) -> MX:
     """
     Minimize the root-mean-square of muscle fatigue.
-
-    Parameters
-    ----------
-    controller: PenaltyController
-        The penalty node elements
-
-    Returns
-    -------
-    The root-mean-square of muscle fatigue
     """
     eps = 1e-8
     muscle_name_list = controller.model.bio_model.muscle_names
     muscle_fatigue = vertcat(
         *[
-            muscle_weights[x] * (controller.model.muscles_dynamics_model[x].a_scale - controller.states["A_" + muscle_name_list[x]].cx) ** 2
+            muscle_weights[x]
+            * (controller.model.muscles_dynamics_model[x].a_scale - controller.states["A_" + muscle_name_list[x]].cx) ** 2
             for x in range(len(muscle_name_list))
         ]
     )
@@ -65,9 +57,7 @@ def set_objective_functions(muscle_fatigue_key, cost_fun_weight):
         elif len(cost_fun_weight) == len(muscle_fatigue_key):
             weights = list(map(float, cost_fun_weight))
         else:
-            raise ValueError(
-                f"cost_fun_weight must be length 1 or {len(muscle_fatigue_key)}, got {len(cost_fun_weight)}"
-            )
+            raise ValueError(f"cost_fun_weight must be length 1 or {len(muscle_fatigue_key)}, got {len(cost_fun_weight)}")
 
     objective_functions.add(
         minimize_root_mean_square_fatigue,
@@ -106,9 +96,7 @@ def prepare_nmpc_bo(
         external_force_dict=cycling_info["resistive_torque"],
         force_name="external_torque",
     )
-    time_series2, _ = model.muscles_dynamics_model[0].get_numerical_data_time_series(
-        window_n_shooting, window_cycle_duration
-    )
+    time_series2, _ = model.muscles_dynamics_model[0].get_numerical_data_time_series(window_n_shooting, window_cycle_duration)
     numerical_time_series.update(time_series2)
 
     # --- Dynamics & states --- #
@@ -127,16 +115,12 @@ def prepare_nmpc_bo(
         ode_solver=ode_solver,
         init_file_path=initial_guess_path,
     )
-    u_bounds, u_init, u_scaling = base.set_u_bounds_and_init(
-        model, window_n_shooting, init_file_path=initial_guess_path
-    )
-    constraints = base.set_constraints(model, x_init["q"].init[2][0] - 2*np.pi, cycle_len, n_cycles_simultaneous)
+    u_bounds, u_init, u_scaling = base.set_u_bounds_and_init(model, window_n_shooting, init_file_path=initial_guess_path)
+    constraints = base.set_constraints(model, x_init["q"].init[2][0] - 2 * np.pi, cycle_len, n_cycles_simultaneous)
 
     # --- Per-muscle fatigue objective --- #
     muscle_fatigue_keys = [f"A_{m.muscle_name}" for m in model.muscles_dynamics_model]
-    objective_functions = set_objective_functions(
-        muscle_fatigue_keys, simulation_conditions["cost_fun_weight"],
-    )
[…]
