### 2026-02-18 2669aeb
Removing bioptim external from cocofest
