+++ .github/workflows/run_tests_win.yml
@@ -40,13 +40,6 @@ jobs:
           conda info
           conda list
 
-      - name: Install bioptim on Windows
-        run: |
-          pwd
-          cd external
-          ./bioptim_install_windows.sh 4 ${{ env.PREFIX_WINDOWS }}
-          cd ..
-
       - name: Install extra dependencies
         run: conda install pytest-cov black pytest pytest-cov codecov packaging -cconda-forge
 
@@ -91,13 +84,6 @@ jobs:
         run: |
           conda info
           conda list
-
-      - name: Install bioptim on Windows
-        run: |
-          pwd
-          cd external
-          ./bioptim_install_windows.sh 4 ${{ env.PREFIX_WINDOWS }}
-          cd ..
 
       - name: Install extra dependencies
         run: pip install coverage
