+++ docs/article/paper.bib
@@ -10,6 +10,17 @@
   doi={10.1371/journal.pone.0149024},
 }
 
+@article{co2025optimal,
+  title = {Optimal control-driven functional electrical stimulation: A PRISMA-ScR scoping review},
+  author = {Kevin Co and Mickaël Begon and François Bailly and Florent Moissenet},
+  journal = {Computers in Biology and Medicine},
+  volume = {199},
+  pages = {111311},
+  year = {2025},
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.compbiomed.2025.111311},
+}
+
 @article{hunt1997feedback,
   title={Feedback control of unsupported standing in paraplegia. I. Optimal control approach},
   author={Hunt, Kenneth J and Munih, Marko and Donaldson, N de N},
@@ -22,12 +33,28 @@
   doi={10.1109/86.650287},
 }
 
-@article{co2025optimal,
-  title={Optimal control driven functional electrical stimulation: A scoping review},
-  author={Co, Kevin and Begon, Micka{\"e}l and Bailly, Fran{\c{c}}ois and Moissenet, Florent},
-  journal={arXiv preprint arXiv:2508.02899},
-  year={2025},
-  doi={10.48550/arXiv.2508.02899},
+@article{ding2003mathematical,
+  title={Mathematical models for fatigue minimization during functional electrical stimulation},
+  author={Ding, Jun and Wexler, Anthony S and Binder-Macleod, Stuart A},
+  journal={Journal of Electromyography and Kinesiology},
+  volume={13},
+  number={6},
+  pages={575--588},
+  year={2003},
+  publisher={Elsevier},
+  doi={10.1016/S1050-6411(03)00102-0},
+}
+
+@article{veltink1992nonlinear,
+  title={Nonlinear joint angle control for artificially stimulated muscle},
+  author={Veltink, Peter H and Chizeck, Howard J and Crago, Patrick E and El-Bialy, Ahmed},
+  journal={IEEE Transactions on biomedical engineering},
+  volume={39},
+  number={4},
+  pages={368--380},
+  year={1992},
+  publisher={IEEE},
+  doi={10.1109/10.126609},
 }
 
 @article{puchaud2023direct,
@@ -64,40 +91,28 @@
   doi={10.1016/j.conengprac.2009.12.007},
 }
 
-@article{veltink1992nonlinear,
-  title={Nonlinear joint angle control for artificially stimulated muscle},
-  author={Veltink, Peter H and Chizeck, Howard J and Crago, Patrick E and El-Bialy, Ahmed},
-  journal={IEEE Transactions on biomedical engineering},
-  volume={39},
-  number={4},
-  pages={368--380},
-  year={1992},
-  publisher={IEEE},
-  doi={10.1109/10.126609},
+@article{dembia2020opensim,
+  title={OpenSim Moco: musculoskeletal optimal control},
+  author={Dembia, Christopher L and Bianco, Nicholas A and Falisse, Antoine and Hicks, Jennifer L and Delp, Scott L},
+  journal={PLOS Computational Biology},
+  volume={16},
+  number={12},
+  pages={e1008493},
+  year={2020},
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1008493},
 }
 
-@article{ding2003mathematical,
-  title={Mathematical models for fatigue minimization during functional electrical stimulation},
-  author={Ding, Jun and Wexler, Anthony S and Binder-Macleod, Stuart A},
-  journal={Journal of Electromyography and Kinesiology},
-  volume={13},
-  number={6},
-  pages={575--588},
-  year={2003},
-  publisher={Elsevier},
[…]
