+++ docs/article/paper.md
@@ -45,12 +45,12 @@ reaching, and grasping. FES rehabilitation mostly relies on empirical settings, 
 populations and muscles. Empirical settings often cause overstimulation and premature fatigue [@ibitoye2016strategies],
 shortening rehabilitation sessions and diminishing therapeutic benefit. Consequently, advanced control approaches like
 optimal control-driven FES are gaining interest in personalizing and improving FES rehabilitation efficiency, meanwhile
-delaying muscle fatigue. To address this need, we designed `Cocofest` (Custom Optimal COntrol for Functional Electrical
-STimulation), an open-source Python package for optimal control-driven FES. `Cocofest` provides a framework to generate
-personalized pulse trains (Fig. 1) based on nonlinear dynamics models for FES (Table. 1), for several musculoskeletal
-models and motor tasks. The package provides more than 10 examples, covering optimization of FES-related pulse train
-parameters (including frequency, pulse width, pulse intensity), FES model parameters identification from in vivo
-measurements, and long duration predictive simulations.
+delaying muscle fatigue [@co2025optimal]. To address this need, we designed `Cocofest` (Custom Optimal COntrol for
+Functional Electrical STimulation), an open-source Python package for optimal control-driven FES. `Cocofest` provides a
+framework to generate personalized pulse trains (Fig. 1) based on nonlinear dynamics models for FES (Table. 1), for
+several musculoskeletal models and motor tasks. The package includes over 10 examples, covering optimization of
+FES-related pulse train parameters (including frequency, pulse width, pulse intensity), FES model parameters
+identification from in-vivo measurements, and long duration predictive simulations.
 
 ![Pulse train parameters that can be optimized in Cocofest](pulse_train.png){ width=90% }
 
@@ -64,23 +64,24 @@ hindering standardization and cumulative progress [@co2025optimal]. To address t
 scientific progress, `Cocofest` fulfills the following four needs:
 
 Firstly, the relationship between the pulse train parameters (e.g., frequency, pulse width and intensity; Fig. 1) and
-the resulting muscle force, joint torque, and muscle fatigue (termed as state variables) can be modeled with different
-nonlinear dynamics [@ding2003mathematical; @veltink1992nonlinear] (Table 1). Gathering them within a unified package
-would facilitate comparison for more informed modelling choices.
+the resulting muscle force, joint torque, and muscle fatigue can be modeled with different nonlinear dynamics
+[@ding2003mathematical; @veltink1992nonlinear]. Gathering them within a unified package would facilitate comparison for
+more informed modelling choices.
 
-Secondly, no study has compared different optimal control problem (OCP) formulations ap-plied to FES, due to OCP
+Secondly, no study has compared different optimal control problem (OCP) formulations applied to FES, due to OCP
 implementation challenges [@co2025optimal]. Easily customizable OCP formulation, involving objective functions, models,
 and transcriptions is required to provide an adequate research framework. Having the possibility to switch between
-several OCP transcriptions, such as direct collocation or direct multiple shooting, is essential when dealing with stiff
-differential equations [@puchaud2023direct], often embed in FES models. Muscle fatigue is the primary challenge in FES.
-Enabling the development and comparison of different OCP formulations could help address research questions, yield novel
-stimulation patterns and enhance fatigue reduction. Moreover, using receding-horizon estimation for longer simulations
-reduces the computational complexity associated with time-varying dynamics (e.g., fatigue) [@ding2003mathematical].
+various OCP transcriptions (e.g., direct collocation or direct multiple shooting) is essential when dealing with stiff
[…]
