+++ external/bioptim_install_windows.sh
@@ -1,12 +1,0 @@
-#!/bin/bash
-
-## Move to the external folder
-cd ${0%/*}
-
-git submodule update --recursive --init
-
-## Move to the bioptim folder
-cd bioptim
-
-# Installing bioptim
-pip install .
