### 2026-02-18 9864047
Updating bioptim installation
