+++ examples/msk_models/model_viewer.py
@@ -5,7 +5,6 @@ The user can interact with the model by changing the joint angles using sliders.
 
 from pyorerun import LiveModelAnimation
 
-
 model_path = "arm26.bioMod"
 animation = LiveModelAnimation(model_path, with_q_charts=True)
 animation.rerun()
