+++ environment.yml
@@ -4,7 +4,6 @@ channels:
 - conda-forge
 dependencies:
 - bioptim=3.4.0
-- biorbd=1.11.2
 - bioviz
 - python-graphviz
 - pyqtgraph
