### 2026-02-18 #8989 @Kev1CO
@arfon - Thank you for keeping us updated.
I adjusted the paper to the JOSS journal's new requests.
We are looking forward to the reviewing process.
