### 2026-02-23 2dc5dad
saving solving time
