+++ examples/fes_multibody/shoulder_abduction/abduction_pw_mhe.py
@@ -1,737 +1,0 @@
-"""
-This example will perform an optimal control program moving time horizon for an abduction/adduction motion driven by FES.
-"""
-
-import pickle
-from sys import platform
-from itertools import product
-from pathlib import Path
-
-import numpy as np
-from numpy.ma.extras import average
-
-from bioptim import (
-    BiorbdModel,
-    BoundsList,
-    ConstraintList,
-    ConstraintFcn,
-    ContactType,
-    CostType,
-    DynamicsList,
-    ExternalForceSetTimeSeries,
-    InitialGuessList,
-    InterpolationType,
-    MultiCyclicCycleSolutions,
-    MultiCyclicNonlinearModelPredictiveControl,
-    ObjectiveFcn,
-    ObjectiveList,
-    OdeSolver,
-    PhaseDynamics,
-    ParameterObjectiveList,
-    SolutionMerge,
-    Solution,
-    Solver,
-    ParameterList,
-    Node,
-    VariableScalingList,
-)
-from cocofest import (
-    DingModelPulseWidthFrequencyWithFatigue,
-    FesMskModel,
-    OcpFesMsk,
-    FesNmpcMsk,
-)
-from examples.fes_multibody.cycling.cost_functions import CustomCostFunctions
-
-
-class MyCyclicNMPC(FesNmpcMsk):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.nodes_per_cycle = self.cycle_len * (
-            self.nlp[0].dynamics_type.ode_solver.polynomial_degree + 1
-            if isinstance(self.nlp[0].dynamics_type.ode_solver, OdeSolver.COLLOCATION)
-            else 1
-        )
-        self.debugg_bounds = False
-        self.previous_bounds = None
-
-    def advance_window_bounds_states(self, sol, n_cycles_simultaneous=None, **extra):
-        # --- Get states results --- #
-        states = sol.decision_states(to_merge=SolutionMerge.NODES)
-        states_keys = states.keys()
-        # --- States are bounded to match the last node of the cycle to ensure continuity between window --- #
-        for key in states_keys:
-            for i in range(states[key].shape[0]):
-                self.nlp[0].x_bounds[key].min[i, 0] = states[key][i][self.nodes_per_cycle]
-                self.nlp[0].x_bounds[key].max[i, 0] = states[key][i][self.nodes_per_cycle]
-        # --- Inform the past cycle stimulation time into the new one --- #
-        self.update_stim()
-        return True
-
-    def advance_window_initial_guess_states(self, sol, n_cycles_simultaneous=None):
-        # --- Get states results --- #
-        states = sol.decision_states(to_merge=SolutionMerge.NODES)
-        states_keys = states.keys()
-        cyclical_keys = [s for s in states if any(s.startswith(prefix) for prefix in ("Cn_", "F_", "q", "qdot"))]
-        continuous_keys = [s for s in states if any(s.startswith(prefix) for prefix in ("A_", "Tau1_", "Km_"))]
-        # --- Set initial guesses for cyclical and continuous states --- #
-        for key in states_keys:
-            for i in range(states[key].shape[0]):
-                if key in cyclical_keys:
-                    self.set_init_cyclical(states, key, i)
-                elif key in continuous_keys:
-                    self.set_init_continuous(states, key, i)
-        self._correct_init_guess_to_fit_bounds(
-            corrected_input="states"
-        )  # This function is called to move init guess within the bounds if not in bounds
-
-        return True
-
-    def advance_window_initial_guess_controls(self, sol, n_cycles_simultaneous=None):
-        # --- Get control results --- #
[…]
