### 2026-02-26 69e30f6
reworked the cose fun
