### 2026-02-27 8297273
remove triceps
