+++ examples/fes_multibody/cycling/cost_functions.py
@@ -151,8 +151,8 @@ class CustomCostFunctions:
                 "power": "2",
                 "state": "fats",
             },
-            "minimize_root_mean_square_fatigue_decay": {
-                "function": self.minimize_root_mean_square_fatigue_decay,
+            "minimize_fatigue_decay": {
+                "function": self.minimize_fatigue_decay,
                 "index": 21,
                 "latex": r"\phi_{21} = \left(\frac{1}{M}\sum_{t=1}^{M}\left(\frac{A_{t+1} - A_t}{-A_{\text{rest}} / alpha_a}\right)^{2}\right)^{\tfrac{1}{2}}",
                 "description": "Minimize the root mean square of muscle fatigue decay",
@@ -698,7 +698,7 @@ class CustomCostFunctions:
         return rms_fatigue
 
     @staticmethod
-    def minimize_root_mean_square_fatigue_decay(controller: PenaltyController) -> MX:
+    def minimize_fatigue_decay(controller: PenaltyController) -> MX:
         """
         Minimize the root-mean-square fatigue decay.
 
@@ -714,15 +714,14 @@ class CustomCostFunctions:
         muscle_name_list = controller.model.bio_model.muscle_names
         dA = CustomCostFunctions.calculate_dA(controller)
 
-        eps = 1e-8
         muscle_fatigue_decay = vertcat(
             *[
-                (dA[x]) ** 2
+                -dA[x]
                 for x in range(len(muscle_name_list))
             ]
         )
-        rms_fatigue = (sum1(muscle_fatigue_decay) / len(muscle_name_list) + eps) ** 0.5
-        return rms_fatigue
+        avg_fatigue = sum1(muscle_fatigue_decay) / len(muscle_name_list)
+        return avg_fatigue
 
     @staticmethod
     def minimize_peak_fatigue_decay(controller: PenaltyController) -> MX:
@@ -743,7 +742,7 @@ class CustomCostFunctions:
 
         muscle_fatigue_decay = vertcat(
             *[
-                (dA[x])
+                (-dA[x])
                 for x in range(len(muscle_name_list))
             ]
         )
@@ -776,7 +775,7 @@ class CustomCostFunctions:
             ]
         )
 
-        rms_fatigue = (sum1(muscle_fatigue_decay) / len(muscle_name_list) + eps) ** 0.5
+        rms_fatigue = (sum1(muscle_fatigue_decay) / (len(muscle_name_list)-1) + eps) ** 0.5
         return rms_fatigue
 
     @staticmethod
@@ -808,12 +807,12 @@ class CustomCostFunctions:
         muscle_fatigue_decay = vertcat(
             *[
                 (1 + tanh(4 * (-dA_nomalized[x]))) ** 2
-                for x in range(len(muscle_name_list))
+                for x in range(len(muscle_name_list)-1)
             ]
         )
 
         eps = 1e-8
-        rms_fatigue = (sum1(muscle_fatigue_decay) / len(muscle_name_list)+ eps) ** 0.5
+        rms_fatigue = (sum1(muscle_fatigue_decay) / (len(muscle_name_list)-1)+ eps) ** 0.5
         return rms_fatigue
 
     @staticmethod
@@ -848,7 +847,7 @@ class CustomCostFunctions:
         dA_nomalized = vertcat(
             *[
                 if_else(dA[x] < 0, dA[x] / max_dA_fatigue[x], dA[x] / max_dA_recovery[x])
-                for x in range(len(muscle_name_list))
+                for x in range(len(muscle_name_list)-1)
             ]
         )
 
@@ -859,7 +858,7 @@ class CustomCostFunctions:
             ]
         )
[…]
