+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -1119,9 +1119,9 @@ if __name__ == "__main__":
             # ["minimize_fatigue_decay"],
             # ["minimize_peak_fatigue_decay"],
             # ["minimize_root_mean_square_tanh_fatigue_decay"],
-            ["minimize_root_mean_square_tanh_fatigue_decay_norm"],
-            ["minimize_root_mean_square_weighted_tanh_fatigue_decay"],
-
+            # ["minimize_root_mean_square_tanh_fatigue_decay_norm"],
+            # ["minimize_root_mean_square_weighted_tanh_fatigue_decay"],
+            ["minimize_average_tanh_fatigue_decay"],
 
             # ["minimize_root_mean_square_weighted_tanh_fatigue_decay"],
 
