### 2026-03-06 d86619c
avg
