### 2026-03-06 #8989 @Kev1CO
Hi @arfon, @rkurchin,
Any news on when an editor will be assigned to our article and enter review?
