### 2026-03-06 441782e
normed fatigue
