+++ examples/fes_multibody/cycling/cycling_bayesian_mhe.py
@@ -13,7 +13,7 @@ from skopt import gp_minimize
 from skopt.space import Real
 from skopt.utils import use_named_args
 from skopt.callbacks import CheckpointSaver
-from casadi import MX, vertcat, sum1
+from casadi import MX, vertcat, sum1, if_else, mmax, tanh
 
 from bioptim import (
     ObjectiveList,
@@ -27,6 +27,8 @@ from bioptim import (
 )
 
 import cycling_pulse_width_mhe as base
+
+from cost_functions import CustomCostFunctions
 
 
 def minimize_root_mean_square_fatigue(controller: PenaltyController, muscle_weights: list) -> MX:
@@ -67,6 +69,45 @@ def minimize_root_mean_pw(controller: PenaltyController, muscle_weights: list) -
     return rms_activation
 
 
+def minimize_avg_fatigue_recovery(controller: PenaltyController, muscle_weights: list) -> MX:
+    """
+    Minimize the average fatigue based on recovery.
+    """
+    muscle_name_list = controller.model.bio_model.muscle_names
+    dA = CustomCostFunctions.calculate_dA(controller)
+    A_rest = [controller.model.muscles_dynamics_model[x].a_scale for x in range(len(muscle_name_list))]
+
+    max_dA_fatigue = [72.2, 61.2, 85.7, 92.3]
+    max_dA_recovery = [2.3, 3.0, 14.8, 35.6]
+    A_min = [41, 70, 379, 932]
+
+    with_triceps = True
+    muscle_range = 4 if with_triceps else 3
+
+    dA_nomalized = vertcat(
+        *[
+            if_else(dA[x] < 0, dA[x] / max_dA_fatigue[x], dA[x] / max_dA_recovery[x])
+            for x in range(muscle_range)
+        ]
+    )
+
+    A_t = vertcat(*[controller.states["A_" + muscle_name_list[x]].cx for x in range(len(muscle_name_list))])
+    fatigue = [((A_rest[i] - A_t[i]) / (A_rest[i] - A_min[i])) for i in range(muscle_range)]
+
+    muscle_fatigue_decay = vertcat(
+        *[
+            muscle_weights[x] *  fatigue[x] * (1 + tanh(-dA_nomalized[x]))
+            for x in range(muscle_range)
+        ]
+    )
+
+    avg_fatigue = sum1(muscle_fatigue_decay) / muscle_range
+    return avg_fatigue
+
+
+
+
+
 def set_objective_functions(muscle_fatigue_key, cost_fun_weight):
     objective_functions = ObjectiveList()
 
@@ -82,7 +123,7 @@ def set_objective_functions(muscle_fatigue_key, cost_fun_weight):
             raise ValueError(f"cost_fun_weight must be length 1 or {len(muscle_fatigue_key)}, got {len(cost_fun_weight)}")
 
     objective_functions.add(
-        minimize_root_mean_pw, #minimize_root_mean_square_fatigue,
+        minimize_avg_fatigue_recovery, #minimize_root_mean_pw, #minimize_root_mean_square_fatigue,
         custom_type=ObjectiveFcn.Lagrange,
         muscle_weights=weights,
         node=Node.ALL,
