### 2026-03-09 19d78fa
fatigue indicator
