+++ examples/fes_multibody/cycling/cost_functions.py
@@ -1,46 +1,43 @@
-from casadi import MX, vertcat, sum1, mmax, fabs, sign, tanh, if_else, log, exp, DM, dot, mmax
+from casadi import MX, vertcat, sum1, fabs, sign, tanh, if_else, log, exp, DM, dot, mmax
 from bioptim import PenaltyController
 from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
 
 class CustomCostFunctions:
     def __init__(self):
         self.dict_functions = {
+            # --- Pulse width --- #
             "minimize_average_activation": {
                 "function": self.minimize_average_activation,
                 "index": 1,
-                "latex": r"\phi_1 = \frac{1}{M}\sum_{m=1}^{M} a^{m}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
-                "description": "Minimize the average muscle activation",
+                "description": "Minimize the average fes activation",
                 "power": "1",
                 "state": "pw",
             },
             "minimize_root_mean_square_activation": {
                 "function": self.minimize_root_mean_square_activation,
                 "index": 2,
-                "latex": r"\phi_2 = \left(\frac{1}{M}\sum_{m=1}^{M} (a^{m})^{2}\right)^{\tfrac{1}{2}}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
-                "description": "Minimize the root mean square of muscle activation",
+                "description": "Minimize the root mean square of fes activation",
                 "power": "2",
                 "state": "pw",
             },
             "minimize_cubic_average_activation": {
                 "function": self.minimize_cubic_average_activation,
                 "index": 3,
-                "latex": r"\phi_3 = \left(\frac{1}{M}\sum_{m=1}^{M} (a^{m})^{3}\right)^{\tfrac{1}{3}}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
-                "description": "Minimize the cubic average of muscle activation",
+                "description": "Minimize the cubic average of fes activation",
                 "power": "3",
                 "state": "pw",
             },
             "minimize_peak_activation": {
                 "function": self.minimize_peak_activation,
                 "index": 4,
-                "latex": r"\phi_4 = \max_{m=1,\ldots,M} \; a^{m}, \quad a^{m}=\frac{f^{m}-f^{m}_{\min}}{f^{m}_{\max}-f^{m}_{\min}}",
-                "description": "Minimize the peak of muscle activation",
+                "description": "Minimize the peak of fes activation",
                 "power": r"\infty",
                 "state": "pw",
             },
+            # --- Force --- #
             "minimize_average_force": {
                 "function": self.minimize_average_force,
                 "index": 5,
-                "latex": r"\phi_5 = \frac{1}{M}\sum_{m=1}^{M} f^{m}",
                 "description": "Minimize the average muscle force",
                 "power": "1",
                 "state": r"F^{m}",
@@ -48,7 +45,6 @@ class CustomCostFunctions:
             "minimize_root_mean_square_force": {
                 "function": self.minimize_root_mean_square_force,
                 "index": 6,
-                "latex": r"\phi_6 = \left(\frac{1}{M}\sum_{m=1}^{M} (f^{m})^{2}\right)^{\tfrac{1}{2}}",
                 "description": "Minimize the root mean square of muscle force",
                 "power": "2",
                 "state": r"F^{m}",
[…]
