+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -40,7 +40,6 @@ from bioptim import (
     VariableScaling,
 )
 from cocofest import (
-    CustomObjective,
     DingModelPulseWidthFrequencyWithFatigue,
     FesMskModel,
     inverse_kinematics_cycling,
@@ -295,8 +294,6 @@ class MyCyclicNMPC(FesNmpcMsk):
 # --------------------#
 #    OCP functions    #
 # --------------------#
-
-
 def prepare_nmpc(
     model: BiorbdModel | FesMskModel,
     mhe_info: dict,
@@ -319,7 +316,7 @@ def prepare_nmpc(
     external_force = cycling_info["resistive_torque"]
     # --- Cost function info --- #
     objective_fun_dict = {"cost_fun_key": simulation_conditions["cost_fun_key"],
-                          "cost_fun_weight": simulation_conditions["cost_fun_weight"],
+                          "cost_fun_weight": 10000,
                           "individual_quadratic": True}
     # --- Pickle file info --- #
     initial_guess_path = simulation_conditions["init_guess_file_path"]
@@ -355,9 +352,6 @@ def prepare_nmpc(
         ode_solver=ode_solver,
         init_file_path=initial_guess_path,
     )
-
-    # --- Set states scaling --- #
-    # x_scaling = set_x_scaling(bio_model=model)  # Less efficient
 
     # --- Set controls --- #
     u_bounds, u_init, u_scaling = set_u_bounds_and_init(model, window_n_shooting, init_file_path=initial_guess_path)
@@ -397,7 +391,6 @@ def prepare_nmpc(
         constraints=constraints,
         x_bounds=x_bounds,
         x_init=x_init,
-        # x_scaling=x_scaling,
         u_bounds=u_bounds,
         u_init=u_init,
         u_scaling=u_scaling,
@@ -416,7 +409,7 @@ def set_external_forces(n_shooting, external_force_dict, force_name):
     reshape_values_array = np.tile(external_force_array[:, np.newaxis], (1, n_shooting))
     external_force_set.add_torque(
         segment=external_force_dict["Segment_application"], values=reshape_values_array, force_name=force_name
-    )  # warning forloop different force name
+    )
     numerical_time_series = {"external_forces": external_force_set.to_numerical_time_series()}
     return numerical_time_series, external_force_set
 
@@ -450,7 +443,6 @@ def set_q_qdot_init(
     else:
         # --- Chose the biorbd model to init the inverse kinematics --- #
         biorbd_model_path = "../../msk_models/Wu/Modified_Wu_Shoulder_Model_Cycling_for_IK.bioMod"
-        # biorbd_model_path = "../../msk_models/Seth/Modified_UL_Seth_2D_Cycling_for_IK.bioMod"
         n_shooting = (
             n_shooting * (ode_solver.polynomial_degree + 1)
             if isinstance(ode_solver, OdeSolver.COLLOCATION)
@@ -546,19 +538,6 @@ def set_x_bounds(
     )
 
     return x_bounds, x_init
-
-
-# Currently not used
-def set_x_scaling(bio_model) -> VariableScalingList:
-    x_scaling = VariableScalingList()
-    model_list = bio_model.muscles_dynamics_model
-    prefix_key_list = ["Cn_", "A_", "Tau1_", "Km_"]
-    scaling_value_list = [1 / 100, 1000, 1 / 100, 1 / 10]
-    for i in range(len(model_list)):
-        for j in range(len(prefix_key_list)):
-            key = prefix_key_list[j] + model_list[i].muscle_name
-            x_scaling.add(key=key, scaling=[scaling_value_list[j]])
-    return x_scaling
 
 
[…]
