+++ examples/fes_multibody/cycling/cost_functions.py
@@ -1,4 +1,4 @@
-from casadi import MX, vertcat, sum1, fabs, sign, tanh, if_else, log, exp, DM, dot, mmax
+from casadi import MX, vertcat, sum1, fabs, sign, tanh, if_else, log, exp, DM, dot, mmax, cos, sin
 from bioptim import PenaltyController
 from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
 
@@ -139,7 +139,13 @@ class CustomCostFunctions:
                 "power": "1",
                 "state": "A_recovery",
             },
-
+            "minimize_useful_torque_fatigue_tradeoff": {
+                "function": self.minimize_useful_torque_fatigue_tradeoff,
+                "index": 21,
+                "description": "Minimize the average fatigue and recovery based on torque efficiency",
+                "power": "1",
+                "state": "tau_eff",
+            },
             "minimize_peak": {
                 "function": self.minimize_peak,
                 "index": 99,
@@ -645,6 +651,79 @@ class CustomCostFunctions:
         avg_fatigue = sum1(muscle_fatigue_decay) / muscle_range
         return avg_fatigue
 
+    @staticmethod
+    def minimize_useful_torque_fatigue_tradeoff(
+            controller: PenaltyController,
+    ) -> MX:
+        """
+        Minimize the fatigue and recuperation based on torque efficiency.
+
+        Parameters
+        ----------
+        controller: PenaltyController
+            The penalty node elements
+
+        Returns
+        -------
+        The average fatigue and recuperation based on torque efficiency
+        """
+        # --- Weights --- #
+        lambda_fatigue = 1.0
+        lambda_dA = 1.0
+        kappa_recovery = 1.0
+
+        # --- Get all information --- #
+        muscle_names, q, qdot, F, A, A_rest, tau_fat, alpha_a, fmax, dA = CustomCostFunctions.get_muscle_quantities(controller)
+
+        # --- Calculate useful torque based on force and muscle efficiency --- #
+        gamma = CustomCostFunctions.useful_gain_from_angle(q[2])
+        tau_use = vertcat(*[
+            if_else(gamma[i] * F[i] > 0, gamma[i] * F[i], 0)
+            for i in range(F.shape[0])
+        ])
+
+        # --- Fatigue and recovery --- #
+        cap = A / A_rest
+        fatigue = 1 - cap
+        dA_norm = CustomCostFunctions.normalized_dA(dA, A_rest, tau_fat, alpha_a, fmax)
+
+        # --- Build the adaptive weight --- #
+        adaptive_weight = vertcat(*[
+            1
+            + fatigue[i]
+            + kappa_recovery * if_else(dA_norm[i] < 0, -dA_norm[i], 0)
+            for i in range(F.shape[0])
+        ])
+
+        # --- Calculate the cost function --- #
+        use_term = vertcat(*[
+            adaptive_weight[i] * (tau_use[i] / fmax[i]) ** 2
+            for i in range(F.shape[0])
+        ])
+
+        fatigue_term = vertcat(*[
+            lambda_fatigue * fatigue[i] ** 2
+            for i in range(F.shape[0])
+        ])
+
+        dA_term = vertcat(*[
+            lambda_dA * if_else(dA_norm[i] < 0, (-dA_norm[i]) ** 2, 0)
+            for i in range(F.shape[0])
+        ])
+
+        return sum1(use_term + fatigue_term + dA_term) / len(muscle_names)
+
+    @staticmethod
+    def minimize_terminal_fatigue_reserve(controller: PenaltyController) -> MX:
+        muscle_names = controller.model.bio_model.muscle_names
[…]
