+++ examples/fes_multibody/cycling/cost_functions.py
@@ -677,10 +677,29 @@ class CustomCostFunctions:
 
         # --- Calculate useful torque based on force and muscle efficiency --- #
         gamma = CustomCostFunctions.useful_gain_from_angle(q[2])
-        tau_use = vertcat(*[
-            if_else(gamma[i] * F[i] > 0, gamma[i] * F[i], 0)
-            for i in range(F.shape[0])
-        ])
+
+        gamma_tanh = vertcat(
+            *[(1 + tanh(10*gamma[x]))
+              for x in range(F.shape[0])
+              ]
+        )
+
+
+        # tau_use = vertcat(*[
+        #     if_else(gamma[i] * F[i] > 0, gamma[i] * F[i], 0)
+        #     for i in range(F.shape[0])
+        # ])
+
+        muscle_motion_contrib = [0.16259391348727756, 0.32644352775103314, 0.06473043181533218, 0.07310528260037236]
+        muscle_motion_contrib_norm = [muscle_contrib/max(muscle_motion_contrib) for muscle_contrib in muscle_motion_contrib]
+
+        muscle_fatigability = [-alpha_a[i]/A_rest[i] for i in range(F.shape[0])]
+        muscle_fatigability_norm = [muscle_fat/max(muscle_fatigability) for muscle_fat in muscle_fatigability]
+
+        muscle_importance_weight = [muscle_motion_contrib_norm[i] * muscle_fatigability_norm[i] for i in range(F.shape[0])]
+
+        torque_distribution = [(muscle_importance_weight[i] / (gamma_tanh[i] + 1e-8)) * F[i] for i in range(F.shape[0])]
+
 
         # --- Fatigue and recovery --- #
         cap = A / A_rest
@@ -696,22 +715,38 @@ class CustomCostFunctions:
         ])
 
         # --- Calculate the cost function --- #
-        use_term = vertcat(*[
-            adaptive_weight[i] * (tau_use[i] / fmax[i]) ** 2
-            for i in range(F.shape[0])
-        ])
+        # use_term = vertcat(*[
+        #     adaptive_weight[i] * (tau_use[i] / fmax[i]) ** 2
+        #     for i in range(F.shape[0])
+        # ])
 
+        # fatigue_term = vertcat(*[
+        #     lambda_fatigue * fatigue[i] ** 2
+        #     for i in range(F.shape[0])
+        # ])
         fatigue_term = vertcat(*[
-            lambda_fatigue * fatigue[i] ** 2
+            lambda_fatigue * (10 ** (4 * fatigue[i]))
             for i in range(F.shape[0])
         ])
 
-        dA_term = vertcat(*[
-            lambda_dA * if_else(dA_norm[i] < 0, (-dA_norm[i]) ** 2, 0)
-            for i in range(F.shape[0])
-        ])
+        # dA_term = vertcat(*[
+        #     lambda_dA * if_else(dA_norm[i] < 0, (-dA_norm[i]) ** 2, 0)
+        #     for i in range(F.shape[0])
+        # ])
+        dA_term = vertcat(
+            *[(1 + tanh(-dA_norm[x]))
+                for x in range(F.shape[0])
+            ]
+        )
 
-        return sum1(use_term + fatigue_term + dA_term) / len(muscle_names)
+        # return sum1(use_term + fatigue_term + dA_term) / len(muscle_names)
+
+        cost_fun = vertcat(
+            *[torque_distribution[i] * fatigue_term[i] * dA_term[i]
+                for i in range(F.shape[0])
+            ]
+        )
+        return sum1(cost_fun) / len(muscle_names)
 
     @staticmethod
     def minimize_terminal_fatigue_reserve(controller: PenaltyController) -> MX:
@@ -842,10 +877,10 @@ class CustomCostFunctions:
[…]
