+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -1130,8 +1130,8 @@ if __name__ == "__main__":
             # ["minimize_root_mean_square_muscle_power"],
 
             # --- Recovery --- #
-            # ["minimize_average_fatigue_and_recovery"],
-            ["minimize_useful_torque_fatigue_tradeoff"],
+            ["minimize_average_fatigue_and_recovery_2"],
+            # ["minimize_useful_torque_fatigue_tradeoff"],
             ]
         },
         init_guess=False,
