+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -1131,6 +1131,8 @@ if __name__ == "__main__":
 
             # --- Recovery --- #
             ["minimize_average_fatigue_and_recovery_3"],
+            ["minimize_average_fatigue_and_recovery_4"],
+            ["minimize_torque_deficit_area"],
             # ["minimize_useful_torque_fatigue_tradeoff"],
             ]
         },
