+++ examples/fes_multibody/cycling/cost_functions.py
@@ -167,6 +167,24 @@ class CustomCostFunctions:
                 "state": "A_recovery",
             },
 
+            "minimize_average_fatigue_and_recovery_4": {
+                "function": self.minimize_average_fatigue_and_recovery_4,
+                "index": 24,
+                "description": "Minimize the average fatigue and recovery",
+                "power": "1",
+                "state": "A_recovery",
+            },
+
+            "minimize_torque_deficit_area": {
+                "function": self.minimize_torque_deficit_area,
+                "index": 25,
+                "latex": r"\phi_{24}=\left(\tau_{\mathrm{thr}}-\sum_{m=1}^{M}[\gamma^m(\theta)F^m]_+\right)_+|\dot{\theta}|",
+                "description": "Minimize the crank-angle area where useful torque is below threshold",
+                "power": "1",
+                "state": r"\tau_{\mathrm{def}}",
+            },
+
+
             "minimize_peak": {
                 "function": self.minimize_peak,
                 "index": 99,
@@ -740,14 +758,57 @@ class CustomCostFunctions:
         # --- Get all information --- #
         muscle_names, q, qdot, F, A, A_rest, tau_fat, alpha_a, fmax, dA = CustomCostFunctions.get_muscle_quantities(
             controller)
+        A_min = [41, 70, 379, 932]
 
         # --- Fatigue --- #
-        weight_fatigue = [1.0, 0.0, 0.7316463404270923, 0.0]
-        cost_fatigue = [(A_rest[i] - A[i])**2 for i in range(F.shape[0])]
+        # weight_fatigue = [1.0, 0.0, 0.7316463404270923, 0.0]
+        weight_fatigue = [0.9536847967328755, 0.9330995253122598, 1.0, 0.0]
+        cost_fatigue = [(1 - ((A_rest[i] - A[i])/A_min[i]))**2 for i in range(F.shape[0])]
 
         # --- Cost function --- #
         cost = vertcat(*[weight_fatigue[i] * cost_fatigue[i] for i in range(F.shape[0])])
-        rms_cost = (sum1(cost) / F.shape[0]) ** (1/2)
+        rms_cost = (sum1(cost) / F.shape[0] + 1e-8) ** (1/2)
+        return rms_cost
+
+    @staticmethod
+    def minimize_average_fatigue_and_recovery_4(controller: PenaltyController) -> MX:
+        """
+        Minimize the average fatigue and recuperation.
+
+        Parameters
+        ----------
+        controller: PenaltyController
+            The penalty node elements
+
+        Returns
+        -------
+        The average fatigue and recuperation
+        """
+
+        # --- Get all information --- #
+        muscle_names, q, qdot, F, A, A_rest, tau_fat, alpha_a, fmax, dA = CustomCostFunctions.get_muscle_quantities(
+            controller)
+        A_min = [41, 70, 379, 932]
+        max_dA_recovery = [A_rest[x] / tau_fat[x] for x in range(F.shape[0])]
+        max_dA_fatigue = [-(alpha_a[x] * fmax[x]) for x in range(F.shape[0])]
+
+        # --- Fatigue --- #
+        weight_fatigue = [0.9536847967328755, 0.9330995253122598, 1.0, 0.0]
+        cost_fatigue = [(1 - ((A_rest[i] - A[i]) / A_min[i])) ** 2 for i in range(F.shape[0])]
+
+
+        # --- Recovery --- #
+        dA_nomalized = vertcat(
+            *[
+                if_else(dA[x] < 0, dA[x] / max_dA_fatigue[x], dA[x] / max_dA_recovery[x])
+                for x in range(F.shape[0])
+            ]
+        )
+        cost_recovery = [(1 + tanh(-dA_nomalized[i])) for i in range(F.shape[0])]
+
[…]
