### 2026-03-16 #8989 @Kev1CO
Hi @borgesaugusto,
Thank you very much for accepting the editor position of our paper.

I do have a few suggestions for the reviewing process.

- Nat Wannawas https://scholar.google.com/citations?user=LGmXgtUAAAAJ
Email: nat.wa@wu.ac.th

- Jason K. Moore https://scholar.google.com/citations?user=i9c-QOYAAAAJ
Email: j.k.moore@tudelft.nl

- Jules Ronné https://scholar.google.com/citations?user=8lfU0h4AAAAJ
Email: jules.ronne@ik.me

- Maël GALLOIS https://scholar.google.com/citations?user=rzRg6UIAAAAJ
Email: mael.gallois@inria.fr

- Hugh Sorby https://profiles.auckland.ac.nz/h-sorby
Email: h.sorby@auckland.ac.nz

Feel free to contact me if needed.
kevin.co@umontreal.ca
