+++ examples/fes_multibody/cycling/cost_functions.py
@@ -761,7 +761,7 @@ class CustomCostFunctions:
         A_min = [41, 70, 379, 932]
 
         # --- Fatigue --- #
-        weight_fatigue = [1.0, 0.6104101922170402, 0.754209800965523, 0.0]
+        weight_fatigue = vertcat([1.0, 0.6104101922170402, 0.754209800965523, 0.0]) * 10000
         cost_fatigue = [((A_rest[i] - A[i]) / (A_rest[i] - A_min[i] + 1e-8)) ** 2 for i in range(F.shape[0])]
 
         # --- Cost function --- #
@@ -793,7 +793,7 @@ class CustomCostFunctions:
 
         # --- Fatigue --- #
         weight_fatigue = [1.0, 0.6104101922170402, 0.754209800965523, 0.0]
-        cost_fatigue = [((A_rest[i] - A[i]) / (A_rest[i] - A_min[i] + 1e-8)) ** 2 for i in range(F.shape[0])]
+        # cost_fatigue = [((A_rest[i] - A[i]) / (A_rest[i] - A_min[i] + 1e-8)) ** 2 for i in range(F.shape[0])]
 
         # --- Recovery --- #
         dA_nomalized = vertcat(
@@ -805,7 +805,8 @@ class CustomCostFunctions:
         cost_recovery = [(1 + tanh(-dA_nomalized[i])) for i in range(F.shape[0])]
 
         # --- Cost function --- #
-        cost = vertcat(*[weight_fatigue[i] * cost_fatigue[i] * cost_recovery[i] for i in range(F.shape[0])])
+        # cost = vertcat(*[weight_fatigue[i] * cost_fatigue[i] * cost_recovery[i] for i in range(F.shape[0])])
+        cost = vertcat(*[weight_fatigue[i] * cost_recovery[i] for i in range(F.shape[0])])
         rms_cost = (sum1(cost) / F.shape[0] + 1e-8) ** (1 / 2)
         return rms_cost
 
