### 2026-03-17 0bacbd5
bayesian cost
