+++ examples/fes_multibody/cycling/cost_functions.py
@@ -175,15 +175,26 @@ class CustomCostFunctions:
                 "state": "A_recovery",
             },
 
-            "minimize_torque_deficit_area": {
-                "function": self.minimize_torque_deficit_area,
+            "minimize_average_fatigue_and_recovery_5": {
+                "function": self.minimize_average_fatigue_and_recovery_5,
                 "index": 25,
-                "latex": r"\phi_{24}=\left(\tau_{\mathrm{thr}}-\sum_{m=1}^{M}[\gamma^m(\theta)F^m]_+\right)_+|\dot{\theta}|",
-                "description": "Minimize the crank-angle area where useful torque is below threshold",
+                "description": "Minimize the average fatigue and recovery",
                 "power": "1",
-                "state": r"\tau_{\mathrm{def}}",
+                "state": "A_recovery",
             },
 
+            # "minimize_torque_deficit_area": {
+            #     "function": self.minimize_torque_deficit_area,
+            #     "index": 25,
+            #     "latex": r"\phi_{24}=\left(\tau_{\mathrm{thr}}-\sum_{m=1}^{M}[\gamma^m(\theta)F^m]_+\right)_+|\dot{\theta}|",
+            #     "description": "Minimize the crank-angle area where useful torque is below threshold",
+            #     "power": "1",
+            #     "state": r"\tau_{\mathrm{def}}",
+            # },
+            #
+            #
+            #
+            #
 
             "minimize_peak": {
                 "function": self.minimize_peak,
@@ -755,20 +766,33 @@ class CustomCostFunctions:
         The average fatigue and recuperation
         """
 
-        # --- Get all information --- #
-        muscle_names, q, qdot, F, A, A_rest, tau_fat, alpha_a, fmax, dA = CustomCostFunctions.get_muscle_quantities(
-            controller)
-        # A_min = [41, 70, 379, 932]
+        # # --- Get all information --- #
+        # muscle_names, q, qdot, F, A, A_rest, tau_fat, alpha_a, fmax, dA = CustomCostFunctions.get_muscle_quantities(
+        #     controller)
+        # # A_min = [41, 70, 379, 932]
+        #
+        # # --- Fatigue --- #
+        # weight_fatigue = vertcat([100.0, 1, 5, 0.0])
+        # # cost_fatigue = [((A_rest[i] - A[i]) / (A_rest[i] - A_min[i] + 1e-8)) ** 2 for i in range(F.shape[0])]
+        # cost_fatigue = [(A_rest[i] - A[i]) ** 2 for i in range(F.shape[0])]
+        #
+        # # --- Cost function --- #
+        # cost = vertcat(*[weight_fatigue[i] * cost_fatigue[i] for i in range(F.shape[0])])
+        # rms_cost = (sum1(cost) / F.shape[0] + 1e-8) ** (1/2)
 
-        # --- Fatigue --- #
-        weight_fatigue = vertcat([100.0, 1, 5, 0.0])
-        # cost_fatigue = [((A_rest[i] - A[i]) / (A_rest[i] - A_min[i] + 1e-8)) ** 2 for i in range(F.shape[0])]
-        cost_fatigue = [(A_rest[i] - A[i]) ** 2 for i in range(F.shape[0])]
-
-        # --- Cost function --- #
-        cost = vertcat(*[weight_fatigue[i] * cost_fatigue[i] for i in range(F.shape[0])])
-        rms_cost = (sum1(cost) / F.shape[0] + 1e-8) ** (1/2)
-        return rms_cost
+        muscle_weights = vertcat([100.0, 1, 5, 0.0])
+        eps = 1e-8
[…]
