+++ examples/fes_multibody/cycling/cost_functions.py
@@ -143,58 +143,13 @@ class CustomCostFunctions:
                 "power": "1",
                 "state": "A_recovery",
             },
-            "minimize_useful_torque_fatigue_tradeoff": {
-                "function": self.minimize_useful_torque_fatigue_tradeoff,
-                "index": 21,
-                "description": "Minimize the average fatigue and recovery based on torque efficiency",
-                "power": "1",
-                "state": "tau_eff",
-            },
-
             "minimize_average_fatigue_and_recovery_2": {
                 "function": self.minimize_average_fatigue_and_recovery_2,
-                "index": 22,
+                "index": 21,
                 "description": "Minimize the average fatigue and recovery",
                 "power": "1",
                 "state": "A_recovery",
             },
-
-            "minimize_average_fatigue_and_recovery_3": {
-                "function": self.minimize_average_fatigue_and_recovery_3,
-                "index": 23,
-                "description": "Minimize the average fatigue and recovery",
-                "power": "1",
-                "state": "A_recovery",
-            },
-
-            "minimize_average_fatigue_and_recovery_4": {
-                "function": self.minimize_average_fatigue_and_recovery_4,
-                "index": 24,
-                "description": "Minimize the average fatigue and recovery",
-                "power": "1",
-                "state": "A_recovery",
-            },
-
-            "minimize_average_fatigue_and_recovery_5": {
-                "function": self.minimize_average_fatigue_and_recovery_5,
-                "index": 25,
-                "description": "Minimize the average fatigue and recovery",
-                "power": "1",
-                "state": "A_recovery",
-            },
-
-            # "minimize_torque_deficit_area": {
-            #     "function": self.minimize_torque_deficit_area,
-            #     "index": 25,
-            #     "latex": r"\phi_{24}=\left(\tau_{\mathrm{thr}}-\sum_{m=1}^{M}[\gamma^m(\theta)F^m]_+\right)_+|\dot{\theta}|",
-            #     "description": "Minimize the crank-angle area where useful torque is below threshold",
-            #     "power": "1",
-            #     "state": r"\tau_{\mathrm{def}}",
-            # },
-            #
-            #
-            #
-            #
 
             "minimize_peak": {
                 "function": self.minimize_peak,
@@ -667,104 +622,6 @@ class CustomCostFunctions:
         -------
         The average fatigue and recuperation
         """
-        muscle_name_list = controller.model.bio_model.muscle_names
-        dA = CustomCostFunctions.calculate_dA(controller)
-        A_rest = [controller.model.muscles_dynamics_model[x].a_scale for x in range(len(muscle_name_list))]
-        tau_fat = [controller.model.muscles_dynamics_model[x].tau_fat for x in range(len(muscle_name_list))]
-        alpha_a = [controller.model.muscles_dynamics_model[x].alpha_a for x in range(len(muscle_name_list))]
-        fmax = [controller.model.muscles_dynamics_model[x].fmax for x in range(len(muscle_name_list))]
-        normed_a = [A / max(A_rest) for A in A_rest]
-
[…]
