+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -1114,8 +1114,9 @@ if __name__ == "__main__":
             # ["minimize_root_mean_square_muscle_power"],
 
             # --- Recovery --- #
-            ["minimize_average_fatigue_and_recovery"],
+            # ["minimize_average_fatigue_and_recovery"],
             # ["minimize_average_fatigue_and_recovery_2"],
+            ["minimize_balanced_fatigue_by_contribution"],
             ]
         },
         init_guess=False,
