+++ examples/fes_multibody/cycling/cost_functions.py
@@ -212,9 +212,11 @@ class CustomCostFunctions:
         """
         eps = 1e-8
         muscle_name_list = controller.model.bio_model.muscle_names
+        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
+
         if isinstance(controller.model.muscles_dynamics_model[0], DingModelPulseWidthFrequency):
             stim_charge = vertcat(
-                *[
+                *[  weight_fatigue[x] *
                     ((controller.controls["last_pulse_width_" + muscle_name_list[x]].cx -
                      controller.ocp.nlp[0].u_bounds["last_pulse_width_" + muscle_name_list[x]].min[0][0])
                     / (controller.ocp.nlp[0].u_bounds["last_pulse_width_" + muscle_name_list[x]].max[0][0] -
@@ -331,8 +333,10 @@ class CustomCostFunctions:
         """
         eps = 1e-8
         muscle_name_list = controller.model.bio_model.muscle_names
+        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
+        
         muscle_force = vertcat(
-            *[
+            *[  weight_fatigue[x] *
                 controller.states["F_" + muscle_name_list[x]].cx ** 2
                 for x in range(len(muscle_name_list))
             ]
@@ -429,8 +433,9 @@ class CustomCostFunctions:
         """
         eps = 1e-8
         muscle_name_list = controller.model.bio_model.muscle_names
+        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
         muscle_stress = vertcat(
-            *[
+            *[  weight_fatigue[x] *
                 (controller.states["F_" + muscle_name_list[x]].cx / controller.model.muscles_dynamics_model[x].pcsa) ** 2
                 for x in range(len(muscle_name_list))
             ]
@@ -603,11 +608,12 @@ class CustomCostFunctions:
         """
         eps = 1e-8
         muscle_name_list = controller.model.bio_model.muscle_names
+        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
         muscle_velocity = controller.model.muscle_velocity()(
             controller.states["q"].cx, controller.states["qdot"].cx, controller.parameters.cx
         )
         muscle_power = vertcat(
-            *[
+            *[  weight_fatigue[x] *
                 (controller.states["F_" + muscle_name_list[x]].cx * muscle_velocity[x]) ** 2
                 for x in range(len(muscle_name_list))
             ]
