### 2026-04-02 #8989 @davidpagnon
Hello, thanks for thinking of me!

I know about some other impressive work by the same lab, such as pyomeca and biorbd: I have no doubt that this toolbox will be equally well made and useful.

Unfortunately, I won't have time to review it in the near future, and I can see the authors have already waited for a while. I can suggest a few of my colleagues: @jc876, @joshcarter97, @andreabraschi, who are also more versed in Optimal Control approaches than I am.

All the best!
