+++ examples/fes_multibody/cycling/cost_functions.py
@@ -212,7 +212,7 @@ class CustomCostFunctions:
         """
         eps = 1e-8
         muscle_name_list = controller.model.bio_model.muscle_names
-        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
+        weight_fatigue = vertcat([1.00000000e+04, 1.55976591e+03, 4.66525639e+03, 1.00000000e-05])
 
         if isinstance(controller.model.muscles_dynamics_model[0], DingModelPulseWidthFrequency):
             stim_charge = vertcat(
@@ -333,7 +333,7 @@ class CustomCostFunctions:
         """
         eps = 1e-8
         muscle_name_list = controller.model.bio_model.muscle_names
-        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
+        weight_fatigue = vertcat([1.00000000e+04, 1.55976591e+03, 4.66525639e+03, 1.00000000e-05])
         
         muscle_force = vertcat(
             *[  weight_fatigue[x] *
@@ -433,7 +433,7 @@ class CustomCostFunctions:
         """
         eps = 1e-8
         muscle_name_list = controller.model.bio_model.muscle_names
-        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
+        weight_fatigue = vertcat([1.00000000e+04, 1.55976591e+03, 4.66525639e+03, 1.00000000e-05])
         muscle_stress = vertcat(
             *[  weight_fatigue[x] *
                 (controller.states["F_" + muscle_name_list[x]].cx / controller.model.muscles_dynamics_model[x].pcsa) ** 2
@@ -608,7 +608,7 @@ class CustomCostFunctions:
         """
         eps = 1e-8
         muscle_name_list = controller.model.bio_model.muscle_names
-        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
+        weight_fatigue = vertcat([1.00000000e+04, 1.55976591e+03, 4.66525639e+03, 1.00000000e-05])
         muscle_velocity = controller.model.muscle_velocity()(
             controller.states["q"].cx, controller.states["qdot"].cx, controller.parameters.cx
         )
@@ -643,7 +643,7 @@ class CustomCostFunctions:
 
         # --- Fatigue --- #
         # weight_fatigue = vertcat([1.0, 0.6104101922170402, 0.754209800965523, 0.0])
-        weight_fatigue = vertcat([10000.0, 2622.984911640199, 3295.935637664999, 1e-05])
+        weight_fatigue = vertcat([1.00000000e+04, 1.55976591e+03, 4.66525639e+03, 1.00000000e-05])
         cost_fatigue = [(A_rest[i] - A[i]) ** 2 for i in range(F.shape[0])]
 
         # --- Cost function --- #
