+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -49,6 +49,9 @@ from cocofest import (
 from examples.fes_multibody.cycling.cost_functions import CustomCostFunctions
 
 
+# --------------------#
+#    MHE functions    #
+# --------------------#
 class MyCyclicNMPC(FesNmpcMsk):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
@@ -309,7 +312,7 @@ def prepare_nmpc(
     ode_solver = mhe_info["ode_solver"]
     use_sx = mhe_info["use_sx"]
 
-    # --- Pickle file info --- #
+    # --- Initial guess file info --- #
     initial_guess_path = simulation_conditions["init_guess_file_path"]
 
     window_n_shooting = cycle_len * n_cycles_simultaneous
@@ -321,9 +324,15 @@ def prepare_nmpc(
     external_force = cycling_info["resistive_torque"]
 
     # --- Cost function info --- #
-    objective_fun_dict = {"cost_fun_key": simulation_conditions["cost_fun_key"],
-                          "cost_fun_weight": 10000,
-                          "individual_quadratic": True}
+    objective_fun_dict = {
+        "cost_fun_key": simulation_conditions["cost_fun_key"],
+        "cost_fun_weight": (
+            0.1
+            if "weight" in simulation_conditions["cost_fun_key"][0]
+            else 1 if "bayesian" in simulation_conditions["cost_fun_key"][0] else 10000
+        ),
+        "individual_quadratic": True,
+    }
 
     # --- Set dynamics --- #
     # --- External force numerical time series --- #
@@ -367,16 +376,26 @@ def prepare_nmpc(
     )
 
     # --- Set parameters (for minmax cost_function) --- #
-    if objective_fun_dict["cost_fun_key"] in [["minimize_peak_force"], ["minimize_peak_activation"], ["minimize_peak_muscle_stress"], ["minimize_peak_fatigue"], ["minimize_peak_fatigue_decay"]]:
-        parameters, parameters_bounds, parameters_init, parameters_objectives = set_parameters(ns=cycle_len, cycle=n_cycles_simultaneous, use_sx=use_sx)
+    if objective_fun_dict["cost_fun_key"] in [
+        ["minimize_peak_force"],
+        ["minimize_peak_activation"],
+        ["minimize_peak_muscle_stress"],
+        ["minimize_peak_fatigue"],
+        ["minimize_peak_fatigue_decay"],
+    ]:
+        parameters, parameters_bounds, parameters_init, parameters_objectives = set_parameters(
+            ns=cycle_len, cycle=n_cycles_simultaneous, use_sx=use_sx
+        )
     else:
         parameters = ParameterList(use_sx=use_sx)
         parameters_bounds = BoundsList()
         parameters_init = InitialGuessList()
         parameters_objectives = ParameterObjectiveList()
 
-     # --- Set constraints --- #
-    constraints = set_constraints(model, x_init["q"].init[2][0] - 2*np.pi, cycle_len, n_cycles_simultaneous, objective_fun_dict["cost_fun_key"])
+    # --- Set constraints --- #
+    constraints = set_constraints(
+        model, x_init["q"].init[2][0] - 2 * np.pi, cycle_len, n_cycles_simultaneous, objective_fun_dict["cost_fun_key"]
+    )
 
     # --- Update model for resistive torque --- #
     model = updating_model(model=model, external_force_set=external_force_set, parameters=parameters)
@@ -590,19 +609,24 @@ def set_constraints(bio_model, one_cycle_bound, cycle_len, n_simultaneous, objec
         axes=[Axis.X, Axis.Y],
     )
 
-    angle_slack = 0.174533/4  # 10 degrees in radiant
[…]
