+++ examples/fes_multibody/cycling/cost_functions.py
@@ -1,39 +1,40 @@
-from casadi import MX, vertcat, sum1, fabs, sign, tanh, if_else, log, exp, DM, dot, mmax, mmin, cos, sin
+from casadi import MX, vertcat, sum1, mmax
 from bioptim import PenaltyController
 from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
-from cocofest.models.hill_coefficients import (muscle_force_length_coefficient,
-                                               muscle_force_velocity_coefficient,
-                                               muscle_passive_force_coefficient)
+
+BAYESIAN_WEIGHT = vertcat([191790, 31609, 117259, 1])
+PHYSIOLOGICAL_WEIGHT = vertcat([191790, 31609, 117259, 1])
 
 
 class CustomCostFunctions:
     def __init__(self):
         self.dict_functions = {
+            # --- UNWEIGHTED --- #
             # --- Pulse width --- #
             "minimize_average_activation": {
                 "function": self.minimize_average_activation,
-                "index": 1,
+                "index": "1",
                 "description": "Minimize the average fes activation",
                 "power": "1",
                 "state": "pw",
             },
             "minimize_root_mean_square_activation": {
                 "function": self.minimize_root_mean_square_activation,
-                "index": 2,
+                "index": "2",
                 "description": "Minimize the root mean square of fes activation",
                 "power": "2",
                 "state": "pw",
             },
             "minimize_cubic_average_activation": {
                 "function": self.minimize_cubic_average_activation,
-                "index": 3,
+                "index": "3",
                 "description": "Minimize the cubic average of fes activation",
                 "power": "3",
                 "state": "pw",
             },
             "minimize_peak_activation": {
                 "function": self.minimize_peak_activation,
-                "index": 4,
+                "index": "4",
                 "description": "Minimize the peak of fes activation",
                 "power": r"\infty",
                 "state": "pw",
@@ -41,28 +42,28 @@ class CustomCostFunctions:
             # --- Force --- #
             "minimize_average_force": {
                 "function": self.minimize_average_force,
-                "index": 5,
+                "index": "5",
                 "description": "Minimize the average muscle force",
                 "power": "1",
                 "state": r"F^{m}",
             },
             "minimize_root_mean_square_force": {
                 "function": self.minimize_root_mean_square_force,
-                "index": 6,
+                "index": "6",
                 "description": "Minimize the root mean square of muscle force",
                 "power": "2",
                 "state": r"F^{m}",
             },
             "minimize_cubic_average_force": {
                 "function": self.minimize_cubic_average_force,
-                "index": 7,
+                "index": "7",
                 "description": "Minimize the cubic average of muscle force",
                 "power": "3",
                 "state": r"F^{m}",
             },
             "minimize_peak_force": {
                 "function": self.minimize_peak_force,
-                "index": 8,
+                "index": "8",
                 "description": "Minimize the peak muscle force",
                 "power": r"\infty",
                 "state": r"F^{m}",
@@ -70,57 +71,57 @@ class CustomCostFunctions:
             # --- Stress --- #
             "minimize_average_muscle_stress": {
                 "function": self.minimize_average_muscle_stress,
-                "index": 9,
-                "description":"Minimize the average muscle stress",
+                "index": "9",
+                "description": "Minimize the average muscle stress",
                 "power": "1",
                 "state": r"\sigma",
             },
             "minimize_root_mean_square_muscle_stress": {
                 "function": self.minimize_root_mean_square_muscle_stress,
[…]
