+++ examples/msk_models/model_viewer.py
@@ -6,6 +6,6 @@ The user can interact with the model by changing the joint angles using sliders.
 from pyorerun import LiveModelAnimation
 
 
-model_path = "arm26.bioMod"
+model_path = "Wu/Modified_Wu_Shoulder_Model_Cycling.bioMod"
 animation = LiveModelAnimation(model_path, with_q_charts=True)
 animation.rerun()

[…]
