+++ examples/fes_multibody/shoulder_abduction/abduction_fes_driven_init_guess.py
@@ -1,0 +1,580 @@
+"""
+This example will perform an optimal control program moving time horizon for an abduction/adduction motion driven by FES.
+"""
+
+import pickle
+from sys import platform
+from itertools import product
+from pathlib import Path
+
+import matplotlib.pyplot as plt
+import numpy as np
+from numpy.ma.extras import average
+
+from bioptim import (
+    BiorbdModel,
+    BoundsList,
+    ConstraintList,
+    ConstraintFcn,
+    CostType,
+    DynamicsList,
+    ExternalForceSetTimeSeries,
+    InitialGuessList,
+    InterpolationType,
+    MultiCyclicCycleSolutions,
+    MultiCyclicNonlinearModelPredictiveControl,
+    ObjectiveFcn,
+    ObjectiveList,
+    OdeSolver,
+    PhaseDynamics,
+    ParameterObjectiveList,
+    SolutionMerge,
+    Solution,
+    Solver,
+    ParameterList,
+    Node,
+    VariableScalingList,
+    OptimalControlProgram,
+)
+from cocofest import (
+    DingModelPulseWidthFrequencyWithFatigue,
+    FesMskModel,
+    OcpFesMsk,
+)
+from examples.fes_multibody.cycling.cost_functions import CustomCostFunctions
+
+
+# --------------------#
+#    OCP functions    #
+# --------------------#
+def prepare_ocp(
+    model: BiorbdModel | FesMskModel,
+    abduction_info: dict,
+    simulation_conditions: dict,
+    initial_guess_data,
+):
+    # --- Data from initial guess --- #
+    q = initial_guess_data["q"]
+    qdot = initial_guess_data["qdot"]
+    tau = initial_guess_data["tau"]
+    force = initial_guess_data["muscles_force"]
+
+    # --- Initialize parameters from dictionaries --- #
+    # --- Abduction info --- #
+    abd_range_config = abduction_info["range_config"]
+    external_force = abduction_info["resistive_torque"]
+    n_shooting = simulation_conditions["stimulation"]
+
+    ode_solver = OdeSolver.COLLOCATION(polynomial_degree=3, method="radau")
+
+    # --- Set dynamics --- #
+    # --- External force numerical time series --- #
+    numerical_time_series, external_force_set = set_external_forces(
+        n_shooting=n_shooting, external_force_dict=external_force, force_name="external_torque"
+    ) # No external torque applied, except to provide FES supplementary assistance
+    # --- Stimulation instant numerical time series --- #
+    numerical_data_time_series, stim_idx_at_node_list = model.muscles_dynamics_model[0].get_numerical_data_time_series(
+        n_shooting, 6
+    )
+    numerical_time_series.update(numerical_data_time_series)
+    # --- Dynamics --- #
+    dynamics = set_dynamics(model=model, numerical_time_series=numerical_time_series, ode_solver=ode_solver)
+
+    # --- Set states --- #
+    x_bounds, x_init = set_x_bounds(
+        model=model,
+        n_shooting=n_shooting,
+        abduction_range=abd_range_config,
+        ode_solver=ode_solver,
+        q = q,
+        qdot = qdot,
+    )
+
+    # --- Set controls --- #
+    u_bounds, u_init, u_scaling = set_u_bounds_and_init(model, n_shooting, tau=tau)
+
+    # --- Set objective --- #
+    objective_functions = set_objective_functions(
+        model=model,
+        n_shooting=n_shooting,
+        force=force,
+    )
+
+    # --- Set parameters (for minmax cost_function) --- #
+    parameters = ParameterList(use_sx=False)
+    parameters_bounds = BoundsList()
[…]
