+++ examples/fes_multibody/shoulder_abduction/abduction_muscle_driven.py
@@ -1,0 +1,398 @@
+"""
+This example will perform an optimal control program moving time horizon for an abduction/adduction motion driven by FES.
+"""
+
+import pickle
+from sys import platform
+
+import matplotlib.pyplot as plt
+import numpy as np
+import biorbd
+
+from bioptim import (
+    BiorbdModel,
+    BoundsList,
+    ConstraintList,
+    DynamicsList,
+    DynamicsFcn,
+    ExternalForceSetTimeSeries,
+    InitialGuessList,
+    InterpolationType,
+    ObjectiveFcn,
+    ObjectiveList,
+    OdeSolver,
+    PhaseDynamics,
+    ParameterObjectiveList,
+    SolutionMerge,
+    Solver,
+    ParameterList,
+    Node,
+    OptimalControlProgram,
+)
+
+
+# --------------------#
+#    OCP functions    #
+# --------------------#
+def prepare_ocp(
+    model: BiorbdModel,
+    abduction_info: dict,
+):
+    # --- Initialize parameters from dictionaries --- #
+    # --- Abduction info --- #
+    abd_range_config = abduction_info["range_config"]
+    n_shooting = 300
+    final_time = 6
+    ode_solver = OdeSolver.COLLOCATION(polynomial_degree=3, method="radau")
+
+    # --- Set dynamics --- #
+    dynamics = DynamicsList()
+    dynamics.add(
+        DynamicsFcn.MUSCLE_DRIVEN,
+        with_residual_torque=True,
+        expand_dynamics=True,
+        phase_dynamics=PhaseDynamics.SHARED_DURING_THE_PHASE,
+        numerical_data_timeseries=None,
+        phase=0,
+        ode_solver=ode_solver,
+    )
+
+    # --- Set states --- #
+    x_bounds, x_init = set_x_bounds(
+        n_shooting=n_shooting,
+        abduction_range=abd_range_config,
+        ode_solver=ode_solver,
+    )
+
+    # --- Set controls --- #
+    u_bounds, u_init = set_u_bounds_and_init(n_shooting)
+
+    # --- Set objective --- #
+    objective_functions = ObjectiveList()
+    objective_functions.add(
+        ObjectiveFcn.Lagrange.MINIMIZE_CONTROL,
+        key="muscles",
+        weight=1e6,
+        quadratic=True,
+        node=Node.ALL_SHOOTING,
+    )
+    objective_functions.add(
+        ObjectiveFcn.Lagrange.MINIMIZE_CONTROL,
+        key="tau",
+        weight=1e-8,
+        quadratic=True,
+        node=Node.ALL_SHOOTING,
+    )
+
+    for i in range(int(n_shooting/24)):
+        objective_functions.add(
+            ObjectiveFcn.Mayer.MINIMIZE_CONTROL, key="tau", weight=1e10, node=i, quadratic=True,
+        )
+
+
+    # --- Set parameters (for minmax cost_function) --- #
+    parameters = ParameterList(use_sx=False)
+    parameters_bounds = BoundsList()
+    parameters_init = InitialGuessList()
+    parameters_objectives = ParameterObjectiveList()
+
+     # --- Set constraints --- #
+    constraints = ConstraintList()
+
+    return OptimalControlProgram(
+        bio_model=[model],
+        dynamics=dynamics,
+        n_shooting=n_shooting,
+        phase_time=final_time,
+        objective_functions=objective_functions,
+        constraints=constraints,
+        x_bounds=x_bounds,
+        x_init=x_init,
+        u_bounds=u_bounds,
+        u_init=u_init,
+        parameters=parameters,
+        parameter_init=parameters_init,
+        parameter_bounds=parameters_bounds,
[…]
