+++ examples/fes_multibody/cycling/cycling_standard_clinics_v2.py
@@ -1,0 +1,861 @@
+"""
+This example demonstrates how long hand cycling could be maintained at 30 Hz with standard clinical settings
+(fixed pulse width and defined stimulation angles). The crank torque is optimized to achieve the target cadence
+(60 rpm). Then, after each cycle, muscle fatigue is updated based on the activation history and tested with an
+optimal control problem. If an optimal solution is found, the standard clinical settings simulation continues.
+If no optimal solution is found, the simulation stops and muscle failure is considered reached.
+
+Important note about the feasibility logic
+------------------------------------------
+The OCP used to assess feasibility intentionally freezes some FES states using prev_state_result[key][0][0].
+This is kept unchanged because the goal is to test whether the clinical condition remains feasible before moving
+one window forward.
+"""
+
+from sys import platform
+import pickle
+import numpy as np
+
+from bioptim import (
+    OptimalControlProgram,
+    ObjectiveList,
+    ObjectiveFcn,
+    ConstraintList,
+    ConstraintFcn,
+    ParameterList,
+    BoundsList,
+    InitialGuessList,
+    ParameterObjectiveList,
+    Node,
+    Axis,
+    OdeSolver,
+    VariableScalingList,
+    InterpolationType,
+    SolutionMerge,
+    Solver,
+    CostType,
+)
+
+from cycling_pulse_width_mhe import (
+    set_fes_model,
+    set_external_forces,
+    set_dynamics,
+    set_q_qdot_init,
+    set_x_bounds,
+    updating_model,
+    set_u_bounds_and_init,
+)
+from cost_functions import CustomCostFunctions
+
+
+# -----------------------------------------------------------------------------
+# Global settings
+# -----------------------------------------------------------------------------
+
+TARGET_RPM = 60
+PHASE_TIME = 60 / TARGET_RPM  # 1 second per revolution at 60 rpm
+N_SHOOTING = 30  # 30 Hz for a 1-second cycle
+N_THREADS = 48
+
+MODEL_PATH = "../../msk_models/Wu/Modified_Wu_Shoulder_Model_Cycling.bioMod"
+RESULT_FILE_NAME = "standard_and_fes_ocp_cycles.pkl"
+
+MUSCLES = ("Delt_ant", "Delt_post", "Biceps", "Triceps")
+STATE_VARS = ("A", "F", "Km", "Tau1")
+CONTROL_VARS = ("last_pulse_width",)
+
+
+# -----------------------------------------------------------------------------
+# Small utilities
+# -----------------------------------------------------------------------------
+
+def target_crank_velocity(turn_number: int, phase_time: float) -> float:
+    """Return the target crank angular velocity in rad/s."""
+    return -2 * np.pi * turn_number / phase_time
+
+
+def compute_net_work(time: np.ndarray, torque: np.ndarray) -> float:
+    """
+    Compute net mechanical work at 60 rpm.
+
+    Parameters
+    ----------
+    time : np.ndarray
+        Time vector in seconds.
+    torque : np.ndarray
+        Torque vector in N.m.
+
+    Returns
+    -------
+    float
+        Net work in joules.
+
+    Notes
+    -----
+    If torque has the same length as time, trapezoidal integration is used.
+    If torque has one fewer point than time, torque is interpreted as a piecewise-constant control per interval.
+    """
+    time = np.asarray(time, dtype=float).squeeze()
+    torque = np.asarray(torque, dtype=float).squeeze()
+
+    if time.ndim != 1:
+        raise ValueError("time must be a 1D array.")
+    if torque.ndim != 1:
+        raise ValueError("torque must be a 1D array.")
+    if time.size < 2:
[…]
