### 2026-05-08 1602f83
similar time/tau shape
