### 2026-05-08 b70be09
Adapt FES dynamics to bioptim 3.4
