+++ examples/fes_multibody/cycling/cycling_with_different_driven_methods.py
@@ -3,6 +3,7 @@ This example will do an optimal control program of a 100 steps hand cycling moti
 muscle driven / FES driven dynamics and includes a resistive torque at the handle.
 """
 
+import argparse
 from sys import platform
 import numpy as np
 
@@ -14,6 +15,7 @@ from bioptim import (
     ConstraintFcn,
     ControlType,
     CostType,
+    DynamicsOptions,
     ExternalForceSetTimeSeries,
     InitialGuessList,
     InterpolationType,
@@ -25,9 +27,10 @@ from bioptim import (
     ParameterList,
     PhaseDynamics,
     Solver,
+    MusclesBiorbdModel,
+    TorqueBiorbdModel,
     VariableScalingList,
-    DynamicsOptionsList,
-    DynamicsOptions,
+    ContactType,
 )
 
 from cocofest import (
@@ -38,6 +41,7 @@ from cocofest import (
     OcpFesMsk,
     DingModelPulseWidthFrequency,
 )
+from cost_functions import CustomCostFunctions
 
 
 def set_external_forces(n_shooting: int, torque: int | float) -> tuple[dict, ExternalForceSetTimeSeries]:
@@ -91,70 +95,64 @@ def update_model(
             previous_stim=model.muscles_dynamics_model[0].previous_stim,
             activate_force_length_relationship=model.activate_force_length_relationship,
             activate_force_velocity_relationship=model.activate_force_velocity_relationship,
-            activate_passive_force_relationship=model.activate_force_velocity_relationship,
+            activate_passive_force_relationship=model.activate_passive_force_relationship,
             activate_residual_torque=model.activate_residual_torque,
             parameters=parameters,
             external_force_set=external_force_set,
-            with_contact=model.with_contact,
+            contact_types=model.contact_types,
         )
     else:
-        model = BiorbdModel(model.path, external_force_set=external_force_set)
+        model = type(model)(
+            model.path,
+            external_force_set=external_force_set,
+            contact_types=(ContactType.RIGID_EXPLICIT,),
+        )
 
     return model
 
 
-# def set_dynamics(
-#     model: BiorbdModel | FesMskModel,
-#     numerical_time_series: dict,
-#     dynamics_type_str: str = "torque_driven",
-#     ode_solver: OdeSolver = OdeSolver.RK4(n_integration_steps=10),
-# ) -> DynamicsList:
-#     """
-#     Set the dynamics of the optimal control program based on the chosen dynamics type.
-#
-#     Parameters
-#     ----------
-#     model: BiorbdModel | FesMskModel
-#         The biomechanical model.
-#     numerical_time_series: dict
-#         External numerical data (e.g., external forces).
-#     dynamics_type_str: str
-#         Type of dynamics ("torque_driven", "muscle_driven", or "fes_driven").
-#
-#     Returns
-#     -------
-#         A DynamicsList configured for the problem.
-#     """
-#     dynamics_type = (
-#         DynamicsFcn.TORQUE_DRIVEN
-#         if dynamics_type_str == "torque_driven"
-#         else (
-#             DynamicsFcn.MUSCLE_DRIVEN
-#             if dynamics_type_str == "muscle_driven"
-#             else model.declare_model_variables if dynamics_type_str == "fes_driven" else None
-#         )
-#     )
-#     if dynamics_type is None:
-#         raise ValueError("Dynamics type not recognized")
-#
-#     dynamics = DynamicsList()
-#     dynamics.add(
-#         dynamics_type=dynamics_type,
-#         dynamic_function=(
[…]
