+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -4,6 +4,7 @@ This example will perform an optimal control program moving time horizon for a h
 
 import os
 import pickle
+from time import perf_counter
 from sys import platform
 from itertools import product
 from pathlib import Path
@@ -18,7 +19,9 @@ from bioptim import (
     BoundsList,
     ConstraintList,
     ConstraintFcn,
+    ContactType,
     CostType,
+    DynamicsOptions,
     ExternalForceSetTimeSeries,
     InitialGuessList,
     InterpolationType,
@@ -27,21 +30,29 @@ from bioptim import (
     ObjectiveFcn,
     ObjectiveList,
     OdeSolver,
+    OptimalControlProgram,
+    PhaseDynamics,
+    ParameterObjectiveList,
     SolutionMerge,
     Solution,
     Solver,
     ParameterList,
     Node,
     VariableScalingList,
+    VariableScaling,
 )
 from cocofest import (
-    CustomObjective,
     DingModelPulseWidthFrequencyWithFatigue,
     FesMskModel,
     inverse_kinematics_cycling,
     OcpFesMsk,
     FesNmpcMsk,
 )
+from examples.fes_multibody.cycling.cost_functions import CustomCostFunctions
+
+DEFAULT_SOLVER_CONFIG = "baseline"
+DEFAULT_TWO_STAGE_LM_ITER = 20
+SOLVER_CONFIG_CHOICES = ("baseline", "exact_jit", "lm", "two_stage")
 
 
 class MyCyclicNMPC(FesNmpcMsk):
@@ -286,11 +297,284 @@ class MyCyclicNMPC(FesNmpcMsk):
             plt.show()
 
 
-# -------------------#
-#   OCP functions   #
-# -------------------#
+def _safe_float(value):
+    if value is None:
+        return None
+    try:
+        return float(value)
+    except (TypeError, ValueError):
+        return None
 
 
+def _safe_int(value):
+    if value is None:
+        return None
+    try:
+        return int(value)
+    except (TypeError, ValueError):
+        return None
+
+
+def _max_abs_constraint(sol: Solution):
+    if sol is None or sol.constraints is None:
+        return None
+    try:
+        return float(np.max(np.abs(np.array(sol.constraints, dtype=float))))
+    except (TypeError, ValueError):
+        return None
+
+
+def _solver_config_file_suffix(config_name: str, n_lm_iter: int) -> str:
+    if config_name == "two_stage":
+        return f"_{config_name}_lm{n_lm_iter}"
+    return f"_{config_name}"
+
+
+def _append_solver_suffix_to_pickle_path(pickle_path: str, config_name: str, n_lm_iter: int) -> str:
+    path = Path(pickle_path)
+    return str(path.with_name(f"{path.stem}{_solver_config_file_suffix(config_name, n_lm_iter)}{path.suffix}"))
+
+
+def configure_casadi_interface_options(nmpc, config_name: str):
+    from bioptim.interfaces.ipopt_interface import IpoptInterface
+
+    if nmpc.ocp_solver is None:
+        nmpc.ocp_solver = IpoptInterface(nmpc)
+
+    for key in ("print_time", "record_time", "jit", "post_expand"):
+        nmpc.ocp_solver.options_common.pop(key, None)
+
+    if config_name in ("exact_jit", "lm", "two_stage"):
+        nmpc.ocp_solver.options_common["print_time"] = True
+        nmpc.ocp_solver.options_common["record_time"] = True
+        nmpc.ocp_solver.options_common["jit"] = True
+        nmpc.ocp_solver.options_common["post_expand"] = True
+
+
+def build_ipopt_solver(
+    config_name: str,
[…]

[…]
