### 2026-05-08 fd74cce
Adapt FES dynamics to bioptim 3.4
