### 2026-05-08 f497555
Pin environment to bioptim 3.4
