### 2026-05-08 224c8ad
Pin environment to bioptim 3.4
