### 2026-05-08 3678d90
Fix CI formatting and solver pins
