+++ examples/fes_multibody/cycling/cycling_bayesian_mhe.py
@@ -1,17 +1,19 @@
 """
 This example will perform a bayesian optimization on an optimal control program moving time horizon for a hand cycling
- motion driven by FES to find the best weight to increase number of cycling before failure.
+motion driven by FES to find the best weight to increase number of cycling before failure.
 """
 
 from pathlib import Path
 from sys import platform
-
 import pickle
+
 import numpy as np
+from numpy.ma.extras import average
 from skopt import gp_minimize
 from skopt.space import Real
 from skopt.utils import use_named_args
-from casadi import MX, vertcat
+from skopt.callbacks import CheckpointSaver
+from casadi import MX, vertcat, sum1, if_else, mmax, tanh
 
 from bioptim import (
     ObjectiveList,
@@ -26,29 +28,86 @@ from bioptim import (
 
 import cycling_pulse_width_mhe as base
 
+from cost_functions import CustomCostFunctions
 
-def minimize_muscle_fatigue(controller: PenaltyController, key) -> MX:
+
+def minimize_root_mean_square_fatigue(controller: PenaltyController, muscle_weights: list) -> MX:
     """
-    Minimize the muscle fatigue.
+    Minimize the root-mean-square of muscle fatigue.
+    """
+    eps = 1e-8
+    muscle_name_list = controller.model.bio_model.muscle_names
+    muscle_fatigue = vertcat(
+        *[
+            muscle_weights[x]
+            * (controller.model.muscles_dynamics_model[x].a_scale - controller.states["A_" + muscle_name_list[x]].cx)
+            ** 2
+            for x in range(len(muscle_name_list))
+        ]
+    )
+    rms_fatigue = (sum1(muscle_fatigue) / len(muscle_name_list) + eps) ** 0.5
+    return rms_fatigue
 
-    Parameters
-    ----------
-    controller: PenaltyController
-        The penalty node elements
 
-    Returns
-    -------
-    The force scaling factor difference
+def minimize_root_mean_pw(controller: PenaltyController, muscle_weights: list) -> MX:
+    """
+    Minimize the root-mean-square of pw.
+    """
+    eps = 1e-8
+    muscle_name_list = controller.model.bio_model.muscle_names
+    stim_charge = vertcat(
+        *[
+            muscle_weights[x]
+            * (
+                (
+                    controller.controls["last_pulse_width_" + muscle_name_list[x]].cx
+                    - controller.ocp.nlp[0].u_bounds["last_pulse_width_" + muscle_name_list[x]].min[0][0]
+                )
+                / (
+                    controller.ocp.nlp[0].u_bounds["last_pulse_width_" + muscle_name_list[x]].max[0][0]
+                    - controller.ocp.nlp[0].u_bounds["last_pulse_width_" + muscle_name_list[x]].min[0][0]
+                )
+            )
+            ** 2
+            for x in range(len(muscle_name_list))
+        ]
+    )
+
+    rms_activation = (sum1(stim_charge) / len(muscle_name_list) + eps) ** 0.5
+    return rms_activation
+
+
+def minimize_avg_fatigue_recovery(controller: PenaltyController, muscle_weights: list) -> MX:
+    """
+    Minimize the average fatigue based on recovery.
     """
     muscle_name_list = controller.model.bio_model.muscle_names
-    muscle_model = controller.model.muscles_dynamics_model
-    muscle_index = muscle_name_list.index(key[2:])  # key is like "A_muscleName"
-    muscle_a_scale_rest = muscle_model[muscle_index].a_scale
-    muscle_fatigue = vertcat(1 - (controller.states[key].cx / muscle_a_scale_rest))
[…]
