+++ cocofest/custom_objectives.py
@@ -4,8 +4,10 @@ This custom objective class regroups all the custom objectives that are used in 
 
 from casadi import MX, vertcat
 from bioptim import PenaltyController
+from .models.fes_model import FesModel
 from .models.ding2007.ding2007 import DingModelPulseWidthFrequency
 from .models.dynamical_model import FesMskModel
+from .models.hmed2018.hmed2018 import DingModelPulseIntensityFrequency
 
 
 class CustomObjective:
@@ -23,7 +25,7 @@ class CustomObjective:
         -------
         The sum of each force scaling factor
         """
-        muscle_name_list = controller.model.muscle_names
+        muscle_name_list = controller.model.bio_model.muscle_names
         muscle_model = controller.model.muscles_dynamics_model
         muscle_fatigue = vertcat(
             *[
@@ -47,7 +49,7 @@ class CustomObjective:
         -------
         The sum of each force
         """
-        muscle_name_list = controller.model.muscle_names
+        muscle_name_list = controller.model.bio_model.muscle_names
         muscle_model = controller.model.muscles_dynamics_model
         muscle_force = vertcat(
             *[
@@ -72,7 +74,7 @@ class CustomObjective:
         The sum of each stimulation control
         """
         if isinstance(controller.model, FesMskModel):
-            muscle_name_list = controller.model.muscle_names
+            muscle_name_list = controller.model.bio_model.muscle_names
             if isinstance(controller.model.muscles_dynamics_model[0], DingModelPulseWidthFrequency):
                 stim_charge = vertcat(
                     *[
@@ -92,3 +94,88 @@ class CustomObjective:
             stim_charge = controller.controls.cx
 
         return stim_charge
+
+    @staticmethod
+    def minimize_muscle_fatigue_normalized(controller: PenaltyController, fes_model: FesModel, power: int = 1) -> MX:
+        """
+        Minimize the normalized muscle fatigue.
+
+        Parameters
+        ----------
+        controller: PenaltyController
+            The penalty node elements
+        fes_model: FesModel
+            The fes model to use
+        power: int
+            The power to use for the cost function
+
+        Returns
+        -------
+        The force scaling factor normalized by the rested capacity
+        """
+        muscle_name = fes_model.muscle_name
+        muscle_fatigue = 1 - (controller.states["A_" + muscle_name].cx / fes_model.a_scale)
+        return muscle_fatigue**power
+
+    @staticmethod
+    def minimize_muscle_force_production_normalized(
+        controller: PenaltyController, fes_model: FesModel, power: int = 1
+    ) -> MX:
+        """
+        Minimize the normalized muscle force production.
+
+        Parameters
+        ----------
+        controller: PenaltyController
+            The penalty node elements
+        fes_model: FesModel
+            The fes model to use
+        power: int
+            The power to use for the cost function
+
+        Returns
+        -------
+        The force normalized by the maximal isometric force
+        """
+        muscle_name = fes_model.muscle_name
+        muscle_force = controller.states["F_" + muscle_name].cx / fes_model.fmax
+        return muscle_force**power
+
+    @staticmethod
+    def minimize_stimulation_charge_normalized(
+        controller: PenaltyController, fes_model: FesModel, power: int = 1
+    ) -> MX:
+        """
+        Minimize the normalized stimulation charge.
+
+        Parameters
+        ----------
+        controller: PenaltyController
+            The penalty node elements
+        fes_model: FesModel
+            The fes model to use
+        power: int
+            The power to use for the cost function
+
+        Returns
+        -------
+        The stimulation control normalized by its maximal value
+        """
+        muscle_name = fes_model.muscle_name
[…]

[…]
