### 2026-05-08 3f5852a
Fix CI formatting and solver pins
