+++ cycling_weight_exploration.py
@@ -1,0 +1,552 @@
+from __future__ import annotations
+
+import json
+import math
+from dataclasses import dataclass
+from pathlib import Path
+
+import matplotlib
+
+matplotlib.use("Agg")
+
+import numpy as np
+import biorbd
+from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
+from matplotlib.figure import Figure
+from scipy.optimize import curve_fit
+
+from cocofest.dynamics.inverse_kinematics_and_dynamics import inverse_kinematics_cycling
+
+ROOT = Path(__file__).resolve().parent
+COCOFEST_ROOT = ROOT if (ROOT / "examples").exists() else ROOT / "cocofest"
+OUTPUT_DIR = ROOT / "analysis_outputs"
+
+MODEL_PATH = COCOFEST_ROOT / "examples" / "msk_models" / "Wu" / "Modified_Wu_Shoulder_Model_Cycling.bioMod"
+IK_MODEL_PATH = COCOFEST_ROOT / "examples" / "msk_models" / "Wu" / "Modified_Wu_Shoulder_Model_Cycling_for_IK.bioMod"
+
+MUSCLE_LIST = ["Delt_ant", "Delt_post", "Biceps", "Triceps"]
+PARAMETERS = {
+    "Biceps": {"Fmax": 149.0, "a_scale": 3314.7, "alpha_a": -5.6e-2, "tau_fat": 179.6, "pcsa": 7.33},
+    "Triceps": {"Fmax": 262.0, "a_scale": 4915.5, "alpha_a": -3.4e-2, "tau_fat": 109.1, "pcsa": 10.87},
+    "Delt_ant": {"Fmax": 48.0, "a_scale": 1148.6, "alpha_a": -1.4e-1, "tau_fat": 445.5, "pcsa": 2.54},
+    "Delt_post": {"Fmax": 51.0, "a_scale": 1234.5, "alpha_a": -1.1e-1, "tau_fat": 342.7, "pcsa": 2.73},
+}
+
+TASK_TORQUE_THRESHOLD = 0.20
+HIGH_DEMAND_FRACTION = 0.80
+
+
+@dataclass
+class FatigueSummary:
+    duty_cycle: float
+    rho_crit: float
+    cycles_to_failure_high_demand: float | None
+    asymptotic_capacity_ratio_at_high_demand: float
+    fit_r2: dict[str, float]
+
+
+def to_numpy(x):
+    if hasattr(x, "to_array"):
+        return np.array(x.to_array(), dtype=float).squeeze()
+    return np.array(x, dtype=float).squeeze()
+
+
+def name_to_str(x):
+    if hasattr(x, "to_string"):
+        return x.to_string()
+    if hasattr(x, "toString"):
+        return x.toString()
+    return str(x)
+
+
+def marker_dict(model, q):
+    markers = model.markers(q)
+    names = [name_to_str(m) for m in model.markerNames()]
+    return {names[i]: to_numpy(markers[i])[:3] for i in range(len(markers))}
+
+
+def hand_position(model, q):
+    return marker_dict(model, q)["hand"][:2]
+
+
+def wheel_center_position(model, q):
+    return marker_dict(model, q)["global_wheel_center"][:2]
+
+
+def hand_jacobian_fd(model, q, eps=1e-7):
+    nq = len(q)
+    jac = np.zeros((2, nq))
+    for j in range(nq):
+        dq = np.zeros(nq)
+        dq[j] = eps
+        jac[:, j] = (hand_position(model, q + dq) - hand_position(model, q - dq)) / (2 * eps)
+    return jac
+
+
[…]
