+++ environment.yml
@@ -11,15 +11,13 @@ name: cocofest-run
 channels:
 - conda-forge
 dependencies:
-- python=3.11
+- python>=3.10,<3.12
 - numpy
 - scipy
 - matplotlib
 - pytest
 - casadi>=3.7
-- ipopt=3.14.16
-- mumps-include=5.7.3
-- mumps-seq=5.7.3
+- ipopt>=3.14,<3.15
 - biorbd>=1.12.0,<1.13.0
 - bioptim=3.4.0
 - bioviz

[…]
