### 2026-05-08 1b5ca25
Use conda bioptim in Windows CI
