### 2026-05-08 e4e8413
Use conda bioptim in Windows CI
