+++ .github/workflows/run_tests_win.yml
@@ -40,8 +40,12 @@ jobs:
           conda info
           conda list
 
+      - name: Verify conda bioptim installation
+        run: |
+          python -c "import bioptim; print(bioptim.__version__, bioptim.__file__)"
+
       - name: Install extra dependencies
-        run: conda install pip pytest-cov black pytest pytest-cov codecov packaging -cconda-forge
+        run: conda install pytest-cov black pytest pytest-cov codecov packaging -cconda-forge
 
       - name: Run tests with code coverage
         run: pytest -v --color=yes --cov-report term-missing --cov=cocofest --cov-report=xml:coverage.xml tests/shard${{ matrix.shard }}
@@ -84,6 +88,10 @@ jobs:
         run: |
           conda info
           conda list
+
+      - name: Verify conda bioptim installation
+        run: |
+          python -c "import bioptim; print(bioptim.__version__, bioptim.__file__)"
 
       - name: Install extra dependencies
         run: pip install coverage
