### 2026-05-08 08e6da1
Fix example compatibility with bioptim 3.4
