+++ examples/getting_started/optimization/pulse_width_optimization_nmpc.py
@@ -50,10 +50,7 @@ def prepare_nmpc(
         total_cycle_len, total_cycle_duration
     )
 
-    # dynamics = OcpFes.declare_dynamics(model, numerical_data_time_series, ode_solver)
-    dynamics_options = OcpFes.declare_dynamics_options(
-        numerical_time_series=numerical_data_time_series, ode_solver=ode_solver
-    )
+    dynamics = OcpFes.declare_dynamics(model, numerical_data_time_series, ode_solver)
 
     x_bounds = OcpFes.set_x_bounds(model)
     x_init_fes = OcpFes.set_x_init(model)
@@ -97,7 +94,7 @@ def prepare_nmpc(
 
     return FesNmpc(
         bio_model=model,
-        dynamics=dynamics_options,
+        dynamics=dynamics,
         cycle_len=cycle_len,
         cycle_duration=cycle_duration,
         n_cycles_simultaneous=n_cycles_simultaneous,
@@ -142,7 +139,6 @@ def main(plot=True):
         minimize_fatigue=False,
         use_sx=False,
         ode_solver=OdeSolver.RK4(n_integration_steps=10),
-        # ode_solver=OdeSolver.COLLOCATION(polynomial_degree=5, method="radau"),
     )
 
     def update_functions(_nmpc: MultiCyclicNonlinearModelPredictiveControl, cycle_idx: int, _sol: Solution):
