### 2026-05-08 759b88b
Fix example compatibility with bioptim 3.4
