### 2026-05-08 f8938fa
Add collocation defects for Ding 2007 NMPC
