+++ cocofest/models/ding2007/ding2007_with_fatigue.py
@@ -3,7 +3,15 @@ from typing import Callable
 from casadi import MX, vertcat
 import numpy as np
 
+from bioptim import (
+    ConfigureProblem,
+    DynamicsEvaluation,
+    NonLinearProgram,
+    OdeSolver,
+    OptimalControlProgram,
+)
 from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
+from cocofest.models.state_configure import StateConfigure
 
 
 class DingModelPulseWidthFrequencyWithFatigue(DingModelPulseWidthFrequency):
@@ -52,8 +60,8 @@ class DingModelPulseWidthFrequencyWithFatigue(DingModelPulseWidthFrequency):
 
     # ---- Absolutely needed methods ---- #
     @property
-    def name_dofs(self) -> list[str]:
-        muscle_name = "_" + self.muscle_name if self.muscle_name is not None else ""
+    def name_dof(self, with_muscle_name: bool = False) -> list[str]:
+        muscle_name = "_" + self.muscle_name if self.muscle_name and with_muscle_name else ""
         return [
             "Cn" + muscle_name,
             "F" + muscle_name,
@@ -117,38 +125,50 @@ class DingModelPulseWidthFrequencyWithFatigue(DingModelPulseWidthFrequency):
 
     def system_dynamics(
         self,
-        time: MX,
-        states: MX,
-        controls: MX,
-        numerical_timeseries: MX,
+        cn: MX,
+        f: MX,
+        a: MX = None,
+        tau1: MX = None,
+        km: MX = None,
+        t: MX = None,
+        t_stim_prev: MX = None,
+        pulse_width: MX = None,
+        force_length_relationship: MX | float = 1,
+        force_velocity_relationship: MX | float = 1,
+        passive_force_relationship: MX | float = 0,
     ) -> MX:
         """
         The system dynamics is the function that describes the models.
 
         Parameters
         ----------
-        time: MX
-            The system's current node time
-        states: MX
-            The state of the system CN, F, A, Tau1, Km
-        controls: MX
-            The controls of the system, pulse_width
-        numerical_timeseries: MX
-            The numerical timeseries of the system
+        cn: MX
+            The value of the ca_troponin_complex (unitless)
+        f: MX
+            The value of the force (N)
+        a: MX
+            The value of the scaling factor (unitless)
+        tau1: MX
+            The value of the time_state_force_no_cross_bridge (ms)
+        km: MX
+            The value of the cross_bridges (unitless)
+        t: MX
+            The current time at which the dynamics is evaluated (s)
+        t_stim_prev: MX
+            The time of the previous stimulation (s)
+        pulse_width: MX
+            The time of the impulse (s)
+        force_length_relationship: MX | float
+            The force length relationship value (unitless)
+        force_velocity_relationship: MX | float
+            The force velocity relationship value (unitless)
+        passive_force_relationship: MX | float
+            The passive force relationship value (unitless)
 
         Returns
         -------
         The value of the derivative of each state dx/dt at the current time t
         """
-        t = time
-        cn = states[0]
-        f = states[1]
-        a = states[2]
-        tau1 = states[3]
-        km = states[4]
-        pulse_width = controls[0]
-        t_stim_prev = numerical_timeseries
-
         cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev)
         a_scale = self.a_calculation(a_scale=a, pulse_width=pulse_width)
 
@@ -158,6 +178,9 @@ class DingModelPulseWidthFrequencyWithFatigue(DingModelPulseWidthFrequency):
             a_scale,
             tau1,
             km,
+            force_length_relationship=force_length_relationship,
+            force_velocity_relationship=force_velocity_relationship,
[…]
