### 2026-05-08 bbe48f3
Add collocation defects for Ding 2007 NMPC
