### 2026-05-08 6f06b8b
Align Ding 2003 model with bioptim 3.4
