+++ cocofest/models/ding2003/ding2003_with_fatigue.py
@@ -3,7 +3,15 @@ from typing import Callable
 from casadi import MX, vertcat
 import numpy as np
 
+from bioptim import (
+    ConfigureProblem,
+    DynamicsEvaluation,
+    NonLinearProgram,
+    OptimalControlProgram,
+)
+
 from cocofest.models.ding2003.ding2003 import DingModelFrequency
+from cocofest.models.state_configure import StateConfigure
 
 
 class DingModelFrequencyWithFatigue(DingModelFrequency):
@@ -93,8 +101,8 @@ class DingModelFrequencyWithFatigue(DingModelFrequency):
 
     # ---- Needed for the example ---- #
     @property
-    def name_dofs(self) -> list[str]:
-        muscle_name = "_" + self.muscle_name if self.muscle_name is not None else ""
+    def name_dof(self, with_muscle_name: bool = False) -> list[str]:
+        muscle_name = "_" + self.muscle_name if self.muscle_name and with_muscle_name else ""
         return [
             "Cn" + muscle_name,
             "F" + muscle_name,
@@ -131,37 +139,47 @@ class DingModelFrequencyWithFatigue(DingModelFrequency):
     # ---- Model's dynamics ---- #
     def system_dynamics(
         self,
-        time: MX,
-        states: MX,
-        controls: MX,
-        numerical_timeseries: MX,
+        cn: MX,
+        f: MX,
+        a: MX = None,
+        tau1: MX = None,
+        km: MX = None,
+        t: MX = None,
+        t_stim_prev: MX | float = None,
+        force_length_relationship: MX | float = 1,
+        force_velocity_relationship: MX | float = 1,
+        passive_force_relationship: MX | float = 0,
     ) -> MX:
         """
         The system dynamics is the function that describes the models.
 
         Parameters
         ----------
-        time: MX
-            The system's current node time
-        states: MX
-            The state of the system CN, F, A, Tau1, Km
-        controls: MX
-            The controls of the system, none
-        numerical_timeseries: MX
-            The numerical timeseries of the system
+        cn: MX
+            The value of the ca_troponin_complex (unitless)
+        f: MX
+            The value of the force (N)
+        a: MX
+            The value of the scaling factor (unitless)
+        tau1: MX
+            The value of the time_state_force_no_cross_bridge (s)
+        km: MX
+            The value of the cross_bridges (unitless)
+        t: MX
+            The current time at which the dynamics is evaluated (s)
+        t_stim_prev: MX | float
+            The previous time at which the stimulation was applied (s)
+        force_length_relationship: MX | float
+            The force length relationship value (unitless)
+        force_velocity_relationship: MX | float
+            The force velocity relationship value (unitless)
+        passive_force_relationship: MX | float
+            The passive force relationship value (unitless)
 
         Returns
         -------
         The value of the derivative of each state dx/dt at the current time t
         """
-        t = time
-        cn = states[0]
-        f = states[1]
-        a = states[2]
-        tau1 = states[3]
-        km = states[4]
-        t_stim_prev = numerical_timeseries
-
         cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev)  # Equation n°1
         f_dot = self.f_dot_fun(
             cn,
@@ -169,6 +187,9 @@ class DingModelFrequencyWithFatigue(DingModelFrequency):
             a,
             tau1,
             km,
+            force_length_relationship=force_length_relationship,
+            force_velocity_relationship=force_velocity_relationship,
+            passive_force_relationship=passive_force_relationship,
         )  # Equation n°2
         a_dot = self.a_dot_fun(a, f)  # Equation n°5
[…]
