### 2026-05-08 bb5357f
Align Ding 2003 model with bioptim 3.4
