+++ cocofest/models/ding2003/ding2003.py
@@ -1,4 +1,4 @@
-from typing import Callable, List
+from typing import Callable
 from math import gcd
 from fractions import Fraction
 
@@ -6,19 +6,17 @@ import numpy as np
 from casadi import MX, exp, vertcat
 
 from bioptim import (
+    ConfigureProblem,
     DynamicsEvaluation,
     NonLinearProgram,
-    StateDynamics,
-    DynamicsFunctions,
-    OdeSolver,
-    States,
+    OptimalControlProgram,
 )
 
 from cocofest.models.state_configure import StateConfigure
 from cocofest.models.fes_model import FesModel
 
 
-class DingModelFrequency(FesModel, StateDynamics):
+class DingModelFrequency(FesModel):
     """
     This is a custom model of the Bioptim package. As CustomModel, some methods are mandatory and must be implemented.
     to make it work with bioptim.
@@ -40,9 +38,8 @@ class DingModelFrequency(FesModel, StateDynamics):
         stim_time: list[float] = None,
         previous_stim: dict = None,
         sum_stim_truncation: int = 20,
-        **kwargs,
     ):
-        super().__init__(name=model_name, **kwargs)
+        super().__init__()
         self._model_name = model_name
         self._muscle_name = muscle_name
         self.sum_stim_truncation = sum_stim_truncation
@@ -71,30 +68,6 @@ class DingModelFrequency(FesModel, StateDynamics):
         self.km_rest = KM_REST_DEFAULT
         self.fmax = 315.5  # Maximum force (N) at 100 Hz
 
-        # ---- Muscle relationship ---- #
-        self.fes_model = None
-        self.force_length_relationship = 1
-        self.force_velocity_relationship = 1
-        self.passive_force_relationship = 0
-
-    # --- Configure variables --- #
-    @property
-    def state_configuration_functions(self) -> List[States | Callable]:
-        return [StateConfigure().configure_all_muscle_states]
-
-    @property
-    def control_configuration_functions(self) -> List[States | Callable]:
-        return []
-
-    @property
-    def algebraic_configuration_functions(self) -> List[States | Callable]:
-        return []
-
-    @property
-    def extra_configuration_functions(self) -> List[States | Callable]:
-        return []
-
-    # --- Set model parameters --- #
     def set_a_rest(self, model, a_rest: MX | float):
         # models is required for bioptim compatibility
         self.a_rest = a_rest
@@ -133,8 +106,8 @@ class DingModelFrequency(FesModel, StateDynamics):
 
     # ---- Needed for the example ---- #
     @property
-    def name_dofs(self) -> list[str]:
-        muscle_name = "_" + self.muscle_name if self.muscle_name is not None else ""
+    def name_dof(self, with_muscle_name: bool = False) -> list[str]:
+        muscle_name = "_" + self.muscle_name if self.muscle_name and with_muscle_name else ""
         return ["Cn" + muscle_name, "F" + muscle_name]
 
     @property
@@ -182,34 +155,38 @@ class DingModelFrequency(FesModel, StateDynamics):
     # ---- Model's dynamics ---- #
     def system_dynamics(
         self,
-        time: MX,
-        states: MX,
-        controls: MX,
-        numerical_timeseries: MX,
+        cn: MX,
+        f: MX,
+        t: MX = None,
+        t_stim_prev: list[MX] = None,
+        force_length_relationship: MX | float = 1,
+        force_velocity_relationship: MX | float = 1,
+        passive_force_relationship: MX | float = 0,
     ) -> MX:
         """
         The system dynamics is the function that describes the models.
 
         Parameters
         ----------
-        time: MX
-            The system's current node time
-        states: MX
-            The state of the system CN, F
-        controls: MX
-            The controls of the system, none
[…]
