### 2026-05-08 48eea2d
Trim migration PR scope and add bioptim 3.4 checks
