+++ plot_offline_weight_intermediates.py
@@ -1,315 +1,0 @@
-from __future__ import annotations
-
-import argparse
-import json
-from pathlib import Path
-
-import matplotlib
-
-matplotlib.use("Agg")
-
-import numpy as np
-from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
-from matplotlib.figure import Figure
-from matplotlib.patches import Patch
-
-import cycling_weight_exploration as cwe
-import derive_offline_task_weights as dotw
-
-OUTPUT_DIR = Path(__file__).resolve().parent / "analysis_outputs"
-
-
-def build_intermediate_summary(
-    target_cycles: int,
-    num_snapshots: int,
-    rho: float,
-    muscle_name: str,
-    pre_risk_width_deg: float,
-    include_gravity: bool,
-) -> dict:
-    if include_gravity:
-        theta, torque_profiles, gravity_torque = cwe.compute_torque_profiles(return_gravity_profile=True)
-    else:
-        theta, torque_profiles = cwe.compute_torque_profiles()
-        gravity_torque = np.zeros_like(theta)
-    theta_deg = np.degrees(theta)
-    positive_torque = np.maximum(torque_profiles, 0.0)
-    combined_positive = gravity_torque + positive_torque.sum(axis=0)
-    risk_mask_fresh = combined_positive < cwe.TASK_TORQUE_THRESHOLD
-    pre_risk_mask_fresh = dotw.build_pre_risk_mask(
-        theta_deg=theta_deg,
-        risk_mask=risk_mask_fresh,
-        pre_risk_width_deg=pre_risk_width_deg,
-    )
-
-    support_mask = positive_torque > 1e-9
-    muscle_index = cwe.MUSCLE_LIST.index(muscle_name)
-    duty_cycle = float(np.mean(support_mask[muscle_index]))
-    fatigue_ratios = {
-        name: dotw.simulate_fatigue_ratios(
-            duty_cycle=float(np.mean(support_mask[i])),
-            parameters=cwe.PARAMETERS[name],
-            total_cycles=target_cycles,
-            rho=rho,
-        )
-        for i, name in enumerate(cwe.MUSCLE_LIST)
-    }
-    intrinsic = {
-        name: dotw.intrinsic_vulnerability(
-            duty_cycle=float(np.mean(support_mask[i])),
-            parameters=cwe.PARAMETERS[name],
-            rho=rho,
-            total_cycles=target_cycles,
-        )
-        for i, name in enumerate(cwe.MUSCLE_LIST)
-    }
-
-    snapshot_cycles = dotw.choose_snapshot_cycles(target_cycles, num_snapshots=num_snapshots)
-    snapshot_rows = []
-    deficit_areas = []
-    criticality_terms_global_max = 0.0
-    for cycle in snapshot_cycles:
-        ratios = np.array([fatigue_ratios[m][cycle - 1] for m in cwe.MUSCLE_LIST], dtype=float)
-        metrics = dotw.feasibility_metrics(
-            theta,
-            torque_profiles,
-            ratios,
-            gravity_torque=gravity_torque,
-            pre_risk_width_deg=pre_risk_width_deg,
-        )
-        deficit_areas.append(metrics["deficit_area"])
-        pm = metrics["per_muscle"][muscle_name]
-        for name in cwe.MUSCLE_LIST:
-            per_muscle_metrics = metrics["per_muscle"][name]
-            criticality_terms_global_max = max(
-                criticality_terms_global_max,
-                per_muscle_metrics["restore_gain"],
-                per_muscle_metrics["support_in_pre_risk_area"],
-                per_muscle_metrics["support_in_risk_area"],
[…]

[…]
