### 2026-05-08 7c54c2b
Trim migration PR scope and add bioptim 3.4 checks
