+++ derive_offline_task_weights.py
@@ -1,342 +1,0 @@
-from __future__ import annotations
-
-import argparse
-import json
-from pathlib import Path
-
-import matplotlib
-
-matplotlib.use("Agg")
-
-import numpy as np
-from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
-from matplotlib.figure import Figure
-
-import cycling_weight_exploration as cwe
-
-OUTPUT_DIR = Path(__file__).resolve().parent / "analysis_outputs"
-GRAVITY_NEGLIGIBLE_TORQUE_NM = 1e-3
-
-
-def simulate_fatigue_ratios(
-    duty_cycle: float,
-    parameters: dict,
-    total_cycles: int,
-    rho: float,
-) -> np.ndarray:
-    a_rest = float(parameters["a_scale"])
-    alpha_a = float(parameters["alpha_a"])
-    tau_fat = float(parameters["tau_fat"])
-    f_max = float(parameters["Fmax"])
-
-    active_duration = duty_cycle
-    rest_duration = 1.0 - duty_cycle
-    force_demand = rho * f_max
-
-    a_value = a_rest
-    ratios = np.zeros(total_cycles, dtype=float)
-    for cycle_idx in range(total_cycles):
-        a_value = cwe.fatigue_discrete_map(
-            a_value=a_value,
-            a_rest=a_rest,
-            alpha_a=alpha_a,
-            force_value=force_demand,
-            tau_fat=tau_fat,
-            active_duration=active_duration,
-            rest_duration=rest_duration,
-        )
-        ratios[cycle_idx] = a_value / a_rest
-    return ratios
-
-
-def intrinsic_vulnerability(duty_cycle: float, parameters: dict, rho: float, total_cycles: int) -> dict[str, float]:
-    fatigue_summary, _, _ = cwe.analyze_fatigue_for_muscle(
-        duty_cycle=duty_cycle, parameters=parameters, rho=rho, max_cycles=total_cycles
-    )
-    cycles_fail = fatigue_summary.cycles_to_failure_high_demand
-    inv_cycles = 1e-4 if cycles_fail is None else 1.0 / float(cycles_fail)
-    low_sustainable_fraction = max(1.0 - float(fatigue_summary.rho_crit), 1e-6)
-    return {
-        "inverse_cycles_to_failure": inv_cycles,
-        "one_minus_rho_crit": low_sustainable_fraction,
-    }
-
-
-def normalize_positive(signal: dict[str, float]) -> dict[str, float]:
-    positive_values = [float(v) for v in signal.values() if v > 0]
-    scale = min(positive_values) if positive_values else 1.0
-    return {k: float(v / scale) for k, v in signal.items()}
-
-
-def choose_snapshot_cycles(total_cycles: int, num_snapshots: int) -> np.ndarray:
-    anchors = np.unique(np.round(np.linspace(1, total_cycles, num_snapshots)).astype(int))
-    return anchors
-
-
-def build_pre_risk_mask(theta_deg: np.ndarray, risk_mask: np.ndarray, pre_risk_width_deg: float = 120.0) -> np.ndarray:
-    theta_deg = np.asarray(theta_deg, dtype=float)
-    risk_mask = np.asarray(risk_mask, dtype=bool)
-    onset_mask = risk_mask & ~np.roll(risk_mask, 1)
-    pre_risk_mask = np.zeros_like(risk_mask, dtype=bool)
-    onset_angles = theta_deg[onset_mask]
-    for onset_angle in onset_angles:
-        angular_distance_to_onset = np.mod(onset_angle - theta_deg, 360.0)
[…]
