+++ analyze_failure_feasibility.py
@@ -1,196 +1,0 @@
-from __future__ import annotations
-
-import argparse
-import json
-from pathlib import Path
-
-import matplotlib
-
-matplotlib.use("Agg")
-
-import numpy as np
-from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
-from matplotlib.figure import Figure
-
-import cycling_weight_exploration as cwe
-
-OUTPUT_DIR = Path(__file__).resolve().parent / "analysis_outputs"
-
-
-def cycle_end_indices(total_points: int, n_shooting_per_cycle: int, polynomial_order: int) -> np.ndarray:
-    points_per_cycle = n_shooting_per_cycle * (polynomial_order + 1)
-    last_indices = np.arange(points_per_cycle, total_points, points_per_cycle, dtype=int)
-    return last_indices
-
-
-def load_run(npz_path: Path) -> dict:
-    data = np.load(npz_path, allow_pickle=True)
-    return {k: np.asarray(data[k]) for k in data.files}
-
-
-def fatigue_ratios_at_index(run: dict, idx: int) -> np.ndarray:
-    ratios = []
-    for muscle_name in cwe.MUSCLE_LIST:
-        a_value = float(run[f"A_{muscle_name}"][0, idx])
-        a_rest = float(cwe.PARAMETERS[muscle_name]["a_scale"])
-        ratios.append(a_value / a_rest)
-    return np.array(ratios, dtype=float)
-
-
-def feasibility_metrics(theta: np.ndarray, torque_profiles: np.ndarray, ratios: np.ndarray) -> dict:
-    scaled_profiles = torque_profiles * ratios[:, None]
-    positive_profiles = np.maximum(scaled_profiles, 0.0)
-    combined_positive = positive_profiles.sum(axis=0)
-    deficit = np.maximum(cwe.TASK_TORQUE_THRESHOLD - combined_positive, 0.0)
-    theta_deg = np.degrees(theta)
-
-    per_muscle = {}
-    for i, muscle_name in enumerate(cwe.MUSCLE_LIST):
-        without_i = combined_positive - positive_profiles[i]
-        deficit_without_i = np.maximum(cwe.TASK_TORQUE_THRESHOLD - without_i, 0.0)
-        restore_ratios = ratios.copy()
-        restore_ratios[i] = 1.0
-        restored_positive = np.maximum(torque_profiles * restore_ratios[:, None], 0.0).sum(axis=0)
-        deficit_if_restored = np.maximum(cwe.TASK_TORQUE_THRESHOLD - restored_positive, 0.0)
-        support_risk_area = float(
-            np.trapezoid(positive_profiles[i] * (combined_positive < cwe.TASK_TORQUE_THRESHOLD), theta)
-        )
-
-        per_muscle[muscle_name] = {
-            "max_positive_torque_nm": float(positive_profiles[i].max()),
-            "positive_area_nm_rad": float(np.trapezoid(positive_profiles[i], theta)),
-            "support_intervals_deg": cwe.wrap_intervals(positive_profiles[i] > 1e-9, theta_deg),
-            "extra_deficit_area_if_removed": float(np.trapezoid(deficit_without_i - deficit, theta)),
-            "deficit_area_reduction_if_restored": float(np.trapezoid(deficit - deficit_if_restored, theta)),
-            "support_in_risk_area_nm_rad": support_risk_area,
-        }
-
-    risk_mask = combined_positive < cwe.TASK_TORQUE_THRESHOLD
-    share_in_risk = {}
-    if np.any(risk_mask):
-        risk_positive = positive_profiles[:, risk_mask]
[…]

[…]
