### 2026-05-08 02bbe7e
Restore bioptim 3.4 compatibility in FES dynamics
