+++ cocofest/models/hmed2018/hmed2018.py
@@ -125,11 +125,15 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
 
     def system_dynamics(
         self,
-        cn: MX,
-        f: MX,
+        cn: MX = None,
+        f: MX = None,
         t: MX = None,
         t_stim_prev: list[MX] | list[float] = None,
         pulse_intensity: list[MX] | list[float] = None,
+        states: MX = None,
+        time: MX = None,
+        controls: MX = None,
+        numerical_timeseries: MX = None,
         force_length_relationship: MX | float = 1,
         force_velocity_relationship: MX | float = 1,
         passive_force_relationship: MX | float = 0,
@@ -160,6 +164,16 @@ class DingModelPulseIntensityFrequency(DingModelFrequency):
         -------
         The value of the derivative of each state dx/dt at the current time t
         """
+        if states is not None:
+            cn = states[0]
+            f = states[1]
+        if time is not None:
+            t = time
+        if numerical_timeseries is not None:
+            t_stim_prev = numerical_timeseries
+        if controls is not None:
+            pulse_intensity = controls
+
         cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev, pulse_intensity)
         f_dot = self.f_dot_fun(
             cn,
