### 2026-05-08 40ffa93
Restore bioptim 3.4 compatibility in FES dynamics
