+++ cocofest/models/ding2007/ding2007_with_fatigue.py
@@ -125,14 +125,18 @@ class DingModelPulseWidthFrequencyWithFatigue(DingModelPulseWidthFrequency):
 
     def system_dynamics(
         self,
-        cn: MX,
-        f: MX,
+        cn: MX = None,
+        f: MX = None,
         a: MX = None,
         tau1: MX = None,
         km: MX = None,
         t: MX = None,
         t_stim_prev: MX = None,
         pulse_width: MX = None,
+        states: MX = None,
+        time: MX = None,
+        controls: MX = None,
+        numerical_timeseries: MX = None,
         force_length_relationship: MX | float = 1,
         force_velocity_relationship: MX | float = 1,
         passive_force_relationship: MX | float = 0,
@@ -169,6 +173,21 @@ class DingModelPulseWidthFrequencyWithFatigue(DingModelPulseWidthFrequency):
         -------
         The value of the derivative of each state dx/dt at the current time t
         """
+        if states is not None:
+            cn = states[0]
+            f = states[1]
+            a = states[2]
+            tau1 = states[3]
+            km = states[4]
+        if time is not None:
+            t = time
+        if numerical_timeseries is not None:
+            t_stim_prev = numerical_timeseries
+        if controls is not None:
+            pulse_width = controls
+        if isinstance(pulse_width, (list, tuple)):
+            pulse_width = pulse_width[0]
+
         cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev)
         a_scale = self.a_calculation(a_scale=a, pulse_width=pulse_width)
 
