+++ cocofest/models/ding2003/ding2003.py
@@ -155,10 +155,14 @@ class DingModelFrequency(FesModel):
     # ---- Model's dynamics ---- #
     def system_dynamics(
         self,
-        cn: MX,
-        f: MX,
+        cn: MX = None,
+        f: MX = None,
         t: MX = None,
         t_stim_prev: list[MX] = None,
+        states: MX = None,
+        time: MX = None,
+        controls: MX = None,
+        numerical_timeseries: MX = None,
         force_length_relationship: MX | float = 1,
         force_velocity_relationship: MX | float = 1,
         passive_force_relationship: MX | float = 0,
@@ -187,6 +191,14 @@ class DingModelFrequency(FesModel):
         -------
         The value of the derivative of each state dx/dt at the current time t
         """
+        if states is not None:
+            cn = states[0]
+            f = states[1]
+        if time is not None:
+            t = time
+        if numerical_timeseries is not None:
+            t_stim_prev = numerical_timeseries
+
         cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev)
         f_dot = self.f_dot_fun(
             cn,

[…]
