### 2026-05-08 0dac401
Restore legacy example compatibility on bioptim 3.4
