+++ cocofest/models/marion2013/marion2013_modified_with_fatigue.py
@@ -58,17 +58,12 @@ class Marion2013ModelPulseWidthFrequencyWithFatigue(Marion2013ModelPulseWidthFre
         self.beta_a = BETA_A_DEFAULT
 
     @property
-    def name_dofs(self, with_muscle_name: bool = False) -> list[str]:
-        muscle_name = "_" + self.muscle_name if self.muscle_name is not None else ""
-        return [
-            "Cn" + muscle_name,
-            "F" + muscle_name,
-            "theta" + muscle_name,
-            "dtheta_dt" + muscle_name,
-            "A" + muscle_name,
-            "Tau1" + muscle_name,
-            "Km" + muscle_name,
-        ]
+    def name_dof(self) -> list[str]:
+        return ["Cn", "F", "theta", "dtheta_dt", "A", "Tau1", "Km"]
+
+    @property
+    def name_dofs(self) -> list[str]:
+        return self.name_dof
 
     @property
     def nb_state(self) -> int:
