### 2026-05-08 bf9681f
Restore legacy example compatibility on bioptim 3.4
