### 2026-05-09 c09a19e
Fix Bioptim 3.4 dynamics compatibility
