### 2026-05-09 763842d
Fix Bioptim 3.4 dynamics compatibility
