+++ .github/workflows/run_tests_win.yml
@@ -25,7 +25,7 @@ jobs:
         if: matrix.os == 'windows-latest'
 
       - name: Checkout code
-        uses: actions/checkout@v3
+        uses: actions/checkout@v6
 
       - name: Setup environment
         uses: conda-incubator/setup-miniconda@v2
@@ -74,7 +74,7 @@ jobs:
           echo "PREFIX=${{ env.PREFIX_WINDOWS }}" >> $GITHUB_ENV
 
       - name: Checkout code
-        uses: actions/checkout@v3
+        uses: actions/checkout@v6
 
       - name: Setup environment
         uses: conda-incubator/setup-miniconda@v2
