### 2026-05-12 ae24eb7
Address PR review cleanup for bioptim 3.4 migration
