+++ cocofest/models/ding2003/ding2003_with_fatigue.py
@@ -184,16 +184,11 @@ class DingModelFrequencyWithFatigue(DingModelFrequency):
         -------
         The value of the derivative of each state dx/dt at the current time t
         """
+        cn, f, t, t_stim_prev = self._legacy_state_inputs(cn, f, t, t_stim_prev, states, time, numerical_timeseries)
         if states is not None:
-            cn = states[0]
-            f = states[1]
             a = states[2]
             tau1 = states[3]
             km = states[4]
-        if time is not None:
-            t = time
-        if numerical_timeseries is not None:
-            t_stim_prev = numerical_timeseries
 
         cn_dot = self.calculate_cn_dot(cn, t, t_stim_prev)  # Equation n°1
         f_dot = self.f_dot_fun(
