+++ examples/fes_multibody/cycling/cost_functions.py
@@ -1,1207 +1,0 @@
-from casadi import MX, vertcat, sum1, fabs, sign, tanh, if_else, log, exp, DM, dot, mmax, mmin, cos, sin, sqrt
-from bioptim import PenaltyController
-from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
-from cocofest.models.hill_coefficients import (
-    muscle_force_length_coefficient,
-    muscle_force_velocity_coefficient,
-    muscle_passive_force_coefficient,
-)
-
-ENDURANCE_1500_FIXED_WEIGHTS = {
-    "Delt_ant": 187386,  # 220.0,
-    "Delt_post": 31609,  # 700.0,
-    "Biceps": 116430,  # 180.0,
-    "Triceps": 1.0,
-}
-
-SIMPLE_WEIGHTED_RMS_FATIGUE_WEIGHTS = {
-    "Delt_ant": 192.0,
-    "Delt_post": 32.0,
-    "Biceps": 117.0,
-    "Triceps": 0.001,
-}
-
-WEIGHTED_SQUARE_FATIGUE_WEIGHTS = {
-    "Delt_ant": 191790.4782397877e-6,
-    "Delt_post": 31608.944067062093e-6,
-    "Biceps": 117259.12134978296e-6,
-    "Triceps": 1.0e-6,
-}
-
-ENDURANCE_1500_A_MIN = {
-    "Delt_ant": 41.0,
-    "Delt_post": 70.0,
-    "Biceps": 379.0,
-    "Triceps": 932.0,
-}
-
-ENDURANCE_RISK_FIXED_CONFIG = {
-    "risk_sharpness": 6.0,
-    "depletion_weight": 1.0,
-    "risk_weight": 0.35,
-    "hazard_cap": 10.0,
-    "eps": 1e-6,
-}
-
-ENDURANCE_RISK_ADAPTIVE_CONFIG = {
-    "risk_sharpness": 6.0,
-    "depletion_weight": 1.0,
-    "risk_weight": 0.20,
-    "adaptive_reserve_gain": 1.5,
-    "adaptive_risk_gain": 3.0,
-    "hazard_cap": 10.0,
-    "eps": 1e-6,
-}
-
-
-class CustomCostFunctions:
-    def __init__(self):
-        self.dict_functions = {
-            # --- Pulse width --- #
-            "minimize_average_activation": {
-                "function": self.minimize_average_activation,
-                "index": 1,
-                "description": "Minimize the average fes activation",
-                "power": "1",
-                "state": "pw",
-            },
-            "minimize_root_mean_square_activation": {
-                "function": self.minimize_root_mean_square_activation,
-                "index": 2,
-                "description": "Minimize the root mean square of fes activation",
-                "power": "2",
-                "state": "pw",
-            },
-            "minimize_cubic_average_activation": {
-                "function": self.minimize_cubic_average_activation,
-                "index": 3,
-                "description": "Minimize the cubic average of fes activation",
-                "power": "3",
-                "state": "pw",
-            },
-            "minimize_peak_activation": {
-                "function": self.minimize_peak_activation,
-                "index": 4,
-                "description": "Minimize the peak of fes activation",
-                "power": r"\infty",
-                "state": "pw",
-            },
-            # --- Force --- #
-            "minimize_average_force": {
-                "function": self.minimize_average_force,
-                "index": 5,
-                "description": "Minimize the average muscle force",
-                "power": "1",
-                "state": r"F^{m}",
-            },
-            "minimize_root_mean_square_force": {
-                "function": self.minimize_root_mean_square_force,
-                "index": 6,
-                "description": "Minimize the root mean square of muscle force",
-                "power": "2",
-                "state": r"F^{m}",
-            },
-            "minimize_cubic_average_force": {
[…]
