### 2026-05-12 107ffe0
Address PR review cleanup for bioptim 3.4 migration
