+++ cocofest/models/state_configure.py
@@ -477,26 +477,27 @@ class StateConfigure:
     def configure_all_muscle_states(self, muscles_dynamics_model, ocp, nlp):
         state_name_list = []
         for muscle_dynamics_model in muscles_dynamics_model:
-            for state_key in muscle_dynamics_model.name_dof:
-                if state_key in self.state_dictionary.keys():
-                    self.state_dictionary[state_key](
-                        ocp=ocp,
-                        nlp=nlp,
-                        as_states=True,
-                        as_controls=False,
-                        muscle_name=muscle_dynamics_model.muscle_name,
-                    )
-                    state_name_list.append(state_key + "_" + muscle_dynamics_model.muscle_name)
-
-        return state_name_list
-
-    def configure_all_fes_model_states(self, ocp, nlp, fes_model):
-        for state_key in fes_model.name_dof:
-            if state_key in self.state_dictionary.keys():
+            for state_key in self._configured_state_keys(muscle_dynamics_model.name_dof):
                 self.state_dictionary[state_key](
                     ocp=ocp,
                     nlp=nlp,
                     as_states=True,
                     as_controls=False,
-                    muscle_name=fes_model.muscle_name,
+                    muscle_name=muscle_dynamics_model.muscle_name,
                 )
+                state_name_list.append(f"{state_key}_{muscle_dynamics_model.muscle_name}")
+
+        return state_name_list
+
+    def configure_all_fes_model_states(self, ocp, nlp, fes_model):
+        for state_key in self._configured_state_keys(fes_model.name_dof):
+            self.state_dictionary[state_key](
+                ocp=ocp,
+                nlp=nlp,
+                as_states=True,
+                as_controls=False,
+                muscle_name=fes_model.muscle_name,
+            )
+
+    def _configured_state_keys(self, state_keys):
+        return [state_key for state_key in state_keys if state_key in self.state_dictionary]
