### 2026-05-16 a75aebf
Refine review feedback on model dynamics
