### 2026-05-16 615daa9
Refine review feedback on model dynamics
