### 2026-05-18 ac0ef5d
Narrow ACADOS MHE PR scope
