+++ examples/fes_multibody/cycling/cycling_mhe_solver_comparison.py
@@ -1,0 +1,171 @@
+"""
+Compare a compact cycling moving-horizon estimator solved with IPOPT and ACADOS.
+
+This script uses the same torque-driven cycling MHE setup for both solvers so their solve time, status, objective, and
+estimated wheel angle can be compared directly. It intentionally avoids the pulse-width FES MHE because bioptim 3.4
+cannot export its numerical time series to ACADOS code generation.
+"""
+
+from __future__ import annotations
+
+import os
+from pathlib import Path
+import sys
+from time import perf_counter
+
+import numpy as np
+from bioptim import MovingHorizonEstimator, SolutionMerge
+from bioptim.misc.enums import SolverType
+from bioptim.optimization.receding_horizon_optimization import RecedingHorizonOptimization
+
+REPO_ROOT = Path(__file__).resolve().parents[3]
+if str(REPO_ROOT) not in sys.path:
+    sys.path.insert(0, str(REPO_ROOT))
+
+from examples.fes_multibody.cycling.cycling_mhe_acados import (
+    configure_acados_solver,
+    configure_ipopt_solver,
+    prepare_mhe,
+)
+
+
+def build_update_function(full_target: np.ndarray, window_len: int):
+    def update_functions(mhe: MovingHorizonEstimator, window_idx: int, _sol):
+        target = full_target[window_idx : window_idx + window_len + 1]
+        mhe.update_objectives_target(target=target[np.newaxis, :], list_index=0)
+        return window_idx < len(full_target) - window_len - 1
+
+    return update_functions
+
+
+def summarize_solution(label: str, sol) -> dict:
+    q_wheel = sol.decision_states(to_merge=SolutionMerge.NODES)["q"][2, :]
+    objective = float(np.nansum(sol.cost)) if getattr(sol, "cost", None) is not None else float("nan")
+    return {
+        "solver": label,
+        "status": sol.status,
+        "objective": objective,
+        "wall_time_s": float(sol.real_time_to_optimize) if sol.real_time_to_optimize is not None else float("nan"),
+        "final_wheel_angle": float(q_wheel[-1]),
+        "wheel_angle_trace": q_wheel,
+    }
+
+
+def solve_windows(mhe: MovingHorizonEstimator, update_function, solver):
+    sol = None
+    total_time = 0.0
+    wall_clock_start = perf_counter()
+    compiled = False
+    window_solutions = []
+
+    mhe.total_optimization_run = 0
+    while update_function(mhe, mhe.total_optimization_run, sol):
+        sol = super(RecedingHorizonOptimization, mhe).solve(solver=solver, warm_start=None)
+        if not compiled and solver.type == SolverType.ACADOS:
+            wall_clock_start = perf_counter()
+            compiled = True
+        if mhe.total_optimization_run == 0 and solver.type == SolverType.IPOPT:
+            solver.online_optim = None
+
+        total_time += sol.real_time_to_optimize or 0.0
+        window_solutions.append(sol)
+        mhe.advance_window(sol)
+        mhe.total_optimization_run += 1
+
+    return {
+        "final_solution": sol,
+        "window_solutions": window_solutions,
+        "solver_time_s": total_time,
+        "wall_time_s": perf_counter() - wall_clock_start,
+    }
+
+
+def run_comparison(
+    model_path: str,
+    n_windows: int = 2,
+    window_len: int = 5,
+    total_angle: float = -1.0,
+    acados_dir: str | None = None,
[…]
