+++ examples/fes_multibody/cycling/cycling_mhe_acados.py
@@ -1,0 +1,169 @@
+"""
+Torque-driven cycling moving horizon example solved with ACADOS.
+
+The FES pulse-width MHE uses time-varying stimulation and external-force numerical series. In bioptim 3.4 these series
+are not exported as ACADOS model parameters, so this compact example keeps the same cycling model family but removes
+the FES numerical time series to exercise the ACADOS moving-horizon pipeline end to end.
+"""
+
+import os
+
+import numpy as np
+
+from bioptim import (
+    BoundsList,
+    DynamicsOptions,
+    InitialGuessList,
+    InterpolationType,
+    MovingHorizonEstimator,
+    Node,
+    ObjectiveFcn,
+    ObjectiveList,
+    OdeSolver,
+    PhaseDynamics,
+    SolutionMerge,
+    Solver,
+    TorqueBiorbdModel,
+)
+
+
+def configure_acados_solver(
+    acados_dir: str | None = None,
+    codegen_dir: str = "result/acados/c_generated_code",
+    model_name: str = "cycling_mhe_acados",
+    max_iter: int = 100,
+):
+    solver = Solver.ACADOS()
+    if acados_dir:
+        solver.set_acados_dir(acados_dir)
+    solver.set_c_generated_code_path(codegen_dir)
+    solver.set_acados_model_name(model_name)
+    solver.set_maximum_iterations(max_iter)
+    solver.set_print_level(0)
+    return solver
+
+
+def configure_ipopt_solver(
+    max_iter: int = 500,
+    linear_solver: str | None = None,
+    tolerance: float = 1e-4,
+):
+    solver = Solver.IPOPT(show_online_optim=False, _max_iter=max_iter, show_options=dict(show_bounds=True))
+    solver.set_convergence_tolerance(tolerance)
+    solver.set_constraint_tolerance(tolerance)
+    solver.set_dual_inf_tol(tolerance)
+    solver.set_hessian_approximation("limited-memory")
+    if linear_solver:
+        solver.set_linear_solver(linear_solver)
+    return solver
+
+
+def prepare_mhe(
+    model_path: str,
+    window_len: int = 5,
+    window_duration: float = 1.0,
+    total_angle: float = -1.0,
+    use_sx: bool = True,
+    n_threads: int = 1,
+):
+    model = TorqueBiorbdModel(model_path)
+    n_nodes = window_len + 1
+
+    x_bounds = BoundsList()
+    q_bounds = model.bounds_from_ranges("q")
+    q_bounds.min[:] = [-50]
+    q_bounds.max[:] = [50]
+    x_bounds.add(
+        "q",
+        bounds=q_bounds,
+        interpolation=InterpolationType.CONSTANT_WITH_FIRST_AND_LAST_DIFFERENT,
+    )
+
+    qdot_bounds = model.bounds_from_ranges("qdot")
+    qdot_bounds.min[:] = [-50]
+    qdot_bounds.max[:] = [50]
+    x_bounds.add(
+        "qdot",
+        bounds=qdot_bounds,
+        interpolation=InterpolationType.CONSTANT_WITH_FIRST_AND_LAST_DIFFERENT,
+    )
+
+    u_bounds = BoundsList()
+    u_bounds.add(
+        "tau",
+        min_bound=np.full((model.nb_tau, 3), -1000),
+        max_bound=np.full((model.nb_tau, 3), 1000),
+        interpolation=InterpolationType.CONSTANT_WITH_FIRST_AND_LAST_DIFFERENT,
+    )
+
+    q_target = np.linspace(0, total_angle, n_nodes)
+    x_init = InitialGuessList()
+    x_init.add(
+        "q",
[…]
