+++ environment.yml
@@ -3,10 +3,15 @@ channels:
 - conda-forge
 dependencies:
 - python>=3.10,<3.12
+- pip
 - numpy
 - scipy
 - matplotlib
 - pytest
+- cmake
+- make
+- c-compiler
+- cxx-compiler
 - casadi>=3.7
 - ipopt>=3.14,<3.15
 - biorbd>=1.12.0,<1.13.0
