### 2026-05-18 9975404
Rebase utility coverage tests onto migration branch
