### 2026-06-08 #10669 @editorialbot
[REVIEW]: Cocofest: an Open-Source Python Package for Functional Electrical Stimulation Optimization in Optimal Control
**Submitting author:** @Kev1CO (<a href="http://orcid.org/0009-0009-0248-3548">Kevin Co</a>)
**Repository:** https://github.com/pyomeca/cocofest
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
