### 2026-06-25 3789763
Harden ACADOS cycling example setup
