### 2026-06-25 dd529cb
Show multi-window cycling MHE updates
