+++ examples/fes_multibody/cycling/cycling_mhe_solver_comparison.py
@@ -9,16 +9,13 @@ cannot export its numerical time series to ACADOS code generation.
 from __future__ import annotations
 
 import os
-from pathlib import Path
 import sys
 from time import perf_counter
 
 import numpy as np
 from bioptim import MovingHorizonEstimator, SolutionMerge
 from bioptim.misc.enums import SolverType
-from bioptim.optimization.receding_horizon_optimization import RecedingHorizonOptimization
 
-REPO_ROOT = Path(__file__).resolve().parents[3]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
@@ -27,19 +24,13 @@ from examples.fes_multibody.cycling.cycling_mhe_acados import (
     REPO_ROOT,
     _resolve_example_path,
     _validate_window_len,
+    build_reference_series,
+    build_update_function,
     configure_acados_solver,
     configure_ipopt_solver,
     prepare_mhe,
+    solve_windows,
 )
-
-
-def build_update_function(full_target: np.ndarray, window_len: int):
-    def update_functions(mhe: MovingHorizonEstimator, window_idx: int, _sol):
-        target = full_target[window_idx : window_idx + window_len + 1]
-        mhe.update_objectives_target(target=target[np.newaxis, :], list_index=0)
-        return window_idx < len(full_target) - window_len - 1
-
-    return update_functions
 
 
 def summarize_solution(label: str, sol) -> dict:
@@ -55,32 +46,18 @@ def summarize_solution(label: str, sol) -> dict:
     }
 
 
-def solve_windows(mhe: MovingHorizonEstimator, update_function, solver):
-    sol = None
-    total_time = 0.0
+def solve_windows_with_timing(mhe: MovingHorizonEstimator, update_function, solver):
     wall_clock_start = perf_counter()
-    compiled = False
-    window_solutions = []
-
-    mhe.total_optimization_run = 0
-    while update_function(mhe, mhe.total_optimization_run, sol):
-        sol = super(RecedingHorizonOptimization, mhe).solve(solver=solver, warm_start=None)
-        if not compiled and solver.type == SolverType.ACADOS:
-            wall_clock_start = perf_counter()
-            compiled = True
-        if mhe.total_optimization_run == 0 and solver.type == SolverType.IPOPT:
-            solver.online_optim = None
-
-        total_time += sol.real_time_to_optimize or 0.0
-        window_solutions.append(sol)
-        mhe.advance_window(sol)
-        mhe.total_optimization_run += 1
-
+    final_solution, window_solutions = solve_windows(mhe, update_function, solver)
+    wall_time = perf_counter() - wall_clock_start
+    solver_time = float(sum((window_sol.real_time_to_optimize or 0.0) for window_sol in window_solutions))
+    if solver.type == SolverType.ACADOS and len(window_solutions) > 1:
+        solver_time = float(sum((window_sol.real_time_to_optimize or 0.0) for window_sol in window_solutions[1:]))
     return {
-        "final_solution": sol,
+        "final_solution": final_solution,
         "window_solutions": window_solutions,
-        "solver_time_s": total_time,
-        "wall_time_s": perf_counter() - wall_clock_start,
+        "solver_time_s": solver_time,
+        "wall_time_s": wall_time,
     }
 
 
@@ -89,17 +66,23 @@ def run_comparison(
     n_windows: int = 2,
     window_len: int = 5,
     total_angle: float = -1.0,
+    perturbation_scale: float = 0.03,
     acados_dir: str | None = None,
     acados_codegen_dir: str = "result/acados/c_generated_code_compare",
[…]
