### 2026-06-25 30f5cce
Refactor cycling MHE solver comparisons
