+++ examples/fes_multibody/cycling/cycling_mhe_solver_comparison.py
@@ -1,171 +1,66 @@
 """
-Compare a compact cycling moving-horizon estimator solved with IPOPT and ACADOS.
+Compare IPOPT and ACADOS on one or several compact solver-compatible cycling MHE problems.
 
-This script uses the same torque-driven cycling MHE setup for both solvers so their solve time, status, objective, and
-estimated wheel angle can be compared directly. It intentionally avoids the pulse-width FES MHE because bioptim 3.4
-cannot export its numerical time series to ACADOS code generation.
+The pulse-width FES MHE remains IPOPT-only in bioptim 3.4 because its time-varying numerical series are not exported
+to ACADOS code generation. This script focuses on a shared torque-driven benchmark family that is easy to maintain and
+extend.
 """
 
 from __future__ import annotations
 
 import os
-import sys
-from time import perf_counter
 
-import numpy as np
-from bioptim import MovingHorizonEstimator, SolutionMerge
-from bioptim.misc.enums import SolverType
-
-if str(REPO_ROOT) not in sys.path:
-    sys.path.insert(0, str(REPO_ROOT))
-
-from examples.fes_multibody.cycling.cycling_mhe_acados import (
+from cycling_mhe_common import (
+    DEFAULT_MODEL_PATH,
+    DEFAULT_PROBLEM_SUITE,
     EXAMPLE_DIR,
+    PROBLEM_LIBRARY,
     REPO_ROOT,
-    _resolve_example_path,
-    _validate_window_len,
-    build_reference_series,
-    build_update_function,
-    configure_acados_solver,
-    configure_ipopt_solver,
-    prepare_mhe,
-    solve_windows,
+    compare_problem,
+    print_problem_comparison,
+    resolve_example_path,
+    run_problem_suite,
 )
 
 
-def summarize_solution(label: str, sol) -> dict:
-    q_wheel = sol.decision_states(to_merge=SolutionMerge.NODES)["q"][2, :]
-    objective = float(np.nansum(sol.cost)) if getattr(sol, "cost", None) is not None else float("nan")
-    return {
-        "solver": label,
-        "status": sol.status,
-        "objective": objective,
-        "wall_time_s": float(sol.real_time_to_optimize) if sol.real_time_to_optimize is not None else float("nan"),
-        "final_wheel_angle": float(q_wheel[-1]),
-        "wheel_angle_trace": q_wheel,
-    }
-
-
-def solve_windows_with_timing(mhe: MovingHorizonEstimator, update_function, solver):
-    wall_clock_start = perf_counter()
-    final_solution, window_solutions = solve_windows(mhe, update_function, solver)
-    wall_time = perf_counter() - wall_clock_start
-    solver_time = float(sum((window_sol.real_time_to_optimize or 0.0) for window_sol in window_solutions))
-    if solver.type == SolverType.ACADOS and len(window_solutions) > 1:
-        solver_time = float(sum((window_sol.real_time_to_optimize or 0.0) for window_sol in window_solutions[1:]))
-    return {
-        "final_solution": final_solution,
-        "window_solutions": window_solutions,
-        "solver_time_s": solver_time,
-        "wall_time_s": wall_time,
-    }
-
-
-def run_comparison(
-    model_path: str,
-    n_windows: int = 2,
-    window_len: int = 5,
-    total_angle: float = -1.0,
-    perturbation_scale: float = 0.03,
-    acados_dir: str | None = None,
-    acados_codegen_dir: str = "result/acados/c_generated_code_compare",
-    ipopt_max_iter: int = 500,
-    acados_max_iter: int = 100,
-):
[…]
