+++ examples/fes_multibody/cycling/cycling_mhe_acados.py
@@ -1,245 +1,60 @@
 """
-Torque-driven cycling moving horizon example solved with ACADOS.
+Run one compact solver-compatible cycling moving-horizon problem with ACADOS.
 
-The FES pulse-width MHE uses time-varying stimulation and external-force numerical series. In bioptim 3.4 these series
-are not exported as ACADOS model parameters, so this compact example keeps the same cycling model family but removes
-the FES numerical time series to exercise the ACADOS moving-horizon pipeline end to end.
+This wrapper stays intentionally small; the reusable setup and comparison logic live in
+`cycling_mhe_common.py` so we can benchmark multiple problem definitions consistently.
 """
 
+from __future__ import annotations
+
 import os
-from pathlib import Path
 
-import numpy as np
-
-from bioptim import (
-    BoundsList,
-    DynamicsOptions,
-    InitialGuessList,
-    InterpolationType,
-    MovingHorizonEstimator,
-    Node,
-    ObjectiveFcn,
-    ObjectiveList,
-    OdeSolver,
-    PhaseDynamics,
-    SolutionMerge,
-    Solver,
-    TorqueBiorbdModel,
+from cycling_mhe_common import (
+    DEFAULT_MODEL_PATH,
+    EXAMPLE_DIR,
+    REPO_ROOT,
+    build_reference_series,
+    build_update_function,
+    configure_acados_solver,
+    prepare_mhe,
+    print_single_solver_result,
+    resolve_example_path,
+    resolve_problem,
+    solve_windows_with_timing,
 )
-from bioptim.misc.enums import SolverType
-from bioptim.optimization.receding_horizon_optimization import RecedingHorizonOptimization
-
-EXAMPLE_DIR = Path(__file__).resolve().parent
-REPO_ROOT = EXAMPLE_DIR.parents[3]
-
-
-def _resolve_example_path(path: str | os.PathLike, root: Path) -> str:
-    path = Path(path)
-    if path.is_absolute():
-        return str(path)
-    return str((root / path).resolve())
-
-
-def _validate_window_len(window_len: int) -> None:
-    if window_len < 2:
-        raise ValueError(
-            "window_len must be >= 2 for the ACADOS cycling MHE. "
-            "Smaller horizons can generate invalid ACADOS code because the intermediate cost uses N - 1 stages."
-        )
-
-
-def build_reference_series(
-    window_len: int,
-    n_windows: int,
-    total_angle: float,
-    perturbation_scale: float = 0.03,
-) -> np.ndarray:
-    n_nodes = window_len + n_windows + 1
-    base_target = np.linspace(0, total_angle, n_nodes)
-    if perturbation_scale == 0:
-        return base_target
-
-    amplitude = perturbation_scale * max(abs(total_angle), 1.0)
-    phase = np.linspace(0, 2 * np.pi, n_nodes)
-    perturbation = amplitude * np.sin(phase)
-    perturbation[0] = 0
-    perturbation[-1] = 0
-    return base_target + perturbation
-
-
-def build_update_function(full_target: np.ndarray, window_len: int):
-    def update_functions(mhe: MovingHorizonEstimator, window_idx: int, _sol):
-        target = full_target[window_idx : window_idx + window_len + 1]
-        mhe.update_objectives_target(target=target[np.newaxis, :], list_index=0)
-        return window_idx < len(full_target) - window_len - 1
-
-    return update_functions
-
-
-def solve_windows(mhe: MovingHorizonEstimator, update_function, solver):
-    sol = None
[…]

[…]
