### 2026-06-25 401a7ba
Add richer cycling MHE benchmark problems
