+++ examples/fes_multibody/cycling/cycling_mhe_common.py
@@ -34,10 +34,15 @@ class CyclingMheProblem:
     name: str
     description: str
     track_wheel_weight: float = 10.0
+    shoulder_posture_weight: float = 0.0
+    elbow_posture_weight: float = 0.0
+    wheel_velocity_tracking_weight: float = 0.0
     control_weight: float = 1e-4
     qdot_weight: float = 0.0
     terminal_tracking_weight: float = 1e-2
     perturbation_scale: float = 0.03
+    shoulder_target: float = 0.75
+    elbow_target: float = 1.5
 
 
 PROBLEM_LIBRARY: dict[str, CyclingMheProblem] = {
@@ -57,6 +62,16 @@ PROBLEM_LIBRARY: dict[str, CyclingMheProblem] = {
         qdot_weight=0.0,
         terminal_tracking_weight=1e-2,
     ),
+    "posture_tracking_control": CyclingMheProblem(
+        name="posture_tracking_control",
+        description="Wheel-angle tracking with shoulder/elbow posture regulation and light torque regularization.",
+        track_wheel_weight=10.0,
+        shoulder_posture_weight=2e-1,
+        elbow_posture_weight=2e-1,
+        control_weight=1e-4,
+        qdot_weight=0.0,
+        terminal_tracking_weight=1e-2,
+    ),
     "tracking_strong_control": CyclingMheProblem(
         name="tracking_strong_control",
         description="Wheel-angle tracking with stronger torque regularization.",
@@ -69,13 +84,16 @@ PROBLEM_LIBRARY: dict[str, CyclingMheProblem] = {
         name="tracking_velocity_control",
         description="Wheel-angle tracking with wheel-speed and torque regularization (experimental).",
         track_wheel_weight=10.0,
+        shoulder_posture_weight=2e-1,
+        elbow_posture_weight=2e-1,
+        wheel_velocity_tracking_weight=5e-4,
         control_weight=1e-4,
-        qdot_weight=5e-4,
+        qdot_weight=0.0,
         terminal_tracking_weight=1e-2,
     ),
 }
 
-DEFAULT_PROBLEM_SUITE = ["tracking", "tracking_control", "tracking_strong_control"]
+DEFAULT_PROBLEM_SUITE = ["tracking", "tracking_control", "posture_tracking_control"]
 
 
 def resolve_problem(problem: str | CyclingMheProblem | None) -> CyclingMheProblem:
@@ -123,10 +141,24 @@ def build_reference_series(
     return base_target + perturbation
 
 
-def build_update_function(full_target: np.ndarray, window_len: int):
+def build_reference_velocity(reference: np.ndarray, window_duration: float, window_len: int) -> np.ndarray:
+    dt = window_duration / window_len
+    return np.gradient(reference, dt)
+
+
+def build_update_function(
+    full_target: np.ndarray,
+    window_len: int,
+    wheel_target_objective_index: int = 0,
+    velocity_target: np.ndarray | None = None,
+    wheel_velocity_objective_index: int | None = None,
+):
     def update_functions(mhe: MovingHorizonEstimator, window_idx: int, _sol):
         target = full_target[window_idx : window_idx + window_len + 1]
-        mhe.update_objectives_target(target=target[np.newaxis, :], list_index=0)
+        mhe.update_objectives_target(target=target[np.newaxis, :], list_index=wheel_target_objective_index)
+        if velocity_target is not None and wheel_velocity_objective_index is not None:
+            qdot_target = velocity_target[window_idx : window_idx + window_len]
+            mhe.update_objectives_target(target=qdot_target[np.newaxis, :], list_index=wheel_velocity_objective_index)
         return window_idx < len(full_target) - window_len - 1
 
     return update_functions
@@ -208,9 +240,14 @@ def configure_ipopt_solver(
[…]
