### 2026-06-26 17e3d0a
Add periodic Ding cycling ACADOS prototype
