+++ cocofest/models/ding2007/ding2007_with_fatigue_periodic.py
@@ -1,0 +1,172 @@
+from typing import Callable
+
+import numpy as np
+from casadi import MX, vertcat, exp
+
+from bioptim import ConfigureProblem, DynamicsEvaluation, NonLinearProgram, OptimalControlProgram
+
+from cocofest.models.ding2007.ding2007_with_fatigue import DingModelPulseWidthFrequencyWithFatigue
+from cocofest.models.state_configure import StateConfigure
+
+
+class DingModelPulseWidthFrequencyWithFatiguePeriodic(DingModelPulseWidthFrequencyWithFatigue):
+    """
+    ACADOS-friendly periodic-stimulation surrogate of the Ding 2007 pulse-width model with fatigue.
+
+    Instead of injecting a truncated list of stimulation times into the dynamics, this variant introduces an
+    auxiliary state `Cn_sum` whose dynamics approximate the calcium summation under a fixed stimulation period.
+    """
+
+    def __init__(
+        self,
+        model_name: str = "ding_2007_with_fatigue_periodic",
+        muscle_name: str = None,
+        stim_time: list[float] = None,
+        previous_stim: dict = None,
+        sum_stim_truncation: int = 20,
+        stim_interval: float | None = None,
+    ):
+        super().__init__(
+            model_name=model_name,
+            muscle_name=muscle_name,
+            stim_time=stim_time,
+            previous_stim=previous_stim,
+            sum_stim_truncation=sum_stim_truncation,
+        )
+        self._stim_interval = stim_interval if stim_interval is not None else self._infer_stim_interval(stim_time)
+
+    @staticmethod
+    def _infer_stim_interval(stim_time: list[float] | None) -> float | None:
+        if stim_time is None or len(stim_time) < 2:
+            return None
+
+        deltas = np.diff(np.array(stim_time, dtype=float))
+        reference = float(deltas[0])
+        if not np.allclose(deltas, reference):
+            raise ValueError("The periodic Ding approximation requires evenly spaced stimulation times.")
+        return reference
+
+    @property
+    def name_dof(self, with_muscle_name: bool = False) -> list[str]:
+        muscle_name = "_" + self.muscle_name if self.muscle_name and with_muscle_name else ""
+        return [
+            "Cn" + muscle_name,
+            "Cn_sum" + muscle_name,
+            "F" + muscle_name,
+            "A" + muscle_name,
+            "Tau1" + muscle_name,
+            "Km" + muscle_name,
+        ]
+
+    @property
+    def nb_state(self) -> int:
+        return 6
+
+    def standard_rest_values(self) -> np.array:
+        return np.array([[0], [0], [0], [self.a_scale], [self.tau1_rest], [self.km_rest]])
+
+    def serialize(self) -> tuple[Callable, dict]:
+        _, base_dict = super().serialize()
+        base_dict["stim_interval"] = self._stim_interval
+        return DingModelPulseWidthFrequencyWithFatiguePeriodic, base_dict
+
+    def stimulation_decay_factor(self) -> float:
+        if self._stim_interval is None:
+            raise ValueError("stim_interval could not be inferred. Provide evenly spaced stim_time or pass stim_interval.")
+        return float(np.exp(-self._stim_interval / self.tauc))
+
+    def periodic_cn_sum_gain(self) -> float:
+        decay = self.stimulation_decay_factor()
+        ri = 1 + (self.get_r0(self.km_rest) - 1) * decay
+        return ri / (self.tauc * (1 - decay))
+
+    def cn_sum_dot_fun(self, cn_sum: MX) -> MX:
+        return -cn_sum / self.tauc + self.periodic_cn_sum_gain()
+
+    def system_dynamics(
+        self,
+        cn: MX = None,
[…]
