### 2026-06-26 1222944
Fine tune ACADOS periodic cycling example
