### 2026-07-01 6828be3
Add cycling solver comparison CLI and import bootstrap
