+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -1,0 +1,219 @@
+"""
+Compare IPOPT and ACADOS on the cycling pulse-width FES MHE using solver-specific but validated configurations.
+
+IPOPT uses the historically robust collocation-based transcription.
+ACADOS uses the solver-compatible periodic Ding surrogate with the lightweight RK4 setup.
+"""
+
+from __future__ import annotations
+
+import argparse
+import os
+from pathlib import Path
+import sys
+
+import numpy as np
+
+REPO_ROOT = Path(__file__).resolve().parents[3]
+if str(REPO_ROOT) not in sys.path:
+    sys.path.insert(0, str(REPO_ROOT))
+
+from cycling_pulse_width_mhe_acados_periodic import build_argument_parser, solve_case
+
+
+EXAMPLE_DIR = Path(__file__).resolve().parent
+
+
+def _namespace_from_cli(**overrides) -> argparse.Namespace:
+    parser = build_argument_parser()
+    args = parser.parse_args([])
+    for key, value in overrides.items():
+        setattr(args, key, value)
+    return args
+
+
+def _format_metric(value) -> str:
+    if value is None:
+        return "None"
+    if isinstance(value, float):
+        return f"{value:.6f}"
+    return str(value)
+
+
+def _solver_config(
+    solver_name: str,
+    objective: str,
+    objective_shape: str,
+    cycles_per_window: int,
+    stimulations_per_cycle: int,
+    n_windows: int,
+    resistive_torque: float,
+    codegen_tag: str | None,
+    ipopt_max_iter: int,
+    acados_max_iter: int,
+) -> argparse.Namespace:
+    if solver_name == "ipopt":
+        return _namespace_from_cli(
+            solver="ipopt",
+            single_shot=False,
+            model_formulation="standard",
+            torque_application="external_forces",
+            cycles_per_window=cycles_per_window,
+            stimulations_per_cycle=stimulations_per_cycle,
+            objective=objective,
+            objective_shape=objective_shape,
+            constant_crank_torque=resistive_torque,
+            max_ipopt_iterations=ipopt_max_iter,
+            n_windows=n_windows,
+            ode_solver="collocation",
+            collocation_degree=3,
+            collocation_method="radau",
+            rk_steps=1,
+            use_sx=False,
+            enforce_start_constraints=True,
+            disable_standard_ipopt_warmup=False,
+            max_consecutive_failing=1,
+            codegen_tag=codegen_tag,
+        )
+
+    if solver_name == "acados":
+        return _namespace_from_cli(
+            solver="acados",
+            single_shot=False,
+            model_formulation="periodic",
+            torque_application="constant",
+            cycles_per_window=cycles_per_window,
+            stimulations_per_cycle=stimulations_per_cycle,
+            objective=objective,
+            objective_shape=objective_shape,
+            constant_crank_torque=resistive_torque,
+            max_acados_iterations=acados_max_iter,
+            n_windows=n_windows,
+            ode_solver="rk4",
+            rk_steps=1,
+            collocation_degree=3,
+            collocation_method="radau",
+            use_sx=True,
+            enforce_start_constraints=False,
+            disable_standard_ipopt_warmup=True,
+            max_consecutive_failing=10,
+            codegen_tag=codegen_tag,
+        )
+
+    raise ValueError(f"Unsupported solver_name '{solver_name}'")
+
+
+def print_comparison(ipopt_result: dict, acados_result: dict) -> None:
[…]
