### 2026-07-01 702fc60
Add physical diagnostics to cycling solver comparison
