### 2026-07-01 1167914
Stabilize cycling MHE solver setup
