+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -4,7 +4,6 @@ This example will perform an optimal control program moving time horizon for a h
 
 import os
 import pickle
-from sys import platform
 from itertools import product
 from pathlib import Path
 import sys
@@ -58,9 +57,44 @@ class MyCyclicNMPC(FesNmpcMsk):
             if isinstance(self.nlp[0].dynamics_type.ode_solver, OdeSolver.COLLOCATION)
             else 1
         )
-        self.pedal_turn_in_one_cycle = 2 * np.pi  # One mhe cycle simulates on pedal turn
+        self.pedal_turn_in_one_cycle = (
+            2 * np.pi
+        )  # One mhe cycle simulates on pedal turn
         self.debugg_bounds = False
         self.previous_bounds = None
+        self.first_node_state_slack = {}
+        self.transfer_debug = False
+        self.bound_first_node_all_states = True
+        self.bound_first_node_wheel_qdot = True
+        self.advance_wheel_q_bounds = False
+        self.wheel_q_path_margin = 2.0
+
+    def _state_slack_for(self, key: str, index: int) -> float:
+        if key in self.first_node_state_slack:
+            configured = self.first_node_state_slack[key]
+        elif any(key.startswith(prefix) for prefix in self.first_node_state_slack):
+            configured = next(
+                self.first_node_state_slack[prefix]
+                for prefix in self.first_node_state_slack
+                if key.startswith(prefix)
+            )
+        else:
+            configured = 0.0
+
+        if isinstance(configured, (list, tuple, np.ndarray)):
+            return float(configured[index])
+        return float(configured)
+
+    def _wheel_cycle_shift(self, states) -> float:
+        wheel_start = float(states["q"][2][0])
+        wheel_next_cycle = float(states["q"][2][self.nodes_per_cycle])
+        delta = wheel_next_cycle - wheel_start
+        wheel_speed = float(states["qdot"][2][self.nodes_per_cycle])
+        if not np.isclose(wheel_speed, 0.0):
+            return float(np.sign(wheel_speed) * abs(self.pedal_turn_in_one_cycle))
+        if not np.isclose(delta, 0.0):
+            return float(np.sign(delta) * abs(self.pedal_turn_in_one_cycle))
+        return self.pedal_turn_in_one_cycle
 
     def advance_window_bounds_states(self, sol, n_cycles_simultaneous=None, **extra):
         # --- Get states results --- #
@@ -82,28 +116,96 @@ class MyCyclicNMPC(FesNmpcMsk):
                 # --- Only doing wheel to prevent over constraining the system --- #
                 if key == "q" or key == "qdot":
                     if i == 2:
-                        self.nlp[0].x_bounds[key].min[i, 0] = states[key][i][self.nodes_per_cycle]
-                        self.nlp[0].x_bounds[key].max[i, 0] = states[key][i][self.nodes_per_cycle]
-                    if key == "q" and i == 2:
-                        self.nlp[0].x_bounds[key].min[i, 0] = (
-                            self.nlp[0].x_bounds["q"].min[i, 0] + self.pedal_turn_in_one_cycle
-                        )
-                        self.nlp[0].x_bounds[key].max[i, 0] = (
-                            self.nlp[0].x_bounds["q"].max[i, 0] + self.pedal_turn_in_one_cycle
-                        )
+                        if key == "qdot" and not self.bound_first_node_wheel_qdot:
+                            continue
+                        center = states[key][i][self.nodes_per_cycle]
+                        slack = self._state_slack_for(key, i)
+                        self.nlp[0].x_bounds[key].min[i, 0] = center - slack
+                        self.nlp[0].x_bounds[key].max[i, 0] = center + slack
[…]
