+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -20,7 +20,6 @@ if str(REPO_ROOT) not in sys.path:
 
 from cycling_pulse_width_mhe_acados_periodic import build_argument_parser, solve_case
 
-
 EXAMPLE_DIR = Path(__file__).resolve().parent
 
 
@@ -35,9 +34,35 @@ def _namespace_from_cli(**overrides) -> argparse.Namespace:
 def _format_metric(value) -> str:
     if value is None:
         return "None"
+    if isinstance(value, bool):
+        return str(value)
     if isinstance(value, float):
         return f"{value:.6f}"
     return str(value)
+
+
+def _resample_trace(trace: np.ndarray, target_len: int) -> np.ndarray:
+    trace = np.asarray(trace, dtype=float).squeeze()
+    if trace.size == target_len:
+        return trace
+    current_grid = np.linspace(0.0, 1.0, trace.size)
+    target_grid = np.linspace(0.0, 1.0, target_len)
+    return np.interp(target_grid, current_grid, trace)
+
+
+def _trace_comparison(ipopt_result: dict, acados_result: dict) -> dict:
+    ipopt_trace = np.asarray(ipopt_result["wheel_angle_trace"], dtype=float).squeeze()
+    acados_trace = np.asarray(acados_result["wheel_angle_trace"], dtype=float).squeeze()
+    common_len = max(ipopt_trace.size, acados_trace.size)
+    ipopt_common = _resample_trace(ipopt_trace, common_len)
+    acados_common = _resample_trace(acados_trace, common_len)
+    diff = acados_common - ipopt_common
+    return {
+        "common_len": common_len,
+        "wheel_rmse": float(np.sqrt(np.mean(diff**2))),
+        "wheel_max_abs_error": float(np.max(np.abs(diff))),
+        "wheel_final_error": float(acados_common[-1] - ipopt_common[-1]),
+    }
 
 
 def _solver_config(
@@ -50,6 +75,7 @@ def _solver_config(
     resistive_torque: float,
     codegen_tag: str | None,
     ipopt_max_iter: int,
+    ipopt_linear_solver: str,
     acados_max_iter: int,
 ) -> argparse.Namespace:
     if solver_name == "ipopt":
@@ -64,6 +90,7 @@ def _solver_config(
             objective_shape=objective_shape,
             constant_crank_torque=resistive_torque,
             max_ipopt_iterations=ipopt_max_iter,
+            ipopt_linear_solver=ipopt_linear_solver,
             n_windows=n_windows,
             ode_solver="collocation",
             collocation_degree=3,
@@ -88,22 +115,25 @@ def _solver_config(
             objective_shape=objective_shape,
             constant_crank_torque=resistive_torque,
             max_acados_iterations=acados_max_iter,
+            ipopt_linear_solver=ipopt_linear_solver,
             n_windows=n_windows,
             ode_solver="rk4",
-            rk_steps=1,
+            rk_steps=5,
             collocation_degree=3,
             collocation_method="radau",
             use_sx=True,
             enforce_start_constraints=False,
-            disable_standard_ipopt_warmup=True,
-            max_consecutive_failing=10,
+            disable_standard_ipopt_warmup=False,
+            max_consecutive_failing=max(n_windows, 10),
             codegen_tag=codegen_tag,
         )
 
     raise ValueError(f"Unsupported solver_name '{solver_name}'")
 
 
-def print_comparison(ipopt_result: dict, acados_result: dict) -> None:
+def print_comparison(
+    ipopt_result: dict, acados_result: dict, print_traces: bool = False
+) -> None:
     print(
         "solver | success | status | objective | solver_time_s | wall_time_s | final_wheel_angle | "
[…]

[…]
