### 2026-07-02 f7349d0
Stabilize cycling MHE solver comparison
