+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -41,6 +41,12 @@ def _format_metric(value) -> str:
     return str(value)
 
 
+def _format_control_metric(value) -> str:
+    if value is None:
+        return "None"
+    return f"{float(value):.6g}"
+
+
 def _resample_trace(trace: np.ndarray, target_len: int) -> np.ndarray:
     trace = np.asarray(trace, dtype=float).squeeze()
     if trace.size == target_len:
@@ -50,19 +56,77 @@ def _resample_trace(trace: np.ndarray, target_len: int) -> np.ndarray:
     return np.interp(target_grid, current_grid, trace)
 
 
+def _wrap_to_pi(values: np.ndarray) -> np.ndarray:
+    return (values + np.pi) % (2 * np.pi) - np.pi
+
+
 def _trace_comparison(ipopt_result: dict, acados_result: dict) -> dict:
     ipopt_trace = np.asarray(ipopt_result["wheel_angle_trace"], dtype=float).squeeze()
     acados_trace = np.asarray(acados_result["wheel_angle_trace"], dtype=float).squeeze()
     common_len = max(ipopt_trace.size, acados_trace.size)
-    ipopt_common = _resample_trace(ipopt_trace, common_len)
-    acados_common = _resample_trace(acados_trace, common_len)
-    diff = acados_common - ipopt_common
+    ipopt_common = _resample_trace(np.unwrap(ipopt_trace), common_len)
+    acados_common = _resample_trace(np.unwrap(acados_trace), common_len)
+    unwrapped_diff = acados_common - ipopt_common
+    unwrapped_turn_offset = int(np.rint(np.median(unwrapped_diff / (2 * np.pi))))
+    turn_aligned_unwrapped_diff = unwrapped_diff - unwrapped_turn_offset * 2 * np.pi
+    phase_diff = _wrap_to_pi(unwrapped_diff)
+    raw_final_error = float(acados_trace[-1] - ipopt_trace[-1])
     return {
         "common_len": common_len,
-        "wheel_rmse": float(np.sqrt(np.mean(diff**2))),
-        "wheel_max_abs_error": float(np.max(np.abs(diff))),
-        "wheel_final_error": float(acados_common[-1] - ipopt_common[-1]),
+        "unwrapped_wheel_rmse": float(np.sqrt(np.mean(unwrapped_diff**2))),
+        "unwrapped_wheel_max_abs_error": float(np.max(np.abs(unwrapped_diff))),
+        "unwrapped_wheel_final_error": float(acados_common[-1] - ipopt_common[-1]),
+        "turn_aligned_unwrapped_rmse": float(
+            np.sqrt(np.mean(turn_aligned_unwrapped_diff**2))
+        ),
+        "turn_aligned_unwrapped_max_abs_error": float(
+            np.max(np.abs(turn_aligned_unwrapped_diff))
+        ),
+        "turn_aligned_unwrapped_final_error": float(turn_aligned_unwrapped_diff[-1]),
+        "unwrapped_turn_offset": unwrapped_turn_offset,
+        "phase_rmse": float(np.sqrt(np.mean(phase_diff**2))),
+        "phase_max_abs_error": float(np.max(np.abs(phase_diff))),
+        "phase_final_error": float(phase_diff[-1]),
+        "raw_final_error": raw_final_error,
+        "raw_final_turn_offset": int(np.rint(raw_final_error / (2 * np.pi))),
     }
+
+
+def _control_comparisons(ipopt_result: dict, acados_result: dict) -> list[dict]:
+    ipopt_controls = ipopt_result.get("control_traces", {})
+    acados_controls = acados_result.get("control_traces", {})
+    common_keys = sorted(set(ipopt_controls).intersection(acados_controls))
+    comparisons = []
+
[…]
