+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -131,6 +131,36 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Shape of the objective terms passed to bioptim.",
     )
     parser.add_argument(
+        "--control-regularization-weight",
+        type=float,
+        default=0.0,
+        help="Optional quadratic MINIMIZE_CONTROL weight on each pulse-width control.",
+    )
+    parser.add_argument(
+        "--control-regularization-target",
+        type=float,
+        default=None,
+        help="Optional pulse-width target, in seconds, for the control regularization.",
+    )
+    parser.add_argument(
+        "--control-regularization-target-source",
+        choices=("constant", "warmup"),
+        default="constant",
+        help="Use a constant pulse-width target or the standard IPOPT warmup controls as the target.",
+    )
+    parser.add_argument(
+        "--wheel-qdot-regularization-weight",
+        type=float,
+        default=0.0,
+        help="Optional quadratic MINIMIZE_STATE weight on the crank/wheel qdot.",
+    )
+    parser.add_argument(
+        "--wheel-qdot-regularization-target",
+        type=float,
+        default=-float(2 * np.pi),
+        help="Target wheel angular velocity, in rad/s, for qdot regularization.",
+    )
+    parser.add_argument(
         "--constant-crank-torque",
         type=float,
         default=-0.2,
@@ -172,6 +202,11 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Skip the one-window IPOPT warmup with the standard Ding formulation before periodic MHE.",
     )
     parser.add_argument(
+        "--disable-historical-ipopt-initial-guess",
+        action="store_true",
+        help="Do not reuse the historical IPOPT initial guess file from result/initial_guess when it exists.",
+    )
+    parser.add_argument(
         "--use-sx",
         dest="use_sx",
         action="store_true",
@@ -205,6 +240,26 @@ def build_ode_solver(args: argparse.Namespace):
             polynomial_degree=args.collocation_degree, method=args.collocation_method
         )
     return OdeSolver.RK4(n_integration_steps=args.rk_steps)
+
+
+def _ode_solver_suffix(ode_solver) -> str:
+    if isinstance(ode_solver, OdeSolver.COLLOCATION):
+        return f"collocation_{ode_solver.polynomial_degree}_{ode_solver.method}"
+    if isinstance(ode_solver, OdeSolver.RK4):
+        return f"rk4_{ode_solver.n_integration_steps}"
+    raise RuntimeError("ode_solver must be COLLOCATION or RK4")
+
+
+def _historical_initial_guess_path(cycles_per_window: int, ode_solver) -> Path | None:
+    filename = f"{cycles_per_window}_initial_guess_{_ode_solver_suffix(ode_solver)}.pkl"
+    candidates = (
+        Path.cwd() / "result" / "initial_guess" / filename,
+        Path(__file__).resolve().parent / "result" / "initial_guess" / filename,
+    )
+    for path in candidates:
+        if path.exists():
+            return path.resolve()
+    return None
 
 
 def ensure_acados_environment(acados_source_dir: str | None = None) -> Path:
@@ -349,6 +404,11 @@ def _codegen_signature(args: argparse.Namespace) -> str:
         "constant_crank_torque": args.constant_crank_torque,
         "use_sx": args.use_sx,
         "enforce_start_constraints": args.enforce_start_constraints,
+        "control_regularization_weight": args.control_regularization_weight,
+        "control_regularization_target": args.control_regularization_target,
+        "control_regularization_target_source": args.control_regularization_target_source,
[…]

[…]
