### 2026-07-03 4e6de37
Tune acados scaling for cycling MHE
