+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -89,7 +89,7 @@ def build_argument_parser() -> argparse.ArgumentParser:
     parser.add_argument(
         "--ode-solver",
         type=str,
-        choices=("rk4", "collocation"),
+        choices=("rk4", "rk8", "irk", "collocation"),
         default="rk4",
         help="Integration scheme used to transcribe the window dynamics.",
     )
@@ -161,6 +161,19 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Target wheel angular velocity, in rad/s, for qdot regularization.",
     )
     parser.add_argument(
+        "--state-scaling",
+        type=str,
+        choices=("none", "fes", "full"),
+        default="none",
+        help="Scale optimization states: none, FES-only, or FES plus q/qdot.",
+    )
+    parser.add_argument(
+        "--pulse-width-scaling",
+        type=float,
+        default=1 / 400,
+        help="Scaling divisor, in seconds, for pulse-width controls.",
+    )
+    parser.add_argument(
         "--constant-crank-torque",
         type=float,
         default=-0.2,
@@ -171,6 +184,50 @@ def build_argument_parser() -> argparse.ArgumentParser:
         type=int,
         default=100,
         help="Maximum number of ACADOS SQP iterations per window.",
+    )
+    parser.add_argument(
+        "--acados-tolerance",
+        type=float,
+        default=None,
+        help="Optional ACADOS NLP convergence tolerance applied to stationarity, constraints and complementarity.",
+    )
+    parser.add_argument(
+        "--acados-collocation-type",
+        type=str,
+        choices=("GAUSS_LEGENDRE", "GAUSS_RADAU_IIA", "EXPLICIT_RUNGE_KUTTA"),
+        default="GAUSS_LEGENDRE",
+        help="Collocation tableau used by the ACADOS integrator.",
+    )
+    parser.add_argument(
+        "--acados-sim-stages",
+        type=int,
+        default=4,
+        help="Number of ACADOS integration stages per integration step.",
+    )
+    parser.add_argument(
+        "--acados-sim-steps",
+        type=int,
+        default=None,
+        help="Number of ACADOS integration substeps per shooting interval. Defaults to max(3, --rk-steps).",
+    )
+    parser.add_argument(
+        "--acados-newton-iter",
+        type=int,
+        default=5,
+        help="Number of Newton iterations used by the ACADOS implicit integrator.",
+    )
+    parser.add_argument(
+        "--acados-newton-tol",
+        type=float,
+        default=None,
+        help="Optional Newton tolerance for the ACADOS implicit integrator.",
+    )
+    parser.add_argument(
+        "--acados-jac-reuse",
+        type=int,
+        choices=(0, 1),
+        default=0,
+        help="Reuse the ACADOS implicit integrator Jacobian across Newton iterations.",
     )
     parser.add_argument(
         "--max-ipopt-iterations",
@@ -239,15 +296,25 @@ def build_ode_solver(args: argparse.Namespace):
         return OdeSolver.COLLOCATION(
             polynomial_degree=args.collocation_degree, method=args.collocation_method
         )
+    if args.ode_solver == "irk":
+        return OdeSolver.IRK(
+            polynomial_degree=args.collocation_degree, method=args.collocation_method
+        )
+    if args.ode_solver == "rk8":
+        return OdeSolver.RK8(n_integration_steps=args.rk_steps)
     return OdeSolver.RK4(n_integration_steps=args.rk_steps)
 
 
 def _ode_solver_suffix(ode_solver) -> str:
+    if isinstance(ode_solver, OdeSolver.IRK):
[…]

[…]
