### 2026-07-03 5d1e90f
Expose acados cycling MHE diagnostics
