+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -25,6 +25,17 @@ from bioptim.optimization.receding_horizon_optimization import (
 from cycling_pulse_width_mhe import prepare_nmpc, set_fes_model
 
 OBJECTIVE_TO_WEIGHT_INDEX = {"force": 0, "fatigue": 1, "control": 2}
+ACADOS_STATUS_NAMES = {
+    0: "ACADOS_SUCCESS",
+    1: "ACADOS_NAN_DETECTED",
+    2: "ACADOS_MAXITER",
+    3: "ACADOS_MINSTEP",
+    4: "ACADOS_QP_FAILURE",
+    5: "ACADOS_READY",
+    6: "ACADOS_UNBOUNDED",
+    7: "ACADOS_TIMEOUT",
+    8: "ACADOS_QPSCALING_BOUNDS_NOT_SATISFIED",
+}
 
 
 def parse_objectives(raw_objective: str) -> set[str]:
@@ -228,6 +239,93 @@ def build_argument_parser() -> argparse.ArgumentParser:
         choices=(0, 1),
         default=0,
         help="Reuse the ACADOS implicit integrator Jacobian across Newton iterations.",
+    )
+    parser.add_argument(
+        "--acados-hessian-approx",
+        choices=("GAUSS_NEWTON", "EXACT"),
+        default="GAUSS_NEWTON",
+        help="Hessian approximation used by ACADOS.",
+    )
+    parser.add_argument(
+        "--acados-nlp-solver-type",
+        choices=("SQP", "SQP_WITH_FEASIBLE_QP"),
+        default="SQP",
+        help="ACADOS NLP solver type.",
+    )
+    parser.add_argument(
+        "--acados-search-direction-mode",
+        choices=("NOMINAL_QP", "BYRD_OMOJOKUN", "FEASIBILITY_QP"),
+        default="NOMINAL_QP",
+        help="Search direction mode used by ACADOS, mainly for SQP_WITH_FEASIBLE_QP.",
+    )
+    parser.add_argument(
+        "--acados-use-constraint-hessian-in-feas-qp",
+        action="store_true",
+        help="Use constraint Hessians in the feasibility QP of SQP_WITH_FEASIBLE_QP.",
+    )
+    parser.add_argument(
+        "--acados-disable-direction-mode-switch-to-nominal",
+        action="store_true",
+        help="Keep ACADOS in the selected non-nominal search direction mode.",
+    )
+    parser.add_argument(
+        "--acados-regularize-method",
+        choices=(
+            "NO_REGULARIZE",
+            "MIRROR",
+            "PROJECT",
+            "PROJECT_REDUC_HESS",
+            "CONVEXIFY",
+            "GERSHGORIN_LEVENBERG_MARQUARDT",
+        ),
+        default="GERSHGORIN_LEVENBERG_MARQUARDT",
+        help="ACADOS Hessian regularization method.",
+    )
+    parser.add_argument(
+        "--acados-levenberg-marquardt",
+        type=float,
+        default=0.0,
+        help="Additional Levenberg-Marquardt diagonal regularization for ACADOS.",
+    )
+    parser.add_argument(
+        "--acados-globalization",
+        choices=("FIXED_STEP", "MERIT_BACKTRACKING", "FUNNEL_L1PEN_LINESEARCH"),
+        default="MERIT_BACKTRACKING",
+        help="ACADOS globalization strategy.",
+    )
+    parser.add_argument(
+        "--acados-fixed-step-length",
+        type=float,
+        default=1.0,
+        help="Step length used when --acados-globalization=FIXED_STEP.",
+    )
+    parser.add_argument(
+        "--acados-nlp-qp-tol-strategy",
[…]
