### 2026-07-03 ffef49c
Project periodic FES warmup states
