+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -357,6 +357,23 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Skip the one-window IPOPT warmup with the standard Ding formulation before periodic MHE.",
     )
     parser.add_argument(
+        "--disable-periodic-fes-warmup-projection",
+        action="store_true",
+        help="Do not project periodic Ding FES initial guesses with a local rollout before ACADOS.",
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-projection-weight",
+        type=float,
+        default=1.0,
+        help="Blend weight between the original warmup FES states and the periodic Ding rollout.",
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-projection-mode",
+        choices=("calcium", "all"),
+        default="all",
+        help="Project only the periodic calcium states or all Ding fatigue states.",
+    )
+    parser.add_argument(
         "--disable-historical-ipopt-initial-guess",
         action="store_true",
         help="Do not reuse the historical IPOPT initial guess file from result/initial_guess when it exists.",
@@ -1168,6 +1185,215 @@ def estimate_periodic_cn_sum_from_cn(
     return cn_values + tauc * cn_dot
 
 
+def _to_numpy_vector(values) -> np.ndarray:
+    return np.asarray(values, dtype=float).reshape((-1,))
+
+
+def _ding_state_keys(muscle_name: str) -> tuple[str, str, str, str, str, str]:
+    return (
+        f"Cn_{muscle_name}",
+        f"Cn_sum_{muscle_name}",
+        f"F_{muscle_name}",
+        f"A_{muscle_name}",
+        f"Tau1_{muscle_name}",
+        f"Km_{muscle_name}",
+    )
+
+
+def _state_trajectory_bounds(
+    periodic_nmpc, key: str, n_nodes: int
+) -> tuple[np.ndarray, np.ndarray]:
+    bounds = periodic_nmpc.nlp[0].x_bounds[key]
+    lower = np.empty(n_nodes)
+    upper = np.empty(n_nodes)
+    for node in range(n_nodes):
+        column = 0 if node == 0 else 2 if node == n_nodes - 1 else 1
+        lower[node] = bounds.min[0, column]
+        upper[node] = bounds.max[0, column]
+    return lower, upper
+
+
+def _clip_state_trajectory_to_bounds(
+    periodic_nmpc, key: str, values: np.ndarray
+) -> np.ndarray:
+    lower, upper = _state_trajectory_bounds(periodic_nmpc, key, values.size)
+    return np.minimum(np.maximum(values, lower), upper)
+
+
+def _periodic_ding_rhs(
+    muscle_model, state: np.ndarray, pulse_width: float
+) -> np.ndarray:
+    return _to_numpy_vector(
+        muscle_model.system_dynamics(
+            cn=state[0],
+            cn_sum=state[1],
+            f=state[2],
+            a=state[3],
+            tau1=state[4],
+            km=state[5],
+            pulse_width=pulse_width,
+        )
+    )
+
+
+def _periodic_calcium_rhs(muscle_model, state: np.ndarray) -> np.ndarray:
+    cn = state[0]
+    cn_sum = state[1]
+    return np.array(
+        [
+            float(muscle_model.cn_dot_fun(cn, cn_sum)),
+            float(muscle_model.cn_sum_dot_fun(cn_sum)),
+        ]
+    )
+
+
+def _rk4_periodic_ding_step(
+    muscle_model, state: np.ndarray, pulse_width: float, dt: float
+) -> np.ndarray:
+    k1 = _periodic_ding_rhs(muscle_model, state, pulse_width)
[…]
