### 2026-07-03 95c97b8
Stabilize periodic FES warmup projection
