+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -374,6 +374,12 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Project only the periodic calcium states or all Ding fatigue states.",
     )
     parser.add_argument(
+        "--periodic-fes-warmup-projection-substeps",
+        type=int,
+        default=10,
+        help="RK4 substeps per shooting interval used by the periodic FES warmup projection.",
+    )
+    parser.add_argument(
         "--disable-historical-ipopt-initial-guess",
         action="store_true",
         help="Do not reuse the historical IPOPT initial guess file from result/initial_guess when it exists.",
@@ -1248,26 +1254,50 @@ def _periodic_calcium_rhs(muscle_model, state: np.ndarray) -> np.ndarray:
 
 
 def _rk4_periodic_ding_step(
-    muscle_model, state: np.ndarray, pulse_width: float, dt: float
+    muscle_model,
+    state: np.ndarray,
+    pulse_width: float,
+    dt: float,
+    n_substeps: int = 1,
 ) -> np.ndarray:
-    k1 = _periodic_ding_rhs(muscle_model, state, pulse_width)
-    k2 = _periodic_ding_rhs(muscle_model, state + 0.5 * dt * k1, pulse_width)
-    k3 = _periodic_ding_rhs(muscle_model, state + 0.5 * dt * k2, pulse_width)
-    k4 = _periodic_ding_rhs(muscle_model, state + dt * k3, pulse_width)
-    return state + dt * (k1 + 2 * k2 + 2 * k3 + k4) / 6
+    if n_substeps < 1:
+        raise ValueError("--periodic-fes-warmup-projection-substeps must be >= 1.")
+
+    sub_dt = dt / n_substeps
+    next_state = np.array(state, dtype=float, copy=True)
+    for _ in range(n_substeps):
+        k1 = _periodic_ding_rhs(muscle_model, next_state, pulse_width)
+        k2 = _periodic_ding_rhs(
+            muscle_model, next_state + 0.5 * sub_dt * k1, pulse_width
+        )
+        k3 = _periodic_ding_rhs(
+            muscle_model, next_state + 0.5 * sub_dt * k2, pulse_width
+        )
+        k4 = _periodic_ding_rhs(muscle_model, next_state + sub_dt * k3, pulse_width)
+        next_state = next_state + sub_dt * (k1 + 2 * k2 + 2 * k3 + k4) / 6
+    return next_state
 
 
-def _rk4_periodic_calcium_step(
+def _exact_periodic_calcium_step(
     muscle_model, state: np.ndarray, dt: float
 ) -> np.ndarray:
-    k1 = _periodic_calcium_rhs(muscle_model, state)
-    k2 = _periodic_calcium_rhs(muscle_model, state + 0.5 * dt * k1)
-    k3 = _periodic_calcium_rhs(muscle_model, state + 0.5 * dt * k2)
-    k4 = _periodic_calcium_rhs(muscle_model, state + dt * k3)
-    return state + dt * (k1 + 2 * k2 + 2 * k3 + k4) / 6
+    tau = muscle_model.tauc
+    decay = np.exp(-dt / tau)
+    cn = state[0]
+    cn_sum = state[1]
+    cn_sum_steady = tau * muscle_model.periodic_cn_sum_gain()
+    next_cn_sum = cn_sum_steady + (cn_sum - cn_sum_steady) * decay
+    next_cn = (
+        decay * cn
+        + cn_sum_steady * (1 - decay)
+        + (cn_sum - cn_sum_steady) * (dt / tau) * decay
+    )
+    return np.array([next_cn, next_cn_sum])
 
 
[…]
