### 2026-07-03 3783f70
Add proximal FES warmup projection
