+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -365,7 +365,7 @@ def build_argument_parser() -> argparse.ArgumentParser:
         "--periodic-fes-warmup-projection-weight",
         type=float,
         default=1.0,
-        help="Blend weight between the original warmup FES states and the periodic Ding rollout.",
+        help="Blend weight between the original warmup FES states and the projected periodic Ding states.",
     )
     parser.add_argument(
         "--periodic-fes-warmup-projection-mode",
@@ -374,10 +374,40 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Project only the periodic calcium states or all Ding fatigue states.",
     )
     parser.add_argument(
+        "--periodic-fes-warmup-projection-strategy",
+        choices=("rollout", "sequential", "least_squares"),
+        default="sequential",
+        help="Project FES warmup states by direct rollout, sequential proximal rollout, or bounded least squares.",
+    )
+    parser.add_argument(
         "--periodic-fes-warmup-projection-substeps",
         type=int,
         default=10,
         help="RK4 substeps per shooting interval used by the periodic FES warmup projection.",
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-projection-proximity-weight",
+        type=float,
+        default=1.0,
+        help="Projection weight that keeps projected FES states close to the IPOPT warmup.",
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-projection-defect-weight",
+        type=float,
+        default=100.0,
+        help="Projection weight applied to normalized periodic FES dynamic defects.",
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-projection-trust-radius",
+        type=float,
+        default=None,
+        help="Optional trust radius, in normalized state units, around the warmup FES states.",
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-projection-max-iterations",
+        type=int,
+        default=200,
+        help="Maximum number of function evaluations for the least-squares FES projection.",
     )
     parser.add_argument(
         "--disable-historical-ipopt-initial-guess",
@@ -1366,15 +1396,279 @@ def _project_periodic_states(
     return projected_states
 
 
+def _projection_state_scales(
+    states: np.ndarray, lower_bounds: np.ndarray, upper_bounds: np.ndarray
+) -> np.ndarray:
+    scales = np.ones((states.shape[0], 1))
+    for idx in range(states.shape[0]):
+        values = [np.abs(states[idx, :])]
+        finite_lower = lower_bounds[idx, np.isfinite(lower_bounds[idx, :])]
+        finite_upper = upper_bounds[idx, np.isfinite(upper_bounds[idx, :])]
+        if finite_lower.size:
+            values.append(np.abs(finite_lower))
+        if finite_upper.size:
+            values.append(np.abs(finite_upper))
+        scale = float(np.max(np.concatenate(values)))
+        scales[idx, 0] = scale if np.isfinite(scale) and scale > 1.0 else 1.0
+    return scales
+
+
+def _strict_projection_bounds(
+    lower_bounds: np.ndarray, upper_bounds: np.ndarray, scales: np.ndarray
+) -> tuple[np.ndarray, np.ndarray]:
+    lower_flat = lower_bounds.reshape((-1,)).astype(float, copy=True)
+    upper_flat = upper_bounds.reshape((-1,)).astype(float, copy=True)
+    scale_flat = np.broadcast_to(scales, lower_bounds.shape).reshape((-1,))
+    eps = np.maximum(1e-12 * scale_flat, 1e-12)
+    locked = upper_flat <= lower_flat
+    centers = 0.5 * (lower_flat[locked] + upper_flat[locked])
[…]
