### 2026-07-03 1f59ef6
Stabilize ACADOS cycling diagnostics
