+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -38,6 +38,87 @@ ACADOS_STATUS_NAMES = {
 }
 
 
+def patch_bioptim_acados_control_bounds() -> None:
+    """
+    Patch the Bioptim 3.4 ACADOS interface control bounds for this example.
+
+    The installed Bioptim interface currently exports control bounds as
+    ``lbu=max`` and ``ubu=min`` during code generation, then later updates the
+    solver with unscaled control bounds. The ACADOS variables are scaled, so both
+    steps make the pulse-width QP infeasible. This local patch keeps the example
+    self-contained until the upstream interface is fixed.
+    """
+    from bioptim.interfaces.acados_interface import AcadosInterface
+
+    if getattr(AcadosInterface, "_cocofest_control_bounds_patch", False):
+        return
+
+    original_set_constraints = AcadosInterface._AcadosInterface__set_constraints
+    original_update_solver = AcadosInterface._AcadosInterface__update_solver
+
+    def scaled_control_bounds(interface) -> tuple[np.ndarray, np.ndarray]:
+        lower = np.empty(interface.acados_ocp.dims.nu)
+        upper = np.empty(interface.acados_ocp.dims.nu)
+        for key in interface.ocp.nlp[0].controls.keys():
+            control_bounds = (
+                interface.ocp.nlp[0]
+                .u_bounds[key]
+                .scale(interface.ocp.nlp[0].u_scaling[key].scaling)
+            )
+            index = interface.ocp.nlp[0].controls[key].index
+            lower[index] = np.asarray(control_bounds.min[:, 0], dtype=float)
+            upper[index] = np.asarray(control_bounds.max[:, 0], dtype=float)
+
+        if np.any(lower > upper):
+            raise RuntimeError(
+                "Scaled ACADOS control bounds are inconsistent "
+                f"(lower={lower}, upper={upper})."
+            )
+        return lower, upper
+
+    def patched_set_constraints(self, ocp):
+        original_set_constraints(self, ocp)
+        lower, upper = scaled_control_bounds(self)
+        self.acados_ocp.constraints.lbu = lower.reshape((-1, 1))
+        self.acados_ocp.constraints.ubu = upper.reshape((-1, 1))
+
+    def patched_update_solver(self):
+        original_update_solver(self)
+        if self.ocp_solver is None:
+            return
+
+        param_init = []
+        for key in self.ocp.nlp[0].parameters.keys():
+            scaled_init = self.ocp.parameter_init[key].scale(
+                self.ocp.parameters[key].scaling.scaling
+            )
+            param_init = np.concatenate((param_init, scaled_init.init[:, 0]))
+
+        terminal_node = self.acados_ocp.solver_options.N_horizon
+        terminal_x = np.empty((self.ocp.nlp[0].states.shape))
+        for key in self.ocp.nlp[0].states.keys():
+            index = self.ocp.nlp[0].states[key].index
+            self.ocp.nlp[0].x_init[key].check_and_adjust_dimensions(
+                self.ocp.nlp[0].states[key].shape, self.ocp.nlp[0].ns
+            )
+            terminal_x[index] = (
+                self.ocp.nlp[0].x_init[key].init.evaluate_at(terminal_node)
+                / self.ocp.nlp[0].x_scaling[key].scaling[:, 0]
+            )
+        self.ocp_solver.set(
+            terminal_node, "x", np.concatenate((param_init, terminal_x))
+        )
+
+        lower, upper = scaled_control_bounds(self)
[…]
