### 2026-07-03 2495459
Add ACADOS FES initial guess diagnostics
