### 2026-07-03 7e05357
Diagnose periodic ACADOS dynamics residuals
