+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -501,9 +501,12 @@ def build_argument_parser() -> argparse.ArgumentParser:
     )
     parser.add_argument(
         "--periodic-fes-warmup-projection-mode",
-        choices=("calcium", "all"),
+        choices=("calcium", "all", "all_except_force"),
         default="all",
-        help="Project only the periodic calcium states or all Ding fatigue states.",
+        help=(
+            "Project only the periodic calcium states, all Ding fatigue states, "
+            "or all states except F to preserve the multibody force trajectory."
+        ),
     )
     parser.add_argument(
         "--periodic-fes-warmup-projection-strategy",
@@ -1227,6 +1230,31 @@ def print_initial_guess_diagnostics(nmpc) -> None:
             f"{_format_named_values(fes_defects['scaled_by_state'])}"
         )
 
+    full_defects = _full_dynamics_rollout_defect_details(nmpc)
+    if full_defects:
+        print(
+            "initial_guess_full_rk4_defect_by_block: "
+            f"{_format_named_values(full_defects['absolute_by_block'])}"
+        )
+        print(
+            "initial_guess_full_rk4_scaled_defect_by_block: "
+            f"{_format_named_values(full_defects['scaled_by_block'])}"
+        )
+        print(
+            "initial_guess_full_rk4_defect_top_keys: "
+            f"{_format_named_values(full_defects['top_keys'])}"
+        )
+        if full_defects["q_by_dof"]:
+            print(
+                "initial_guess_full_rk4_defect_q_by_dof: "
+                f"{_format_named_values(full_defects['q_by_dof'])}"
+            )
+        if full_defects["qdot_by_dof"]:
+            print(
+                "initial_guess_full_rk4_defect_qdot_by_dof: "
+                f"{_format_named_values(full_defects['qdot_by_dof'])}"
+            )
+
 
 def project_qdot_initial_guess_from_q(nmpc) -> None:
     if "q" not in nmpc.nlp[0].x_init.keys() or "qdot" not in nmpc.nlp[0].x_init.keys():
@@ -1655,15 +1683,172 @@ def _periodic_fes_rollout_defect_details(
     }
 
 
+def _stack_initial_guess_values(container, variables, n_nodes: int) -> np.ndarray:
+    values = np.zeros((variables.shape, n_nodes))
+    for key in variables.keys():
+        values[variables[key].index, :] = np.asarray(container[key].init, dtype=float)
+    return values
+
+
+def _full_dynamics_rhs(
+    nlp, time: float, dt: float, state: np.ndarray, control: np.ndarray
+) -> np.ndarray:
+    return _to_numpy_vector(
+        nlp.dynamics_func(
+            np.array([time, dt], dtype=float),
+            state,
+            control,
+            np.array([], dtype=float),
+            np.array([], dtype=float),
+            np.array([], dtype=float),
+        )
+    )
+
+
+def _rk4_full_dynamics_step(
+    nlp,
+    state: np.ndarray,
+    control: np.ndarray,
+    time: float,
+    dt: float,
+    n_substeps: int,
+) -> np.ndarray:
+    sub_dt = dt / n_substeps
+    next_state = np.array(state, dtype=float, copy=True)
+    for substep in range(n_substeps):
+        sub_time = time + substep * sub_dt
+        k1 = _full_dynamics_rhs(nlp, sub_time, sub_dt, next_state, control)
+        k2 = _full_dynamics_rhs(
+            nlp,
+            sub_time + 0.5 * sub_dt,
+            sub_dt,
+            next_state + 0.5 * sub_dt * k1,
+            control,
+        )
+        k3 = _full_dynamics_rhs(
+            nlp,
[…]
