### 2026-07-03 efe2a72
Add isolated periodic IPOPT refinement warmstart
