+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -4,6 +4,7 @@ CLI-friendly ACADOS example for the periodic Ding pulse-width cycling MHE with a
 
 import argparse
 import ctypes
+from copy import deepcopy
 import hashlib
 import json
 import os
@@ -492,6 +493,25 @@ def build_argument_parser() -> argparse.ArgumentParser:
         "--disable-periodic-fes-warmup-projection",
         action="store_true",
         help="Do not project periodic Ding FES initial guesses with a local rollout before ACADOS.",
+    )
+    parser.add_argument(
+        "--periodic-ipopt-refinement",
+        action="store_true",
+        help=(
+            "Run a one-window IPOPT refinement on the periodic formulation "
+            "before handing the initial guess to ACADOS."
+        ),
+    )
+    parser.add_argument(
+        "--disable-periodic-ipopt-refinement",
+        action="store_true",
+        help=argparse.SUPPRESS,
+    )
+    parser.add_argument(
+        "--periodic-ipopt-refinement-iterations",
+        type=int,
+        default=300,
+        help="Maximum IPOPT iterations for the periodic warmstart refinement.",
     )
     parser.add_argument(
         "--periodic-fes-warmup-projection-weight",
@@ -2433,6 +2453,174 @@ def apply_standard_warmup_to_periodic_nmpc(
     return adapted_solution
 
 
+def apply_solution_directly_to_periodic_nmpc_initial_guess(periodic_nmpc, solution):
+    adapted_solution = _adapt_warmup_solution_to_periodic_nodes(periodic_nmpc, solution)
+    states = adapted_solution.decision_states(to_merge=SolutionMerge.NODES)
+    controls = adapted_solution.decision_controls(to_merge=SolutionMerge.NODES)
+
+    for key in periodic_nmpc.nlp[0].x_init.keys():
+        if key not in states:
+            continue
+        values = np.asarray(states[key], dtype=float)
+        target = periodic_nmpc.nlp[0].x_init[key].init
+        if values.shape != target.shape:
+            raise ValueError(
+                f"Cannot copy refined state '{key}' with shape {values.shape} "
+                f"into initial guess shape {target.shape}."
+            )
+        target[:, :] = values
+
+    for key in periodic_nmpc.nlp[0].u_init.keys():
+        if key not in controls:
+            continue
+        values = np.asarray(controls[key], dtype=float)
+        target = periodic_nmpc.nlp[0].u_init[key].init
+        if values.shape != target.shape:
+            raise ValueError(
+                f"Cannot copy refined control '{key}' with shape {values.shape} "
+                f"into initial guess shape {target.shape}."
+            )
+        target[:, :] = values
+
+    periodic_nmpc._correct_init_guess_to_fit_bounds(corrected_input="states")
+    periodic_nmpc._correct_init_guess_to_fit_bounds(corrected_input="controls")
+    periodic_nmpc._sync_acados_state_bounds()
+    return adapted_solution
+
+
+def _copy_list_values(source, target, attribute_name: str) -> None:
+    source_keys = set(source.keys())
+    for key in target.keys():
+        if key not in source_keys:
+            continue
+        source_values = np.asarray(getattr(source[key], attribute_name), dtype=float)
+        target_values = getattr(target[key], attribute_name)
+        if source_values.shape != target_values.shape:
+            raise ValueError(
+                f"Cannot copy '{key}' {attribute_name} with shape {source_values.shape} "
+                f"into shape {target_values.shape}."
+            )
[…]
