### 2026-07-04 6736993
Improve cycling solver comparison diagnostics
