+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -18,7 +18,11 @@ REPO_ROOT = Path(__file__).resolve().parents[3]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
-from cycling_pulse_width_mhe_acados_periodic import build_argument_parser, solve_case
+from cycling_pulse_width_mhe_acados_periodic import (
+    ACADOS_STATUS_NAMES,
+    build_argument_parser,
+    solve_case,
+)
 
 EXAMPLE_DIR = Path(__file__).resolve().parent
 
@@ -45,6 +49,30 @@ def _format_control_metric(value) -> str:
     if value is None:
         return "None"
     return f"{float(value):.6g}"
+
+
+def _status_label(status) -> str:
+    if status is None:
+        return "None"
+    return ACADOS_STATUS_NAMES.get(status, str(status))
+
+
+def _solver_status_label(solver_name: str, status) -> str:
+    if solver_name == "ACADOS":
+        return _status_label(status)
+    return "None" if status is None else str(status)
+
+
+def _format_array(values) -> str:
+    if values is None:
+        return "None"
+    if isinstance(values, dict):
+        return f"error={values.get('error')}"
+    return np.array2string(
+        np.asarray(values, dtype=float),
+        precision=3,
+        suppress_small=False,
+    )
 
 
 def _resample_trace(trace: np.ndarray, target_len: int) -> np.ndarray:
@@ -149,6 +177,9 @@ def _solver_config(
     state_scaling: str,
     pulse_width_scaling: float,
     acados_tolerance: float | None,
+    acados_diagnostics: bool,
+    periodic_ipopt_refinement: bool,
+    periodic_ipopt_refinement_iterations: int,
 ) -> argparse.Namespace:
     if solver_name == "ipopt":
         return _namespace_from_cli(
@@ -181,6 +212,9 @@ def _solver_config(
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_tolerance=acados_tolerance,
+            acados_diagnostics=False,
+            periodic_ipopt_refinement=False,
+            periodic_ipopt_refinement_iterations=periodic_ipopt_refinement_iterations,
         )
 
     if solver_name == "acados":
@@ -214,6 +248,9 @@ def _solver_config(
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_tolerance=acados_tolerance,
+            acados_diagnostics=acados_diagnostics,
+            periodic_ipopt_refinement=periodic_ipopt_refinement,
+            periodic_ipopt_refinement_iterations=periodic_ipopt_refinement_iterations,
         )
 
     raise ValueError(f"Unsupported solver_name '{solver_name}'")
@@ -223,7 +260,7 @@ def print_comparison(
     ipopt_result: dict, acados_result: dict, print_traces: bool = False
 ) -> None:
     print(
-        "solver | success | solver_success | physical_success | status | objective | solver_time_s | wall_time_s | final_wheel_angle | "
+        "solver | success | solver_success | physical_success | status | status_label | objective | solver_time_s | wall_time_s | final_wheel_angle | "
         "requested_cycles | attempted_windows | successful_windows | exported_cycles | covered_cycles"
     )
     for label, result in (("IPOPT", ipopt_result), ("ACADOS", acados_result)):
@@ -233,6 +270,7 @@ def print_comparison(
             f"{_format_metric(result.get('solver_success'))} | "
             f"{_format_metric(result.get('physical_success'))} | "
             f"{_format_metric(result['status'])} | "
+            f"{_solver_status_label(label, result['status'])} | "
             f"{_format_metric(result['objective'])} | "
[…]
