### 2026-07-04 1f0111a
Reduce memory for periodic IPOPT refinement
