### 2026-07-04 1ec4d37
Expose robust ACADOS comparison options
