+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -75,6 +75,15 @@ def _format_array(values) -> str:
     )
 
 
+def _effective_status(result: dict):
+    if result.get("status") is not None:
+        return result["status"]
+    for status in result.get("window_statuses") or []:
+        if status != 0:
+            return status
+    return result.get("status")
+
+
 def _resample_trace(trace: np.ndarray, target_len: int) -> np.ndarray:
     trace = np.asarray(trace, dtype=float).squeeze()
     if trace.size == target_len:
@@ -177,6 +186,14 @@ def _solver_config(
     state_scaling: str,
     pulse_width_scaling: float,
     acados_tolerance: float | None,
+    acados_qp_iter_max: int,
+    acados_levenberg_marquardt: float,
+    acados_regularize_method: str,
+    acados_hessian_approx: str,
+    acados_nlp_solver_type: str,
+    acados_search_direction_mode: str,
+    acados_globalization: str,
+    acados_nlp_qp_tol_strategy: str,
     acados_diagnostics: bool,
     periodic_ipopt_refinement: bool,
     periodic_ipopt_refinement_iterations: int,
@@ -213,6 +230,14 @@ def _solver_config(
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_tolerance=acados_tolerance,
+            acados_qp_iter_max=acados_qp_iter_max,
+            acados_levenberg_marquardt=acados_levenberg_marquardt,
+            acados_regularize_method=acados_regularize_method,
+            acados_hessian_approx=acados_hessian_approx,
+            acados_nlp_solver_type=acados_nlp_solver_type,
+            acados_search_direction_mode=acados_search_direction_mode,
+            acados_globalization=acados_globalization,
+            acados_nlp_qp_tol_strategy=acados_nlp_qp_tol_strategy,
             acados_diagnostics=False,
             periodic_ipopt_refinement=False,
             periodic_ipopt_refinement_iterations=periodic_ipopt_refinement_iterations,
@@ -250,6 +275,14 @@ def _solver_config(
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_tolerance=acados_tolerance,
+            acados_qp_iter_max=acados_qp_iter_max,
+            acados_levenberg_marquardt=acados_levenberg_marquardt,
+            acados_regularize_method=acados_regularize_method,
+            acados_hessian_approx=acados_hessian_approx,
+            acados_nlp_solver_type=acados_nlp_solver_type,
+            acados_search_direction_mode=acados_search_direction_mode,
+            acados_globalization=acados_globalization,
+            acados_nlp_qp_tol_strategy=acados_nlp_qp_tol_strategy,
             acados_diagnostics=acados_diagnostics,
             periodic_ipopt_refinement=periodic_ipopt_refinement,
             periodic_ipopt_refinement_iterations=periodic_ipopt_refinement_iterations,
@@ -267,13 +300,14 @@ def print_comparison(
         "requested_cycles | attempted_windows | successful_windows | exported_cycles | covered_cycles"
     )
     for label, result in (("IPOPT", ipopt_result), ("ACADOS", acados_result)):
+        status = _effective_status(result)
         print(
             f"{label} | "
             f"{_format_metric(result.get('success'))} | "
             f"{_format_metric(result.get('solver_success'))} | "
             f"{_format_metric(result.get('physical_success'))} | "
[…]
