### 2026-07-04 dab5a12
Expose ACADOS initialization comparison options
