+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -193,7 +193,18 @@ def _solver_config(
     acados_nlp_solver_type: str,
     acados_search_direction_mode: str,
     acados_globalization: str,
+    acados_fixed_step_length: float,
     acados_nlp_qp_tol_strategy: str,
+    acados_qpscaling_scale_objective: str,
+    acados_qpscaling_scale_constraints: str,
+    acados_ext_qp_res: bool,
+    acados_project_qdot_from_q: bool,
+    disable_periodic_fes_warmup_projection: bool,
+    periodic_fes_warmup_projection_weight: float,
+    periodic_fes_warmup_projection_mode: str,
+    periodic_fes_warmup_projection_strategy: str,
+    periodic_fes_warmup_projection_substeps: int,
+    periodic_fes_warmup_projection_defect_weight: float,
     acados_diagnostics: bool,
     periodic_ipopt_refinement: bool,
     periodic_ipopt_refinement_iterations: int,
@@ -237,7 +248,24 @@ def _solver_config(
             acados_nlp_solver_type=acados_nlp_solver_type,
             acados_search_direction_mode=acados_search_direction_mode,
             acados_globalization=acados_globalization,
+            acados_fixed_step_length=acados_fixed_step_length,
             acados_nlp_qp_tol_strategy=acados_nlp_qp_tol_strategy,
+            acados_qpscaling_scale_objective=acados_qpscaling_scale_objective,
+            acados_qpscaling_scale_constraints=acados_qpscaling_scale_constraints,
+            acados_ext_qp_res=False,
+            acados_project_qdot_from_q=False,
+            disable_periodic_fes_warmup_projection=True,
+            periodic_fes_warmup_projection_weight=periodic_fes_warmup_projection_weight,
+            periodic_fes_warmup_projection_mode=periodic_fes_warmup_projection_mode,
+            periodic_fes_warmup_projection_strategy=(
+                periodic_fes_warmup_projection_strategy
+            ),
+            periodic_fes_warmup_projection_substeps=(
+                periodic_fes_warmup_projection_substeps
+            ),
+            periodic_fes_warmup_projection_defect_weight=(
+                periodic_fes_warmup_projection_defect_weight
+            ),
             acados_diagnostics=False,
             periodic_ipopt_refinement=False,
             periodic_ipopt_refinement_iterations=periodic_ipopt_refinement_iterations,
@@ -282,7 +310,26 @@ def _solver_config(
             acados_nlp_solver_type=acados_nlp_solver_type,
             acados_search_direction_mode=acados_search_direction_mode,
             acados_globalization=acados_globalization,
+            acados_fixed_step_length=acados_fixed_step_length,
             acados_nlp_qp_tol_strategy=acados_nlp_qp_tol_strategy,
+            acados_qpscaling_scale_objective=acados_qpscaling_scale_objective,
+            acados_qpscaling_scale_constraints=acados_qpscaling_scale_constraints,
+            acados_ext_qp_res=acados_ext_qp_res,
+            acados_project_qdot_from_q=acados_project_qdot_from_q,
+            disable_periodic_fes_warmup_projection=(
+                disable_periodic_fes_warmup_projection
+            ),
+            periodic_fes_warmup_projection_weight=periodic_fes_warmup_projection_weight,
+            periodic_fes_warmup_projection_mode=periodic_fes_warmup_projection_mode,
+            periodic_fes_warmup_projection_strategy=(
+                periodic_fes_warmup_projection_strategy
+            ),
+            periodic_fes_warmup_projection_substeps=(
[…]
