### 2026-07-04 e240cb8
Add state trace comparison diagnostics
