+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -166,6 +166,51 @@ def _control_comparisons(ipopt_result: dict, acados_result: dict) -> list[dict]:
     return comparisons
 
 
+def _state_comparisons(ipopt_result: dict, acados_result: dict) -> list[dict]:
+    ipopt_states = ipopt_result.get("state_traces", {})
+    acados_states = acados_result.get("state_traces", {})
+    common_keys = sorted(set(ipopt_states).intersection(acados_states))
+    comparisons = []
+
+    for key in common_keys:
+        ipopt_values = np.asarray(ipopt_states[key], dtype=float)
+        acados_values = np.asarray(acados_states[key], dtype=float)
+        if ipopt_values.ndim == 1:
+            ipopt_values = ipopt_values[np.newaxis, :]
+        if acados_values.ndim == 1:
+            acados_values = acados_values[np.newaxis, :]
+        if ipopt_values.size == 0 or acados_values.size == 0:
+            continue
+
+        for row in range(min(ipopt_values.shape[0], acados_values.shape[0])):
+            common_len = max(ipopt_values.shape[1], acados_values.shape[1])
+            ipopt_common = _resample_trace(ipopt_values[row, :], common_len)
+            acados_common = _resample_trace(acados_values[row, :], common_len)
+            diff = acados_common - ipopt_common
+            comparisons.append(
+                {
+                    "key": key if ipopt_values.shape[0] == 1 else f"{key}[{row}]",
+                    "common_len": common_len,
+                    "rmse": float(np.sqrt(np.mean(diff**2))),
+                    "mae": float(np.mean(np.abs(diff))),
+                    "max_abs_error": float(np.max(np.abs(diff))),
+                    "final_error": float(diff[-1]),
+                    "ipopt_mean": float(np.mean(ipopt_common)),
+                    "acados_mean": float(np.mean(acados_common)),
+                    "ipopt_range": (
+                        float(np.min(ipopt_common)),
+                        float(np.max(ipopt_common)),
+                    ),
+                    "acados_range": (
+                        float(np.min(acados_common)),
+                        float(np.max(acados_common)),
+                    ),
+                }
+            )
+
+    return sorted(comparisons, key=lambda item: item["rmse"], reverse=True)
+
+
 def _solver_config(
     solver_name: str,
     objective: str,
@@ -340,7 +385,10 @@ def _solver_config(
 
 
 def print_comparison(
-    ipopt_result: dict, acados_result: dict, print_traces: bool = False
+    ipopt_result: dict,
+    acados_result: dict,
+    print_traces: bool = False,
+    state_comparison_limit: int = 12,
 ) -> None:
     print(
         "solver | success | solver_success | physical_success | status | status_label | objective | solver_time_s | wall_time_s | final_wheel_angle | "
@@ -444,6 +492,31 @@ def print_comparison(
     else:
         print("control comparison warning: no common control keys were found.")
 
+    state_metrics = _state_comparisons(ipopt_result, acados_result)
+    if state_metrics and state_comparison_limit:
+        print(
+            "state comparison | key | common_len | rmse | mae | max_abs_error | final_error | "
+            "ipopt_mean | acados_mean | ipopt_range | acados_range"
+        )
+        for metric in state_metrics[:state_comparison_limit]:
+            ipopt_min, ipopt_max = metric["ipopt_range"]
+            acados_min, acados_max = metric["acados_range"]
+            print(
+                "state comparison | "
+                f"{metric['key']} | "
+                f"{metric['common_len']} | "
+                f"{_format_control_metric(metric['rmse'])} | "
[…]
