### 2026-07-04 44476ab
Add warmup state diagnostics for ACADOS cycling
