+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -461,6 +461,12 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Project the ACADOS qdot initial guess from finite differences of q before solving.",
     )
     parser.add_argument(
+        "--warmup-state-comparison-limit",
+        type=int,
+        default=12,
+        help="Number of warmup-vs-initial-guess state rows to print when ACADOS diagnostics are enabled.",
+    )
+    parser.add_argument(
         "--max-ipopt-iterations",
         type=int,
         default=2000,
@@ -2405,6 +2411,157 @@ def _adapt_warmup_solution_to_periodic_nodes(
     return _WarmupSolutionAdapter(adapted_states, adapted_controls)
 
 
+def _warmup_state_is_directly_comparable(key: str) -> bool:
+    return key in ("q", "qdot") or key.startswith(("F_", "A_", "Tau1_", "Km_"))
+
+
+def _resample_trace(values: np.ndarray, target_len: int) -> np.ndarray:
+    values = np.asarray(values, dtype=float).reshape(-1)
+    if values.size == target_len:
+        return values
+    current_grid = np.linspace(0.0, 1.0, values.size)
+    target_grid = np.linspace(0.0, 1.0, target_len)
+    return np.interp(target_grid, current_grid, values)
+
+
+def _expected_receding_state_initial_guess(
+    periodic_nmpc, states: dict, key: str, row: int
+) -> np.ndarray:
+    state_values = np.asarray(states[key], dtype=float)
+    if state_values.ndim == 1:
+        state_values = state_values[np.newaxis, :]
+    values = state_values[row, :]
+    nodes_per_cycle = periodic_nmpc.nodes_per_cycle
+
+    if key == "q" and row == state_values.shape[0] - 1:
+        if not getattr(periodic_nmpc, "use_signed_wheel_shift", False):
+            shifted_n_plus_one_cycles = (
+                values[nodes_per_cycle:-1] + periodic_nmpc.pedal_turn_in_one_cycle
+            )
+            last_cycle = values[-nodes_per_cycle - 1 :]
+            if shifted_n_plus_one_cycles.size == 0:
+                return last_cycle
+            return np.concatenate((shifted_n_plus_one_cycles, last_cycle))
+
+        wheel_cycle_shift = periodic_nmpc._wheel_cycle_shift(states)
+        n_plus_one_cycles = values[nodes_per_cycle:-1]
+        shifted_last_cycle = values[-nodes_per_cycle - 1 :] + wheel_cycle_shift
+        if n_plus_one_cycles.size == 0:
+            return shifted_last_cycle
+        return np.concatenate((n_plus_one_cycles, shifted_last_cycle))
+
+    if key in ("q", "qdot") or key.startswith(("Cn_", "Cn_sum_", "F_")):
+        n_plus_one_cycles = values[nodes_per_cycle:-1]
+        last_cycle = values[-nodes_per_cycle - 1 :]
+        if n_plus_one_cycles.size == 0:
+            return last_cycle
+        return np.concatenate((n_plus_one_cycles, last_cycle))
+
+    if key.startswith(("A_", "Tau1_", "Km_")):
+        n_plus_one_cycles = values[nodes_per_cycle:-1]
+        last_cycle = values[-nodes_per_cycle - 1 :]
+        if n_plus_one_cycles.size == 0:
+            return last_cycle
+        delta = n_plus_one_cycles[-1] - last_cycle[0]
+        shifted_last_cycle = last_cycle + delta
+        return np.concatenate((n_plus_one_cycles, shifted_last_cycle))
+
+    return values
+
+
+def _warmup_state_comparisons(reference_solution, periodic_nmpc) -> list[dict]:
[…]
