### 2026-07-06 7690adb
Compare ACADOS solution against initial guess
