### 2026-07-06 d29b2ca
Add ACADOS pulse width trust region option
