### 2026-07-06 1a6d040
Add ACADOS initial defect node diagnostics
