+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -248,6 +248,7 @@ def _solver_config(
     state_scaling: str,
     pulse_width_scaling: float,
     acados_pulse_width_trust_radius: float | None,
+    acados_fatigue_warmstart_mode: str,
     acados_tolerance: float | None,
     acados_qp_iter_max: int,
     acados_levenberg_marquardt: float,
@@ -305,6 +306,7 @@ def _solver_config(
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_pulse_width_trust_radius=None,
+            acados_fatigue_warmstart_mode="continuous",
             acados_tolerance=acados_tolerance,
             acados_qp_iter_max=acados_qp_iter_max,
             acados_levenberg_marquardt=acados_levenberg_marquardt,
@@ -369,6 +371,7 @@ def _solver_config(
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_pulse_width_trust_radius=acados_pulse_width_trust_radius,
+            acados_fatigue_warmstart_mode=acados_fatigue_warmstart_mode,
             acados_tolerance=acados_tolerance,
             acados_qp_iter_max=acados_qp_iter_max,
             acados_levenberg_marquardt=acados_levenberg_marquardt,
@@ -612,6 +615,7 @@ def main(
     pulse_width_scaling: float = 1 / 400,
     acados_pulse_width_scaling: float | None = None,
     acados_pulse_width_trust_radius: float | None = None,
+    acados_fatigue_warmstart_mode: str = "continuous",
     acados_tolerance: float | None = None,
     acados_qp_iter_max: int = 50,
     acados_levenberg_marquardt: float = 0.0,
@@ -664,6 +668,7 @@ def main(
         state_scaling=state_scaling,
         pulse_width_scaling=pulse_width_scaling,
         acados_pulse_width_trust_radius=None,
+        acados_fatigue_warmstart_mode="continuous",
         acados_tolerance=acados_tolerance,
         acados_qp_iter_max=acados_qp_iter_max,
         acados_levenberg_marquardt=acados_levenberg_marquardt,
@@ -730,6 +735,7 @@ def main(
             else pulse_width_scaling
         ),
         acados_pulse_width_trust_radius=acados_pulse_width_trust_radius,
+        acados_fatigue_warmstart_mode=acados_fatigue_warmstart_mode,
         acados_tolerance=acados_tolerance,
         acados_qp_iter_max=acados_qp_iter_max,
         acados_levenberg_marquardt=acados_levenberg_marquardt,
@@ -822,6 +828,11 @@ def build_cli() -> argparse.ArgumentParser:
     parser.add_argument("--pulse-width-scaling", type=float, default=1 / 400)
     parser.add_argument("--acados-pulse-width-scaling", type=float, default=None)
     parser.add_argument("--acados-pulse-width-trust-radius", type=float, default=None)
+    parser.add_argument(
+        "--acados-fatigue-warmstart-mode",
+        choices=("continuous", "cyclical"),
+        default="continuous",
+    )
     parser.add_argument("--acados-tolerance", type=float, default=None)
     parser.add_argument("--acados-qp-iter-max", type=int, default=50)
     parser.add_argument("--acados-levenberg-marquardt", type=float, default=0.0)
@@ -948,6 +959,7 @@ if __name__ == "__main__":
         pulse_width_scaling=args.pulse_width_scaling,
         acados_pulse_width_scaling=args.acados_pulse_width_scaling,
         acados_pulse_width_trust_radius=args.acados_pulse_width_trust_radius,
[…]
