### 2026-07-06 5fbedb1
Add blended force warmup projection diagnostics
