+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -553,11 +553,21 @@ def build_argument_parser() -> argparse.ArgumentParser:
     )
     parser.add_argument(
         "--periodic-fes-warmup-projection-mode",
-        choices=("calcium", "all", "all_except_force"),
+        choices=("calcium", "all", "all_except_force", "all_force_blend"),
         default="all",
         help=(
             "Project only the periodic calcium states, all Ding fatigue states, "
-            "or all states except F to preserve the multibody force trajectory."
+            "all states except F to preserve the multibody force trajectory, "
+            "or all states with a separate blend weight for F."
+        ),
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-force-projection-weight",
+        type=float,
+        default=0.25,
+        help=(
+            "Blend weight used only for F states when "
+            "--periodic-fes-warmup-projection-mode=all_force_blend."
         ),
     )
     parser.add_argument(
@@ -2024,10 +2034,11 @@ def _projection_state_keys(muscle_name: str, projection_mode: str) -> tuple[str,
     state_keys = _ding_state_keys(muscle_name)
     if projection_mode == "calcium":
         return state_keys[:2]
-    if projection_mode in ("all", "all_except_force"):
+    if projection_mode in ("all", "all_except_force", "all_force_blend"):
         return state_keys
     raise ValueError(
-        "--periodic-fes-warmup-projection-mode must be 'calcium', 'all' or 'all_except_force'."
+        "--periodic-fes-warmup-projection-mode must be 'calcium', 'all', "
+        "'all_except_force' or 'all_force_blend'."
     )
 
 
@@ -2332,10 +2343,15 @@ def project_periodic_fes_initial_guess(
     projection_defect_weight: float = 100.0,
     projection_trust_radius: float | None = None,
     projection_max_iterations: int = 200,
+    force_projection_weight: float = 0.25,
 ) -> dict[str, float]:
     if projection_weight < 0.0 or projection_weight > 1.0:
         raise ValueError(
             "--periodic-fes-warmup-projection-weight must be between 0 and 1."
+        )
+    if force_projection_weight < 0.0 or force_projection_weight > 1.0:
+        raise ValueError(
+            "--periodic-fes-warmup-force-projection-weight must be between 0 and 1."
         )
     if projection_strategy not in ("rollout", "sequential", "least_squares"):
         raise ValueError(
@@ -2428,7 +2444,14 @@ def project_periodic_fes_initial_guess(
         for state_idx, key in enumerate(state_keys):
             if key not in write_state_keys:
                 continue
-            values = blended_states[state_idx, :]
+            if projection_mode == "all_force_blend" and key.startswith("F_"):
+                state_projection_weight = force_projection_weight
+                values = (
+                    state_projection_weight * projected_states[state_idx, :]
+                    + (1.0 - state_projection_weight) * original_states[state_idx, :]
+                )
+            else:
+                values = blended_states[state_idx, :]
             lower, upper = _state_trajectory_bounds(periodic_nmpc, key, values.size)
             lower_violation = np.maximum(lower - values, 0.0)
             upper_violation = np.maximum(values - upper, 0.0)
@@ -2454,6 +2477,7 @@ def project_periodic_fes_initial_guess(
         "projection_mode": projection_mode,
         "projection_strategy": projection_strategy,
         "projection_substeps": projection_substeps,
+        "force_projection_weight": force_projection_weight,
         "projection_proximity_weight": projection_proximity_weight,
         "projection_defect_weight": projection_defect_weight,
         "projection_trust_radius": (
@@ -2745,6 +2769,7 @@ def apply_standard_warmup_to_periodic_nmpc(
[…]
