+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -268,7 +268,11 @@ def _solver_config(
     periodic_fes_warmup_projection_mode: str,
     periodic_fes_warmup_projection_strategy: str,
     periodic_fes_warmup_projection_substeps: int,
+    periodic_fes_warmup_projection_proximity_weight: float,
     periodic_fes_warmup_projection_defect_weight: float,
+    periodic_fes_warmup_projection_trust_radius: float | None,
+    periodic_fes_warmup_projection_max_iterations: int,
+    periodic_fes_warmup_force_projection_weight: float,
     acados_diagnostics: bool,
     periodic_ipopt_refinement: bool,
     periodic_ipopt_refinement_iterations: int,
@@ -330,8 +334,20 @@ def _solver_config(
             periodic_fes_warmup_projection_substeps=(
                 periodic_fes_warmup_projection_substeps
             ),
+            periodic_fes_warmup_projection_proximity_weight=(
+                periodic_fes_warmup_projection_proximity_weight
+            ),
             periodic_fes_warmup_projection_defect_weight=(
                 periodic_fes_warmup_projection_defect_weight
+            ),
+            periodic_fes_warmup_projection_trust_radius=(
+                periodic_fes_warmup_projection_trust_radius
+            ),
+            periodic_fes_warmup_projection_max_iterations=(
+                periodic_fes_warmup_projection_max_iterations
+            ),
+            periodic_fes_warmup_force_projection_weight=(
+                periodic_fes_warmup_force_projection_weight
             ),
             acados_diagnostics=False,
             periodic_ipopt_refinement=False,
@@ -397,8 +413,20 @@ def _solver_config(
             periodic_fes_warmup_projection_substeps=(
                 periodic_fes_warmup_projection_substeps
             ),
+            periodic_fes_warmup_projection_proximity_weight=(
+                periodic_fes_warmup_projection_proximity_weight
+            ),
             periodic_fes_warmup_projection_defect_weight=(
                 periodic_fes_warmup_projection_defect_weight
+            ),
+            periodic_fes_warmup_projection_trust_radius=(
+                periodic_fes_warmup_projection_trust_radius
+            ),
+            periodic_fes_warmup_projection_max_iterations=(
+                periodic_fes_warmup_projection_max_iterations
+            ),
+            periodic_fes_warmup_force_projection_weight=(
+                periodic_fes_warmup_force_projection_weight
             ),
             acados_diagnostics=acados_diagnostics,
             periodic_ipopt_refinement=periodic_ipopt_refinement,
@@ -635,7 +663,11 @@ def main(
     periodic_fes_warmup_projection_mode: str = "all",
     periodic_fes_warmup_projection_strategy: str = "sequential",
     periodic_fes_warmup_projection_substeps: int = 10,
+    periodic_fes_warmup_projection_proximity_weight: float = 1.0,
     periodic_fes_warmup_projection_defect_weight: float = 100.0,
+    periodic_fes_warmup_projection_trust_radius: float | None = None,
+    periodic_fes_warmup_projection_max_iterations: int = 200,
+    periodic_fes_warmup_force_projection_weight: float = 0.25,
     acados_diagnostics: bool = False,
     periodic_ipopt_refinement: bool = False,
     periodic_ipopt_refinement_iterations: int = 300,
@@ -688,8 +720,20 @@ def main(
         periodic_fes_warmup_projection_mode=periodic_fes_warmup_projection_mode,
         periodic_fes_warmup_projection_strategy=periodic_fes_warmup_projection_strategy,
         periodic_fes_warmup_projection_substeps=periodic_fes_warmup_projection_substeps,
[…]
