### 2026-07-08 50d744e
Add adaptive force warmup projection
