+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -248,6 +248,7 @@ def _solver_config(
     state_scaling: str,
     pulse_width_scaling: float,
     acados_pulse_width_trust_radius: float | None,
+    acados_fes_state_trust_radius: float | None,
     acados_fatigue_warmstart_mode: str,
     acados_tolerance: float | None,
     acados_qp_iter_max: int,
@@ -273,6 +274,8 @@ def _solver_config(
     periodic_fes_warmup_projection_trust_radius: float | None,
     periodic_fes_warmup_projection_max_iterations: int,
     periodic_fes_warmup_force_projection_weight: float,
+    periodic_fes_warmup_force_qdot_defect_limit: float,
+    periodic_fes_warmup_force_adaptive_steps: int,
     acados_diagnostics: bool,
     periodic_ipopt_refinement: bool,
     periodic_ipopt_refinement_iterations: int,
@@ -310,6 +313,7 @@ def _solver_config(
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_pulse_width_trust_radius=None,
+            acados_fes_state_trust_radius=None,
             acados_fatigue_warmstart_mode="continuous",
             acados_tolerance=acados_tolerance,
             acados_qp_iter_max=acados_qp_iter_max,
@@ -349,6 +353,12 @@ def _solver_config(
             periodic_fes_warmup_force_projection_weight=(
                 periodic_fes_warmup_force_projection_weight
             ),
+            periodic_fes_warmup_force_qdot_defect_limit=(
+                periodic_fes_warmup_force_qdot_defect_limit
+            ),
+            periodic_fes_warmup_force_adaptive_steps=(
+                periodic_fes_warmup_force_adaptive_steps
+            ),
             acados_diagnostics=False,
             periodic_ipopt_refinement=False,
             periodic_ipopt_refinement_iterations=periodic_ipopt_refinement_iterations,
@@ -387,6 +397,7 @@ def _solver_config(
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_pulse_width_trust_radius=acados_pulse_width_trust_radius,
+            acados_fes_state_trust_radius=acados_fes_state_trust_radius,
             acados_fatigue_warmstart_mode=acados_fatigue_warmstart_mode,
             acados_tolerance=acados_tolerance,
             acados_qp_iter_max=acados_qp_iter_max,
@@ -427,6 +438,12 @@ def _solver_config(
             ),
             periodic_fes_warmup_force_projection_weight=(
                 periodic_fes_warmup_force_projection_weight
+            ),
+            periodic_fes_warmup_force_qdot_defect_limit=(
+                periodic_fes_warmup_force_qdot_defect_limit
+            ),
+            periodic_fes_warmup_force_adaptive_steps=(
+                periodic_fes_warmup_force_adaptive_steps
             ),
             acados_diagnostics=acados_diagnostics,
             periodic_ipopt_refinement=periodic_ipopt_refinement,
@@ -643,6 +660,7 @@ def main(
     pulse_width_scaling: float = 1 / 400,
     acados_pulse_width_scaling: float | None = None,
     acados_pulse_width_trust_radius: float | None = None,
+    acados_fes_state_trust_radius: float | None = None,
     acados_fatigue_warmstart_mode: str = "continuous",
     acados_tolerance: float | None = None,
     acados_qp_iter_max: int = 50,
@@ -668,6 +686,8 @@ def main(
     periodic_fes_warmup_projection_trust_radius: float | None = None,
     periodic_fes_warmup_projection_max_iterations: int = 200,
[…]
