+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -276,6 +276,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--acados-fes-state-trust-radius",
+        type=float,
+        default=None,
+        help=(
+            "Optional normalized trust-region radius around the current ACADOS FES state "
+            "initial guess range. Applies to Cn, Cn_sum, F, A, Tau1 and Km states."
+        ),
+    )
+    parser.add_argument(
         "--acados-fatigue-warmstart-mode",
         choices=("continuous", "cyclical"),
         default="continuous",
@@ -553,12 +562,19 @@ def build_argument_parser() -> argparse.ArgumentParser:
     )
     parser.add_argument(
         "--periodic-fes-warmup-projection-mode",
-        choices=("calcium", "all", "all_except_force", "all_force_blend"),
+        choices=(
+            "calcium",
+            "all",
+            "all_except_force",
+            "all_force_blend",
+            "all_force_adaptive_blend",
+        ),
         default="all",
         help=(
             "Project only the periodic calcium states, all Ding fatigue states, "
             "all states except F to preserve the multibody force trajectory, "
-            "or all states with a separate blend weight for F."
+            "all states with a separate blend weight for F, or all states with "
+            "an automatically reduced F blend weight constrained by a qdot defect limit."
         ),
     )
     parser.add_argument(
@@ -567,8 +583,24 @@ def build_argument_parser() -> argparse.ArgumentParser:
         default=0.25,
         help=(
             "Blend weight used only for F states when "
-            "--periodic-fes-warmup-projection-mode=all_force_blend."
+            "--periodic-fes-warmup-projection-mode is all_force_blend or "
+            "all_force_adaptive_blend."
         ),
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-force-qdot-defect-limit",
+        type=float,
+        default=3.0,
+        help=(
+            "Maximum full-dynamics qdot defect allowed while selecting the F blend weight "
+            "when --periodic-fes-warmup-projection-mode=all_force_adaptive_blend."
+        ),
+    )
+    parser.add_argument(
+        "--periodic-fes-warmup-force-adaptive-steps",
+        type=int,
+        default=10,
+        help="Number of candidate F blend weights tested in all_force_adaptive_blend mode.",
     )
     parser.add_argument(
         "--periodic-fes-warmup-projection-strategy",
@@ -2034,11 +2066,16 @@ def _projection_state_keys(muscle_name: str, projection_mode: str) -> tuple[str,
     state_keys = _ding_state_keys(muscle_name)
     if projection_mode == "calcium":
         return state_keys[:2]
-    if projection_mode in ("all", "all_except_force", "all_force_blend"):
+    if projection_mode in (
+        "all",
+        "all_except_force",
+        "all_force_blend",
+        "all_force_adaptive_blend",
+    ):
         return state_keys
     raise ValueError(
         "--periodic-fes-warmup-projection-mode must be 'calcium', 'all', "
-        "'all_except_force' or 'all_force_blend'."
+        "'all_except_force', 'all_force_blend' or 'all_force_adaptive_blend'."
     )
 
 
@@ -2050,6 +2087,71 @@ def _projection_write_state_keys(
         force_key = f"F_{muscle_name}"
         return tuple(key for key in state_keys if key != force_key)
     return state_keys
+
+
+def _qdot_defect_max_from_full_dynamics(nmpc) -> float:
+    defects = _full_dynamics_rollout_defect_details(nmpc)
+    qdot_defects = defects.get("qdot_by_dof", {}) if defects else {}
[…]
