### 2026-07-08 2123489
Add IPOPT ACADOS-like comparison profile
