+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -26,6 +26,35 @@ from cycling_pulse_width_mhe_acados_periodic import (
 
 EXAMPLE_DIR = Path(__file__).resolve().parent
 
+IPOPT_PROFILE_DEFAULTS = {
+    "historical": {
+        "model_formulation": "standard",
+        "torque_application": "external_forces",
+        "ode_solver": "collocation",
+        "rk_steps": 1,
+        "collocation_degree": 3,
+        "collocation_method": "radau",
+        "use_sx": False,
+        "enforce_start_constraints": True,
+        "disable_standard_ipopt_warmup": False,
+        "disable_periodic_fes_warmup_projection": True,
+        "fatigue_warmstart_mode": "continuous",
+    },
+    "acados_like": {
+        "model_formulation": "periodic",
+        "torque_application": "constant",
+        "ode_solver": "rk4",
+        "rk_steps": 5,
+        "collocation_degree": 3,
+        "collocation_method": "radau",
+        "use_sx": True,
+        "enforce_start_constraints": False,
+        "disable_standard_ipopt_warmup": False,
+        "disable_periodic_fes_warmup_projection": None,
+        "fatigue_warmstart_mode": None,
+    },
+}
+
 
 def _namespace_from_cli(**overrides) -> argparse.Namespace:
     parser = build_argument_parser()
@@ -228,6 +257,14 @@ def _initial_guess_state_comparisons(acados_result: dict) -> list[dict]:
     )
 
 
+def _normalize_ipopt_profile(profile: str) -> str:
+    return profile.replace("-", "_")
+
+
+def _pick(value, fallback):
+    return fallback if value is None else value
+
+
 def _solver_config(
     solver_name: str,
     objective: str,
@@ -281,13 +318,43 @@ def _solver_config(
     periodic_ipopt_refinement_iterations: int,
     periodic_ipopt_refinement_use_sx: bool,
     warmup_state_comparison_limit: int,
+    ipopt_profile: str = "historical",
+    ipopt_model_formulation: str | None = None,
+    ipopt_torque_application: str | None = None,
+    ipopt_ode_solver: str | None = None,
+    ipopt_rk_steps: int | None = None,
+    ipopt_collocation_degree: int | None = None,
+    ipopt_collocation_method: str | None = None,
+    ipopt_use_sx: bool | None = None,
+    ipopt_enforce_start_constraints: bool | None = None,
+    ipopt_disable_standard_warmup: bool | None = None,
+    ipopt_disable_periodic_fes_warmup_projection: bool | None = None,
+    ipopt_fatigue_warmstart_mode: str | None = None,
+    ipopt_disable_historical_initial_guess: bool = False,
 ) -> argparse.Namespace:
     if solver_name == "ipopt":
+        normalized_profile = _normalize_ipopt_profile(ipopt_profile)
+        if normalized_profile not in IPOPT_PROFILE_DEFAULTS:
+            raise ValueError("--ipopt-profile must be 'historical' or 'acados_like'.")
+
+        defaults = IPOPT_PROFILE_DEFAULTS[normalized_profile]
+        disable_projection_default = defaults["disable_periodic_fes_warmup_projection"]
+        if disable_projection_default is None:
+            disable_projection_default = disable_periodic_fes_warmup_projection
+
+        fatigue_warmstart_default = defaults["fatigue_warmstart_mode"]
+        if fatigue_warmstart_default is None:
+            fatigue_warmstart_default = acados_fatigue_warmstart_mode
+
         return _namespace_from_cli(
             solver="ipopt",
             single_shot=False,
-            model_formulation="standard",
-            torque_application="external_forces",
+            model_formulation=_pick(
+                ipopt_model_formulation, defaults["model_formulation"]
+            ),
[…]
