+++ examples/fes_multibody/cycling/cycling_standard_clinics.py
@@ -1,540 +1,0 @@
-"""
-This example demonstrates how long hand cycling could be maintained at 30 Hz with standard clinical settings (fixed
-pulse width and defined stimulation angles). The crank torque will be optimized to achieve the target cadence (60 rpm).
-Then, after each cycle, the muscle fatigue will be updated based on the activation history and tested with an optimal
-control problem. If an optimal solution is found, the standard clinical settings simulation continues considering a
-cycle could be performed. When an optimal solution is not found, this ends meaning it reached muscle failure.
-"""
-
-from sys import platform
-import pickle
-import numpy as np
-
-from bioptim import (
-    OptimalControlProgram,
-    ObjectiveList,
-    ObjectiveFcn,
-    ConstraintList,
-    ConstraintFcn,
-    ParameterList,
-    BoundsList,
-    InitialGuessList,
-    ParameterObjectiveList,
-    Node,
-    Axis,
-    OdeSolver,
-    VariableScalingList,
-    InterpolationType,
-    SolutionMerge,
-    Solver,
-    CostType,
-)
-
-from cycling_pulse_width_mhe import (
-    set_fes_model,
-    set_external_forces,
-    set_dynamics,
-    set_q_qdot_init,
-    set_x_bounds,
-    updating_model,
-    set_u_bounds_and_init,
-)
-from cost_functions import CustomCostFunctions
-
-from collections import defaultdict
-
-MUSCLES = ("Delt_ant", "Delt_post", "Biceps", "Triceps")
-STATE_VARS = ("A", "F", "Km", "Tau1")
-CONTROL_VARS = ("last_pulse_width",)
-
-
-def prepare_standard_fes_cycling(optim_info, cycling_info, model, previous_problem=None, previous_sol=None):
-    # --- Optimization info --- #
-    phase_time = optim_info["phase_time"]
-    n_shooting = optim_info["n_shooting"]
-    ode_solver = optim_info["ode_solver"]
-    use_sx = optim_info["use_sx"]
-
-    # --- Cycling info --- #
-    turn_number = cycling_info["turn_number"]
-    pedal_config = cycling_info["pedal_config"]
-    external_force = cycling_info["resistive_torque"]
-
-    # --- Set dynamics --- #
-    # --- External force numerical time series --- #
-    numerical_time_series, external_force_set = set_external_forces(
-        n_shooting=n_shooting, external_force_dict=external_force, force_name="external_torque"
-    )
-    # --- Stimulation instant numerical time series --- #
-    numerical_data_time_series, stim_idx_at_node_list = model.muscles_dynamics_model[0].get_numerical_data_time_series(
-        n_shooting, phase_time
-    )
-    numerical_time_series.update(numerical_data_time_series)
-    # --- Dynamics --- #
-    dynamics = set_dynamics(model=model, numerical_time_series=numerical_time_series, ode_solver=ode_solver)
-
-    # --- Set states --- #
-    if previous_problem is None:
-        # --- Set q (position and speed) initial guesses --- #
-        x_init = set_q_qdot_init(
-            n_shooting=n_shooting,
-            pedal_config=pedal_config,
-            turn_number=turn_number,
-            ode_solver=ode_solver,
-            init_file_path=None,
-        )
-
-        # --- Set bounds and FES initial guesses --- #
-        x_bounds, x_init = set_x_bounds(
-            model=model,
-            x_init=x_init,
-            n_shooting=n_shooting,
-            ode_solver=ode_solver,
-            init_file_path=None,
-        )
-    else:
[…]
