+++ examples/fes_multibody/cycling/cost_functions.py
@@ -2,8 +2,8 @@ from casadi import MX, vertcat, sum1, mmax
 from bioptim import PenaltyController
 from cocofest.models.ding2007.ding2007 import DingModelPulseWidthFrequency
 
-BAYESIAN_WEIGHT = vertcat([191790, 31609, 117259, 1])
-PHYSIOLOGICAL_WEIGHT = vertcat([191790, 31609, 117259, 1])
+BAYESIAN_WEIGHT = vertcat([10000, 1560, 4665, 0.0001])  # example values from minimize_root_mean_square_fatigue bayesian weight optimization
+PHYSIOLOGICAL_WEIGHT = vertcat([1, 0.0943160798625659, 0.389268746606572, 0])  # example values from physiological weight calculation
 
 
 class CustomCostFunctions:
@@ -45,28 +45,28 @@ class CustomCostFunctions:
                 "index": "5",
                 "description": "Minimize the average muscle force",
                 "power": "1",
-                "state": r"F^{m}",
+                "state": r"f^{m}",
             },
             "minimize_root_mean_square_force": {
                 "function": self.minimize_root_mean_square_force,
                 "index": "6",
                 "description": "Minimize the root mean square of muscle force",
                 "power": "2",
-                "state": r"F^{m}",
+                "state": r"f^{m}",
             },
             "minimize_cubic_average_force": {
                 "function": self.minimize_cubic_average_force,
                 "index": "7",
                 "description": "Minimize the cubic average of muscle force",
                 "power": "3",
-                "state": r"F^{m}",
+                "state": r"f^{m}",
             },
             "minimize_peak_force": {
                 "function": self.minimize_peak_force,
                 "index": "8",
                 "description": "Minimize the peak muscle force",
                 "power": r"\infty",
-                "state": r"F^{m}",
+                "state": r"f^{m}",
             },
             # --- Stress --- #
             "minimize_average_muscle_stress": {
@@ -103,28 +103,28 @@ class CustomCostFunctions:
                 "index": "13",
                 "description": "Minimize the average muscle fatigue",
                 "power": "1",
-                "state": "A",
+                "state": "a'",
             },
             "minimize_root_mean_square_fatigue": {
                 "function": self.minimize_root_mean_square_fatigue,
                 "index": "14",
                 "description": "Minimize the root mean square of muscle fatigue",
                 "power": "2",
-                "state": "A",
+                "state": "a'",
             },
             "minimize_cubic_average_fatigue": {
                 "function": self.minimize_cubic_average_fatigue,
                 "index": "15",
                 "description": "Minimize the cubic average of muscle fatigue",
                 "power": "3",
-                "state": "A",
+                "state": "a'",
             },
             "minimize_peak_fatigue": {
                 "function": self.minimize_peak_fatigue,
                 "index": "16",
                 "description": "Minimize the peak muscle fatigue",
                 "power": r"\infty",
-                "state": "A",
+                "state": "a'",
             },
             # --- Power --- #
             "minimize_root_mean_square_muscle_power": {
@@ -133,7 +133,7 @@ class CustomCostFunctions:
                 "latex": r"\phi_{17} = \left(\frac{1}{M}\sum_{m=1}^{M} (f^{m} v^{m})^{2}\right)^{\tfrac{1}{2}}",
                 "description": "Minimize the root mean square of muscle power",
                 "power": "2",
-                "state": "W",
+                "state": "w",
             },
             # --- BAYESIAN --- #
             # --- Pulse width --- #
[…]
