+++ examples/fes_multibody/cycling/cycling_standard_clinics.py
@@ -36,7 +36,7 @@ from bioptim import (
     CostType,
 )
 
-from cycling_pulse_width_mhe import (
+from cycling.cycling_pulse_width_mhe import (
     set_fes_model,
     set_external_forces,
     set_dynamics,
@@ -45,20 +45,19 @@ from cycling_pulse_width_mhe import (
     updating_model,
     set_u_bounds_and_init,
 )
-from cost_functions import CustomCostFunctions
+from cycling.cost_functions import CustomCostFunctions
 
 
 # -----------------------------------------------------------------------------
 # Global settings
 # -----------------------------------------------------------------------------
-
 TARGET_RPM = 60
 PHASE_TIME = 60 / TARGET_RPM  # 1 second per revolution at 60 rpm
 N_SHOOTING = 30  # 30 Hz for a 1-second cycle
 N_THREADS = 48
 
-MODEL_PATH = "../../../msk_models/Wu/Modified_Wu_Shoulder_Model_Cycling.bioMod"
-RESULT_FILE_NAME = "standard_and_fes_ocp_cycles.pkl"
+MODEL_PATH = "../../msk_models/Wu/Modified_Wu_Shoulder_Model_Cycling.bioMod"
+RESULT_FILE_NAME = "temp/standard_and_fes_ocp_cycles.pkl"
 
 MUSCLES = ("Delt_ant", "Delt_post", "Biceps", "Triceps")
 STATE_VARS = ("A", "F", "Km", "Tau1")
@@ -66,13 +65,11 @@ CONTROL_VARS = ("last_pulse_width",)
 
 
 # -----------------------------------------------------------------------------
-# Small utilities
+# Utilities
 # -----------------------------------------------------------------------------
-
 def target_crank_velocity(turn_number: int, phase_time: float) -> float:
     """Return the target crank angular velocity in rad/s."""
     return -2 * np.pi * turn_number / phase_time
-
 
 def compute_net_work(time: np.ndarray, torque: np.ndarray) -> float:
     """
@@ -118,15 +115,13 @@ def compute_net_work(time: np.ndarray, torque: np.ndarray) -> float:
         f"Got len(time)={time.size}, len(torque)={torque.size}."
     )
 
-
 def compute_constant_torque_work(time: np.ndarray, constant_torque: float) -> float:
     """Compute net work for a constant torque over the provided time vector."""
     time = np.asarray(time, dtype=float).squeeze()
     torque = np.full(time.shape, constant_torque, dtype=float)
     return compute_net_work(time=time, torque=torque)
 
-
-def prepare_solver(max_iter: int = 10000):
+def prepare_solver(max_iter: int = 10):
     """Prepare the IPOPT solver."""
     solver = Solver.IPOPT(
         show_online_optim=False,
@@ -143,7 +138,6 @@ def prepare_solver(max_iter: int = 10000):
 # -----------------------------------------------------------------------------
 # Pulse-width settings for the clinical condition
 # -----------------------------------------------------------------------------
-
 def is_angle_in_range(angle_vector: np.ndarray, start: float, end: float) -> np.ndarray:
     """
     Return a Boolean vector indicating whether each angle is inside a stimulation range.
@@ -155,7 +149,6 @@ def is_angle_in_range(angle_vector: np.ndarray, start: float, end: float) -> np.
         return (angle_vector >= start) & (angle_vector <= end)
 
     return (angle_vector >= start) | (angle_vector <= end)
-
 
 def set_pw_dictionary(muscle_models, n_shooting: int, active_pw: float = 0.0003):
     """Create the fixed pulse-width dictionary for the standard clinical stimulation pattern."""
[…]

[…]
