+++ examples/sensitivity/truncation/summation_truncation_graph.py
@@ -1,267 +1,0 @@
-"""
-Obsolete example need to be updated
-"""
-
-# import pickle
-# import numpy as np
-# import matplotlib.pyplot as plt
-# from matplotlib.ticker import MaxNLocator, FixedLocator, IndexLocator
-# from matplotlib import colors
-#
-#
-# # --- Extracting the data from the files --- #
-# mode = ["single"]
-# for desired_mode in mode:
-#     if desired_mode == "single":
-#         with open(r"../../data/truncation_single.pkl", "rb") as f:
-#             data = pickle.load(f)
-#             single_parameter_list = data["parameter_list"]
-#             single_force_total_results = data["force_total_results"]
-#             single_calcium_total_results = data["calcium_total_results"]
-#             single_a_total_results = data["a_total_results"]
-#             single_km_total_results = data["km_total_results"]
-#             single_tau1_total_results = data["tau1_total_results"]
-#             single_computations_time = data["computations_time"]
-#             single_computations_time_avg = data["computations_time_avg"]
-#             single_repetition = data["repetition"]
-#             single_ocp_time = data["creation_ocp_time"]
-#
-#     elif desired_mode == "doublet":
-#         with open(r"../../data/truncation_doublet.pkl", "rb") as f:
-#             data = pickle.load(f)
-#             doublet_parameter_list = data["parameter_list"]
-#             doublet_force_total_results = data["force_total_results"]
-#             doublet_calcium_total_results = data["calcium_total_results"]
-#             doublet_a_total_results = data["a_total_results"]
-#             doublet_km_total_results = data["km_total_results"]
-#             doublet_tau1_total_results = data["tau1_total_results"]
-#             doublet_computations_time = data["computations_time"]
-#             doublet_computations_time_avg = data["computations_time_avg"]
-#             doublet_repetition = data["repetition"]
-#             doublet_ocp_time = data["creation_ocp_time"]
-#
-#     elif desired_mode == "triplet":
-#         with open(r"../../data/truncation_triplet.pkl", "rb") as f:
-#             data = pickle.load(f)
-#             triplet_parameter_list = data["parameter_list"]
-#             triplet_force_total_results = data["force_total_results"]
-#             triplet_calcium_total_results = data["calcium_total_results"]
-#             triplet_a_total_results = data["a_total_results"]
-#             triplet_km_total_results = data["km_total_results"]
-#             triplet_tau1_total_results = data["tau1_total_results"]
-#             triplet_computations_time = data["computations_time"]
-#             triplet_computations_time_avg = data["computations_time_avg"]
-#             triplet_repetition = data["repetition"]
-#             triplet_ocp_time = data["creation_ocp_time"]
-#     else:
-#         raise ValueError("Not available pulse mode")
-#
-#
-# # --- Getting the results --- #
-#
-# name_error_list = [single_force_total_results]
-# name_error_list_cn = [single_calcium_total_results]
-# computations_time_list = [single_computations_time]
-# parameter_list = [single_parameter_list]
-#
-#
-# list_max_error = []
-# list_min_error = []
[…]
