### 2026-07-14 379f791
updating examples to bioptim 3.4.0 ver
