+++ examples/fes_multibody/shoulder_abduction/abduction_muscle_driven.py
@@ -9,6 +9,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import biorbd
 
+import cocofest._matplotlib_compat  # Temporary fix, see cocofest/_matplotlib_compat.py
 from bioptim import (
     BoundsList,
     ConstraintList,
