+++ examples/other_fes_models/marion2013_example.py
@@ -7,6 +7,7 @@ end of the last node while minimizing the muscle force state.
 """
 
 import numpy as np
+import cocofest._matplotlib_compat  # Temporary fix, see cocofest/_matplotlib_compat.py
 from bioptim import (
     Solver,
     ObjectiveList,
