+++ cocofest/__init__.py
@@ -1,4 +1,5 @@
 from . import _matplotlib_compat  # Temporary fix
+from .misc.__version__ import __version__
 from .custom_objectives import CustomObjective
 from .custom_constraints import CustomConstraint
 from .models.fes_model import FesModel
