+++ pyproject.toml
@@ -31,6 +31,11 @@ conda_only = [
     "rerun-sdk==0.21.0",
     "scikit-optimize",
 ]
+dev = [
+    "pytest",
+    "pytest-cov",
+    "black",
+]
 
 [project.urls]
 Homepage = "https://github.com/pyomeca/cocofest"
