+++ examples/fes_multibody/shoulder_abduction/abduction_muscle_driven.py
@@ -84,11 +84,14 @@ def prepare_ocp(
         node=Node.ALL_SHOOTING,
     )
 
-    for i in range(int(n_shooting/24)):
+    for i in range(int(n_shooting / 24)):
         objective_functions.add(
-            ObjectiveFcn.Mayer.MINIMIZE_CONTROL, key="tau", weight=1e10, node=i, quadratic=True,
+            ObjectiveFcn.Mayer.MINIMIZE_CONTROL,
+            key="tau",
+            weight=1e10,
+            node=i,
+            quadratic=True,
         )
-
 
     # --- Set parameters (for minmax cost_function) --- #
     parameters = ParameterList(use_sx=False)
@@ -96,7 +99,7 @@ def prepare_ocp(
     parameters_init = InitialGuessList()
     parameters_objectives = ParameterObjectiveList()
 
-     # --- Set constraints --- #
+    # --- Set constraints --- #
     constraints = ConstraintList()
 
     return OptimalControlProgram(
@@ -139,14 +142,17 @@ def set_x_bounds(n_shooting: int, abduction_range: dict, ode_solver) -> tuple[Bo
 
     # q from 0.523599 rad (30°) to 1.91986 rad (110°) in 3s than back to 0.523599 rad (30°) in 3s
     q_init = np.concatenate(
-        (np.linspace(abduction_range["min"], abduction_range["max"], int(n_shooting / 2) + 1), np.linspace(abduction_range["max"], abduction_range["min"], int(n_shooting / 2) + 1)[1:]), axis=0
+        (
+            np.linspace(abduction_range["min"], abduction_range["max"], int(n_shooting / 2) + 1),
+            np.linspace(abduction_range["max"], abduction_range["min"], int(n_shooting / 2) + 1)[1:],
+        ),
+        axis=0,
     )
     x_init.add(key="q", initial_guess=np.array([q_init]), phase=0, interpolation=interpolation_type)
 
     # qdot initial guess: linearly increasing from 0.0 rad/s to 4.0 rad/s at 1.5 s, then decreasing to -4 rad/s at 4.5 s, back to 0 rad/s at 6 s.
     qdot_init = np.concatenate(
-        (np.linspace(0.0, 4.0, int(n_shooting / 4) + 1), np.linspace(4.0, -4.0,
-            int(n_shooting / 2) + 1)[1:]), axis=0
+        (np.linspace(0.0, 4.0, int(n_shooting / 4) + 1), np.linspace(4.0, -4.0, int(n_shooting / 2) + 1)[1:]), axis=0
     )
     qdot_init = np.concatenate((qdot_init, np.linspace(-4.0, 0.0, int(n_shooting / 4) + 1)[1:]), axis=0)
     x_init.add(key="qdot", initial_guess=np.array([qdot_init]), phase=0, interpolation=interpolation_type)
@@ -155,33 +161,43 @@ def set_x_bounds(n_shooting: int, abduction_range: dict, ode_solver) -> tuple[Bo
     x_bounds = BoundsList()
 
     # --- Setting q bounds --- #
-    slack = 0.349066 # 10° slack
+    slack = 0.349066  # 10° slack
     humerus_q_min = np.array([abduction_range["min"]] * (n_shooting + 1))
     humerus_q_max = np.array([abduction_range["max"]] * (n_shooting + 1))
     humerus_q_min -= slack
     humerus_q_max += slack
 
     # linear bounds around the turning point to help the solver with the change of direction
-    humerus_q_max[:n_shooting // 2 - 100] = np.linspace(abduction_range["min"] + slack, abduction_range["max"] + slack/4, n_shooting // 2 - 100)
[…]
