### 2026-07-20 b76b2c0
Addressing issue #33 + removing obsolete gitmodules
