+++ docs/article/paper.bib
@@ -18,7 +18,7 @@
   pages = {111311},
   year = {2025},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/j.compbiomed.2025.111311},
+  doi = {10.1016/j.compbiomed.2025.111311},
 }
 
 @article{hunt1997feedback,
