+++ environment.yml
@@ -3,6 +3,7 @@ name: cocofest
 channels:
 - conda-forge
 dependencies:
+- python=3.11
 - bioptim=3.4.0
 - bioviz
 - python-graphviz
