+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -43,7 +43,7 @@ from cocofest import (
     FesMskModel,
     inverse_kinematics_cycling,
     OcpFesMsk,
-    FesNmpcMsk,
+    FesMheMsk,
 )
 from examples.fes_multibody.cycling.cost_functions import CustomCostFunctions
 
@@ -51,7 +51,7 @@ from examples.fes_multibody.cycling.cost_functions import CustomCostFunctions
 # --------------------#
 #    MHE functions    #
 # --------------------#
-class MyCyclicNMPC(FesNmpcMsk):
+class MyCyclicMHE(FesMheMsk):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.nodes_per_cycle = self.cycle_len * (
@@ -296,7 +296,7 @@ class MyCyclicNMPC(FesNmpcMsk):
 # --------------------#
 #    OCP functions    #
 # --------------------#
-def prepare_nmpc(
+def prepare_mhe(
     model: BiorbdModel | FesMskModel,
     mhe_info: dict,
     cycling_info: dict,
@@ -399,7 +399,7 @@ def prepare_nmpc(
     # --- Update model for resistive torque --- #
     model = updating_model(model=model, external_force_set=external_force_set, parameters=parameters)
 
-    return MyCyclicNMPC(
+    return MyCyclicMHE(
         bio_model=[model],
         dynamics=dynamics_options,
         cycle_len=cycle_len,
@@ -854,7 +854,7 @@ def create_simulation_list(
     return sims
 
 
-def save_sol_in_pkl(sol, simulation_conditions, nmpc, is_initial_guess=False, torque=None):
+def save_sol_in_pkl(sol, simulation_conditions, mhe, is_initial_guess=False, torque=None):
     solution = sol[0] if not is_initial_guess else sol[1][0]
     time = solution.stepwise_time(to_merge=[SolutionMerge.NODES]).T[0]
     states = solution.stepwise_states(to_merge=[SolutionMerge.NODES])
@@ -897,7 +897,7 @@ def save_sol_in_pkl(sol, simulation_conditions, nmpc, is_initial_guess=False, to
 
     recalculate_objective = False
     if recalculate_objective:
-        recalculate_objective_dict = recalculate_objective_fun(sol[1], nmpc, sim_cond=simulation_conditions)
+        recalculate_objective_dict = recalculate_objective_fun(sol[1], mhe, sim_cond=simulation_conditions)
         similar_cost_values = [
             True if objective_values_per_kept_cycle == recalculate_objective_dict[key] else False
             for key in recalculate_objective_dict
@@ -925,7 +925,7 @@ def save_sol_in_pkl(sol, simulation_conditions, nmpc, is_initial_guess=False, to
     print(simulation_conditions["pickle_file_path"])
 
 
-def recalculate_objective_fun(cycle_solutions: list[Solution], nmpc, sim_cond) -> dict:
+def recalculate_objective_fun(cycle_solutions: list[Solution], mhe, sim_cond) -> dict:
     import time
 
     recalculated_cost_functions = {}
@@ -939,12 +939,12 @@ def recalculate_objective_fun(cycle_solutions: list[Solution], nmpc, sim_cond) -
             "cost_fun_weight": sim_cond["cost_fun_weight"],
         }
         objective = set_objective_functions(obj_fun_dict, recalculate=True)
-        nmpc.common_objective_functions = objective
+        mhe.common_objective_functions = objective
         cost_function_values = []
         for i in range(len(cycle_solutions)):
-            _states, _controls, _parameters = nmpc.export_cycles(cycle_solutions[i])
+            _states, _controls, _parameters = mhe.export_cycles(cycle_solutions[i])
[…]
