### 2026-07-20 e30758b
mhe reference for reduced complexity
