+++ docs/article/paper.bib
@@ -68,6 +68,18 @@
   doi={10.1016/j.cma.2023.116162},
 }
 
+@article{qu2009computation,
+  title={Computation of arrival cost for moving horizon estimation via unscented Kalman filtering},
+  author={Qu, Cheryl C and Hahn, Juergen},
+  journal={Journal of Process Control},
+  volume={19},
+  number={2},
+  pages={358--363},
+  year={2009},
+  publisher={Elsevier}
+  doi={10.1016/j.jprocont.2008.04.005},
+}
+
 @article{wakeling2023review,
   title={A review of the efforts to develop muscle and musculoskeletal models for biomechanics in the last 50 years},
   author={Wakeling, James M and Febrer-Nafr{\'\i}a, M{\'\i}riam and De Groote, Friedl},
