### 2026-07-20 c9ab80f
Enabling examples to be runned outside of the repository using Path
