+++ examples/getting_started/multibody/pulse_intensity_optimization_multibody.py
@@ -5,6 +5,8 @@ The stimulation frequency will be optimized between 10 and 100 Hz and pulse inte
 threshold and 130mA to satisfy the flexion and minimizing required elbow torque control.
 """
 
+from pathlib import Path
+
 import numpy as np
 
 import cocofest._matplotlib_compat  # Temporary fix, see cocofest/_matplotlib_compat.py
@@ -17,6 +19,10 @@ from bioptim import (
     ControlType,
 )
 from cocofest import DingModelPulseIntensityFrequencyWithFatigue, OcpFesMsk, FesMskModel, CustomObjective
+
+DEFAULT_BIORBD_PATH = str(
+    Path(__file__).resolve().parent.parent.parent / "msk_models" / "Arm26" / "arm26_biceps_1dof.bioMod"
+)
 
 
 def prepare_ocp(model: FesMskModel, final_time: float, external_force: dict, msk_info: dict, pi_max: int = 130):
@@ -94,7 +100,7 @@ def prepare_ocp(model: FesMskModel, final_time: float, external_force: dict, msk
     )
 
 
-def main(plot=True, biorbd_path="../../msk_models/Arm26/arm26_biceps_1dof.bioMod"):
+def main(plot=True, biorbd_path=DEFAULT_BIORBD_PATH):
     simulation_ending_time = 1
     model = FesMskModel(
         name=None,
