+++ examples/msk_models/model_viewer.py
@@ -3,8 +3,10 @@ This example is a simple example of how to use the LiveModelAnimation class to a
 The user can interact with the model by changing the joint angles using sliders.
 """
 
+from pathlib import Path
+
 from pyorerun import LiveModelAnimation
 
-model_path = "arm26.bioMod"
+model_path = str(Path(__file__).resolve().parent / "Arm26" / "arm26.bioMod")
 animation = LiveModelAnimation(model_path, with_q_charts=True)
 animation.rerun()
