+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -1080,7 +1080,12 @@ def main(
     get_initial_guess = init_guess
 
     # --- Model choice --- #
-    model_path = "../../msk_models/Wu/Modified_Wu_Shoulder_Model_Cycling.bioMod"
+    model_path = str(
+        Path(__file__).resolve().parent.parent.parent
+        / "msk_models"
+        / "Wu"
+        / "Modified_Wu_Shoulder_Model_Cycling.bioMod"
+    )
 
     # --- MHE parameters --- #
     ode_solver = OdeSolver.COLLOCATION(polynomial_degree=3, method="radau")
