### 2026-07-20 249a711
Readme enhancement
