+++ docs/article/paper.bib
@@ -76,7 +76,7 @@
   number={2},
   pages={358--363},
   year={2009},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.jprocont.2008.04.005},
 }
 
