+++ README.md
@@ -30,8 +30,8 @@ Supports predictive musculoskeletal simulation driven by FES, moving time horizo
   - [Available FES models](#available-fes-models)
   - [Musculoskeletal model driven by FES](#musculoskeletal-model-driven-by-fes)
   - [Moving time horizons](#moving-time-horizons)
-  - [Identification](#identification)
   - [Initial value problem](#initial-value-problem)
+  - [Identification](#identification)
   - [Summation truncation](#summation-truncation)
 - [Other](#other)
   - [Want to contribute?](#want-to-contribute)
@@ -42,11 +42,6 @@ Supports predictive musculoskeletal simulation driven by FES, moving time horizo
   - [Acknowledgements](#acknowledgements)
 
 </details>
-
-
-<p align="center"> 
-  <img src="" alt="">
-</p>
 
 <a id="about"></a>
 <h1 align="center">
@@ -69,7 +64,7 @@ and robust solver like [Ipopt](https://github.com/coin-or/Ipopt).
 
 > \[!IMPORTANT]
 >
-> `Cocofest` as no clinical clearance and should not be used for rehabilitation purposes. </br>
+> `Cocofest` has no clinical clearance and should not be used for rehabilitation purposes. </br>
 > Don't forget to <a href="https://github.com/pyomeca/cocofest/stargazers"><img src="https://media2.dev.to/dynamic/image/width=1000,height=420,fit=cover,gravity=auto,format=auto/https%3A%2F%2Fthepracticaldev.s3.amazonaws.com%2Fi%2F2nn6mhp57inp6rdxarzt.png" align="center" width="80" alt="Star the repository"> </a>
 > the repository to show your support and help us grow the community!
 
@@ -142,6 +137,7 @@ script under `getting_started` can be run with:
 ```bash
 python examples/getting_started/optimization/pulse_width_optimization.py
 ```
+See [`examples/README.md`](examples/README.md) for an index of every example folder and what each script demonstrates.
 
 <p align="center"> 
   <img src="https://i.imgur.com/zXE9tC6.png" alt="">
@@ -195,7 +191,7 @@ model = ModelMaker.create_model("ding2007_with_fatigue", stim_time=[0, 0.1, 0.2]
 > \[!NOTE]
 >
 > It is possible to implement more FES models into Cocofest.
-> Adventurous enough to code it by yourself, we are looking forward to read your [pull request](how-to-contribute).
+> Adventurous enough to code it by yourself, we are looking forward to read your [pull request](docs/contributing.md).
 > Feel free to reach out on discord or submit an issue if you need help.
 
 <a id="musculoskeletal-model-driven-by-fes"></a>
@@ -250,7 +246,39 @@ You can find more examples of musculoskeletal model driven by FES in the followi
 <a id="moving-time-horizons"></a>
 ## ⏳ Moving time horizons
 
-For longer time span simulation and apprehend muscle fatigue apparition, `Cocofest` implements moving time horizons (MHE).
+For longer time span simulation and apprehend muscle fatigue apparition, `Cocofest` implements moving time horizons (MHE)
+through the [`FesMhe`](cocofest/optimization/fes_mhe.py) (single muscle) and [`FesMheMsk`](cocofest/optimization/fes_mhe_multibody.py)
+(musculoskeletal) classes. Each window is solved, then the horizon slides forward by re-using the previous stimulation
+history to keep fatigue state continuous across windows.
+
+```python
+from cocofest import DingModelPulseWidthFrequencyWithFatigue, OcpFes, FesMhe
+
+model = DingModelPulseWidthFrequencyWithFatigue(stim_time=[...], sum_stim_truncation=10)
[…]
