### 2026-07-20 1496a23
Add periodic stimulation history validation
