+++ examples/fes_multibody/cycling/validate_periodic_stimulation_history.py
@@ -1,0 +1,685 @@
+"""
+Validate the periodic Ding stimulation-history surrogate against the base Ding model.
+
+The base Ding 2007 fatigue model evaluates the calcium driving term from the
+truncated list of previous stimulation times. The periodic variant replaces this
+history by the additional state Cn_sum. This script compares both formulations
+with the same stimulation times and pulse-width command history.
+"""
+
+from __future__ import annotations
+
+import argparse
+from dataclasses import dataclass
+import os
+from pathlib import Path
+
+os.environ.setdefault("MPLBACKEND", "Agg")
+
+import matplotlib
+
+matplotlib.use("Agg", force=True)
+import matplotlib.pyplot as plt
+import numpy as np
+
+STATE_LABELS = ("Cn", "Cn_sum", "F", "A", "Tau1", "Km")
+COMPARABLE_STATE_LABELS = ("Cn", "F", "A", "Tau1", "Km")
+
+
+@dataclass(frozen=True)
+class Ding2007Parameters:
+    tauc: float = 0.011
+    r0_km_relationship: float = 1.04
+    a_scale: float = 4920.0
+    tau1_rest: float = 0.060601
+    tau2: float = 0.001
+    km_rest: float = 0.137
+    pd0: float = 0.000131405
+    pdt: float = 0.000194138
+    alpha_a: float = -4.0 * 10e-2
+    tau_fat: float = 127.0
+    alpha_tau1: float = 2.1 * 10e-6
+    alpha_km: float = 1.9 * 10e-6
+
+
+def build_argument_parser() -> argparse.ArgumentParser:
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument("--cycle-duration", type=float, default=1.0)
+    parser.add_argument("--n-cycles", type=int, default=2)
+    parser.add_argument("--prehistory-cycles", type=int, default=5)
+    parser.add_argument("--stimulations-per-cycle", type=int, default=30)
+    parser.add_argument("--sum-stim-truncation", type=int, default=20)
+    parser.add_argument("--substeps-per-stimulation", type=int, default=20)
+    parser.add_argument("--pulse-width", type=float, default=0.000365702)
+    parser.add_argument(
+        "--pulse-width-profile",
+        choices=("constant", "sinusoidal", "staircase"),
+        default="constant",
+    )
+    parser.add_argument(
+        "--output-dir",
+        type=Path,
+        default=Path(__file__).resolve().parent
+        / "result"
+        / "periodic_stimulation_history_validation",
+    )
+    parser.add_argument("--show", action="store_true")
+    return parser
+
+
+def pulse_width_at_time(
+    time: float,
+    cycle_duration: float,
+    base_pulse_width: float,
+    profile: str,
+) -> float:
+    phase = (time % cycle_duration) / cycle_duration
+    if profile == "sinusoidal":
+        return float(base_pulse_width * (1.0 + 0.25 * np.sin(2.0 * np.pi * phase)))
+    if profile == "staircase":
+        return float(base_pulse_width * (0.85 if phase < 0.5 else 1.15))
+    return float(base_pulse_width)
+
+
+def finite_stimulation_history(
+    stim_times: np.ndarray,
+    time: float,
+    truncation: int,
+) -> np.ndarray:
+    history = stim_times[stim_times <= time + 1e-12]
+    if history.size == 0:
+        return np.array([-1e7], dtype=float)
+    return history[-truncation:]
+
+
+def base_cn_sum(
+    parameters: Ding2007Parameters,
+    stim_times: np.ndarray,
+    time: float,
+    truncation: int,
+) -> float:
+    history = finite_stimulation_history(stim_times, time, truncation)
[…]
