### 2026-07-20 04569bc
Add variable stimulation history validation
