+++ examples/fes_multibody/cycling/validate_periodic_stimulation_history.py
@@ -57,6 +57,18 @@ def build_argument_parser() -> argparse.ArgumentParser:
         default="constant",
     )
     parser.add_argument(
+        "--activation-profile",
+        choices=("constant", "sinusoidal", "staircase", "alternating", "random"),
+        default="constant",
+        help=(
+            "Effective stimulation intensity profile used as lambda_i in the base "
+            "calcium history. The current periodic surrogate uses only its mean."
+        ),
+    )
+    parser.add_argument("--activation-low", type=float, default=0.2)
+    parser.add_argument("--activation-high", type=float, default=1.8)
+    parser.add_argument("--random-seed", type=int, default=42)
+    parser.add_argument(
         "--output-dir",
         type=Path,
         default=Path(__file__).resolve().parent
@@ -92,23 +104,83 @@ def finite_stimulation_history(
     return history[-truncation:]
 
 
-def base_cn_sum(
-    parameters: Ding2007Parameters,
+def finite_stimulation_history_indices(
     stim_times: np.ndarray,
     time: float,
     truncation: int,
+) -> np.ndarray:
+    indices = np.where(stim_times <= time + 1e-12)[0]
+    if indices.size == 0:
+        return np.array([], dtype=int)
+    return indices[-truncation:]
+
+
+def build_activation_values(
+    stim_times: np.ndarray,
+    cycle_duration: float,
+    profile: str,
+    low: float,
+    high: float,
+    random_seed: int,
+) -> np.ndarray:
+    if low < 0.0 or high < 0.0:
+        raise ValueError("--activation-low and --activation-high must be non-negative.")
+    if high < low:
+        raise ValueError("--activation-high must be >= --activation-low.")
+
+    if profile == "constant":
+        return np.ones(stim_times.shape)
+
+    phase = (stim_times % cycle_duration) / cycle_duration
+    mid = 0.5 * (low + high)
+    amplitude = 0.5 * (high - low)
+    if profile == "sinusoidal":
+        return mid + amplitude * np.sin(2.0 * np.pi * phase)
+    if profile == "staircase":
+        return np.where(phase < 0.5, low, high)
+    if profile == "alternating":
+        return np.where(np.arange(stim_times.size) % 2 == 0, low, high)
+
+    rng = np.random.default_rng(random_seed)
+    return rng.uniform(low, high, size=stim_times.shape)
+
+
+def activation_at_time(
+    stim_times: np.ndarray,
+    activation_values: np.ndarray,
+    time: float,
 ) -> float:
-    history = finite_stimulation_history(stim_times, time, truncation)
+    previous_indices = np.where(stim_times <= time + 1e-12)[0]
+    if previous_indices.size == 0:
+        return float(activation_values[0])
+    return float(activation_values[previous_indices[-1]])
+
+
+def base_cn_sum(
+    parameters: Ding2007Parameters,
+    stim_times: np.ndarray,
+    activation_values: np.ndarray,
+    time: float,
+    truncation: int,
+) -> float:
+    history_indices = finite_stimulation_history_indices(stim_times, time, truncation)
+    if history_indices.size == 0:
+        return 0.0
+
+    history = stim_times[history_indices]
+    history_activation = activation_values[history_indices]
     r0 = parameters.km_rest + parameters.r0_km_relationship
     value = 0.0
-    for idx, stim_time in enumerate(history):
+    for idx, (stim_time, activation) in enumerate(
+        zip(history, history_activation, strict=True)
+    ):
         if idx == 0:
             ri = 1.0
         else:
             ri = 1.0 + (r0 - 1.0) * np.exp(
                 -(stim_time - history[idx - 1]) / parameters.tauc
             )
[…]
