### 2026-07-20 429d584
Add IPOPT-driven pulse-aware stimulation validation
