+++ examples/fes_multibody/cycling/validate_periodic_stimulation_history.py
@@ -1,10 +1,12 @@
 """
-Validate the periodic Ding stimulation-history surrogate against the base Ding model.
+Validate periodic Ding stimulation-history surrogates against the base Ding model.
 
 The base Ding 2007 fatigue model evaluates the calcium driving term from the
 truncated list of previous stimulation times. The periodic variant replaces this
-history by the additional state Cn_sum. This script compares both formulations
-with the same stimulation times and pulse-width command history.
+history by the additional state Cn_sum. This script compares a mean-intensity
+periodic approximation and a pulse-aware periodic state update with the same
+stimulation times, pulse-width command history, and optional IPOPT-derived
+effective stimulation intensity history.
 """
 
 from __future__ import annotations
@@ -58,16 +60,44 @@ def build_argument_parser() -> argparse.ArgumentParser:
     )
     parser.add_argument(
         "--activation-profile",
-        choices=("constant", "sinusoidal", "staircase", "alternating", "random"),
+        choices=(
+            "constant",
+            "sinusoidal",
+            "staircase",
+            "alternating",
+            "random",
+            "ipopt",
+        ),
         default="constant",
         help=(
             "Effective stimulation intensity profile used as lambda_i in the base "
-            "calcium history. The current periodic surrogate uses only its mean."
+            "calcium history. Use 'ipopt' to derive lambda_i from a saved IPOPT "
+            "solution."
         ),
     )
     parser.add_argument("--activation-low", type=float, default=0.2)
     parser.add_argument("--activation-high", type=float, default=1.8)
+    parser.add_argument("--activation-clip-low", type=float, default=None)
+    parser.add_argument("--activation-clip-high", type=float, default=None)
     parser.add_argument("--random-seed", type=int, default=42)
+    parser.add_argument(
+        "--ipopt-data-file",
+        type=Path,
+        default=None,
+        help=(
+            "Saved IPOPT .npz solution used when --activation-profile ipopt. "
+            "Defaults to result/initial_guess/2_initial_guess_collocation_3_radau.npz "
+            "when it exists."
+        ),
+    )
+    parser.add_argument("--ipopt-muscle", default="Triceps")
+    parser.add_argument("--ipopt-control-key", default=None)
+    parser.add_argument(
+        "--ipopt-activation-source",
+        choices=("pulse_width", "fatigue", "pulse_width_and_fatigue"),
+        default="pulse_width_and_fatigue",
+    )
+    parser.add_argument("--fatigue-activation-gain", type=float, default=8.0)
     parser.add_argument(
         "--output-dir",
         type=Path,
@@ -118,18 +148,31 @@ def finite_stimulation_history_indices(
 def build_activation_values(
     stim_times: np.ndarray,
     cycle_duration: float,
-    profile: str,
-    low: float,
-    high: float,
-    random_seed: int,
+    args: argparse.Namespace,
+    parameters: Ding2007Parameters,
 ) -> np.ndarray:
+    profile = args.activation_profile
+    low = args.activation_low
+    high = args.activation_high
     if low < 0.0 or high < 0.0:
         raise ValueError("--activation-low and --activation-high must be non-negative.")
     if high < low:
         raise ValueError("--activation-high must be >= --activation-low.")
+    if (
+        args.activation_clip_low is not None
+        and args.activation_clip_high is not None
+        and args.activation_clip_high < args.activation_clip_low
+    ):
[…]
