### 2026-07-20 b7ed9e1
Add IPOPT-driven periodic stimulation intensity
