+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -139,6 +139,181 @@ def build_cost_fun_weight(objectives: set[str]) -> list[int]:
     return weights
 
 
+def effective_periodic_cn_sum_stimulation_mode(args: argparse.Namespace) -> str:
+    if args.stimulation_intensity_source != "none":
+        return "stimulation_intensity"
+    return args.periodic_cn_sum_stimulation_mode
+
+
+def default_stimulation_intensity_ipopt_file(args: argparse.Namespace) -> Path:
+    return (
+        Path(__file__).resolve().parent
+        / "result"
+        / "initial_guess"
+        / f"{args.cycles_per_window}_initial_guess_collocation_3_radau.npz"
+    )
+
+
+def pulse_width_recruitment(muscle_model, pulse_width_values: np.ndarray) -> np.ndarray:
+    pulse_width_values = np.asarray(pulse_width_values, dtype=float)
+    effective_pulse_width = np.maximum(0.0, pulse_width_values - muscle_model.pd0)
+    return 1.0 - np.exp(-effective_pulse_width / muscle_model.pdt)
+
+
+def pulse_width_activation_from_ipopt(
+    muscle_model,
+    pulse_width_values: np.ndarray,
+    reference_pulse_width: float,
+) -> np.ndarray:
+    denominator = pulse_width_recruitment(
+        muscle_model, np.array([reference_pulse_width])
+    )[0]
+    if denominator <= 0:
+        raise ValueError(
+            "--stimulation-intensity-reference-pulse-width must be greater than pd0."
+        )
+    return pulse_width_recruitment(muscle_model, pulse_width_values) / denominator
+
+
+def ipopt_control_times(
+    data, value_count: int, cycle_duration: float, args: argparse.Namespace
+) -> np.ndarray:
+    if "stim_time" in data.files:
+        stim_time = np.asarray(data["stim_time"], dtype=float).squeeze()
+        if stim_time.ndim == 1 and stim_time.size == value_count:
+            return stim_time
+
+    if "n_shooting_per_cycle" in data.files:
+        stimulations_per_cycle = int(np.asarray(data["n_shooting_per_cycle"]).item())
+    else:
+        stimulations_per_cycle = args.stimulations_per_cycle
+    return np.arange(value_count, dtype=float) * cycle_duration / stimulations_per_cycle
+
+
+def sample_periodic_values_by_index(
+    values: np.ndarray,
+    n_nodes: int,
+    node_convention: str,
+) -> np.ndarray:
+    if values.size == 0:
+        raise ValueError("Cannot sample an empty IPOPT stimulation intensity series.")
+
+    offset = 1 if node_convention == "next" else 0
+    indices = (np.arange(n_nodes) + offset) % values.size
+    return values[indices]
+
+
+def fatigue_activation_from_ipopt(
+    data,
+    muscle_name: str,
+    source_times: np.ndarray,
+    fatigue_gain: float,
+) -> np.ndarray:
+    key = f"A_{muscle_name}"
+    if key not in data.files or "time" not in data.files:
+        return np.ones(source_times.shape)
+
+    time = np.asarray(data["time"], dtype=float).squeeze()
+    a_values = np.asarray(data[key], dtype=float).squeeze()
+    if time.ndim != 1 or a_values.ndim != 1 or time.size != a_values.size:
+        return np.ones(source_times.shape)
+
+    sampled_a = np.interp(source_times, time, a_values)
+    rest_a = max(float(sampled_a[0]), 1e-12)
+    fatigue_index = np.maximum(0.0, (rest_a - sampled_a) / rest_a)
+    return 1.0 + fatigue_gain * fatigue_index
+
+
+def build_ipopt_stimulation_intensity_timeseries(
[…]

[…]
