### 2026-07-20 831d77f
Fix ACADOS stage-varying stimulation inputs
