+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -39,23 +39,38 @@ ACADOS_STATUS_NAMES = {
 }
 
 
-def patch_bioptim_acados_control_bounds() -> None:
+def patch_bioptim_acados_interface() -> None:
     """
-    Patch the Bioptim 3.4 ACADOS interface control bounds for this example.
+    Patch the Bioptim 3.4 ACADOS interface for this example.
 
     The installed Bioptim interface currently exports control bounds as
     ``lbu=max`` and ``ubu=min`` during code generation, then later updates the
     solver with unscaled control bounds. The ACADOS variables are scaled, so both
-    steps make the pulse-width QP infeasible. This local patch keeps the example
-    self-contained until the upstream interface is fixed.
+    steps make the pulse-width QP infeasible. The interface also passes
+    ``numerical_data_timeseries`` to the exported dynamics without declaring it
+    as an ACADOS model parameter, which leaves stage-varying inputs as free
+    CasADi symbols. This local patch keeps the example self-contained until the
+    upstream interface is fixed.
     """
     from bioptim.interfaces.acados_interface import AcadosInterface
 
-    if getattr(AcadosInterface, "_cocofest_control_bounds_patch", False):
+    if getattr(AcadosInterface, "_cocofest_interface_patch", False):
         return
 
+    from bioptim.interfaces.interface_utils import get_numerical_timeseries
+
+    original_export_model = AcadosInterface._AcadosInterface__acados_export_model
     original_set_constraints = AcadosInterface._AcadosInterface__set_constraints
     original_update_solver = AcadosInterface._AcadosInterface__update_solver
+
+    def patched_export_model(self, ocp):
+        original_export_model(self, ocp)
+        numerical_timeseries = ocp.nlp[0].numerical_timeseries.cx_start
+        if numerical_timeseries.shape[0] == 0:
+            return
+
+        self.acados_model.p = numerical_timeseries
+        self.acados_ocp.parameter_values = np.zeros(numerical_timeseries.shape[0])
 
     def scaled_control_bounds(interface) -> tuple[np.ndarray, np.ndarray]:
         lower = np.empty(interface.acados_ocp.dims.nu)
@@ -88,6 +103,15 @@ def patch_bioptim_acados_control_bounds() -> None:
         if self.ocp_solver is None:
             return
 
+        if self.ocp.nlp[0].numerical_timeseries.shape:
+            terminal_node = self.acados_ocp.solver_options.N_horizon
+            for stage in range(terminal_node + 1):
+                stage_values = np.asarray(
+                    get_numerical_timeseries(self.ocp, 0, stage, slice(None)),
+                    dtype=float,
+                ).reshape(-1)
+                self.ocp_solver.set(stage, "p", stage_values)
+
         param_init = []
         for key in self.ocp.nlp[0].parameters.keys():
             scaled_init = self.ocp.parameter_init[key].scale(
@@ -115,9 +139,10 @@ def patch_bioptim_acados_control_bounds() -> None:
             self.ocp_solver.constraints_set(stage, "lbu", lower)
             self.ocp_solver.constraints_set(stage, "ubu", upper)
 
+    AcadosInterface._AcadosInterface__acados_export_model = patched_export_model
     AcadosInterface._AcadosInterface__set_constraints = patched_set_constraints
     AcadosInterface._AcadosInterface__update_solver = patched_update_solver
-    AcadosInterface._cocofest_control_bounds_patch = True
+    AcadosInterface._cocofest_interface_patch = True
 
 
 def parse_objectives(raw_objective: str) -> set[str]:
@@ -1955,7 +1980,10 @@ def _clip_state_trajectory_to_bounds(
 
 
 def _periodic_ding_rhs(
[…]
