### 2026-07-20 d8d4326
Restore multibody FES collocation defects
