### 2026-07-20 4905651
Restore multibody FES collocation defects
