### 2026-07-20 44eee99
Improve periodic ACADOS warmstart diagnostics
