+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -23,7 +23,10 @@ from bioptim.optimization.receding_horizon_optimization import (
     RecedingHorizonOptimization,
 )
 
-from cycling_pulse_width_mhe import prepare_nmpc, set_fes_model
+try:
+    from .cycling_pulse_width_mhe import prepare_nmpc, set_fes_model
+except ImportError:
+    from cycling_pulse_width_mhe import prepare_nmpc, set_fes_model
 
 OBJECTIVE_TO_WEIGHT_INDEX = {"force": 0, "fatigue": 1, "control": 2}
 ACADOS_STATUS_NAMES = {
@@ -200,6 +203,77 @@ def pulse_width_activation_from_ipopt(
     return pulse_width_recruitment(muscle_model, pulse_width_values) / denominator
 
 
+def cyclic_moving_average(values: np.ndarray, window: int) -> np.ndarray:
+    values = np.asarray(values, dtype=float)
+    if window < 1:
+        raise ValueError("--stimulation-intensity-smoothing-window must be >= 1.")
+    if window % 2 == 0:
+        raise ValueError("--stimulation-intensity-smoothing-window must be odd.")
+    if window == 1:
+        return values.copy()
+
+    offsets = np.arange(window) - window // 2
+    return np.mean([np.roll(values, offset) for offset in offsets], axis=0)
+
+
+def fatigue_activation_from_values(
+    a_values: np.ndarray,
+    rest_a: float,
+    fatigue_gain: float,
+) -> np.ndarray:
+    a_values = np.asarray(a_values, dtype=float)
+    rest_a = max(float(rest_a), 1e-12)
+    fatigue_index = np.maximum(0.0, (rest_a - a_values) / rest_a)
+    return 1.0 + fatigue_gain * fatigue_index
+
+
+def stimulation_intensity_from_trajectories(
+    args: argparse.Namespace,
+    muscle_model,
+    pulse_width_values: np.ndarray,
+    a_values: np.ndarray | None,
+) -> np.ndarray:
+    pulse_width_values = np.asarray(pulse_width_values, dtype=float).reshape(-1)
+    pulse_factor = np.ones(pulse_width_values.shape)
+    fatigue_factor = np.ones(pulse_width_values.shape)
+
+    if args.stimulation_intensity_ipopt_source in (
+        "pulse_width",
+        "pulse_width_and_fatigue",
+    ):
+        pulse_factor = pulse_width_activation_from_ipopt(
+            muscle_model,
+            pulse_width_values,
+            args.stimulation_intensity_reference_pulse_width,
+        )
+        pulse_factor = cyclic_moving_average(
+            pulse_factor, args.stimulation_intensity_smoothing_window
+        )
+        pulse_factor = 1.0 + args.stimulation_intensity_homotopy_alpha * (
+            pulse_factor - 1.0
+        )
+
+    if (
+        args.stimulation_intensity_ipopt_source
+        in ("fatigue", "pulse_width_and_fatigue")
+        and a_values is not None
+    ):
+        fatigue_reference = (
+            float(a_values[0]) if np.asarray(a_values).size else muscle_model.a_scale
+        )
+        fatigue_factor = fatigue_activation_from_values(
+            a_values,
+            rest_a=fatigue_reference,
+            fatigue_gain=args.stimulation_intensity_fatigue_gain,
+        )
+
+    return np.clip(
+        pulse_factor * fatigue_factor,
+        args.stimulation_intensity_clip_low,
+        args.stimulation_intensity_clip_high,
+    )
+
+
 def ipopt_control_times(
     data, value_count: int, cycle_duration: float, args: argparse.Namespace
 ) -> np.ndarray:
[…]
