### 2026-07-20 877eb38
Test full multibody collocation defects
