### 2026-07-20 f982606
Test full multibody collocation defects
