### 2026-07-20 3c4c770
Keep periodic Ding stimulation intensity fixed
