+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -167,326 +167,6 @@ def build_cost_fun_weight(objectives: set[str]) -> list[int]:
     return weights
 
 
-def effective_periodic_cn_sum_stimulation_mode(args: argparse.Namespace) -> str:
-    if args.stimulation_intensity_source != "none":
-        return "stimulation_intensity"
-    return args.periodic_cn_sum_stimulation_mode
-
-
-def default_stimulation_intensity_ipopt_file(args: argparse.Namespace) -> Path:
-    return (
-        Path(__file__).resolve().parent
-        / "result"
-        / "initial_guess"
-        / f"{args.cycles_per_window}_initial_guess_collocation_3_radau.npz"
-    )
-
-
-def pulse_width_recruitment(muscle_model, pulse_width_values: np.ndarray) -> np.ndarray:
-    pulse_width_values = np.asarray(pulse_width_values, dtype=float)
-    effective_pulse_width = np.maximum(0.0, pulse_width_values - muscle_model.pd0)
-    return 1.0 - np.exp(-effective_pulse_width / muscle_model.pdt)
-
-
-def pulse_width_activation_from_ipopt(
-    muscle_model,
-    pulse_width_values: np.ndarray,
-    reference_pulse_width: float,
-) -> np.ndarray:
-    denominator = pulse_width_recruitment(
-        muscle_model, np.array([reference_pulse_width])
-    )[0]
-    if denominator <= 0:
-        raise ValueError(
-            "--stimulation-intensity-reference-pulse-width must be greater than pd0."
-        )
-    return pulse_width_recruitment(muscle_model, pulse_width_values) / denominator
-
-
-def cyclic_moving_average(values: np.ndarray, window: int) -> np.ndarray:
-    values = np.asarray(values, dtype=float)
-    if window < 1:
-        raise ValueError("--stimulation-intensity-smoothing-window must be >= 1.")
-    if window % 2 == 0:
-        raise ValueError("--stimulation-intensity-smoothing-window must be odd.")
-    if window == 1:
-        return values.copy()
-
-    offsets = np.arange(window) - window // 2
-    return np.mean([np.roll(values, offset) for offset in offsets], axis=0)
-
-
-def fatigue_activation_from_values(
-    a_values: np.ndarray,
-    rest_a: float,
-    fatigue_gain: float,
-) -> np.ndarray:
-    a_values = np.asarray(a_values, dtype=float)
-    rest_a = max(float(rest_a), 1e-12)
-    fatigue_index = np.maximum(0.0, (rest_a - a_values) / rest_a)
-    return 1.0 + fatigue_gain * fatigue_index
-
-
-def stimulation_intensity_from_trajectories(
-    args: argparse.Namespace,
-    muscle_model,
-    pulse_width_values: np.ndarray,
-    a_values: np.ndarray | None,
-) -> np.ndarray:
-    pulse_width_values = np.asarray(pulse_width_values, dtype=float).reshape(-1)
-    pulse_factor = np.ones(pulse_width_values.shape)
-    fatigue_factor = np.ones(pulse_width_values.shape)
-
-    if args.stimulation_intensity_ipopt_source in (
-        "pulse_width",
-        "pulse_width_and_fatigue",
-    ):
-        pulse_factor = pulse_width_activation_from_ipopt(
-            muscle_model,
-            pulse_width_values,
-            args.stimulation_intensity_reference_pulse_width,
-        )
-        pulse_factor = cyclic_moving_average(
-            pulse_factor, args.stimulation_intensity_smoothing_window
-        )
-        pulse_factor = 1.0 + args.stimulation_intensity_homotopy_alpha * (
-            pulse_factor - 1.0
-        )
-
-    if (
[…]
