### 2026-07-20 390c863
Fix bioptim 3.4 compat test import after rebase
