### 2026-07-20 71af9ef
Add ACADOS horizon continuation warmstart
