+++ tests/shard1/test_periodic_pulse_width.py
@@ -1,11 +1,14 @@
 import numpy as np
 from types import SimpleNamespace
+from bioptim import Solver
 
 from cocofest.models.ding2007.ding2007_with_fatigue_periodic import (
     DingModelPulseWidthFrequencyWithFatiguePeriodic,
 )
 from examples.fes_multibody.cycling.cycling_pulse_width_mhe_acados_periodic import (
     pulse_width_initial_guess_summary,
+    set_acados_unsafe_option,
+    tile_one_cycle_solution_to_periodic_nmpc,
 )
 
 
@@ -80,3 +83,86 @@ def test_pulse_width_summary_preserves_ipopt_control_variation():
             "span": 0.00045,
         }
     ]
+
+
+def test_one_cycle_solution_is_tiled_with_wheel_and_fatigue_drift():
+    source = SimpleNamespace(
+        decision_states=lambda to_merge=None: {
+            "q": np.array(
+                [
+                    [1.0, 1.1, 1.1, 1.0],
+                    [2.0, 2.1, 2.1, 2.0],
+                    [0.0, -2.0, -4.0, -2 * np.pi],
+                ]
+            ),
+            "qdot": np.array(
+                [
+                    [0.0, 0.1, -0.1, 0.0],
+                    [0.0, 0.2, -0.2, 0.0],
+                    [-2 * np.pi, -2 * np.pi, -2 * np.pi, -2 * np.pi],
+                ]
+            ),
+            "F_Biceps": np.array([[0.0, 2.0, 1.0, 0.0]]),
+            "A_Biceps": np.array([[10.0, 9.0, 8.0, 7.0]]),
+        },
+        decision_controls=lambda to_merge=None: {
+            "last_pulse_width_Biceps": np.array([[0.0002, 0.0004, 0.0003]])
+        },
+    )
+
+    def guess(shape):
+        return SimpleNamespace(init=np.zeros(shape))
+
+    def bounds(rows):
+        return SimpleNamespace(
+            min=np.full((rows, 3), -100.0),
+            max=np.full((rows, 3), 100.0),
+        )
+
+    nlp = SimpleNamespace(
+        x_init={
+            "q": guess((3, 7)),
+            "qdot": guess((3, 7)),
+            "F_Biceps": guess((1, 7)),
+            "A_Biceps": guess((1, 7)),
+        },
+        u_init={"last_pulse_width_Biceps": guess((1, 6))},
+        x_bounds={"q": bounds(3), "qdot": bounds(3)},
+    )
+    nmpc = SimpleNamespace(
+        nlp=[nlp],
+        _correct_init_guess_to_fit_bounds=lambda corrected_input: None,
+        _sync_acados_state_bounds=lambda: None,
+    )
+
+    summary = tile_one_cycle_solution_to_periodic_nmpc(nmpc, source)
+
+    np.testing.assert_allclose(
+        nlp.u_init["last_pulse_width_Biceps"].init,
+        [[0.0002, 0.0004, 0.0003, 0.0002, 0.0004, 0.0003]],
+    )
+    np.testing.assert_allclose(
+        nlp.x_init["q"].init[2],
+        [0.0, -2.0, -4.0, -2 * np.pi, -2.0 - 2 * np.pi, -4.0 - 2 * np.pi, -4 * np.pi],
+    )
+    np.testing.assert_allclose(
+        nlp.x_init["F_Biceps"].init,
+        [[0.0, 2.0, 1.0, 0.0, 2.0, 1.0, 0.0]],
+    )
+    np.testing.assert_allclose(
+        nlp.x_init["A_Biceps"].init,
+        [[10.0, 9.0, 8.0, 7.0, 6.0, 5.0, 4.0]],
+    )
[…]
