+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -331,6 +331,20 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--acados-horizon-continuation",
+        action="store_true",
+        help=(
+            "Solve and cache a one-cycle ACADOS problem, then tile that solution as the "
+            "initial guess for a longer horizon."
+        ),
+    )
+    parser.add_argument(
+        "--acados-continuation-source-max-iterations",
+        type=int,
+        default=500,
+        help="Maximum SQP iterations used to obtain the one-cycle ACADOS continuation source.",
+    )
+    parser.add_argument(
         "--constant-crank-torque",
         type=float,
         default=-0.2,
@@ -897,6 +911,70 @@ def _load_warmup_cache(cache_path: Path) -> "_WarmupSolutionAdapter":
     return _WarmupSolutionAdapter(states, controls)
 
 
+def _continuation_cache_signature(args: argparse.Namespace) -> str:
+    repository_root = Path(__file__).resolve().parents[3]
+    payload = {
+        "kind": "acados_one_cycle_continuation",
+        "cache_version": 1,
+        "objective": args.objective,
+        "objective_shape": args.objective_shape,
+        "stimulations_per_cycle": args.stimulations_per_cycle,
+        "constant_crank_torque": args.constant_crank_torque,
+        "torque_application": args.torque_application,
+        "state_scaling": args.state_scaling,
+        "pulse_width_scaling": args.pulse_width_scaling,
+        "control_regularization_weight": args.control_regularization_weight,
+        "control_regularization_target": args.control_regularization_target,
+        "wheel_qdot_regularization_weight": args.wheel_qdot_regularization_weight,
+        "wheel_qdot_regularization_target": args.wheel_qdot_regularization_target,
+        "integrator_type": args.acados_integrator_type,
+        "collocation_type": args.acados_collocation_type,
+        "sim_stages": args.acados_sim_stages,
+        "sim_steps": args.acados_sim_steps,
+        "newton_iter": args.acados_newton_iter,
+        "hessian_approx": args.acados_hessian_approx,
+        "nlp_solver_type": args.acados_nlp_solver_type,
+        "qp_solver": args.acados_qp_solver,
+        "max_iterations": args.acados_continuation_source_max_iterations,
+        "source_convergence_tolerance": None,
+        "source_stationarity_tolerance": 1e-3,
+        "standard_warmup_transfer": args.acados_standard_warmup_transfer,
+        "fatigue_warmstart_mode": args.acados_fatigue_warmstart_mode,
+        "periodic_projection": not args.disable_periodic_fes_warmup_projection,
+        "periodic_projection_mode": args.periodic_fes_warmup_projection_mode,
+        "periodic_projection_strategy": args.periodic_fes_warmup_projection_strategy,
+        "periodic_projection_weight": args.periodic_fes_warmup_projection_weight,
+        "periodic_ipopt_refinement": (
+            args.periodic_ipopt_refinement
+            and not args.disable_periodic_ipopt_refinement
+        ),
+        "pulse_width_trust_radius": args.acados_pulse_width_trust_radius,
+        "fes_state_trust_radius": args.acados_fes_state_trust_radius,
+        "sources": [
+            _source_stamp(
+                Path(__file__).resolve().parent / "cycling_pulse_width_mhe.py"
+            ),
+            _source_stamp(
+                repository_root
+                / "cocofest"
+                / "models"
[…]
