### 2026-07-21 9aa3bb0
Stabilize ACADOS inter-window continuation
