+++ tests/shard1/test_periodic_pulse_width.py
@@ -2,9 +2,11 @@ import numpy as np
 from types import SimpleNamespace
 from bioptim import Solver
 
+import examples.fes_multibody.cycling.cycling_pulse_width_mhe_acados_periodic as periodic_example
 from cocofest.models.ding2007.ding2007_with_fatigue_periodic import (
     DingModelPulseWidthFrequencyWithFatiguePeriodic,
 )
+from examples.fes_multibody.cycling.cycling_pulse_width_mhe import MyCyclicNMPC
 from examples.fes_multibody.cycling.cycling_pulse_width_mhe_acados_periodic import (
     pulse_width_initial_guess_summary,
     set_acados_unsafe_option,
@@ -166,3 +168,177 @@ def test_unsafe_acados_option_is_initialized_on_each_solver_instance():
 
     assert first_solver._test_repeated_option == 0.5
     assert second_solver._test_repeated_option == 0.25
+
+
+def test_acados_diagnostics_snapshot_is_independent_from_shared_solver(monkeypatch):
+    live_diagnostics = {"status": 0, "residuals": np.array([1.0, 2.0])}
+    monkeypatch.setattr(
+        periodic_example,
+        "collect_acados_diagnostics",
+        lambda solution: live_diagnostics,
+    )
+
+    snapshot = periodic_example.snapshot_acados_diagnostics(SimpleNamespace())
+    live_diagnostics["status"] = 2
+    live_diagnostics["residuals"][0] = 99.0
+
+    assert snapshot["status"] == 0
+    np.testing.assert_allclose(snapshot["residuals"], [1.0, 2.0])
+
+
+def test_acados_diagnostics_must_meet_strict_cache_tolerances():
+    strict = {"residuals": np.array([0.1, 0.009, 0.0, 0.008])}
+    relaxed_only = {"residuals": np.array([3.0, 0.009, 0.0, 0.008])}
+
+    assert periodic_example.acados_diagnostics_meet_tolerances(
+        strict, convergence_tolerance=1e-2, stationarity_tolerance=0.15
+    )
+    assert not periodic_example.acados_diagnostics_meet_tolerances(
+        relaxed_only, convergence_tolerance=1e-2, stationarity_tolerance=0.15
+    )
+
+
+def test_codegen_signature_ignores_run_only_options():
+    parser = periodic_example.build_argument_parser()
+    reference = parser.parse_args([])
+    longer_diagnostic_run = parser.parse_args(
+        ["--n-windows", "20", "--acados-diagnostics", "--codegen-tag", "diagnostic"]
+    )
+
+    assert periodic_example._codegen_signature(
+        reference
+    ) == periodic_example._codegen_signature(longer_diagnostic_run)
+    assert periodic_example._horizon_seed_cache_signature(
+        reference
+    ) == periodic_example._horizon_seed_cache_signature(longer_diagnostic_run)
+
+    longer_diagnostic_run.stimulations_per_cycle += 1
+    assert periodic_example._codegen_signature(
+        reference
+    ) != periodic_example._codegen_signature(longer_diagnostic_run)
+    assert periodic_example._horizon_seed_cache_signature(
+        reference
+    ) != periodic_example._horizon_seed_cache_signature(longer_diagnostic_run)
+
+
+def test_acados_qp_warm_start_cli_options_are_explicit():
+    parser = periodic_example.build_argument_parser()
+    default_args = parser.parse_args([])
+    warm_args = parser.parse_args(
+        [
+            "--acados-qp-warm-start-level",
+            "1",
+            "--acados-warm-start-first-qp",
[…]
