### 2026-07-21 01f68df
Guard ACADOS transfer rollout
