+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -574,6 +574,23 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Project the ACADOS qdot initial guess from finite differences of q before solving.",
     )
     parser.add_argument(
+        "--acados-transfer-full-dynamics-rollout",
+        action="store_true",
+        help="Reintegrate the appended cycle with the complete dynamics after each window transfer.",
+    )
+    parser.add_argument(
+        "--acados-transfer-rollout-substeps",
+        type=int,
+        default=5,
+        help="RK4 substeps used by the full-dynamics inter-window rollout.",
+    )
+    parser.add_argument(
+        "--acados-transfer-rollout-max-bound-violation",
+        type=float,
+        default=1.0,
+        help="Reject the inter-window rollout when a state exceeds its bounds by more than this value.",
+    )
+    parser.add_argument(
         "--warmup-state-comparison-limit",
         type=int,
         default=12,
@@ -2435,6 +2452,96 @@ def _full_dynamics_rollout_defect_details(
     }
 
 
+def rollout_transferred_cycle_full_dynamics(
+    periodic_nmpc,
+    n_substeps: int = 5,
+    max_allowed_bound_violation: float | None = None,
+) -> dict:
+    if n_substeps < 1:
+        raise ValueError("ACADOS transfer rollout substeps must be >= 1.")
+
+    nlp = periodic_nmpc.nlp[0]
+    first_state_key = next(iter(nlp.x_init.keys()))
+    first_control_key = next(iter(nlp.u_init.keys()))
+    n_state_nodes = nlp.x_init[first_state_key].init.shape[1]
+    n_control_nodes = nlp.u_init[first_control_key].init.shape[1]
+    if n_state_nodes != n_control_nodes + 1:
+        raise ValueError(
+            "The transfer rollout requires one more state node than controls."
+        )
+
+    start_node = periodic_nmpc.nodes_per_cycle - 1
+    if start_node < 0 or start_node >= n_control_nodes:
+        raise ValueError("The transfer rollout start node is outside the horizon.")
+
+    states = _stack_initial_guess_values(nlp.x_init, nlp.states, n_state_nodes)
+    original_states = states.copy()
+    controls = _stack_initial_guess_values(nlp.u_init, nlp.controls, n_control_nodes)
+    dt = periodic_nmpc.cycle_duration / periodic_nmpc.cycle_len
+    stage_numerical_timeseries = [
+        _numerical_timeseries_at_node(nlp, node) for node in range(n_control_nodes)
+    ]
+
+    for node in range(start_node, n_control_nodes):
+        states[:, node + 1] = _rk4_full_dynamics_step(
+            nlp,
+            states[:, node],
+            controls[:, node],
+            node * dt,
+            dt,
+            n_substeps=n_substeps,
+            numerical_timeseries=stage_numerical_timeseries[node],
+        )
+        if not np.all(np.isfinite(states[:, node + 1])):
+            return {
+                "applied": False,
+                "start_node": start_node,
+                "nonfinite_node": node + 1,
+            }
+
+    max_bound_violation = 0.0
+    max_bound_violation_by_key = {}
+    terminal_delta = {}
+    max_delta_by_key = {}
+    state_values = {}
+    for key in nlp.states.keys():
+        indexes = np.asarray(nlp.states[key].index).reshape((-1,)).tolist()
+        values = states[indexes, :]
+        state_values[key] = values
+        original_values = original_states[indexes, :]
+        lower, upper = _trajectory_bounds_for_guess(nlp.x_bounds[key], n_state_nodes)
+        violations = np.maximum(lower - values, 0.0) + np.maximum(values - upper, 0.0)
[…]
