### 2026-07-21 cace6b6
Add exact periodic-node Ding forcing for ACADOS
