+++ cocofest/models/ding2007/ding2007_with_fatigue_periodic_node.py
@@ -1,0 +1,222 @@
+from typing import Callable
+
+import numpy as np
+from casadi import MX, exp, vertcat
+
+from bioptim import (
+    ConfigureProblem,
+    DynamicsEvaluation,
+    NonLinearProgram,
+    OptimalControlProgram,
+)
+
+from cocofest.models.ding2007.ding2007_with_fatigue import (
+    DingModelPulseWidthFrequencyWithFatigue,
+)
+from cocofest.models.state_configure import StateConfigure
+
+
+class DingModelPulseWidthFrequencyWithFatiguePeriodicNode(
+    DingModelPulseWidthFrequencyWithFatigue
+):
+    """Fixed-frequency Ding model with an exact within-interval calcium history.
+
+    Each shooting interval starts at a stimulation. Numerical data provide the
+    steady truncated post-stimulation calcium-history amplitude and the interval
+    start time. The history then decays exactly inside the interval; pulse width
+    only affects force recruitment and the stimulation intensity remains fixed
+    to one. This formulation targets receding-horizon windows after their warmup
+    cycle; it does not model the initial buildup from an unstimulated muscle.
+    """
+
+    def __init__(
+        self,
+        model_name: str = "ding_2007_with_fatigue_periodic_node",
+        muscle_name: str = None,
+        stim_time: list[float] = None,
+        previous_stim: dict = None,
+        sum_stim_truncation: int = 20,
+        stim_interval: float | None = None,
+    ):
+        super().__init__(
+            model_name=model_name,
+            muscle_name=muscle_name,
+            stim_time=stim_time,
+            previous_stim=previous_stim,
+            sum_stim_truncation=sum_stim_truncation,
+        )
+        self._stim_interval = (
+            float(stim_interval)
+            if stim_interval is not None
+            else self._infer_stim_interval(stim_time)
+        )
+
+    @staticmethod
+    def _infer_stim_interval(stim_time: list[float] | None) -> float | None:
+        if stim_time is None or len(stim_time) < 2:
+            return None
+        intervals = np.diff(np.asarray(stim_time, dtype=float))
+        if not np.allclose(intervals, intervals[0]):
+            raise ValueError(
+                "The periodic-node Ding model requires evenly spaced stimulations."
+            )
+        return float(intervals[0])
+
+    def serialize(self) -> tuple[Callable, dict]:
+        return DingModelPulseWidthFrequencyWithFatiguePeriodicNode, {
+            "model_name": self.model_name,
+            "muscle_name": self.muscle_name,
+            "stim_time": self.stim_time,
+            "previous_stim": self.previous_stim,
+            "sum_stim_truncation": self.sum_stim_truncation,
+            "stim_interval": self._stim_interval,
+        }
+
+    def stimulation_increment(self) -> float:
+        if self._stim_interval is None:
+            raise ValueError(
+                "stim_interval is required for periodic-node calcium forcing."
+            )
+        decay = np.exp(-self._stim_interval / self.tauc)
+        return float(1.0 + (self.get_r0(self.km_rest) - 1.0) * decay)
+
+    def post_stimulation_amplitude(
+        self, retained_stimulations: int | None = None
+    ) -> float:
+        """Return the history sum immediately after a stimulation.
+
+        This reproduces the historical truncated sum: the oldest retained pulse
+        has coefficient one and subsequent pulses use Ding's ``R_i`` factor.
+        """
+
+        if self._stim_interval is None:
+            raise ValueError(
+                "stim_interval is required for periodic-node calcium forcing."
+            )
+        count = (
+            self.sum_stim_truncation
+            if retained_stimulations is None
[…]
