+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -42,6 +42,66 @@ ACADOS_STATUS_NAMES = {
 }
 
 
+def build_time_dependent_rk4_map(
+    rhs,
+    state,
+    control,
+    stage_parameters,
+    local_time,
+    interval_duration: float,
+    n_substeps: int,
+):
+    """Build a node-to-node RK4 map while retaining the local time dependence."""
+    from casadi import Function
+
+    if n_substeps < 1:
+        raise ValueError("n_substeps must be strictly positive.")
+    rhs_function = Function(
+        "cocofest_discrete_rhs",
+        [state, control, stage_parameters, local_time],
+        [rhs],
+    )
+    step = float(interval_duration) / n_substeps
+    next_state = state
+    for substep in range(n_substeps):
+        time = substep * step
+        k1 = rhs_function(next_state, control, stage_parameters, time)
+        k2 = rhs_function(
+            next_state + step * k1 / 2,
+            control,
+            stage_parameters,
+            time + step / 2,
+        )
+        k3 = rhs_function(
+            next_state + step * k2 / 2,
+            control,
+            stage_parameters,
+            time + step / 2,
+        )
+        k4 = rhs_function(
+            next_state + step * k3,
+            control,
+            stage_parameters,
+            time + step,
+        )
+        next_state = next_state + step * (k1 + 2 * k2 + 2 * k3 + k4) / 6
+    return next_state
+
+
+def _periodic_node_interval_duration(ocp) -> float:
+    muscle_models = getattr(ocp.nlp[0].model, "muscles_dynamics_model", ())
+    intervals = {
+        float(model._stim_interval)
+        for model in muscle_models
+        if getattr(model, "_stim_interval", None) is not None
+    }
+    if len(intervals) != 1:
+        raise RuntimeError(
+            "The periodic-node DISCRETE map requires one shared stimulation interval."
+        )
+    return intervals.pop()
+
+
 def patch_bioptim_acados_interface() -> None:
     """
     Patch the Bioptim 3.4 ACADOS interface for this example.
@@ -72,8 +132,52 @@ def patch_bioptim_acados_interface() -> None:
         if numerical_timeseries.shape[0] == 0:
             return
 
+        from casadi import SX, substitute
+
+        original_time = ocp.nlp[0].time_cx
+        local_time = SX.sym("cocofest_local_time", 1, 1)
+        is_periodic_node = "periodic_calcium" in str(
+            ocp.nlp[0].numerical_data_timeseries
+        )
+        interval_duration = (
+            _periodic_node_interval_duration(ocp) if is_periodic_node else None
+        )
+        if is_periodic_node and self.opts.integrator_type == "DISCRETE":
+            timed_rhs = substitute(
+                self.acados_model.f_expl_expr,
+                original_time,
+                local_time + numerical_timeseries[-1],
+            )
+            self.acados_model.disc_dyn_expr = build_time_dependent_rk4_map(
+                rhs=timed_rhs,
+                state=self.acados_model.x,
+                control=self.acados_model.u,
+                stage_parameters=numerical_timeseries,
+                local_time=local_time,
+                interval_duration=interval_duration,
+                n_substeps=getattr(ocp, "_cocofest_discrete_substeps", 5),
+            )
+        else:
+            if is_periodic_node and self.opts.integrator_type == "ERK":
+                # The bundled explicit-ODE generator has no time input. Use the
+                # interval midpoint as a documented approximation for ERK.
+                dynamics_time = numerical_timeseries[-1] + interval_duration / 2
[…]

[…]
