### 2026-07-21 2402106
Fix scaled ACADOS warm starts for periodic cycling
