+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -102,6 +102,35 @@ def _periodic_node_interval_duration(ocp) -> float:
     return intervals.pop()
 
 
+def _periodic_node_dynamics_time(
+    integrator_type: str, acados_time, interval_start, interval_duration: float
+):
+    """Return the absolute time used by the periodic-node Ding dynamics."""
+
+    if integrator_type == "ERK":
+        # The bundled explicit-ODE generator does not expose the stage time.
+        return interval_start + interval_duration / 2
+    if integrator_type == "IRK":
+        # OCP shooting intervals initialize the IRK simulator at t0=0.
+        return acados_time + interval_start
+    raise ValueError(f"Unsupported time-dependent integrator: {integrator_type}")
+
+
+def _scaled_acados_dynamics_rhs(rhs, state_scaling: np.ndarray, n_parameters: int):
+    """Convert a physical derivative to the derivative of ACADOS scaled states."""
+
+    from casadi import DM, vertcat
+
+    state_scaling = np.asarray(state_scaling, dtype=float).reshape(-1)
+    complete_scaling = np.concatenate((np.ones(n_parameters), state_scaling))
+    if rhs.shape[0] != complete_scaling.size:
+        raise ValueError(
+            "ACADOS dynamics and scaling dimensions differ "
+            f"({rhs.shape[0]} != {complete_scaling.size})."
+        )
+    return rhs / vertcat(*DM(complete_scaling).elements())
+
+
 def patch_bioptim_acados_interface() -> None:
     """
     Patch the Bioptim 3.4 ACADOS interface for this example.
@@ -128,6 +157,19 @@ def patch_bioptim_acados_interface() -> None:
 
     def patched_export_model(self, ocp):
         original_export_model(self, ocp)
+        state_scaling = np.ones(ocp.nlp[0].states.shape)
+        for key in ocp.nlp[0].states.keys():
+            state_scaling[ocp.nlp[0].states[key].index] = np.asarray(
+                ocp.nlp[0].x_scaling[key].scaling[:, 0], dtype=float
+            )
+        self.acados_model.f_expl_expr = _scaled_acados_dynamics_rhs(
+            self.acados_model.f_expl_expr,
+            state_scaling,
+            ocp.nlp[0].parameters.shape,
+        )
+        self.acados_model.f_impl_expr = (
+            self.acados_model.xdot - self.acados_model.f_expl_expr
+        )
         numerical_timeseries = ocp.nlp[0].numerical_timeseries.cx_start
         if numerical_timeseries.shape[0] == 0:
             return
@@ -158,12 +200,17 @@ def patch_bioptim_acados_interface() -> None:
                 n_substeps=getattr(ocp, "_cocofest_discrete_substeps", 5),
             )
         else:
-            if is_periodic_node and self.opts.integrator_type == "ERK":
-                # The bundled explicit-ODE generator has no time input. Use the
-                # interval midpoint as a documented approximation for ERK.
-                dynamics_time = numerical_timeseries[-1] + interval_duration / 2
+            if is_periodic_node and self.opts.integrator_type in ("ERK", "IRK"):
+                dynamics_time = _periodic_node_dynamics_time(
+                    self.opts.integrator_type,
+                    local_time,
+                    numerical_timeseries[-1],
+                    interval_duration,
+                )
+                if self.opts.integrator_type == "IRK":
+                    self.acados_model.t = local_time
             elif is_periodic_node:
-                dynamics_time = local_time + numerical_timeseries[-1]
+                dynamics_time = local_time
                 self.acados_model.t = local_time
             else:
                 dynamics_time = original_time
@@ -242,6 +289,7 @@ def patch_bioptim_acados_interface() -> None:
         )
 
[…]
