### 2026-07-21 dc09c08
Build endurance benchmark and fix NMPC transfer
