+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -478,6 +478,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--acados-transfer-pulse-width-trust-radius",
+        type=float,
+        default=None,
+        help=(
+            "Optional pulse-width trust-region radius after the first solved window. "
+            "Defaults to --acados-pulse-width-trust-radius."
+        ),
+    )
+    parser.add_argument(
         "--acados-fes-state-trust-radius",
         type=float,
         default=None,
@@ -894,6 +903,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--periodic-ipopt-refinement-ode-solver",
+        choices=("target", "rk4", "irk"),
+        default="target",
+        help=(
+            "Integrator used by the periodic IPOPT bridge. 'target' preserves "
+            "the main Bioptim transcription; 'irk' can match an ACADOS IRK map."
+        ),
+    )
+    parser.add_argument(
         "--periodic-fes-warmup-projection-weight",
         type=float,
         default=1.0,
@@ -1182,6 +1200,10 @@ def _periodic_ipopt_refinement_cache_path(
         "standard_warmup_transfer": args.acados_standard_warmup_transfer,
         "fatigue_warmstart_mode": args.acados_fatigue_warmstart_mode,
         "use_sx": args.periodic_ipopt_refinement_use_sx,
+        "ode_solver": args.periodic_ipopt_refinement_ode_solver,
+        "acados_sim_stages": args.acados_sim_stages,
+        "acados_sim_steps": args.acados_sim_steps,
+        "acados_collocation_type": args.acados_collocation_type,
         "ipopt_linear_solver": args.ipopt_linear_solver,
         "sources": [
             _source_stamp(model_path),
@@ -1719,6 +1741,22 @@ def _initial_guess_traces(container) -> dict[str, np.ndarray]:
         key: np.array(container[key].init, dtype=float, copy=True)
         for key in container.keys()
     }
+
+
+def _control_bounds_summary(nmpc) -> dict[str, dict[str, float]]:
+    summary = {}
+    original_bounds = getattr(nmpc, "_cocofest_original_control_bounds", {})
+    for key in nmpc.nlp[0].controls.keys():
+        bounds = original_bounds.get(key, nmpc.nlp[0].u_bounds[key])
+        if isinstance(bounds, tuple):
+            lower, upper = bounds
+        else:
+            lower, upper = bounds.min, bounds.max
+        summary[key] = {
+            "lower": float(np.min(np.asarray(lower, dtype=float))),
+            "upper": float(np.max(np.asarray(upper, dtype=float))),
+        }
+    return summary
 
 
 def _status_is_success(status) -> bool:
@@ -4461,6 +4499,7 @@ def build_periodic_ipopt_refinement_nmpc(
     cycling_info: dict,
     simulation_conditions: dict,
     model_formulation: str,
+    refinement_ode_solver=None,
 ):
     refinement_model = set_fes_model(
         str(model_path),
@@ -4469,6 +4508,8 @@ def build_periodic_ipopt_refinement_nmpc(
         periodic_node_forcing=model_formulation == "periodic_node",
     )
     refinement_mhe_info = dict(mhe_info)
+    if refinement_ode_solver is not None:
+        refinement_mhe_info["ode_solver"] = refinement_ode_solver
     refinement_nmpc = prepare_nmpc(
         refinement_model,
         refinement_mhe_info,
@@ -4578,6 +4619,20 @@ def apply_pulse_width_control_trust_region(
     if radius < 0:
[…]

[…]
