### 2026-07-21 d7dc71e
Preserve historical IPOPT horizon transfer
