### 2026-07-21 f81a9b4
Compare solver trajectories at shooting nodes
