+++ tests/shard1/test_periodic_pulse_width.py
@@ -296,6 +296,26 @@ def test_benchmark_compares_only_the_successful_prefix():
     assert limited["control_traces"]["last_pulse_width_Biceps"].shape == (1, 4)
 
 
+def test_benchmark_extracts_collocation_shooting_nodes_without_interpolation():
+    result = _benchmark_result([0, 0], solver_success=True, success=True)
+    result["args"] = SimpleNamespace(
+        stimulations_per_cycle=2,
+        ode_solver="collocation",
+        collocation_degree=3,
+    )
+    result["exported_cycles"] = 2
+    collocation_values = np.arange(17, dtype=float)
+    result["wheel_angle_trace"] = collocation_values
+    result["state_traces"] = {"q": collocation_values[np.newaxis, :]}
+    result["control_traces"] = {"u": np.arange(4, dtype=float)[np.newaxis, :]}
+
+    limited = comparison_example._truncate_result_to_cycles(result, 2)
+
+    np.testing.assert_array_equal(limited["wheel_angle_trace"], [0, 4, 8, 12, 16])
+    np.testing.assert_array_equal(limited["state_traces"]["q"], [[0, 4, 8, 12, 16]])
+    np.testing.assert_array_equal(limited["control_traces"]["u"], [[0, 1, 2, 3]])
+
+
 def test_endurance_metrics_report_fatigue_and_control_saturation():
     result = _benchmark_result([0, 0, 4])
 
