+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -137,19 +137,89 @@ def _validated_cycle_count(result: dict) -> int:
     return prefix
 
 
-def _truncate_result_to_cycles(result: dict, cycle_count: int) -> dict:
+def _configured_state_node_stride(result: dict) -> int:
+    args = result["args"]
+    ode_solver = str(getattr(args, "ode_solver", "")).lower()
+    if ode_solver in {"collocation", "irk"}:
+        return int(getattr(args, "collocation_degree", 3)) + 1
+    return 1
+
+
+def _exported_cycle_count(result: dict) -> int:
+    if result.get("exported_cycles") is not None:
+        return int(result["exported_cycles"])
+
     shooting_per_cycle = int(result["args"].stimulations_per_cycle)
-    state_nodes = cycle_count * shooting_per_cycle + 1
-    control_nodes = cycle_count * shooting_per_cycle
+    state_stride = _configured_state_node_stride(result)
+    state_node_count = np.asarray(result["wheel_angle_trace"]).size
+    interval_count, remainder = divmod(state_node_count - 1, state_stride)
+    cycle_count, cycle_remainder = divmod(interval_count, shooting_per_cycle)
+    if remainder or cycle_remainder:
+        raise ValueError(
+            "Cannot infer the number of exported cycles from the state trace."
+        )
+    return cycle_count
+
+
+def _shooting_node_state_trace(
+    values: np.ndarray, result: dict, cycle_count: int
+) -> np.ndarray:
+    values = np.asarray(values)
+    shooting_per_cycle = int(result["args"].stimulations_per_cycle)
+    exported_cycles = _exported_cycle_count(result)
+    exported_intervals = exported_cycles * shooting_per_cycle
+    if exported_intervals <= 0:
+        return values[..., :0]
+
+    available_intervals, remainder = divmod(values.shape[-1] - 1, exported_intervals)
+    if remainder:
+        raise ValueError(
+            "State trace cannot be mapped exactly to the shooting-node grid: "
+            f"{values.shape[-1]} values for {exported_intervals} intervals."
+        )
+    stride = available_intervals
+    expected_stride = _configured_state_node_stride(result)
+    if stride != expected_stride:
+        raise ValueError(
+            "Unexpected number of state points per shooting interval: "
+            f"observed {stride}, expected {expected_stride}."
+        )
+
+    requested_intervals = cycle_count * shooting_per_cycle
+    return values[..., : requested_intervals * stride + 1 : stride]
+
+
+def _shooting_node_control_trace(
+    values: np.ndarray, result: dict, cycle_count: int
+) -> np.ndarray:
+    values = np.asarray(values)
+    shooting_per_cycle = int(result["args"].stimulations_per_cycle)
+    exported_intervals = _exported_cycle_count(result) * shooting_per_cycle
+    if exported_intervals <= 0:
+        return values[..., :0]
+
+    stride, remainder = divmod(values.shape[-1], exported_intervals)
+    if remainder or stride <= 0:
+        raise ValueError(
+            "Control trace cannot be mapped exactly to the shooting-node grid: "
+            f"{values.shape[-1]} values for {exported_intervals} intervals."
+        )
+    requested_intervals = cycle_count * shooting_per_cycle
+    return values[..., : requested_intervals * stride : stride]
+
+
+def _truncate_result_to_cycles(result: dict, cycle_count: int) -> dict:
     return {
         **result,
-        "wheel_angle_trace": np.asarray(result["wheel_angle_trace"])[:state_nodes],
+        "wheel_angle_trace": _shooting_node_state_trace(
+            result["wheel_angle_trace"], result, cycle_count
+        ),
         "state_traces": {
[…]
