### 2026-07-21 22dfe4d
Diagnose ACADOS MHE residuals
