+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -257,6 +257,14 @@ def patch_bioptim_acados_interface() -> None:
         if self.ocp_solver is None:
             return
 
+        dual_summary = apply_acados_dual_warm_start(
+            self.ocp_solver,
+            horizon=self.acados_ocp.solver_options.N_horizon,
+            mode=getattr(self.ocp, "_cocofest_dual_warm_start_mode", "preserve"),
+            shift_stages=getattr(self.ocp, "_cocofest_dual_shift_stages", 0),
+        )
+        self.ocp._cocofest_last_dual_warm_start_summary = dual_summary
+
         if self.ocp.nlp[0].numerical_timeseries.shape:
             terminal_node = self.acados_ocp.solver_options.N_horizon
             for stage in range(terminal_node + 1):
@@ -317,6 +325,55 @@ def patch_bioptim_acados_interface() -> None:
     AcadosInterface._AcadosInterface__set_constraints = patched_set_constraints
     AcadosInterface._AcadosInterface__update_solver = patched_update_solver
     AcadosInterface._cocofest_interface_patch = True
+
+
+def apply_acados_dual_warm_start(
+    acados_solver, horizon: int, mode: str, shift_stages: int
+) -> dict[str, int | str]:
+    if mode not in {"preserve", "reset", "shift"}:
+        raise ValueError(f"Unsupported ACADOS dual warm-start mode '{mode}'.")
+    if shift_stages < 0:
+        raise ValueError("ACADOS dual shift must be non-negative.")
+    if mode == "preserve":
+        return {"mode": mode, "shift_stages": 0, "zeroed_tail_stages": 0}
+
+    lam = [
+        np.asarray(acados_solver.get(stage, "lam"), dtype=float)
+        for stage in range(horizon + 1)
+    ]
+    pi = [
+        np.asarray(acados_solver.get(stage, "pi"), dtype=float)
+        for stage in range(horizon)
+    ]
+    if mode == "reset":
+        for stage, values in enumerate(lam):
+            acados_solver.set(stage, "lam", np.zeros_like(values))
+        for stage, values in enumerate(pi):
+            acados_solver.set(stage, "pi", np.zeros_like(values))
+        return {
+            "mode": mode,
+            "shift_stages": 0,
+            "zeroed_tail_stages": horizon + 1,
+        }
+
+    shift = min(shift_stages, horizon)
+    for stage in range(horizon + 1):
+        source = stage + shift
+        values = (
+            lam[source]
+            if source <= horizon and lam[source].shape == lam[stage].shape
+            else np.zeros_like(lam[stage])
+        )
+        acados_solver.set(stage, "lam", values)
+    for stage in range(horizon):
+        source = stage + shift
+        values = pi[source] if source < horizon else np.zeros_like(pi[stage])
+        acados_solver.set(stage, "pi", values)
+    return {
+        "mode": mode,
+        "shift_stages": shift,
+        "zeroed_tail_stages": shift,
+    }
 
 
 def parse_objectives(raw_objective: str) -> set[str]:
@@ -709,6 +766,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Initialize the first QP warm start from the current NLP iterate.",
     )
     parser.add_argument(
+        "--acados-dual-warm-start-mode",
+        choices=("preserve", "reset", "shift"),
+        default="reset",
+        help=(
+            "Treatment of ACADOS inequality and dynamics multipliers between MHE windows. "
+            "Reset is the robust default; shift is experimental."
+        ),
+    )
+    parser.add_argument(
         "--acados-qpscaling-scale-objective",
[…]
