### 2026-07-21 0aa084e
Warm start IPOPT receding horizons with bound duals
