### 2026-07-21 6261647
Stabilize Acados MHE objective and transfers
