+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -533,6 +533,10 @@ def _solver_config(
     control_regularization_target_source: str,
     wheel_qdot_regularization_weight: float,
     wheel_qdot_regularization_target: float,
+    wheel_qdot_bound_margin: float,
+    terminal_qdot_regularization_weight: float,
+    terminal_qdot_regularization_target_source: str,
+    acados_terminal_wheel_q_slack: float,
     state_scaling: str,
     pulse_width_scaling: float,
     acados_pulse_width_trust_radius: float | None,
@@ -640,6 +644,12 @@ def _solver_config(
             control_regularization_target_source=control_regularization_target_source,
             wheel_qdot_regularization_weight=wheel_qdot_regularization_weight,
             wheel_qdot_regularization_target=wheel_qdot_regularization_target,
+            wheel_qdot_bound_margin=wheel_qdot_bound_margin,
+            terminal_qdot_regularization_weight=(terminal_qdot_regularization_weight),
+            terminal_qdot_regularization_target_source=(
+                terminal_qdot_regularization_target_source
+            ),
+            acados_terminal_wheel_q_slack=acados_terminal_wheel_q_slack,
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_pulse_width_trust_radius=None,
@@ -734,6 +744,12 @@ def _solver_config(
             control_regularization_target_source=control_regularization_target_source,
             wheel_qdot_regularization_weight=wheel_qdot_regularization_weight,
             wheel_qdot_regularization_target=wheel_qdot_regularization_target,
+            wheel_qdot_bound_margin=wheel_qdot_bound_margin,
+            terminal_qdot_regularization_weight=(terminal_qdot_regularization_weight),
+            terminal_qdot_regularization_target_source=(
+                terminal_qdot_regularization_target_source
+            ),
+            acados_terminal_wheel_q_slack=acados_terminal_wheel_q_slack,
             state_scaling=state_scaling,
             pulse_width_scaling=pulse_width_scaling,
             acados_pulse_width_trust_radius=acados_pulse_width_trust_radius,
@@ -1063,6 +1079,10 @@ def main(
     wheel_qdot_regularization_weight: float = 0.0,
     acados_wheel_qdot_regularization_weight: float | None = None,
     wheel_qdot_regularization_target: float = -float(2 * np.pi),
+    wheel_qdot_bound_margin: float = 3.0,
+    terminal_qdot_regularization_weight: float = 0.0,
+    terminal_qdot_regularization_target_source: str = "previous",
+    acados_terminal_wheel_q_slack: float = 0.2,
     state_scaling: str = "none",
     acados_state_scaling: str | None = None,
     pulse_width_scaling: float = 1 / 400,
@@ -1080,7 +1100,7 @@ def main(
     acados_hessian_approx: str = "GAUSS_NEWTON",
     acados_nlp_solver_type: str = "SQP",
     acados_search_direction_mode: str = "NOMINAL_QP",
-    acados_globalization: str = "MERIT_BACKTRACKING",
+    acados_globalization: str = "FUNNEL_L1PEN_LINESEARCH",
     acados_fixed_step_length: float = 1.0,
     acados_nlp_qp_tol_strategy: str = "ADAPTIVE_QPSCALING",
     acados_qpscaling_scale_objective: str = "OBJECTIVE_GERSHGORIN",
@@ -1148,6 +1168,12 @@ def main(
         control_regularization_target_source=control_regularization_target_source,
         wheel_qdot_regularization_weight=wheel_qdot_regularization_weight,
         wheel_qdot_regularization_target=wheel_qdot_regularization_target,
+        wheel_qdot_bound_margin=wheel_qdot_bound_margin,
[…]
