+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -18,7 +18,7 @@ REPO_ROOT = Path(__file__).resolve().parents[3]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
-from bioptim import MultiCyclicCycleSolutions, OdeSolver, SolutionMerge, Solver
+from bioptim import MultiCyclicCycleSolutions, Node, OdeSolver, SolutionMerge, Solver
 from bioptim.optimization.receding_horizon_optimization import (
     RecedingHorizonOptimization,
 )
@@ -558,9 +558,12 @@ def build_argument_parser() -> argparse.ArgumentParser:
     )
     parser.add_argument(
         "--control-regularization-target-source",
-        choices=("constant", "warmup"),
+        choices=("constant", "warmup", "previous"),
         default="constant",
-        help="Use a constant pulse-width target or the standard IPOPT warmup controls as the target.",
+        help=(
+            "Use a constant target, the standard IPOPT warmup controls, or the "
+            "shifted controls from the previous MHE window."
+        ),
     )
     parser.add_argument(
         "--wheel-qdot-regularization-weight",
@@ -573,6 +576,24 @@ def build_argument_parser() -> argparse.ArgumentParser:
         type=float,
         default=-float(2 * np.pi),
         help="Target wheel angular velocity, in rad/s, for qdot regularization.",
+    )
+    parser.add_argument(
+        "--wheel-qdot-bound-margin",
+        type=float,
+        default=3.0,
+        help="Symmetric wheel-speed bound around -2*pi, in rad/s.",
+    )
+    parser.add_argument(
+        "--terminal-qdot-regularization-weight",
+        type=float,
+        default=0.0,
+        help="Quadratic Mayer weight retaining terminal joint velocities near their shifted reference.",
+    )
+    parser.add_argument(
+        "--terminal-qdot-regularization-target-source",
+        choices=("initial", "previous"),
+        default="previous",
+        help="Keep the initial terminal-velocity target or update it from the previous MHE window.",
     )
     parser.add_argument(
         "--state-scaling",
@@ -789,7 +810,7 @@ def build_argument_parser() -> argparse.ArgumentParser:
     parser.add_argument(
         "--acados-globalization",
         choices=("FIXED_STEP", "MERIT_BACKTRACKING", "FUNNEL_L1PEN_LINESEARCH"),
-        default="MERIT_BACKTRACKING",
+        default="FUNNEL_L1PEN_LINESEARCH",
         help="ACADOS globalization strategy.",
     )
     parser.add_argument(
@@ -868,7 +889,13 @@ def build_argument_parser() -> argparse.ArgumentParser:
         "--acados-wheel-q-slack",
         type=float,
         default=0.02,
-        help="First and terminal node slack, in rad, for the ACADOS wheel/crank angle transfer bounds.",
+        help="First-node slack, in rad, for the ACADOS wheel/crank angle transfer bound.",
+    )
+    parser.add_argument(
+        "--acados-terminal-wheel-q-slack",
+        type=float,
+        default=0.2,
+        help="Terminal crank-angle slack in rad; independent from the first-node transfer slack.",
     )
     parser.add_argument(
         "--acados-wheel-qdot-slack",
@@ -1444,8 +1471,17 @@ def _continuation_cache_signature(args: argparse.Namespace) -> str:
         "pulse_width_scaling": args.pulse_width_scaling,
         "control_regularization_weight": args.control_regularization_weight,
         "control_regularization_target": args.control_regularization_target,
+        "control_regularization_target_source": args.control_regularization_target_source,
         "wheel_qdot_regularization_weight": args.wheel_qdot_regularization_weight,
[…]

[…]
