### 2026-07-21 bc07230
Add per-window IPOPT refinement diagnostics
