+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1067,6 +1067,14 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Maximum IPOPT iterations for the periodic warmstart refinement.",
     )
     parser.add_argument(
+        "--periodic-ipopt-refinement-each-window",
+        action="store_true",
+        help=(
+            "Diagnostically rerun the periodic IPOPT refinement after each MHE "
+            "transfer before calling ACADOS. Its time is warmstart overhead."
+        ),
+    )
+    parser.add_argument(
         "--periodic-ipopt-refinement-use-sx",
         action="store_true",
         help=(
@@ -4825,11 +4833,15 @@ def _copy_periodic_runtime_settings(source_nmpc, target_nmpc) -> None:
     target_nmpc.n_cycles_simultaneous = source_nmpc.n_cycles_simultaneous
     for attribute_name in (
         "first_node_state_slack",
+        "terminal_state_slack",
         "bound_first_node_all_states",
         "bound_first_node_wheel_qdot",
         "advance_wheel_q_bounds",
         "wheel_q_path_margin",
         "use_signed_wheel_shift",
+        "transfer_initial_guess_mode",
+        "repeat_cyclical_state_initial_guess",
+        "continuous_state_initial_guess_mode",
     ):
         if hasattr(source_nmpc, attribute_name):
             setattr(
@@ -5330,6 +5342,14 @@ def solve_case(args: argparse.Namespace, echo: bool = True) -> dict:
     periodic_ipopt_refinement_enabled = (
         args.periodic_ipopt_refinement and not args.disable_periodic_ipopt_refinement
     )
+    if (
+        args.periodic_ipopt_refinement_each_window
+        and not periodic_ipopt_refinement_enabled
+    ):
+        raise ValueError(
+            "--periodic-ipopt-refinement-each-window requires "
+            "--periodic-ipopt-refinement."
+        )
     periodic_ipopt_reference_solution = None
 
     continuation_source = None
@@ -5583,6 +5603,10 @@ def solve_case(args: argparse.Namespace, echo: bool = True) -> dict:
             )
             print("periodic_ipopt_refinement: " f"{periodic_ipopt_refinement_enabled}")
             print(
+                "periodic_ipopt_refinement_each_window: "
+                f"{args.periodic_ipopt_refinement_each_window}"
+            )
+            print(
                 "periodic_ipopt_refinement_iterations: "
                 f"{args.periodic_ipopt_refinement_iterations}"
             )
@@ -5758,6 +5782,7 @@ def solve_case(args: argparse.Namespace, echo: bool = True) -> dict:
                 f"l2_rhs_difference={periodicity['l2_rhs_difference']:.12g}"
             )
 
+    refinement_nmpc = None
     if (
         args.solver == "acados"
         and periodic_cn_sum_approximation
@@ -5781,7 +5806,10 @@ def solve_case(args: argparse.Namespace, echo: bool = True) -> dict:
                     f"hit ({refinement_cache_path.name})"
                 )
                 print("periodic_ipopt_refinement_applied: True")
-        else:
+        if (
+            not refinement_cache_path.exists()
+            or args.periodic_ipopt_refinement_each_window
+        ):
             refinement_nmpc = build_periodic_ipopt_refinement_nmpc(
                 source_nmpc=nmpc,
                 model_path=model_path,
@@ -5810,6 +5838,7 @@ def solve_case(args: argparse.Namespace, echo: bool = True) -> dict:
                     )
                 ),
             )
+        if not refinement_cache_path.exists():
             periodic_ipopt_reference_solution = run_periodic_ipopt_refinement(
                 refinement_nmpc,
[…]
