### 2026-07-21 a15d342
Align Acados MHE transfer with IRK map
