+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -920,6 +920,14 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Reintegrate the appended cycle with the complete dynamics after each window transfer.",
     )
     parser.add_argument(
+        "--acados-transfer-irk-rollout",
+        action="store_true",
+        help=(
+            "Reintegrate the appended cycle with the exact generated ACADOS IRK simulator "
+            "after each window transfer."
+        ),
+    )
+    parser.add_argument(
         "--acados-transfer-phase-one",
         action="store_true",
         help=(
@@ -2504,9 +2512,11 @@ def summarize_windows(sol, requested_windows: int, cycles_per_window: int) -> No
     def _fmt(value) -> str:
         return "None" if value is None else f"{value:.6f}"
 
-    merged_solution, source_window_solutions, exported_cycle_solutions = (
-        _split_receding_solution(sol)
-    )
+    (
+        merged_solution,
+        source_window_solutions,
+        exported_cycle_solutions,
+    ) = _split_receding_solution(sol)
     accounting = _window_accounting(
         source_window_solutions, exported_cycle_solutions, cycles_per_window
     )
@@ -2554,9 +2564,11 @@ def summarize_windows(sol, requested_windows: int, cycles_per_window: int) -> No
 
 
 def build_window_summary(sol, requested_windows: int, cycles_per_window: int) -> dict:
-    merged_solution, source_window_solutions, exported_cycle_solutions = (
-        _split_receding_solution(sol)
-    )
+    (
+        merged_solution,
+        source_window_solutions,
+        exported_cycle_solutions,
+    ) = _split_receding_solution(sol)
     accounting = _window_accounting(
         source_window_solutions, exported_cycle_solutions, cycles_per_window
     )
@@ -3450,7 +3462,9 @@ def rollout_transferred_cycle_full_dynamics(
             "The transfer rollout requires one more state node than controls."
         )
 
-    start_node = periodic_nmpc.nodes_per_cycle - 1
+    # Nodes 0..nodes_per_cycle are retained from the solved horizon. Only the
+    # intervals after that shared endpoint belong to the appended cycle.
+    start_node = periodic_nmpc.nodes_per_cycle
     if start_node < 0 or start_node >= n_control_nodes:
         raise ValueError("The transfer rollout start node is outside the horizon.")
 
@@ -3519,6 +3533,190 @@ def rollout_transferred_cycle_full_dynamics(
         "max_bound_violation_by_key": max_bound_violation_by_key,
         "terminal_delta": terminal_delta,
         "max_delta_by_key": max_delta_by_key,
+    }
+
+
+def _acados_variable_scaling(nlp, variables, scaling_container) -> np.ndarray:
+    scaling = np.ones(variables.shape)
+    for key in variables.keys():
+        indexes = np.asarray(variables[key].index).reshape((-1,)).tolist()
+        key_scaling = np.asarray(
+            scaling_container[key].scaling[:, 0], dtype=float
+        ).reshape(-1)
+        if key_scaling.size != len(indexes):
+            raise ValueError(
+                f"ACADOS scaling for '{key}' has {key_scaling.size} rows, "
+                f"expected {len(indexes)}."
+            )
+        scaling[indexes] = key_scaling
+    if not np.all(np.isfinite(scaling)) or np.any(scaling <= 0.0):
+        raise ValueError(
+            "ACADOS transfer scaling must be finite and strictly positive."
+        )
+    return scaling
+
+
+def _get_or_create_acados_sim_solver(periodic_nmpc):
+    cached_solver = getattr(periodic_nmpc, "_cocofest_acados_sim_solver", None)
+    if cached_solver is not None:
+        return cached_solver, False
+
[…]
