+++ tests/shard1/test_periodic_pulse_width.py
@@ -1153,8 +1153,8 @@ def test_full_dynamics_transfer_rollout_reintegrates_appended_cycle():
     )
     control_variables = Variables({"acceleration": SimpleNamespace(index=[0])}, shape=1)
     x_init = {
-        "q": SimpleNamespace(init=np.array([[0.0, 0.5, 9.0, 9.0, 9.0]])),
-        "qdot": SimpleNamespace(init=np.array([[1.0, 1.0, 9.0, 9.0, 9.0]])),
+        "q": SimpleNamespace(init=np.array([[0.0, 0.5, 1.0, 9.0, 9.0]])),
+        "qdot": SimpleNamespace(init=np.array([[1.0, 1.0, 1.0, 9.0, 9.0]])),
     }
     loose_bounds = SimpleNamespace(
         min=np.full((1, 3), -100.0), max=np.full((1, 3), 100.0)
@@ -1179,13 +1179,13 @@ def test_full_dynamics_transfer_rollout_reintegrates_appended_cycle():
     )
 
     assert summary["applied"] is True
-    assert summary["start_node"] == 1
+    assert summary["start_node"] == 2
     assert summary["max_bound_violation"] == 0.0
     np.testing.assert_allclose(x_init["q"].init, [[0.0, 0.5, 1.0, 1.5, 2.0]])
     np.testing.assert_allclose(x_init["qdot"].init, [[1.0, 1.0, 1.0, 1.0, 1.0]])
 
-    x_init["q"].init[:, 2:] = 9.0
-    x_init["qdot"].init[:, 2:] = 9.0
+    x_init["q"].init[:, 3:] = 9.0
+    x_init["qdot"].init[:, 3:] = 9.0
     nlp.x_bounds["q"] = SimpleNamespace(
         min=np.full((1, 3), -100.0), max=np.full((1, 3), 1.0)
     )
@@ -1195,7 +1195,137 @@ def test_full_dynamics_transfer_rollout_reintegrates_appended_cycle():
 
     assert rejected["applied"] is False
     assert rejected["max_bound_violation"] == 1.0
-    np.testing.assert_allclose(x_init["q"].init[:, 2:], 9.0)
+    np.testing.assert_allclose(x_init["q"].init[:, 3:], 9.0)
+
+
+def test_acados_irk_transfer_rollout_uses_scaled_variables_and_stage_data():
+    class Variables(dict):
+        def __init__(self, *args, shape, **kwargs):
+            super().__init__(*args, **kwargs)
+            self.shape = shape
+
+    class FakeSimulator:
+        def __init__(self):
+            self.acados_sim = SimpleNamespace(dims=SimpleNamespace(nx=2, nu=1))
+            self.calls = []
+
+        def set(self, field, value):
+            assert field == "t0"
+            np.testing.assert_allclose(value, [0.0])
+
+        def simulate(self, x, u, p):
+            self.calls.append((x.copy(), u.copy(), p.copy()))
+            return x + np.array([u[0], p[0]])
+
+        def get(self, field):
+            assert field == "time_tot"
+            return 0.001
+
+    states = Variables(
+        {"q": SimpleNamespace(index=[0]), "qdot": SimpleNamespace(index=[1])},
+        shape=2,
+    )
+    controls = Variables({"acceleration": SimpleNamespace(index=[0])}, shape=1)
+    x_init = {
+        "q": SimpleNamespace(init=np.array([[0.0, 1.0, 2.0, 9.0, 9.0]])),
+        "qdot": SimpleNamespace(init=np.array([[2.0, 2.0, 2.0, 9.0, 9.0]])),
+    }
+    scaling = lambda value: SimpleNamespace(  # noqa: E731 - compact test fixture.
+        scaling=np.array([[value]])
+    )
[…]
