### 2026-07-21 b4010a1
Add Acados control homotopy diagnostics
