+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -449,6 +449,23 @@ def parse_objectives(raw_objective: str) -> set[str]:
     return values or {"force"}
 
 
+def parse_control_homotopy_radii(raw_radii: str) -> tuple[float, ...]:
+    radii = tuple(float(item.strip()) for item in raw_radii.split(",") if item.strip())
+    if not radii:
+        raise argparse.ArgumentTypeError(
+            "The control homotopy requires at least one radius."
+        )
+    if any(not np.isfinite(radius) or radius <= 0.0 for radius in radii):
+        raise argparse.ArgumentTypeError(
+            "Control homotopy radii must be finite and strictly positive."
+        )
+    if any(next_radius <= radius for radius, next_radius in zip(radii, radii[1:])):
+        raise argparse.ArgumentTypeError(
+            "Control homotopy radii must be strictly increasing."
+        )
+    return radii
+
+
 def build_cost_fun_weight(objectives: set[str]) -> list[int]:
     weights = [0, 0, 0]
     for objective in objectives:
@@ -936,6 +953,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--acados-cyclical-transfer-mode",
+        choices=("extrapolate", "repeat"),
+        default="extrapolate",
+        help=(
+            "Construct cyclic states in the appended MHE cycle by preserving the "
+            "last observed cycle drift or by repeating the last cycle verbatim."
+        ),
+    )
+    parser.add_argument(
         "--acados-transfer-rollout-substeps",
         type=int,
         default=5,
@@ -1017,6 +1043,53 @@ def build_argument_parser() -> argparse.ArgumentParser:
         type=float,
         default=1e-5,
         help="Half-width of fixed ACADOS control bounds in scaled control units.",
+    )
+    parser.add_argument(
+        "--acados-control-homotopy-radii",
+        type=parse_control_homotopy_radii,
+        default=None,
+        help=(
+            "Comma-separated, strictly increasing pulse-width radii in seconds. "
+            "ACADOS first solves with controls fixed to the IPOPT seed, then solves "
+            "one proximal subproblem per radius before restoring the original bounds."
+        ),
+    )
+    parser.add_argument(
+        "--acados-control-homotopy-tolerance",
+        type=float,
+        default=5e-4,
+        help="KKT tolerance used only by the ACADOS control-homotopy subproblems.",
+    )
+    parser.add_argument(
+        "--acados-control-homotopy-max-restarts",
+        type=int,
+        default=2,
+        help=(
+            "Maximum number of same-radius SQP restarts when a homotopy stage has "
+            "small feasibility residuals but has not reached stationarity."
+        ),
+    )
+    parser.add_argument(
+        "--acados-control-homotopy-keep-final-radius",
+        action="store_true",
+        help=(
+            "Keep the largest accepted homotopy radius for the first MHE window "
+            "instead of immediately restoring unrestricted control bounds."
+        ),
+    )
+    parser.add_argument(
+        "--acados-control-homotopy-each-window",
+        action="store_true",
+        help=(
+            "Repeat the fixed-control and radius continuation after each MHE "
+            "window transfer before solving the next window."
+        ),
+    )
+    parser.add_argument(
+        "--acados-control-homotopy-window-growth",
+        type=float,
+        default=1.0,
+        help=("Factor applied to the retained homotopy radius after each MHE window."),
     )
     parser.add_argument(
         "--ipopt-linear-solver",
@@ -2258,6 +2331,119 @@ def acados_diagnostics_meet_tolerances(
     )
 
 
+def acados_homotopy_stage_is_restartable(
+    diagnostics: dict, convergence_tolerance: float
+) -> bool:
[…]
