### 2026-07-22 5d5f670
Bound Acados homotopy attempt length
