### 2026-07-22 c854c49
Cache and stabilize Acados window refinements
