+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1165,6 +1165,14 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--periodic-ipopt-refinement-window-cache",
+        action="store_true",
+        help=(
+            "Cache successful diagnostic IPOPT refinements by window so repeated "
+            "ACADOS transfer experiments reuse the same feasible seed."
+        ),
+    )
+    parser.add_argument(
         "--periodic-ipopt-refinement-use-sx",
         action="store_true",
         help=(
@@ -1490,6 +1498,15 @@ def _periodic_ipopt_refinement_cache_path(
         ],
     }
     return _cache_root() / f"periodic_ipopt_{_short_hash(payload)}.npz"
+
+
+def _periodic_ipopt_window_refinement_cache_path(
+    args: argparse.Namespace, model_path: Path, window: int
+) -> Path:
+    if window < 0:
+        raise ValueError("The IPOPT refinement cache window must be non-negative.")
+    base_path = _periodic_ipopt_refinement_cache_path(args, model_path)
+    return base_path.with_name(f"{base_path.stem}_window_{window:04d}.npz")
 
 
 def _acados_seed_cache_path(args: argparse.Namespace, model_path: Path) -> Path | None:
@@ -2356,6 +2373,18 @@ def acados_homotopy_stage_is_restartable(
     )
 
 
+def set_acados_runtime_max_iterations(periodic_nmpc, max_iterations: int) -> bool:
+    """Update the generated capsule when Bioptim reuses an existing Acados solver."""
+
+    acados_interface = getattr(periodic_nmpc, "ocp_solver", None)
+    acados_solver = getattr(acados_interface, "ocp_solver", None)
+    options_set = getattr(acados_solver, "options_set", None)
+    if options_set is None:
+        return False
+    options_set("nlp_solver_max_iter", int(max_iterations))
+    return True
+
+
 def run_acados_control_homotopy(
     periodic_nmpc,
     solver,
@@ -2371,8 +2400,17 @@ def run_acados_control_homotopy(
 
     stage_solver = deepcopy(solver)
     stage_solver.set_convergence_tolerance(convergence_tolerance)
-    if stage_iterations is not None:
+    if (
+        stage_iterations is not None
+        and getattr(stage_solver, "nlp_solver_max_iter", None) != stage_iterations
+    ):
         stage_solver.set_maximum_iterations(stage_iterations)
+    acados_interface = getattr(periodic_nmpc, "ocp_solver", None)
+    if getattr(acados_interface, "ocp_solver", None) is not None:
+        # The generated solver already owns this immutable maximum. The initial
+        # homotopy compiles it with stage_iterations; subsequent windows must not
+        # ask Bioptim to reconfigure the same Acados capsule.
+        stage_solver.set_only_first_options_has_changed(False)
     summaries = []
 
     if solve_stage is None:
@@ -2394,6 +2432,8 @@ def run_acados_control_homotopy(
 
             stage_accepted = False
             for attempt in range(max_restarts + 1):
+                if stage_iterations is not None:
+                    set_acados_runtime_max_iterations(periodic_nmpc, stage_iterations)
                 solution = solve_stage()
                 diagnostics = snapshot_acados_diagnostics(solution)
                 accepted = _status_is_success(
@@ -6121,6 +6161,10 @@ def solve_case(args: argparse.Namespace, echo: bool = True) -> dict:
                 f"{args.periodic_ipopt_refinement_each_window}"
             )
             print(
+                "periodic_ipopt_refinement_window_cache: "
+                f"{args.periodic_ipopt_refinement_window_cache}"
+            )
+            print(
[…]
