### 2026-07-22 e554ebd
Recover Acados after failed homotopy stages
