### 2026-07-22 3fcecc4
Add strict inter-window FES continuity mode
