### 2026-07-22 047ab17
Add Acados transfer recovery diagnostics
