+++ tests/shard1/test_periodic_pulse_width.py
@@ -284,6 +284,47 @@ def test_control_homotopy_radii_are_parsed_as_an_increasing_sequence():
     assert args.acados_control_homotopy_stage_iterations == 40
 
 
+def test_transfer_bound_homotopy_fractions_are_parsed():
+    args = periodic_example.build_argument_parser().parse_args(
+        [
+            "--acados-transfer-irk-rollout",
+            "--acados-transfer-bound-homotopy",
+            "--acados-transfer-bound-homotopy-fractions",
+            "0,0.5,1",
+            "--acados-transfer-bound-homotopy-padding",
+            "0.1",
+            "--acados-transfer-bound-homotopy-iterations",
+            "12",
+        ]
+    )
+
+    assert args.acados_transfer_bound_homotopy is True
+    assert args.acados_transfer_bound_homotopy_fractions == (0.0, 0.5, 1.0)
+    assert args.acados_transfer_bound_homotopy_padding == 0.1
+    assert args.acados_transfer_bound_homotopy_iterations == 12
+
+
+def test_transfer_sqp_restart_options_are_parsed():
+    parser = periodic_example.build_argument_parser()
+    args = parser.parse_args(
+        [
+            "--acados-transfer-sqp-restarts",
+            "4",
+            "--acados-transfer-sqp-restart-iterations",
+            "2",
+            "--acados-transfer-sqp-restart-feasibility-tolerance",
+            "0.02",
+        ]
+    )
+
+    assert args.acados_transfer_sqp_restarts == 4
+    assert args.acados_transfer_sqp_restart_iterations == 2
+    assert args.acados_transfer_sqp_restart_feasibility_tolerance == 0.02
+    assert periodic_example._codegen_signature(
+        parser.parse_args([])
+    ) != periodic_example._codegen_signature(args)
+
+
 def test_acados_cyclical_transfer_extrapolates_by_default():
     args = periodic_example.build_argument_parser().parse_args([])
 
@@ -1549,6 +1590,29 @@ def test_full_dynamics_transfer_rollout_reintegrates_appended_cycle():
     np.testing.assert_allclose(x_init["q"].init[:, 3:], 9.0)
 
 
+def test_appended_pulse_width_scaling_preserves_retained_cycle_and_clips():
+    values = np.array([[0.1, 0.2, 0.2, 0.3]])
+    nmpc = SimpleNamespace(
+        nodes_per_cycle=2,
+        nlp=[
+            SimpleNamespace(
+                u_init={"last_pulse_width_Biceps": SimpleNamespace(init=values)},
+                u_bounds={
+                    "last_pulse_width_Biceps": SimpleNamespace(
+                        min=np.zeros((1, 3)), max=np.full((1, 3), 0.6)
+                    )
+                },
+            )
+        ],
+    )
+
+    summary = periodic_example.scale_appended_pulse_width_controls(nmpc, 2.5)
+
+    np.testing.assert_allclose(values, [[0.1, 0.2, 0.5, 0.6]])
+    assert summary["start_node"] == 2
+    assert summary["controls"]["last_pulse_width_Biceps"]["clipped_count"] == 1
+
+
 def test_acados_irk_transfer_rollout_uses_scaled_variables_and_stage_data():
     class Variables(dict):
         def __init__(self, *args, shape, **kwargs):
@@ -1559,10 +1623,11 @@ def test_acados_irk_transfer_rollout_uses_scaled_variables_and_stage_data():
         def __init__(self):
             self.acados_sim = SimpleNamespace(dims=SimpleNamespace(nx=2, nu=1))
             self.calls = []
[…]
