+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -466,6 +466,37 @@ def parse_control_homotopy_radii(raw_radii: str) -> tuple[float, ...]:
     return radii
 
 
+def parse_transfer_bound_homotopy_fractions(
+    raw_fractions: str,
+) -> tuple[float, ...]:
+    fractions = tuple(
+        float(item.strip()) for item in raw_fractions.split(",") if item.strip()
+    )
+    if not fractions:
+        raise argparse.ArgumentTypeError(
+            "The transfer-bound homotopy requires at least one fraction."
+        )
+    if any(
+        not np.isfinite(fraction) or fraction < 0.0 or fraction > 1.0
+        for fraction in fractions
+    ):
+        raise argparse.ArgumentTypeError(
+            "Transfer-bound homotopy fractions must be finite and between 0 and 1."
+        )
+    if any(
+        next_fraction <= fraction
+        for fraction, next_fraction in zip(fractions, fractions[1:])
+    ):
+        raise argparse.ArgumentTypeError(
+            "Transfer-bound homotopy fractions must be strictly increasing."
+        )
+    if not np.isclose(fractions[-1], 1.0):
+        raise argparse.ArgumentTypeError(
+            "Transfer-bound homotopy fractions must end at 1."
+        )
+    return fractions
+
+
 def build_cost_fun_weight(objectives: set[str]) -> list[int]:
     weights = [0, 0, 0]
     for objective in objectives:
@@ -962,6 +993,68 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--acados-transfer-bound-homotopy",
+        action="store_true",
+        help=(
+            "Temporarily enlarge path and terminal state bounds around the IRK "
+            "transfer, then tighten them back to the physical bounds with ACADOS."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-bound-homotopy-fractions",
+        type=parse_transfer_bound_homotopy_fractions,
+        default=(0.0, 0.25, 0.5, 0.75, 1.0),
+        help=(
+            "Strictly increasing interpolation fractions ending at 1 for the "
+            "transfer-bound homotopy."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-bound-homotopy-padding",
+        type=float,
+        default=0.05,
+        help=(
+            "Relative padding around the transferred state trajectory in the "
+            "fully relaxed bound stage."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-bound-homotopy-iterations",
+        type=int,
+        default=30,
+        help="Maximum SQP iterations for each transfer-bound homotopy stage.",
+    )
+    parser.add_argument(
+        "--acados-transfer-bound-homotopy-tolerance",
+        type=float,
+        default=1e-4,
+        help="KKT tolerance used by transfer-bound homotopy stages.",
+    )
+    parser.add_argument(
+        "--acados-transfer-sqp-restarts",
+        type=int,
+        default=0,
+        help=(
+            "Maximum number of short ACADOS restarts used to repair a transferred "
+            "initial guess before solving the next MHE window."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-sqp-restart-iterations",
+        type=int,
+        default=1,
+        help="SQP iterations performed by each transfer-repair attempt.",
+    )
+    parser.add_argument(
+        "--acados-transfer-sqp-restart-feasibility-tolerance",
+        type=float,
+        default=1e-2,
+        help=(
[…]
