### 2026-07-22 ff474b3
Refactor shared receding-horizon warm starts
