+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -11,6 +11,7 @@ import os
 from pathlib import Path
 import sys
 from sys import platform as sys_platform
+from time import perf_counter
 
 import numpy as np
 
@@ -21,6 +22,13 @@ if str(REPO_ROOT) not in sys.path:
 from bioptim import MultiCyclicCycleSolutions, Node, OdeSolver, SolutionMerge, Solver
 from bioptim.optimization.receding_horizon_optimization import (
     RecedingHorizonOptimization,
+)
+
+from cocofest.optimization.receding_horizon_initial_guess import (
+    audit_initial_guess,
+    copy_container_values,
+    project_initial_guess_to_bounds,
+    snapshot_container,
 )
 
 try:
@@ -972,9 +980,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Project the ACADOS qdot initial guess from finite differences of q before solving.",
     )
     parser.add_argument(
+        "--transfer-full-dynamics-rollout",
         "--acados-transfer-full-dynamics-rollout",
+        dest="acados_transfer_full_dynamics_rollout",
         action="store_true",
-        help="Reintegrate the appended cycle with the complete dynamics after each window transfer.",
+        help=(
+            "Reintegrate the appended cycle with the complete dynamics after each "
+            "window transfer. This solver-independent rollout is available to IPOPT "
+            "and ACADOS."
+        ),
     )
     parser.add_argument(
         "--acados-transfer-irk-rollout",
@@ -985,11 +999,13 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--transfer-phase-one",
         "--acados-transfer-phase-one",
+        dest="acados_transfer_phase_one",
         action="store_true",
         help=(
             "Apply the bounded proximal complete-dynamics phase I after each MHE window "
-            "transfer. Uses the --full-dynamics-phase-one-* settings."
+            "transfer for either solver. Uses the --full-dynamics-phase-one-* settings."
         ),
     )
     parser.add_argument(
@@ -1064,13 +1080,17 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--transfer-rollout-substeps",
         "--acados-transfer-rollout-substeps",
+        dest="acados_transfer_rollout_substeps",
         type=int,
         default=5,
         help="RK4 substeps used by the full-dynamics inter-window rollout.",
     )
     parser.add_argument(
+        "--transfer-pulse-width-scale",
         "--acados-transfer-pulse-width-scale",
+        dest="acados_transfer_pulse_width_scale",
         type=float,
         default=1.0,
         help=(
@@ -1079,7 +1099,9 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--transfer-rollout-max-bound-violation",
         "--acados-transfer-rollout-max-bound-violation",
+        dest="acados_transfer_rollout_max_bound_violation",
         type=float,
         default=1.0,
         help="Reject the inter-window rollout when a state exceeds its bounds by more than this value.",
@@ -2161,13 +2183,6 @@ def _state_traces_from_exported_cycles(
         )
 
     return state_traces
-
-
-def _initial_guess_traces(container) -> dict[str, np.ndarray]:
-    return {
-        key: np.array(container[key].init, dtype=float, copy=True)
-        for key in container.keys()
-    }
 
 
 def _control_bounds_summary(nmpc) -> dict[str, dict[str, float]]:
@@ -4393,8 +4408,8 @@ def run_acados_transfer_bound_homotopy(
                 warm_start=None,
             )
 
[…]

[…]
