### 2026-07-22 cf8c04f
Stabilize Acados cyclic MHE transfers
