+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -474,6 +474,44 @@ def parse_control_homotopy_radii(raw_radii: str) -> tuple[float, ...]:
     return radii
 
 
+def parse_proximal_control_weights(raw_weights: str) -> tuple[float, ...]:
+    weights = tuple(
+        float(item.strip()) for item in raw_weights.split(",") if item.strip()
+    )
+    if not weights:
+        raise argparse.ArgumentTypeError(
+            "The proximal control continuation requires at least one weight."
+        )
+    if any(not np.isfinite(weight) or weight <= 0.0 for weight in weights):
+        raise argparse.ArgumentTypeError(
+            "Proximal control weights must be finite and strictly positive."
+        )
+    if any(next_weight >= weight for weight, next_weight in zip(weights, weights[1:])):
+        raise argparse.ArgumentTypeError(
+            "Proximal control weights must be strictly decreasing."
+        )
+    return weights
+
+
+def parse_terminal_wheel_q_slacks(raw_slacks: str) -> tuple[float, ...]:
+    slacks = tuple(
+        float(item.strip()) for item in raw_slacks.split(",") if item.strip()
+    )
+    if not slacks:
+        raise argparse.ArgumentTypeError(
+            "Terminal wheel-bound continuation requires at least one slack."
+        )
+    if any(not np.isfinite(slack) or slack < 0.0 for slack in slacks):
+        raise argparse.ArgumentTypeError(
+            "Terminal wheel-bound slacks must be finite and non-negative."
+        )
+    if any(next_slack >= slack for slack, next_slack in zip(slacks, slacks[1:])):
+        raise argparse.ArgumentTypeError(
+            "Terminal wheel-bound slacks must be strictly decreasing."
+        )
+    return slacks
+
+
 def parse_transfer_bound_homotopy_fractions(
     raw_fractions: str,
 ) -> tuple[float, ...]:
@@ -647,9 +685,12 @@ def build_argument_parser() -> argparse.ArgumentParser:
     )
     parser.add_argument(
         "--terminal-qdot-regularization-target-source",
-        choices=("initial", "previous"),
+        choices=("initial", "previous", "first_node"),
         default="previous",
-        help="Keep the initial terminal-velocity target or update it from the previous MHE window.",
+        help=(
+            "Keep the initial terminal-velocity target, update it from the previous "
+            "MHE window, or enforce cyclic velocity relative to the current first node."
+        ),
     )
     parser.add_argument(
         "--state-scaling",
@@ -954,6 +995,20 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Terminal crank-angle slack in rad; independent from the first-node transfer slack.",
     )
     parser.add_argument(
+        "--acados-terminal-wheel-q-homotopy-slacks",
+        type=parse_terminal_wheel_q_slacks,
+        default=None,
+        help=(
+            "Comma-separated decreasing terminal crank-angle slacks. ACADOS "
+            "solves each bound stage around the same one-turn target."
+        ),
+    )
+    parser.add_argument(
+        "--acados-terminal-wheel-q-homotopy-each-window",
+        action="store_true",
+        help="Repeat terminal crank-bound tightening after every MHE transfer.",
+    )
+    parser.add_argument(
         "--acados-wheel-qdot-slack",
         type=float,
         default=0.5,
@@ -1099,6 +1154,26 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--transfer-ding-force-compensation",
+        action="store_true",
+        help=(
+            "Choose each appended-cycle pulse width with a scalar Ding rollout "
+            "so its next-node force follows the preceding cycle."
+        ),
+    )
+    parser.add_argument(
+        "--transfer-ding-force-compensation-substeps",
+        type=int,
[…]
