+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -23,12 +23,16 @@ try:
     from .cycling_pulse_width_mhe_acados_periodic import (
         ACADOS_STATUS_NAMES,
         build_argument_parser,
+        parse_proximal_control_weights,
+        parse_terminal_wheel_q_slacks,
         solve_case,
     )
 except ImportError:
     from cycling_pulse_width_mhe_acados_periodic import (
         ACADOS_STATUS_NAMES,
         build_argument_parser,
+        parse_proximal_control_weights,
+        parse_terminal_wheel_q_slacks,
         solve_case,
     )
 
@@ -1147,12 +1151,19 @@ def main(
     terminal_qdot_regularization_weight: float = 0.0,
     terminal_qdot_regularization_target_source: str = "previous",
     acados_terminal_wheel_q_slack: float = 0.2,
+    acados_terminal_wheel_q_homotopy_slacks: tuple[float, ...] | None = None,
+    acados_terminal_wheel_q_homotopy_each_window: bool = False,
     state_scaling: str = "none",
     acados_state_scaling: str | None = None,
     pulse_width_scaling: float = 1 / 400,
     acados_pulse_width_scaling: float | None = None,
     acados_pulse_width_trust_radius: float | None = None,
     acados_transfer_pulse_width_trust_radius: float | None = None,
+    acados_proximal_control_weights: tuple[float, ...] | None = None,
+    acados_proximal_control_each_window: bool = False,
+    acados_proximal_control_tolerance: float = 5e-4,
+    acados_proximal_control_stage_iterations: int = 50,
+    acados_proximal_control_max_restarts: int = 1,
     acados_fes_state_trust_radius: float | None = None,
     acados_fatigue_warmstart_mode: str = "continuous",
     acados_tolerance: float | None = None,
@@ -1176,6 +1187,9 @@ def main(
     shared_initial_phase_one: bool = False,
     shared_transfer_rollout_substeps: int = 5,
     shared_transfer_rollout_max_bound_violation: float = 1.0,
+    shared_transfer_ding_force_compensation: bool = False,
+    shared_transfer_ding_force_compensation_substeps: int = 5,
+    shared_transfer_ding_force_compensation_iterations: int = 20,
     acados_integrator_type: str = "IRK",
     acados_collocation_type: str = "GAUSS_LEGENDRE",
     acados_sim_stages: int = 4,
@@ -1417,6 +1431,23 @@ def main(
     )
     acados_args.acados_stationarity_tolerance = acados_stationarity_tolerance
     acados_args.acados_dual_warm_start_mode = acados_dual_warm_start_mode
+    acados_args.acados_proximal_control_weights = acados_proximal_control_weights
+    acados_args.acados_proximal_control_each_window = (
+        acados_proximal_control_each_window
+    )
+    acados_args.acados_proximal_control_tolerance = acados_proximal_control_tolerance
+    acados_args.acados_proximal_control_stage_iterations = (
+        acados_proximal_control_stage_iterations
+    )
+    acados_args.acados_proximal_control_max_restarts = (
+        acados_proximal_control_max_restarts
+    )
+    acados_args.acados_terminal_wheel_q_homotopy_slacks = (
+        acados_terminal_wheel_q_homotopy_slacks
+    )
+    acados_args.acados_terminal_wheel_q_homotopy_each_window = (
+        acados_terminal_wheel_q_homotopy_each_window
+    )
     for solver_args in (ipopt_args, acados_args):
         solver_args.acados_transfer_full_dynamics_rollout = (
             shared_transfer_full_dynamics_rollout
@@ -1426,6 +1457,15 @@ def main(
         solver_args.acados_transfer_rollout_substeps = shared_transfer_rollout_substeps
[…]
