### 2026-07-22 203e9b8
Separate Acados transfer compensation in benchmarks
