### 2026-07-22 e8d83d1
Add periodic collocation benchmark bridge
