### 2026-07-23 a987450
Warm-start Acados from periodic collocation
