### 2026-07-23 56be433
Expose robust Acados transfer restarts
