### 2026-07-23 e36a0db
Expose HPIPM benchmark tuning
