### 2026-07-23 9955482
Compare exported solver trajectories
