+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -491,10 +491,20 @@ def _state_trace_comparisons(
             common_len = min(reference_values.shape[1], compared_values.shape[1])
             reference_common = reference_values[row, :common_len]
             compared_common = compared_values[row, :common_len]
+            metric_key = key if reference_values.shape[0] == 1 else f"{key}[{row}]"
+            if metric_key == "q[2]":
+                reference_common = np.unwrap(reference_common)
+                compared_common = np.unwrap(compared_common)
+                turn_offset = int(
+                    np.rint(
+                        np.median((compared_common - reference_common) / (2 * np.pi))
+                    )
+                )
+                compared_common = compared_common - turn_offset * 2 * np.pi
             diff = compared_common - reference_common
             comparisons.append(
                 {
-                    "key": (key if reference_values.shape[0] == 1 else f"{key}[{row}]"),
+                    "key": metric_key,
                     "common_len": common_len,
                     "rmse": float(np.sqrt(np.mean(diff**2))),
                     "mae": float(np.mean(np.abs(diff))),
@@ -514,6 +524,89 @@ def _state_trace_comparisons(
             )
 
     return sorted(comparisons, key=lambda item: item["rmse"], reverse=True)
+
+
+def _print_solution_metrics(
+    ipopt_result: dict,
+    acados_result: dict,
+    *,
+    line_prefix: str = "",
+    state_comparison_limit: int = 12,
+) -> None:
+    trace_metrics = _trace_comparison(ipopt_result, acados_result)
+    print(
+        f"{line_prefix}wheel trace comparison | "
+        f"common_len={trace_metrics['common_len']} | "
+        f"unwrapped_rmse={trace_metrics['unwrapped_wheel_rmse']:.6f} | "
+        f"unwrapped_max_abs_error={trace_metrics['unwrapped_wheel_max_abs_error']:.6f} | "
+        f"unwrapped_final_error={trace_metrics['unwrapped_wheel_final_error']:.6f} | "
+        f"turn_aligned_unwrapped_rmse={trace_metrics['turn_aligned_unwrapped_rmse']:.6f} | "
+        f"turn_aligned_unwrapped_max_abs_error={trace_metrics['turn_aligned_unwrapped_max_abs_error']:.6f} | "
+        f"turn_aligned_unwrapped_final_error={trace_metrics['turn_aligned_unwrapped_final_error']:.6f} | "
+        f"unwrapped_turn_offset={trace_metrics['unwrapped_turn_offset']} | "
+        f"phase_rmse={trace_metrics['phase_rmse']:.6f} | "
+        f"phase_max_abs_error={trace_metrics['phase_max_abs_error']:.6f} | "
+        f"phase_final_error={trace_metrics['phase_final_error']:.6f}"
+    )
+    print(
+        f"{line_prefix}raw wheel final-angle representation | "
+        f"raw_final_error={trace_metrics['raw_final_error']:.6f} | "
+        f"raw_final_turn_offset={trace_metrics['raw_final_turn_offset']}"
+    )
+
+    control_metrics = _control_comparisons(ipopt_result, acados_result)
+    if control_metrics:
+        print(
+            f"{line_prefix}control comparison | key | common_len | rmse | mae | max_abs_error | final_error | "
+            "ipopt_mean | acados_mean | ipopt_sum | acados_sum | ipopt_range | acados_range"
+        )
+        for metric in control_metrics:
+            print(
+                f"{line_prefix}control comparison | "
+                f"{metric['key']} | "
+                f"{metric['common_len']} | "
+                f"{_format_control_metric(metric['rmse'])} | "
+                f"{_format_control_metric(metric['mae'])} | "
+                f"{_format_control_metric(metric['max_abs_error'])} | "
[…]
