### 2026-07-23 e9ef1e6
Report exported fatigue separately
