### 2026-07-23 73aa9c4
Stabilize Acados transfer warm starts
