+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1122,7 +1122,16 @@ def build_argument_parser() -> argparse.ArgumentParser:
         "--acados-transfer-bound-homotopy-tolerance",
         type=float,
         default=1e-4,
-        help="KKT tolerance used by transfer-bound homotopy stages.",
+        help="KKT residual threshold used to accept transfer-bound homotopy stages.",
+    )
+    parser.add_argument(
+        "--acados-transfer-bound-homotopy-solver-tolerance",
+        type=float,
+        default=None,
+        help=(
+            "Optional stricter KKT tolerance requested from ACADOS during each "
+            "transfer-bound homotopy stage."
+        ),
     )
     parser.add_argument(
         "--acados-transfer-sqp-restarts",
@@ -1374,6 +1383,14 @@ def build_argument_parser() -> argparse.ArgumentParser:
         type=int,
         default=1,
         help="Maximum same-weight restarts for a nearly feasible proximal stage.",
+    )
+    parser.add_argument(
+        "--acados-proximal-control-try-next-weight-on-failure",
+        action="store_true",
+        help=(
+            "After resetting a failed proximal stage, try the next lower "
+            "configured weight from the unchanged transferred primal."
+        ),
     )
     parser.add_argument(
         "--ipopt-linear-solver",
@@ -2895,6 +2912,7 @@ def run_acados_proximal_control_continuation(
     convergence_tolerance: float,
     max_restarts: int = 1,
     stage_iterations: int | None = 50,
+    try_next_weight_on_failure: bool = False,
     echo: bool = True,
     solve_stage=None,
 ) -> list[dict]:
@@ -3001,7 +3019,7 @@ def run_acados_proximal_control_continuation(
                     periodic_nmpc, solution
                 )
             summary["solver_reset"] = reset_acados_solver_memory(periodic_nmpc)
-        if not stage_accepted:
+        if not stage_accepted and not try_next_weight_on_failure:
             break
 
     if accepted_weight is not None:
@@ -3288,12 +3306,15 @@ def print_initial_guess_diagnostics(nmpc) -> None:
             )
 
 
-def project_qdot_initial_guess_from_q(nmpc) -> None:
+def project_qdot_initial_guess_from_q(nmpc, start_node: int = 0) -> dict:
     if "q" not in nmpc.nlp[0].x_init.keys() or "qdot" not in nmpc.nlp[0].x_init.keys():
-        return
+        return {"applied": False, "max_change": 0.0, "clipped_count": 0}
 
     dt = nmpc.cycle_duration / nmpc.cycle_len
     q = np.asarray(nmpc.nlp[0].x_init["q"].init, dtype=float)
+    previous_qdot = np.asarray(nmpc.nlp[0].x_init["qdot"].init, dtype=float).copy()
+    if start_node < 0 or start_node >= q.shape[1]:
+        raise ValueError("qdot projection start_node must index the state trajectory.")
     qdot = np.empty_like(q)
     qdot[:, :-1] = np.diff(q, axis=1) / dt
     qdot[:, -1] = qdot[:, -2]
@@ -3301,8 +3322,18 @@ def project_qdot_initial_guess_from_q(nmpc) -> None:
     lower, upper = _trajectory_bounds_for_guess(
         nmpc.nlp[0].x_bounds["qdot"], qdot.shape[1]
     )
-    nmpc.nlp[0].x_init["qdot"].init[:, :] = np.minimum(np.maximum(qdot, lower), upper)
+    projected_qdot = np.minimum(np.maximum(qdot, lower), upper)
+    nmpc.nlp[0].x_init["qdot"].init[:, start_node:] = projected_qdot[:, start_node:]
     nmpc._sync_acados_state_bounds()
[…]

[…]
