### 2026-07-23 3369ec0
Adapt Acados qdot transfer projection
