+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1393,6 +1393,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--continue-after-acados-transfer-failure",
+        action="store_true",
+        help=(
+            "Diagnostic mode that keeps advancing after an auxiliary ACADOS "
+            "transfer solve fails. Endurance runs stop at the first failed "
+            "transfer by default."
+        ),
+    )
+    parser.add_argument(
         "--ipopt-linear-solver",
         type=str,
         default="ma57",
@@ -3306,9 +3315,19 @@ def print_initial_guess_diagnostics(nmpc) -> None:
             )
 
 
-def project_qdot_initial_guess_from_q(nmpc, start_node: int = 0) -> dict:
+def project_qdot_initial_guess_from_q(
+    nmpc,
+    start_node: int = 0,
+    select_by_dynamics: bool = False,
+    dynamics_substeps: int = 5,
+    max_backtracking_steps: int = 5,
+) -> dict:
     if "q" not in nmpc.nlp[0].x_init.keys() or "qdot" not in nmpc.nlp[0].x_init.keys():
         return {"applied": False, "max_change": 0.0, "clipped_count": 0}
+    if dynamics_substeps < 1:
+        raise ValueError("qdot projection dynamics_substeps must be positive.")
+    if max_backtracking_steps < 0:
+        raise ValueError("qdot projection backtracking steps must be non-negative.")
 
     dt = nmpc.cycle_duration / nmpc.cycle_len
     q = np.asarray(nmpc.nlp[0].x_init["q"].init, dtype=float)
@@ -3323,15 +3342,60 @@ def project_qdot_initial_guess_from_q(nmpc, start_node: int = 0) -> dict:
         nmpc.nlp[0].x_bounds["qdot"], qdot.shape[1]
     )
     projected_qdot = np.minimum(np.maximum(qdot, lower), upper)
-    nmpc.nlp[0].x_init["qdot"].init[:, start_node:] = projected_qdot[:, start_node:]
-    nmpc._sync_acados_state_bounds()
     projected_slice = projected_qdot[:, start_node:]
     previous_slice = previous_qdot[:, start_node:]
     raw_slice = qdot[:, start_node:]
+    accepted_step = 1.0
+    scaled_defect_before = None
+    scaled_defect_after = None
+
+    if select_by_dynamics:
+        before = _full_dynamics_rollout_defect_details(
+            nmpc, n_substeps=dynamics_substeps
+        )
+
+        def maximum_scaled_defect(details):
+            return max(details.get("scaled_by_block", {}).values(), default=np.inf)
+
+        scaled_defect_before = maximum_scaled_defect(before)
+        if np.isfinite(scaled_defect_before):
+            accepted_step = 0.0
+            scaled_defect_after = scaled_defect_before
+            tolerance = max(1e-12, abs(scaled_defect_before) * 1e-12)
+            for step_index in range(max_backtracking_steps + 1):
+                step = 0.5**step_index
+                candidate = previous_slice + step * (projected_slice - previous_slice)
+                nmpc.nlp[0].x_init["qdot"].init[:, start_node:] = candidate
+                details = _full_dynamics_rollout_defect_details(
+                    nmpc, n_substeps=dynamics_substeps
+                )
+                scaled_defect = maximum_scaled_defect(details)
+                if (
+                    np.isfinite(scaled_defect)
+                    and scaled_defect < scaled_defect_after - tolerance
+                ):
+                    accepted_step = step
+                    scaled_defect_after = scaled_defect
+            accepted_qdot = previous_slice + accepted_step * (
+                projected_slice - previous_slice
+            )
+            nmpc.nlp[0].x_init["qdot"].init[:, start_node:] = accepted_qdot
+        else:
[…]
