### 2026-07-23 5831419
Restore Acados transfer mechanics before retries
