+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1058,6 +1058,32 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Project the ACADOS qdot initial guess from finite differences of q before solving.",
     )
     parser.add_argument(
+        "--acados-transfer-mechanical-restoration",
+        action="store_true",
+        help=(
+            "Jointly restore appended-cycle qdot and one pulse-width offset per "
+            "muscle using a reduced rollout sensitivity problem."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-mechanical-control-radius",
+        type=float,
+        default=5e-5,
+        help="Maximum pulse-width correction per muscle in the reduced restoration.",
+    )
+    parser.add_argument(
+        "--acados-transfer-mechanical-regularization",
+        type=float,
+        default=1e-2,
+        help="Tikhonov regularization applied to reduced restoration parameters.",
+    )
+    parser.add_argument(
+        "--acados-transfer-mechanical-substeps",
+        type=int,
+        default=5,
+        help="RK4 substeps used by reduced appended-cycle rollout sensitivities.",
+    )
+    parser.add_argument(
         "--transfer-full-dynamics-rollout",
         "--acados-transfer-full-dynamics-rollout",
         dest="acados_transfer_full_dynamics_rollout",
@@ -1383,6 +1409,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         type=int,
         default=1,
         help="Maximum same-weight restarts for a nearly feasible proximal stage.",
+    )
+    parser.add_argument(
+        "--acados-proximal-control-restart-feasibility-factor",
+        type=float,
+        default=1.0,
+        help=(
+            "Multiplier applied only to the feasibility threshold that permits "
+            "another proximal SQP call; final acceptance remains unchanged."
+        ),
     )
     parser.add_argument(
         "--acados-proximal-control-try-next-weight-on-failure",
@@ -2920,6 +2955,7 @@ def run_acados_proximal_control_continuation(
     weights: tuple[float, ...],
     convergence_tolerance: float,
     max_restarts: int = 1,
+    restart_feasibility_factor: float = 1.0,
     stage_iterations: int | None = 50,
     try_next_weight_on_failure: bool = False,
     echo: bool = True,
@@ -2975,7 +3011,8 @@ def run_acados_proximal_control_continuation(
                 and solution.status in (2, 4)
                 and attempt < max_restarts
                 and acados_homotopy_stage_is_restartable(
-                    {"residuals": restart_residuals}, convergence_tolerance
+                    {"residuals": restart_residuals},
+                    convergence_tolerance * restart_feasibility_factor,
                 )
             )
             residuals = diagnostics.get("residuals")
@@ -3397,6 +3434,231 @@ def project_qdot_initial_guess_from_q(
         "scaled_defect_after": scaled_defect_after,
         "max_change": float(np.max(np.abs(accepted_slice - previous_slice))),
         "clipped_count": int(np.count_nonzero(projected_slice != raw_slice)),
+    }
+
+
+def _appended_mechanical_rollout_residual(
+    nmpc,
+    start_node: int,
+    n_substeps: int,
+) -> np.ndarray:
+    """Roll out the appended cycle and compare its mechanical state trajectory."""
+
+    nlp = nmpc.nlp[0]
+    first_state_key = next(iter(nlp.x_init.keys()))
+    first_control_key = next(iter(nlp.u_init.keys()))
+    n_state_nodes = nlp.x_init[first_state_key].init.shape[1]
+    n_control_nodes = nlp.u_init[first_control_key].init.shape[1]
+    if n_state_nodes != n_control_nodes + 1:
+        raise ValueError(
[…]

[…]
