### 2026-07-24 d3e9bac
Scope Acados phase-one transfer projections
