+++ tests/shard1/test_periodic_pulse_width.py
@@ -1354,6 +1354,33 @@ def test_state_comparison_aligns_wheel_turn_representation():
     np.testing.assert_allclose(wheel["final_error"], 0.01)
 
 
+def test_full_dynamics_q_defect_scale_is_independent_of_unwrapped_turns():
+    class Variables(dict):
+        pass
+
+    nlp = SimpleNamespace(
+        states=Variables(
+            q=SimpleNamespace(index=[0, 1]),
+            qdot=SimpleNamespace(index=[2]),
+            F_Test=SimpleNamespace(index=[3]),
+        )
+    )
+    states = np.array(
+        [
+            [0.0, 100.0 * 2.0 * np.pi],
+            [1.0, 1.5],
+            [-2.0, -3.0],
+            [10.0, 20.0],
+        ]
+    )
+
+    scales = periodic_example._full_dynamics_defect_state_scales(nlp, states)
+
+    np.testing.assert_allclose(scales[[0, 1], :], 2.0 * np.pi)
+    assert scales[2, 0] == 3.0
+    assert scales[3, 0] == 20.0
+
+
 def test_endurance_metrics_report_fatigue_and_control_saturation():
     result = _benchmark_result([0, 0, 4])
 
@@ -2031,7 +2058,7 @@ def test_proximal_phase_one_backtracks_to_respect_state_change_limit(monkeypatch
         defect_weight=1.0,
         n_substeps=1,
         max_backtracking_steps=2,
-        max_state_change=5.0,
+        max_state_change_by_block={"q": 5.0},
     )
 
     assert summary["accepted"] is True
@@ -2040,7 +2067,94 @@ def test_proximal_phase_one_backtracks_to_respect_state_change_limit(monkeypatch
     assert summary["scaled_defect_after"] == 0.8
     assert summary["max_state_change"] == 4.5
     assert summary["candidate_max_state_change"] == 9.0
+    assert summary["state_change_by_block"]["q"] == 4.5
     np.testing.assert_allclose(nmpc.nlp[0].x_init["q"].init, [[0.0, 5.5]])
+
+
+def test_proximal_phase_one_preserves_retained_nodes_and_immutable_blocks(
+    monkeypatch,
+):
+    class StateVariables(dict):
+        shape = 2
+
+    class ControlVariables(dict):
+        shape = 1
+
+    defects = iter(
+        [
+            {
+                "scaled_by_block": {"q": 1.0, "fes": 1.0},
+                "absolute_by_block": {"q": 1.0, "fes": 1.0},
+            },
+            {
+                "scaled_by_block": {"q": 0.5, "fes": 1.0},
+                "absolute_by_block": {"q": 0.5, "fes": 1.0},
+            },
+        ]
+    )
+    q = np.array([[0.0, 1.0, 2.0]])
+    force = np.array([[10.0, 11.0, 12.0]])
+    nmpc = SimpleNamespace(
+        cycle_duration=2.0,
+        cycle_len=2,
+        nlp=[
+            SimpleNamespace(
+                x_init={
+                    "q": SimpleNamespace(init=q.copy()),
+                    "F_Test": SimpleNamespace(init=force.copy()),
+                },
+                u_init={"u": SimpleNamespace(init=np.zeros((1, 2)))},
+                x_bounds={
+                    "q": SimpleNamespace(
+                        min=np.full((1, 3), -100.0),
+                        max=np.full((1, 3), 100.0),
+                    ),
+                    "F_Test": SimpleNamespace(
+                        min=np.full((1, 3), -100.0),
+                        max=np.full((1, 3), 100.0),
+                    ),
+                },
+                states=StateVariables(
+                    q=SimpleNamespace(index=[0]),
[…]

[…]
