### 2026-07-25 6404980
Benchmark assisted cycling with optional NLP solvers
