+++ tests/shard1/test_solver_backends.py
@@ -1,0 +1,197 @@
+from types import SimpleNamespace
+
+import pytest
+
+from cocofest.optimization.solver_backends import (
+    SolverBackendUnavailable,
+    configure_nlp_solver,
+    nlp_solver_availability,
+)
+
+
+class _FakeSolver:
+    def __init__(self, **constructor_options):
+        self.constructor_options = constructor_options
+        self.calls = []
+
+    def __getattr__(self, name):
+        if not name.startswith("set_"):
+            raise AttributeError(name)
+
+        def record(*args):
+            self.calls.append((name, *args))
+
+        return record
+
+
+class _Factory:
+    def __init__(self):
+        self.instances = []
+
+    def __call__(self, **kwargs):
+        instance = _FakeSolver(**kwargs)
+        self.instances.append(instance)
+        return instance
+
+
+def _solver_namespace(*names):
+    return SimpleNamespace(**{name: _Factory() for name in names})
+
+
+def test_solver_availability_checks_bioptim_factory_and_casadi_plugin():
+    namespace = _solver_namespace("IPOPT", "MADNLP")
+    plugins = SimpleNamespace(has_nlpsol=lambda name: name == "madnlp")
+
+    assert nlp_solver_availability(
+        "madnlp", solver_namespace=namespace, casadi_module=plugins
+    ) == (True, None)
+    available, reason = nlp_solver_availability(
+        "alpaqa", solver_namespace=namespace, casadi_module=plugins
+    )
+    assert available is False
+    assert "Solver.ALPAQA" in reason
+
+
+def test_configure_madnlp_enables_validated_dual_warm_start():
+    namespace = _solver_namespace("MADNLP")
+
+    solver = configure_nlp_solver(
+        "madnlp",
+        max_iterations=321,
+        tolerance=2e-6,
+        madnlp_mu_init=1e-2,
+        madnlp_linear_solver="UmfpackSolver",
+        madnlp_max_wall_time=12.5,
+        madnlp_nlp_scaling=True,
+        madnlp_acceptable_tolerance=1e-4,
+        madnlp_acceptable_iterations=3,
+        solver_namespace=namespace,
+        check_availability=False,
+    )
+
+    assert ("set_convergence_tolerance", 2e-6) in solver.calls
+    assert ("set_constraint_tolerance", 2e-6) in solver.calls
+    assert ("set_maximum_iterations", 321) in solver.calls
+    assert ("set_print_level", "ERROR") in solver.calls
+    assert ("set_warm_start_options", 1e-2) in solver.calls
+    assert ("set_option_unsafe", "UmfpackSolver", "linear_solver") in solver.calls
+    assert ("set_option_unsafe", 12.5, "max_wall_time") in solver.calls
+    assert ("set_option_unsafe", True, "nlp_scaling") in solver.calls
+    assert ("set_option_unsafe", 1e-4, "acceptable_tol") in solver.calls
+    assert ("set_option_unsafe", 3, "acceptable_iter") in solver.calls
+    assert ("set_c_compile", False) in solver.calls
+
+
+def test_configure_alpaqa_sets_both_iteration_budgets_and_lbfgs():
+    namespace = _solver_namespace("ALPAQA")
+
+    solver = configure_nlp_solver(
+        "alpaqa",
+        max_iterations=500,
+        tolerance=1e-5,
+        alpaqa_alm_max_iterations=80,
+        alpaqa_lbfgs_memory=30,
+        alpaqa_max_wall_time=0.75,
+        alpaqa_initial_penalty=5.0,
+        alpaqa_initial_tolerance=1e-3,
+        alpaqa_penalty_update_factor=5.0,
[…]
