### 2026-07-26 25fc206
Fix CasADi discovery in Linux benchmark
