### 2026-07-26 f7f5168
Align Alpaqa build with CasADi C++ ABI
