### 2026-07-26 d06e1a9
Support headless benchmark plotting imports
