### 2026-07-27 c28b248
Support CasADi biorbd vector conversion
