+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -119,6 +119,7 @@ jobs:
             --first-node-wheel-q-slack 0 \
             --terminal-wheel-q-slack "${{ inputs.terminal_wheel_q_slack }}" \
             --compact-rho-output \
+            --print-traces \
             --output-json benchmark-seed-result/seed-check.json
           seed_path="$(find examples/fes_multibody/cycling/result/cache -name 'warmup_*.npz' -print -quit)"
           if [[ -z "$seed_path" ]]; then
@@ -257,7 +258,10 @@ jobs:
           if target == "alpaqa":
               assert hasattr(bioptim.Solver, "ALPAQA")
           PY
-          python -m pytest -q tests/shard1/test_solver_backends.py
+          python -m pytest -q \
+            tests/shard1/test_solver_backends.py \
+            tests/shard1/test_biorbd_backend_conversion.py \
+            tests/shard1/test_plot_backend.py
 
       - name: Run the ${{ matrix.solver }} assisted benchmark
         env:
@@ -316,6 +320,7 @@ jobs:
             --first-node-wheel-q-slack 0 \
             --terminal-wheel-q-slack "$BENCHMARK_Q_SLACK" \
             --compact-rho-output \
+            --print-traces \
             --output-json "benchmark-results/${BENCHMARK_SOLVER}.json" \
             "${solver_options[@]}" \
             2>&1 | tee "benchmark-results/${BENCHMARK_SOLVER}.log"
