### 2026-07-27 38d489b
Fix CasADi-backed inverse kinematics
