### 2026-07-27 97bceb0
Separate warmup and RHO iteration budgets
