### 2026-07-27 3c29708
Stabilize assisted solver benchmark seed
