### 2026-07-27 48975f8
Fix Linux solver benchmark validation
