+++ .github/scripts/install_casadi_alpaqa_linux.sh
@@ -46,6 +46,7 @@ cmake \
   -DWITH_PYTHON3=ON \
   -DWITH_IPOPT=ON \
   -DWITH_ALPAQA=ON \
+  -DWITH_BUILD_EIGEN3=ON \
   -DWITH_BUILD_ALPAQA=ON \
   -DBUILD_ALPAQA_GIT_REPO="$ALPAQA_REPOSITORY" \
   -DBUILD_ALPAQA_GIT_SHALLOW=OFF \
