+++ cocofest/optimization/solver_backends.py
@@ -179,9 +179,11 @@ def configure_nlp_solver(
 
     if solver_name == "madnlp":
         solver.set_print_level("ERROR" if print_level == 0 else print_level)
-        # This enables CasADi's lam_g0/lam_x0 inputs without asking Bioptim to
-        # replace Cocofest's already shifted primal initial guess.
-        solver.set_warm_start_options(madnlp_mu_init)
+        # The pinned CasADi MadNLP plugin accepts ``mu_init`` but rejects the
+        # ``dual_initialized`` option added by Bioptim's convenience method.
+        # Dual transfer is configured independently by Cocofest and is off by
+        # default, so keep this backend configuration primal-only.
+        solver.set_option_unsafe(madnlp_mu_init, "mu_init")
         for name, value in (
             ("linear_solver", madnlp_linear_solver),
             ("max_wall_time", madnlp_max_wall_time),
