### 2026-07-27 dba8cf6
Build Alpaqa before linking CasADi
