### 2026-07-27 d1d42be
Use supported MadNLP runtime options
