### 2026-07-27 738f862
Match pinned optional solver runtimes
