### 2026-07-27 955403b
Remove unsupported MadNLP barrier option
