### 2026-07-27 6586b77
Harden optional solver benchmark runtimes
