### 2026-07-27 f32b3bd
Relax Alpaqa runtime smoke tolerance
