### 2026-07-27 afb645d
Enable threaded CasADi for Alpaqa benchmarks
