### 2026-07-27 faf0a0e
Keep Alpaqa hot starts primal after limits
