### 2026-07-27 687af9e
Document final assisted solver benchmark
