+++ docs/message_kevin_benchmark.md
@@ -2,46 +2,118 @@
 
 Salut Kevin,
 
-J’ai adapté le benchmark des solveurs au nouveau problème de pédalage. Le
-couple externe est maintenant une **assistance de 0,20 N·m** : dans la
-convention du modèle, le couple signé vaut `-0.20 N.m` et fournit environ
-`+1.2566 W` à la cadence nominale. L’objectif reste uniquement la minimisation
-de la fatigue. Le problème commun à IPOPT, MadNLP et Alpaqa utilise une
-collocation Radau de degré 3, un cycle par RHO, 30 stimulations par cycle et une
-tolérance de `0.002 rad` sur l’angle terminal du pédalier.
+J’ai adapté et durci le benchmark du MHE de pédalage pour comparer IPOPT,
+MadNLP et Alpaqa sur exactement le même problème assisté. Le couple externe
+est une **assistance de 0,20 N·m** : comme le pédalier tourne avec
+`qdot < 0`, le couple généralisé constant vaut `-0.20 N.m`, soit environ
+`+1.2566 W` de puissance mécanique à la cadence nominale. L’objectif est
+uniquement la minimisation de la fatigue.
 
-Le workflow GitHub
-`.github/workflows/cycling_solver_benchmark_linux.yml` prépare d’abord un seed
-IPOPT assisté, vérifié physiquement, puis donne exactement le même artifact aux
-trois solveurs. MadNLP et Alpaqa font ensuite un hot-start IPOPT périodique
-certifié avant leur premier RHO. Le coût de cette préparation est publié
-séparément : il ne faut pas le confondre avec le temps chaud des résolutions
-suivantes.
+Le NLP commun utilise :
 
-Le run par défaut contient 30 RHO et utilise tous les cœurs visibles par
-`nproc`, avec les pools BLAS/OpenMP/Julia internes limités à un thread. Deux
-échecs consécutifs sont autorisés pour confirmer une non-convergence. Le
-préfixe scientifique validé s’arrête néanmoins au premier RHO échoué ou
-infaisable.
+- la dynamique `periodic_node` ;
+- une collocation directe Radau de degré 3 ;
+- un horizon MHE d’un cycle par RHO, avec 30 stimulations par cycle ;
+- le scaling complet des états et `0.002 rad` de tolérance sur l’angle
+  terminal du pédalier ;
+- 30 RHO demandés et deux échecs physiques consécutifs avant l’arrêt.
 
-Le rapport final compare :
+Le workflow prépare d’abord un seed IPOPT sur le **problème assisté cible** et
+le certifie physiquement. Le vieux seed à `+0.22 N.m` résistif n’est utilisé
+que comme trajectoire primale de continuation vers `-0.20 N.m`; il n’est
+jamais présenté comme une solution assistée. Les trois jobs téléchargent
+ensuite exactement le même artifact immuable.
 
-- la convergence globale, le nombre de cycles validés et le premier RHO
-  échoué ;
-- la durée mur-à-mur, la préparation, la somme des temps RHO, puis la médiane
-  et le P90 des temps chauds ;
-- pour chaque RHO, le statut générique et natif, les itérations, le temps
-  solveur, le temps mur et l’infaisabilité primale recalculée à partir des
-  bornes du NLP ;
-- les patrons de largeur d’impulsion aux **cycles/RHO 10 et 30** du même run,
-  en secondes et microsecondes, avec les bornes, la phase réelle du pédalier et
-  sa vitesse. Ce ne sont pas deux OCP contenant respectivement 10 et 30 cycles.
[…]
