### 2026-07-27 31fa37d
Screen MadNLP and Alpaqa solver options
