+++ .github/scripts/summarize_solver_option_screen.py
@@ -1,0 +1,292 @@
+#!/usr/bin/env python3
+"""Summarize the focused Alpaqa and MadNLP option screen."""
+
+from __future__ import annotations
+
+import argparse
+import csv
+import json
+import math
+import statistics
+from pathlib import Path
+
+DEFAULT_EXPECTED_VARIANTS = (
+    "alpaqa-panoc60-default",
+    "alpaqa-panoc20-default",
+    "alpaqa-panoc5-default",
+    "alpaqa-panoc20-auto-penalty",
+    "alpaqa-panoc20-penalty0p01",
+    "alpaqa-panoc20-penalty0p01-factor2",
+    "madnlp-tol1e-6-default",
+    "madnlp-tol1e-7-default",
+    "madnlp-tol1e-8-default",
+    "madnlp-tol1e-7-mumps",
+    "madnlp-tol1e-7-umfpack",
+)
+
+
+def _finite(value) -> float | None:
+    try:
+        value = float(value)
+    except (TypeError, ValueError):
+        return None
+    return value if math.isfinite(value) else None
+
+
+def _median(values) -> float | None:
+    finite = [item for value in values if (item := _finite(value)) is not None]
+    return statistics.median(finite) if finite else None
+
+
+def _sum_evaluations(stats: dict) -> int:
+    return sum(
+        int(value)
+        for key, value in stats.items()
+        if key.startswith("n_call_") and _finite(value) is not None
+    )
+
+
+def load_rows(paths: list[Path], common_feasibility_threshold: float) -> list[dict]:
+    rows = []
+    variants = set()
+    for path in paths:
+        variant = path.stem
+        if variant in variants:
+            raise ValueError(f"Duplicate screen variant: {variant}")
+        variants.add(variant)
+
+        payload = json.loads(path.read_text(encoding="utf-8"))
+        results = payload.get("results") or []
+        if len(results) != 1:
+            raise ValueError(f"{path} must contain exactly one solver result")
+        result = results[0]
+        solver = str(result.get("solver", "")).lower()
+        configuration = (payload.get("configurations") or {}).get(solver) or {}
+        windows = result.get("windows") or []
+        infeasibilities = [
+            _finite(
+                (window.get("feasibility") or {}).get("effective_primal_infeasibility")
+            )
+            for window in windows
+        ]
+        common_feasible = [
+            value is not None and value <= common_feasibility_threshold
+            for value in infeasibilities
+        ]
+        common_validated = [
+            feasible and bool(window.get("solver_converged"))
+            for window, feasible in zip(windows, common_feasible)
+        ]
+        common_prefix = 0
+        for validated in common_validated:
+            if not validated:
+                break
+            common_prefix += 1
+        stats = result.get("nlp_solver_stats") or []
+        evaluation_counts = [_sum_evaluations(item) for item in stats]
+        rows.append(
+            {
+                "variant": variant,
+                "solver": solver,
+                "configured_tolerance": _finite(configuration.get("nlp_tolerance")),
+                "linear_solver": configuration.get("madnlp_linear_solver"),
+                "panoc_max_wall_time_s": _finite(
+                    configuration.get("alpaqa_panoc_max_wall_time")
+                ),
+                "initial_penalty": _finite(configuration.get("alpaqa_initial_penalty")),
+                "penalty_update_factor": _finite(
+                    configuration.get("alpaqa_penalty_update_factor")
+                ),
+                "success": bool(result.get("success")),
+                "validated_cycles_internal": int(result.get("validated_cycles") or 0),
+                "attempted_windows": int(result.get("attempted_windows") or 0),
+                "common_feasible_windows": sum(common_feasible),
+                "common_validated_prefix": common_prefix,
[…]
