### 2026-07-27 ec2eba0
Apply screened NLP solver settings
