### 2026-07-27 5df58ec
Use tolerant float assertion in benchmark test
