### 2026-07-27 30fec8c
Document final assisted solver benchmark
