### 2026-07-27 b5371d2
Retire Alpaqa from endurance benchmark
