### 2026-07-27 7d5b0df
Document 100-cycle solver benchmark
