### 2026-07-27 6c37027
Add Fatrop to cycling endurance benchmark
