+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -21,7 +21,17 @@ REPO_ROOT = Path(__file__).resolve().parents[3]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
-from bioptim import MultiCyclicCycleSolutions, Node, OdeSolver, SolutionMerge, Solver
+from bioptim import (
+    MultiCyclicCycleSolutions,
+    Node,
+    OdeSolver,
+    SolutionMerge,
+    Solver,
+)
+try:
+    from bioptim import OrderingStrategy
+except ImportError:  # Bioptim releases before the Fatrop time-major interface
+    OrderingStrategy = None
 from bioptim.optimization.receding_horizon_optimization import (
     RecedingHorizonOptimization,
 )
@@ -1689,6 +1699,46 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--fatrop-dual-warm-start-mode",
+        choices=("off", "constraints", "bounds", "all"),
+        default="off",
+        help=(
+            "Reuse Fatrop multipliers after shifting the primal MHE guess. "
+            "Disabled by default until time-major multiplier blocks are shifted "
+            "and independently validated."
+        ),
+    )
+    parser.add_argument(
+        "--max-fatrop-iterations",
+        type=int,
+        default=1000,
+        help="Maximum number of Fatrop iterations per MHE window (native maximum: 1000).",
+    )
+    parser.add_argument(
+        "--fatrop-structure-detection",
+        choices=("auto", "manual"),
+        default="auto",
+        help=(
+            "CasADi Fatrop OCP-structure detection. Automatic detection is the "
+            "portable benchmark default."
+        ),
+    )
+    parser.add_argument(
+        "--fatrop-bound-tightening-factor",
+        type=float,
+        default=1e-8,
+        help=(
+            "Compensatory tightening of interval bounds before the Fatrop "
+            "call. The default offsets its native relative relaxation."
+        ),
+    )
+    parser.add_argument(
+        "--fatrop-c-compile",
+        action="store_true",
+        help="Experimentally generate and compile the CasADi NLP used by Fatrop.",
+    )
+    parser.add_argument("--fatrop-print-level", type=int, default=0)
+    parser.add_argument(
         "--madnlp-dual-warm-start-mode",
         choices=("off", "constraints", "bounds", "all"),
         default="off",
@@ -1708,7 +1758,7 @@ def build_argument_parser() -> argparse.ArgumentParser:
         "--nlp-tolerance",
         type=float,
         default=1e-6,
-        help="Common convergence and constraint tolerance for IPOPT, MadNLP, and Alpaqa.",
+        help="Common convergence and constraint tolerance for IPOPT, Fatrop, MadNLP, and Alpaqa.",
     )
     parser.add_argument(
         "--primal-feasibility-threshold",
@@ -1716,7 +1766,7 @@ def build_argument_parser() -> argparse.ArgumentParser:
         default=None,
         help=(
             "Optional common absolute primal-feasibility threshold used to "
-            "validate IPOPT, MadNLP, and Alpaqa windows independently of each "
+            "validate IPOPT, Fatrop, MadNLP, and Alpaqa windows independently of each "
             "solver's internal tolerance. The legacy default is ten times "
             "--nlp-tolerance."
         ),
@@ -3292,6 +3342,16 @@ def configure_cycle_nlp_solver(args: argparse.Namespace):
             tolerance=args.nlp_tolerance,
             madnlp_c_compile=args.madnlp_c_compile,
             madnlp_linear_solver=args.madnlp_linear_solver,
+        )
+    if args.solver == "fatrop":
[…]
