+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -45,7 +45,7 @@ except ImportError:
     )
 
 EXAMPLE_DIR = Path(__file__).resolve().parent
-BENCHMARK_SOLVERS = ("ipopt", "acados", "madnlp", "alpaqa")
+BENCHMARK_SOLVERS = ("ipopt", "acados", "fatrop", "madnlp")
 BENCHMARK_STIMULATION_PATTERN_CYCLES = (10, 30, 100)
 BENCHMARK_CONFIGURATION_FIELDS = (
     "solver",
@@ -57,6 +57,7 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "collocation_degree",
     "collocation_method",
     "use_sx",
+    "nlp_ordering_strategy",
     "state_scaling",
     "pulse_width_scaling",
     "pulse_width_active_set",
@@ -98,6 +99,12 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "max_ipopt_iterations",
     "standard_warmup_max_iterations",
     "disable_historical_ipopt_initial_guess",
+    "fatrop_dual_warm_start_mode",
+    "fatrop_c_compile",
+    "fatrop_structure_detection",
+    "fatrop_bound_tightening_factor",
+    "fatrop_print_level",
+    "max_fatrop_iterations",
     "madnlp_dual_warm_start_mode",
     "madnlp_c_compile",
     "madnlp_linear_solver",
@@ -1319,18 +1326,29 @@ def _nlp_solver_config(
     alpaqa_maximum_penalty: float | None = None,
     alpaqa_panoc_max_wall_time: float | None = None,
     alpaqa_max_no_progress: int | None = None,
+    fatrop_c_compile: bool = False,
+    fatrop_structure_detection: str = "auto",
+    fatrop_bound_tightening_factor: float = 1e-8,
+    fatrop_print_level: int = 0,
     periodic_ipopt_hot_start: bool = False,
 ) -> argparse.Namespace:
     """Clone the IPOPT transcription so only the nonlinear solver changes."""
 
-    if solver_name not in {"madnlp", "alpaqa"}:
-        raise ValueError("The optional NLP solver must be 'madnlp' or 'alpaqa'.")
+    if solver_name not in {"fatrop", "madnlp", "alpaqa"}:
+        raise ValueError(
+            "The optional NLP solver must be 'fatrop', 'madnlp' or 'alpaqa'."
+        )
     args = argparse.Namespace(**vars(reference_args))
     args.solver = solver_name
     args.nlp_tolerance = tolerance
     setattr(args, f"max_{solver_name}_iterations", max_iterations)
     setattr(args, f"{solver_name}_dual_warm_start_mode", dual_warm_start_mode)
     args.nlp_periodic_ipopt_hot_start = periodic_ipopt_hot_start
+    if solver_name == "fatrop":
+        args.fatrop_c_compile = fatrop_c_compile
+        args.fatrop_structure_detection = fatrop_structure_detection
+        args.fatrop_bound_tightening_factor = fatrop_bound_tightening_factor
+        args.fatrop_print_level = fatrop_print_level
     if solver_name == "madnlp":
         args.madnlp_linear_solver = madnlp_linear_solver
     if solver_name == "alpaqa":
@@ -1624,7 +1642,11 @@ def _run_benchmark_case(
     try:
         uses_c_codegen = (
             solver_name == "ipopt" and getattr(args, "ipopt_c_compile", False)
-        ) or (solver_name == "madnlp" and getattr(args, "madnlp_c_compile", False))
+        ) or (
+            solver_name == "fatrop" and getattr(args, "fatrop_c_compile", False)
+        ) or (
[…]
