### 2026-07-27 d45bb5e
Document Fatrop 100-cycle benchmark
