+++ docs/message_kevin_benchmark.md
@@ -63,54 +63,58 @@ de ce runtime, tandis qu’UMFPACK a été environ 64 % plus lent.
 
 Le benchmark Linux 100 RHO est ici :
 
-<https://github.com/mickaelbegon/cocofest/actions/runs/30304318862>
+<https://github.com/mickaelbegon/cocofest/actions/runs/30309452077>
 
-Les deux jobs ont utilisé les quatre cœurs exposés par le runner pour
+Les trois jobs ont utilisé les quatre cœurs exposés par le runner pour
 l’évaluation CasADi/Bioptim. Les pools OpenMP, BLAS, MKL, NumExpr et Julia
 imbriqués restent à un thread pour éviter la sur-souscription.
 
-## Résultats à 100 RHO déjà établis
+## Résultats à 100 RHO
 
 | Solveur | RHO résolus | Préfixe strict | Préparation | Somme des RHO | Médiane chaude | P90 chaud | Mur-à-mur |
 |---|---:|---:|---:|---:|---:|---:|---:|
-| IPOPT-MUMPS (`tol=1e-6`) | 100/100 | 100/100 | 23.03 s | 818.74 s | 8.241 s | 11.130 s | 866.22 s |
-| MadNLP-MUMPS (`tol=1e-8`) | 99/100 | 85/100 | 45.72 s | 815.33 s | 5.623 s | 7.745 s | 881.48 s |
+| IPOPT-MUMPS (`tol=1e-6`) | 100/100 | 100/100 | 23.82 s | 830.20 s | 8.311 s | 11.190 s | 878.90 s |
+| Fatrop (`tol=1e-6`) | 100/100 | 100/100 | 42.96 s | 1148.17 s | 11.979 s | 14.016 s | 1209.02 s |
+| MadNLP-MUMPS (`tol=1e-8`) | 100/100 | 100/100 | 54.90 s | 799.46 s | 6.252 s | 9.278 s | 877.35 s |
 
-Il faut distinguer les deux colonnes MadNLP. Le **préfixe strict** s’arrête au
-premier échec afin de ne pas considérer comme exécutée une trajectoire passant
-par un RHO non certifié. MadNLP n’a toutefois pas cessé de résoudre : son
-RHO 86 atteint `SOLVER_RET_LIMITED` après 2000 itérations et 187.12 s, avec
-`4.85e-2` d’infaisabilité, puis les RHO 87 à 100 convergent et sont
-physiquement faisables. Il y a donc 99 solutions certifiées sur 100 et jamais
-deux échecs consécutifs.
+Les trois solveurs convergent et passent l’audit physique sur les 100 fenêtres.
+Les infaisabilités maximales sont `9.64e-7` pour IPOPT, `9.55e-7` pour
+Fatrop et `1.00e-6` pour MadNLP, donc nettement sous le seuil commun `1e-5`.
+Fatrop ne viole aucune borne physique après compensation de sa relaxation
+interne.
 
-IPOPT résout le même RHO 86 en 76 itérations, 8.59 s et avec une infaisabilité
-de `7.34e-7`. L’échec MadNLP ressemble donc davantage à un accident de
-robustesse locale ou de hot start qu’à une impossibilité physique créée par la
-fatigue. Comme les 14 fenêtres suivantes récupèrent, on ne peut pas non plus
-présenter le cycle 86 comme le point de défaillance physiologique.
+Fatrop est régulier dans ce run, sans échec ni outlier extrême, mais il est ici
+nettement le plus lent : +44 % sur la médiane chaude
+et +38 % au mur-à-mur par rapport à IPOPT. Son coût augmente avec l’horizon :
+77 itérations et 7.60 s de médiane sur les RHO 1–10, contre 122 itérations et
[…]
