### 2026-07-27 3d51dc6
Use absolute crank targets and MadNLP PARDISO
