+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1331,7 +1331,11 @@ def build_argument_parser() -> argparse.ArgumentParser:
         "--acados-terminal-wheel-q-slack",
         type=float,
         default=0.002,
-        help="Terminal crank-angle slack in rad; independent from the first-node transfer slack.",
+        help=(
+            "Terminal crank-angle slack in rad around the absolute initial-angle "
+            "plus signed-cycle-count reference; independent from the first-node "
+            "transfer slack."
+        ),
     )
     parser.add_argument(
         "--terminal-wheel-q-slack",
@@ -1339,8 +1343,9 @@ def build_argument_parser() -> argparse.ArgumentParser:
         default=None,
         help=(
             "Backend-independent terminal crank-angle slack in rad. When set, "
-            "it overrides --acados-terminal-wheel-q-slack for IPOPT, MadNLP, "
-            "Alpaqa, and ACADOS."
+            "it overrides --acados-terminal-wheel-q-slack for IPOPT, Fatrop, "
+            "MadNLP, Alpaqa, and ACADOS. The center is an absolute cycle-count "
+            "reference, so this tolerance cannot accumulate between RHO windows."
         ),
     )
     parser.add_argument(
@@ -1784,13 +1789,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
             "mumps",
             "umfpack",
             "lapack_cpu",
+            "pardiso_mkl",
             "cudss",
             "lapack_gpu",
             "cucholesky",
         ),
         help=(
-            "MadNLP C-runtime linear solver. The default is mumps; GPU choices "
-            "require a compatible MadNLP runtime and runner."
+            "MadNLP C-runtime linear solver. The default is mumps; pardiso_mkl "
+            "requires the x86-64 libMad MKL runtime, and GPU choices require a "
+            "compatible runtime and runner."
         ),
     )
     parser.add_argument(
@@ -2741,7 +2748,7 @@ def _warmup_cache_path(
 def _periodic_ipopt_refinement_cache_path(
     args: argparse.Namespace,
     model_path: Path,
-    cache_version: int = 3,
+    cache_version: int = 4,
 ) -> Path:
     repository_root = Path(__file__).resolve().parents[3]
     payload = {
@@ -2790,6 +2797,8 @@ def _periodic_ipopt_refinement_cache_path(
         payload["wheel_cycle_boundary_slack"] = (
             args.acados_terminal_wheel_q_slack
         )
+    if cache_version >= 4:
+        payload["terminal_wheel_q_reference_mode"] = "absolute_initial"
     return _cache_root() / f"periodic_ipopt_{_short_hash(payload)}.npz"
 
 
@@ -2811,7 +2820,7 @@ def _acados_seed_cache_path(args: argparse.Namespace, model_path: Path) -> Path 
     )
     payload = {
         "kind": "acados_seed",
-        "cache_version": 2,
+        "cache_version": 3,
         "model_formulation": args.model_formulation,
         "cycles_per_window": args.cycles_per_window,
         "stimulations_per_cycle": args.stimulations_per_cycle,
@@ -2824,6 +2833,7 @@ def _acados_seed_cache_path(args: argparse.Namespace, model_path: Path) -> Path 
         "torque_application": args.torque_application,
         "state_scaling": args.state_scaling,
         "pulse_width_scaling": args.pulse_width_scaling,
+        "terminal_wheel_q_reference_mode": "absolute_initial",
         "integrator_type": args.acados_integrator_type,
         "sim_stages": args.acados_sim_stages,
         "sim_steps": args.acados_sim_steps,
[…]
