### 2026-07-28 6f8f162
Fix absolute crank cycle advance
