### 2026-07-28 413d5bc
Run MadNLP benchmark on GCC 13
