### 2026-07-28 7b4acad
Build CasADi against the libMad ABI
