### 2026-07-28 704570c
Verify pinned CasADi MadNLP revision
