### 2026-07-28 495db09
Persist certified MadNLP runtime cache
