### 2026-07-28 cc580d8
Use valid MadNLP quiet log level
