+++ docs/message_kevin_benchmark.md
@@ -169,7 +169,9 @@ Attention : le premier essai Linux a révélé que le wheel officiel CasADi 3.7.
 charge encore l’ancienne ABI `libmadnlp_c.so`, incompatible avec la nouvelle
 ABI `libMad.so`. Le workflow construit donc maintenant la révision CasADi
 3.7.2 post-release épinglée qui intègre cette nouvelle ABI ; un simple lien
-symbolique aurait été incorrect. Cette voie est limitée à Linux x86-64 et
+symbolique aurait été incorrect. Ce runtime embarque MadNLP 0.9.2, où le
+niveau silencieux est l’énumération `ERROR=6` (`0` est invalide) ; ce réglage
+est maintenant appliqué au smoke test et à l’OCP. Cette voie est limitée à Linux x86-64 et
 n’est pas disponible nativement sur les Mac Apple Silicon. Elle n’a pas encore
 de résultat sur l’OCP complet : le prochain benchmark doit donc être
 interprété comme la première comparaison MadNLP-PARDISO face à IPOPT-MUMPS et
