+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -580,7 +580,8 @@ jobs:
               }
           elif target == "madnlp":
               plugin_options["madnlp"] = {
-                  "print_level": 0,
+                  # MadNLP 0.9.2 LogLevels: TRACE=1, ..., ERROR=6.
+                  "print_level": 6,
                   "max_iter": 50,
                   "tol": 1e-8,
                   "linear_solver": "PardisoMKLSolver",
