### 2026-07-28 e23a6eb
Exercise Cocofest MadNLP adapter in CI
