### 2026-07-28 b5e1f6d
Fix absolute crank angle benchmark audit
