+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -5280,6 +5280,52 @@ def _window_feasibility_tolerance(args: argparse.Namespace) -> float | None:
     return args.nlp_tolerance if threshold is None else float(threshold) / 10.0
 
 
+def _wheel_cycle_diagnostic_tolerances(
+    args: argparse.Namespace,
+    wheel_q_scaling: float = 1.0,
+) -> tuple[float, float]:
+    """Return progress and fixed-reference tolerances for crank-angle audits."""
+
+    if not np.isfinite(wheel_q_scaling) or wheel_q_scaling <= 0.0:
+        raise ValueError("wheel_q_scaling must be finite and strictly positive.")
+    solver_tolerance = (
+        args.acados_tolerance
+        if args.solver == "acados" and args.acados_tolerance is not None
+        else args.nlp_tolerance
+    )
+    scaled_feasibility_threshold = getattr(
+        args, "primal_feasibility_threshold", None
+    )
+    if args.solver == "acados" or scaled_feasibility_threshold is None:
+        scaled_feasibility_threshold = 10.0 * solver_tolerance
+    # The independent feasibility audit operates on the scaled decision
+    # vector. Convert its accepted q violation back to physical radians.
+    numerical_margin = float(scaled_feasibility_threshold) * wheel_q_scaling
+    terminal_slack = args.acados_terminal_wheel_q_slack
+    first_node_slack = args.acados_wheel_q_slack
+    progress_tolerance = (
+        max(2.0 * terminal_slack, first_node_slack + terminal_slack)
+        + 2.0 * numerical_margin
+    )
+    absolute_tolerance = max(first_node_slack, terminal_slack) + numerical_margin
+    return progress_tolerance, absolute_tolerance
+
+
+def _wheel_q_state_scaling(nmpc) -> float:
+    """Return the wheel-angle decision scaling used by the NLP."""
+
+    try:
+        scaling = np.asarray(
+            nmpc.nlp[0].x_scaling["q"].scaling, dtype=float
+        ).reshape((-1,))
+        wheel_q_scaling = float(scaling[2])
+    except (AttributeError, IndexError, KeyError, TypeError):
+        return 1.0
+    if not np.isfinite(wheel_q_scaling) or wheel_q_scaling <= 0.0:
+        raise RuntimeError("The crank-position state scaling must be positive.")
+    return wheel_q_scaling
+
+
 def _native_solver_status(nmpc) -> str | None:
     """Extract the native status retained by the optional Bioptim interfaces."""
 
@@ -5326,6 +5372,8 @@ def summarize_windows(
     feasibility_tolerance: float | None = None,
     expected_cycle_shift: float = -2.0 * np.pi,
     cycle_progress_tolerance: float = 0.5,
+    absolute_cycle_reference: float | None = None,
+    absolute_cycle_tolerance: float | None = None,
 ) -> None:
     def _fmt(value) -> str:
         return "None" if value is None else f"{value:.6f}"
@@ -5346,6 +5394,8 @@ def summarize_windows(
         requested_windows=requested_windows,
         expected_cycle_shift=expected_cycle_shift,
         cycle_progress_tolerance=cycle_progress_tolerance,
+        absolute_cycle_reference=absolute_cycle_reference,
+        absolute_cycle_tolerance=absolute_cycle_tolerance,
     )
     window_feasibility = [
         _solution_feasibility_summary(solution, feasibility_tolerance)
@@ -5411,6 +5461,8 @@ def build_window_summary(
     feasibility_tolerance: float | None = None,
     expected_cycle_shift: float = -2.0 * np.pi,
     cycle_progress_tolerance: float = 0.5,
+    absolute_cycle_reference: float | None = None,
+    absolute_cycle_tolerance: float | None = None,
 ) -> dict:
     (
         merged_solution,
[…]
