### 2026-07-28 502a81e
Docstring
