+++ cocofest/models/veltink1992/__init__.py
@@ -1,0 +1,6 @@
+"""
+Veltink1992 model: nonlinear joint angle control for artificially stimulated muscle.
+
+Veltink, P. H., Chizeck, H. J., Crago, P. E., & El-Bialy, A. (1992). Nonlinear joint angle control
+for artificially stimulated muscle. IEEE Transactions on Biomedical Engineering, 39(4), 368-380.
+"""
