+++ cocofest/models/marion2009/__init__.py
@@ -1,0 +1,6 @@
+"""
+Marion2009 model: effect of muscle length on fatigue during electrical stimulation.
+
+Marion, M. S., Wexler, A. S., Hull, M. L., & Binder-Macleod, S. A. (2009). Predicting the effect of
+muscle length on fatigue during electrical stimulation. Muscle & Nerve, 40(4), 573-581.
+"""
