+++ cocofest/misc/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package version and small general-purpose utilities not tied to FES models or optimal control.
+"""
