+++ cocofest/optimization/__init__.py
@@ -1,0 +1,3 @@
+"""
+Optimal control problem builders for FES optimization, identification and moving horizon estimation.
+"""
