+++ docs/sphinx/requirements.txt
@@ -1,0 +1,4 @@
+sphinx
+myst-parser
+furo
+sphinxcontrib-mermaid
