+++ cocofest/models/marion2013/marion2013_modified_with_fatigue.py
@@ -6,6 +6,8 @@ electrical stimulation pulse trains as a function of pulse duration. Journal of 
 and Rehabilitation, 10, 1-16.
 """
 
+from __future__ import annotations
+
 from typing import Callable
 from casadi import MX, vertcat
 import numpy as np
