+++ .readthedocs.yaml
@@ -11,7 +11,6 @@ conda:
 
 python:
   install:
-    - requirements: docs/sphinx/requirements.txt
     - method: pip
       path: .
 
