### 2026-07-28 88543b9
Align crank audit with warmup cycle
