### 2026-07-28 360df91
Document absolute-angle solver benchmark
