+++ docs/message_kevin_benchmark.md
@@ -3,133 +3,141 @@
 Salut Kevin,
 
 J’ai adapté et durci le benchmark du MHE de pédalage pour comparer IPOPT,
-Fatrop et MadNLP sur le même problème assisté. Alpaqa est désormais retiré de la matrice
-d’endurance : le meilleur réglage testé n’a validé aucun RHO, a atteint deux
-limites de 600 s et a fini le second RHO à `4.57e-2` d’infaisabilité. Son
-intégration et ses résultats historiques restent documentés pour la
-reproductibilité, mais il ne consomme plus de temps dans le benchmark normal.
+Fatrop et MadNLP sur le même problème assisté. L’objectif du NLP est uniquement
+la minimisation de la fatigue.
 
 Le couple externe est une **assistance de 0,20 N·m**. Comme le pédalier tourne
 avec `qdot < 0`, le couple généralisé constant vaut `-0.20 N.m`, soit environ
-`+1.26 W` de puissance mécanique à la cadence nominale. L’objectif du NLP est
-uniquement la minimisation de la fatigue.
+`+1.26 W` de puissance mécanique à la cadence nominale.
 
 Le problème commun utilise :
 
-- la dynamique `periodic_node` ;
+- la dynamique musculaire `periodic_node` ;
 - une collocation directe Radau de degré 3 ;
 - un horizon MHE d’un cycle par RHO, avec 30 stimulations par cycle ;
-- le scaling complet des états pour IPOPT et MadNLP ;
-- une tolérance de `0.002 rad` sur la progression angulaire terminale ;
-- 100 RHO demandés et un arrêt après deux échecs consécutifs.
+- 100 RHO demandés et un arrêt après deux échecs consécutifs ;
+- une cible angulaire absolue
+  `q_initial + k*(-2*pi)`, avec une bande terminale de `0.002 rad` ;
+- le scaling complet des états pour IPOPT et MadNLP.
+
+Fatrop utilise actuellement un ordre temporel des variables et aucun scaling
+des états. Avec le scaling complet, sa détection structurée refuse les gaps de
+collocation. Le rapport marque explicitement cet écart : ses temps ne
+constituent donc pas encore une comparaison purement « backend contre
+backend ».
 
 Le workflow construit d’abord un seed IPOPT sur le problème assisté cible et
-le certifie physiquement. IPOPT, Fatrop et MadNLP téléchargent ensuite le même artifact
-immuable et utilisent le même commit Bioptim. Le vieux seed résistif n’est
-qu’une trajectoire de continuation et n’est jamais présenté comme solution du
-problème assisté.
+le certifie physiquement. Les trois solveurs téléchargent ensuite exactement
+le même artefact et le même commit Bioptim. Le vieux seed résistif n’est
+utilisé que comme point de départ d’une continuation vers le problème
+assisté ; il n’est jamais présenté comme une solution du nouveau problème.
 
-MadNLP reçoit un hot start primal complet : états et contrôles sont décalés
-d’un cycle, extrapolés puis projetés dans leurs bornes, avec continuité des
-états de fatigue. Un raffinement IPOPT périodique certifié prépare son premier
-RHO. La réutilisation des multiplicateurs MadNLP reste désactivée, car le
-runtime épinglé ne la supporte pas proprement. IPOPT réutilise les
-multiplicateurs de bornes.
+IPOPT réutilise les multiplicateurs de bornes. Fatrop et MadNLP reçoivent un
[…]
