### 2026-07-28 f68a3e3
Label cycling benchmark with requested horizon
