### 2026-07-28 1d4f50a
Add reduced cycling solver benchmark
