+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -52,6 +52,7 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "objective",
     "objective_shape",
     "model_formulation",
+    "mechanical_formulation",
     "torque_application",
     "ode_solver",
     "collocation_degree",
@@ -552,14 +553,23 @@ def _external_crank_power_metrics(result: dict, cycle_count: int) -> dict:
     metrics["torque_nm"] = float(torque)
 
     limited = _truncate_result_to_cycles(result, cycle_count)
-    qdot = limited.get("state_traces", {}).get("qdot")
+    state_traces = limited.get("state_traces", {})
+    qdot = state_traces.get("qdot")
+    velocity_index = 2
+    if qdot is None:
+        qdot = state_traces.get("omega")
+        velocity_index = 0
     if qdot is None:
         return metrics
     qdot = np.asarray(qdot, dtype=float)
-    if qdot.ndim != 2 or qdot.shape[0] < 3 or qdot.shape[1] == 0:
+    if (
+        qdot.ndim != 2
+        or qdot.shape[0] <= velocity_index
+        or qdot.shape[1] == 0
+    ):
         return metrics
 
-    power = float(torque) * qdot[2, :]
+    power = float(torque) * qdot[velocity_index, :]
     if not np.all(np.isfinite(power)):
         return metrics
     if power.size == 1:
@@ -1808,11 +1818,20 @@ def _stimulation_pattern_snapshot(result: dict, cycle: int) -> dict:
         crank_direction * (cycle_wheel_angle - wheel_angle[start]),
         2.0 * np.pi,
     ).tolist()
-    qdot = limited.get("state_traces", {}).get("qdot")
+    state_traces = limited.get("state_traces", {})
+    qdot = state_traces.get("qdot")
+    velocity_index = 2
+    if qdot is None:
+        qdot = state_traces.get("omega")
+        velocity_index = 0
     if qdot is not None:
         qdot = np.asarray(qdot, dtype=float)
-        if qdot.ndim == 2 and qdot.shape[0] >= 3 and qdot.shape[1] >= stop:
-            crank_velocity = qdot[2, start:stop]
+        if (
+            qdot.ndim == 2
+            and qdot.shape[0] > velocity_index
+            and qdot.shape[1] >= stop
+        ):
+            crank_velocity = qdot[velocity_index, start:stop]
             if np.all(np.isfinite(crank_velocity)):
                 snapshot["crank_velocity_rad_s"] = crank_velocity.tolist()
 
@@ -1953,6 +1972,7 @@ def solver_overview_rows(results: dict[str, dict]) -> list[dict]:
                 "solver_success": bool(result.get("solver_success")),
                 "physical_success": bool(result.get("physical_success")),
                 "status": None if status is None else int(status),
+                "validated_windows": performance["successful_prefix_windows"],
                 "validated_cycles": performance["validated_cycles"],
                 "attempted_windows": result.get("attempted_windows"),
                 "successful_windows": result.get("successful_windows"),
@@ -1975,6 +1995,9 @@ def solver_overview_rows(results: dict[str, dict]) -> list[dict]:
                     result.get("end_to_end_wall_time_s")
                 ),
                 "initial_guess_preparation_time_s": preparation_time,
+                "reduced_profile_build_time_s": _finite_float(
+                    result.get("reduced_profile_build_time_s")
+                ),
                 "attempted_rho_wall_time_sum_s": attempted_rho_wall_time,
                 "unattributed_wall_time_s": unattributed_wall_time,
                 "validated_solver_time_s": performance["successful_solver_time_s"],
@@ -2140,6 +2163,8 @@ def main(
     cycles_per_window: int = 1,
     stimulations_per_cycle: int = 30,
     n_windows: int = 2,
[…]
