### 2026-07-28 71e69a5
Unify cycling solver benchmark initialization
