+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -83,6 +83,8 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "standard_warmup_seed",
     "standard_warmup_seed_continuation",
     "legacy_standard_warmup_seed_signed_torque",
+    "common_initial_solution",
+    "common_initial_solution_output",
     "ipopt_hsl_library",
     "ipopt_c_compile",
     "ipopt_print_level",
@@ -562,11 +564,7 @@ def _external_crank_power_metrics(result: dict, cycle_count: int) -> dict:
     if qdot is None:
         return metrics
     qdot = np.asarray(qdot, dtype=float)
-    if (
-        qdot.ndim != 2
-        or qdot.shape[0] <= velocity_index
-        or qdot.shape[1] == 0
-    ):
+    if qdot.ndim != 2 or qdot.shape[0] <= velocity_index or qdot.shape[1] == 0:
         return metrics
 
     power = float(torque) * qdot[velocity_index, :]
@@ -581,9 +579,7 @@ def _external_crank_power_metrics(result: dict, cycle_count: int) -> dict:
     role = (
         "driving"
         if mean_power > power_tolerance
-        else "resistive"
-        if mean_power < -power_tolerance
-        else "neutral"
+        else "resistive" if mean_power < -power_tolerance else "neutral"
     )
     metrics.update(
         mean_power_w=mean_power,
@@ -1658,11 +1654,9 @@ def _run_benchmark_case(
     start = perf_counter()
     try:
         uses_c_codegen = (
-            solver_name == "ipopt" and getattr(args, "ipopt_c_compile", False)
-        ) or (
-            solver_name == "fatrop" and getattr(args, "fatrop_c_compile", False)
-        ) or (
-            solver_name == "madnlp" and getattr(args, "madnlp_c_compile", False)
+            (solver_name == "ipopt" and getattr(args, "ipopt_c_compile", False))
+            or (solver_name == "fatrop" and getattr(args, "fatrop_c_compile", False))
+            or (solver_name == "madnlp" and getattr(args, "madnlp_c_compile", False))
         )
         if uses_c_codegen:
             previous_cwd = Path.cwd()
@@ -1715,9 +1709,7 @@ def _benchmark_window_rows(result: dict) -> list[dict]:
             if index < len(statuses) and statuses[index] is not None
             else None
         )
-        window_feasibility = (
-            feasibility[index] if index < len(feasibility) else None
-        )
+        window_feasibility = feasibility[index] if index < len(feasibility) else None
         solver_stats = stats_by_window.get(index) or {}
         madnlp_stats = solver_stats.get("madnlp") or {}
         native_status = next(
@@ -1826,11 +1818,7 @@ def _stimulation_pattern_snapshot(result: dict, cycle: int) -> dict:
         velocity_index = 0
     if qdot is not None:
         qdot = np.asarray(qdot, dtype=float)
-        if (
-            qdot.ndim == 2
-            and qdot.shape[0] > velocity_index
-            and qdot.shape[1] >= stop
-        ):
+        if qdot.ndim == 2 and qdot.shape[0] > velocity_index and qdot.shape[1] >= stop:
             crank_velocity = qdot[velocity_index, start:stop]
             if np.all(np.isfinite(crank_velocity)):
                 snapshot["crank_velocity_rad_s"] = crank_velocity.tolist()
@@ -1853,12 +1841,8 @@ def _stimulation_pattern_snapshot(result: dict, cycle: int) -> dict:
         if lower is not None and upper is not None and upper > lower:
             normalized = (cycle_values - lower) / (upper - lower)
             bound_tolerance = max(1e-12, (upper - lower) * 1e-3)
-            lower_fraction = float(
-                np.mean(cycle_values <= lower + bound_tolerance)
-            )
-            upper_fraction = float(
[…]
