+++ tests/shard1/test_periodic_pulse_width.py
@@ -503,6 +503,34 @@ def test_acados_example_defaults_to_the_assisted_periodic_profile():
     assert ipopt_args.acados_control_homotopy_keep_final_radius is False
 
 
+def test_common_target_seed_enables_the_robust_acados_reference_preparation():
+    parser = periodic_example.build_argument_parser()
+    args = parser.parse_args(["--common-initial-solution", "common.npz"])
+
+    periodic_example.apply_assisted_hot_start_defaults(args)
+
+    assert args.periodic_fes_warmup_projection_strategy == "rollout"
+    assert args.full_dynamics_phase_one is True
+    assert args.acados_transfer_full_dynamics_rollout is True
+    assert args.acados_bind_first_node_fes_states is True
+    assert args.acados_dual_warm_start_mode == "reset"
+    assert args.acados_nlp_solver_type == "SQP"
+    assert args.acados_hessian_approx == "GAUSS_NEWTON"
+
+    disabled = parser.parse_args(
+        [
+            "--common-initial-solution",
+            "common.npz",
+            "--disable-acados-assisted-hot-start",
+        ]
+    )
+    periodic_example.apply_assisted_hot_start_defaults(disabled)
+    assert disabled.periodic_fes_warmup_projection_strategy == "sequential"
+    assert disabled.full_dynamics_phase_one is False
+    assert disabled.acados_transfer_full_dynamics_rollout is False
+    assert disabled.acados_bind_first_node_fes_states is False
+
+
 def test_solver_clis_distinguish_assistance_magnitude_from_signed_torque():
     periodic_parser = periodic_example.build_argument_parser()
     comparison_parser = comparison_example.build_cli()
@@ -576,6 +604,70 @@ def test_standard_warmup_cache_metadata_rejects_a_resistance_seed(tmp_path):
     assert loaded.metadata == metadata
     periodic_example._validate_standard_warmup_seed(loaded, args, assisted_path)
 
+
+def test_common_initial_solution_metadata_rejects_an_incompatible_horizon(tmp_path):
+    args = SimpleNamespace(
+        model_formulation="periodic_node",
+        mechanical_formulation="reduced",
+        cycles_per_window=1,
+        stimulations_per_cycle=30,
+        constant_crank_torque=-0.2,
+        torque_application="constant",
+        ode_solver="collocation",
+        nlp_ordering_strategy="time_major",
+        solver="ipopt",
+    )
+    metadata = periodic_example._common_initial_solution_metadata(args)
+    metadata["cycles_per_window"] = 2
+    seed = periodic_example._WarmupSolutionAdapter({}, {}, metadata=metadata)
+
+    with pytest.raises(ValueError, match="cycles_per_window"):
+        periodic_example._validate_common_initial_solution_metadata(
+            seed, args, tmp_path / "common.npz"
+        )
+
+
+def test_common_initial_solution_metadata_allows_a_transcription_change(tmp_path):
+    args = SimpleNamespace(
+        model_formulation="periodic_node",
+        mechanical_formulation="full",
+        cycles_per_window=1,
+        stimulations_per_cycle=30,
+        constant_crank_torque=-0.2,
+        torque_application="constant",
+        ode_solver="rk4",
+        nlp_ordering_strategy="time_major",
+        solver="fatrop",
+    )
+    metadata = periodic_example._common_initial_solution_metadata(args)
+    metadata["ode_solver"] = "collocation"
[…]

[…]
