### 2026-07-28 26e8a02
Harden reduced ACADOS preparation
