+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -688,7 +688,13 @@ def apply_assisted_hot_start_defaults(args: argparse.Namespace) -> None:
         if args.periodic_fes_warmup_projection_strategy == "sequential":
             args.periodic_fes_warmup_projection_strategy = "rollout"
         args.full_dynamics_phase_one = True
-        args.acados_transfer_full_dynamics_rollout = True
+        if not (
+            args.acados_transfer_full_dynamics_rollout
+            or args.acados_transfer_irk_rollout
+        ):
+            # Once the ACADOS OCP exists, advance the shifted window with the
+            # exact generated IRK map used by the solver.
+            args.acados_transfer_irk_rollout = True
         args.acados_bind_first_node_fes_states = True
     if assisted_hot_start and args.acados_control_homotopy_radii is None:
         args.acados_control_homotopy_radii = DEFAULT_ASSISTED_CONTROL_HOMOTOPY_RADII
@@ -2972,8 +2978,18 @@ def _common_initial_solution_metadata(args: argparse.Namespace) -> dict:
         "mechanical_formulation": args.mechanical_formulation,
         "cycles_per_window": int(args.cycles_per_window),
         "stimulations_per_cycle": int(args.stimulations_per_cycle),
+        "objective": sorted(parse_objectives(args.objective)),
+        "objective_shape": args.objective_shape,
         "constant_crank_torque": float(args.constant_crank_torque),
         "torque_application": args.torque_application,
+        "enforce_start_constraints": bool(args.enforce_start_constraints),
+        "first_node_wheel_q_slack": float(args.acados_wheel_q_slack),
+        "terminal_wheel_q_slack": float(args.acados_terminal_wheel_q_slack),
+        "terminal_wheel_q_reference_mode": args.terminal_wheel_q_reference_mode,
+        "pulse_width_scaling": float(args.pulse_width_scaling),
+        "pulse_width_active_set": args.pulse_width_active_set,
+        "pulse_width_minimum_policy": "model_pd0",
+        "pulse_width_maximum_s": 0.0006,
         "ode_solver": args.ode_solver,
         "nlp_ordering_strategy": getattr(args, "nlp_ordering_strategy", None),
         "producer_solver": args.solver,
@@ -3000,7 +3016,17 @@ def _validate_common_initial_solution_metadata(
         "mechanical_formulation",
         "cycles_per_window",
         "stimulations_per_cycle",
+        "objective",
+        "objective_shape",
         "torque_application",
+        "enforce_start_constraints",
+        "first_node_wheel_q_slack",
+        "terminal_wheel_q_slack",
+        "terminal_wheel_q_reference_mode",
+        "pulse_width_scaling",
+        "pulse_width_active_set",
+        "pulse_width_minimum_policy",
+        "pulse_width_maximum_s",
     ):
         if metadata.get(field) != expected[field]:
             raise ValueError(
@@ -6205,13 +6231,37 @@ def _rk4_full_dynamics_step(
     return next_state
 
 
+def _phase_one_state_keys(nlp) -> dict[str, tuple[str, ...]]:
+    """Map generic phase-I blocks onto full or reduced mechanical states."""
+
+    state_keys = tuple(nlp.states.keys())
+    position_keys = (
+        ("q",) if "q" in state_keys else (("theta",) if "theta" in state_keys else ())
+    )
+    velocity_keys = (
+        ("qdot",)
+        if "qdot" in state_keys
+        else (("omega",) if "omega" in state_keys else ())
+    )
+    mechanical_keys = set(position_keys + velocity_keys)
+    return {
+        # Keep the historical public block names so existing CLI options such
[…]
