+++ tests/shard1/test_solver_backends.py
@@ -121,6 +121,18 @@ def test_configure_madnlp_maps_pardiso_mkl_to_native_libmad_type():
     ) in solver.calls
 
 
+def test_configure_madnlp_keeps_mumps_runtime_name():
+    solver = configure_nlp_solver(
+        "madnlp",
+        max_iterations=321,
+        madnlp_linear_solver="mumps",
+        solver_namespace=_solver_namespace("MADNLP"),
+        check_availability=False,
+    )
+
+    assert ("set_option_unsafe", "mumps", "linear_solver") in solver.calls
+
+
 def test_configure_fatrop_uses_time_structured_native_options():
     namespace = _solver_namespace("FATROP")
 
