### 2026-07-29 5a47578
Cache cycling benchmark numerical stacks
