+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -62,6 +62,8 @@ env:
   LIBMAD_PARDISO_COMMIT: 5529f23a6bff33c566ad954da38d352f1f172356
   JULIAC_COMMIT: 73be8587a80bbb65dab7acd71d406f72867a3571
   JULIA_VERSION: "1.12.6"
+  CONDA_ENV_NAME: cocofest-cycling-benchmark
+  NUMERICAL_STACK_CACHE_VERSION: v1
   OMP_NUM_THREADS: "1"
   OPENBLAS_NUM_THREADS: "1"
   MKL_NUM_THREADS: "1"
@@ -95,9 +97,23 @@ jobs:
         uses: conda-incubator/setup-miniconda@v4
         with:
           miniforge-version: latest
-          activate-environment: cocofest-cycling-benchmark
-          environment-file: .github/cycling-benchmark-linux-environment.yml
+          activate-environment: ${{ env.CONDA_ENV_NAME }}
           conda-remove-defaults: true
+
+      - name: Restore the official numerical stack
+        id: official-stack-cache
+        uses: actions/cache/restore@v6
+        with:
+          path: ${{ env.CONDA }}/envs/${{ env.CONDA_ENV_NAME }}
+          key: numerical-stack-${{ env.NUMERICAL_STACK_CACHE_VERSION }}-${{ inputs.runner_label }}-${{ runner.arch }}-official-casadi-${{ env.CASADI_VERSION }}-${{ hashFiles('.github/cycling-benchmark-linux-environment.yml', '.github/scripts/install_biorbd_casadi_linux.sh') }}
+
+      - name: Install the benchmark dependencies
+        if: steps.official-stack-cache.outputs.cache-hit != 'true'
+        run: |
+          conda env update \
+            --name "$CONDA_ENV_NAME" \
+            --file .github/cycling-benchmark-linux-environment.yml \
+            --prune
 
       - name: Detect the CPU allocation
         run: |
@@ -109,10 +125,24 @@ jobs:
           lscpu
 
       - name: Install the reference numerical stack
+        if: steps.official-stack-cache.outputs.cache-hit != 'true'
         run: |
           python -m pip install --no-deps "casadi==$CASADI_VERSION"
           bash .github/scripts/install_biorbd_casadi_linux.sh
-          python -m pip install --no-deps -e .benchmark-deps/bioptim
+
+      - name: Validate the restored official stack
+        run: |
+          python -c 'import casadi, biorbd_casadi, numpy, pytest'
+
+      - name: Save the official numerical stack
+        if: steps.official-stack-cache.outputs.cache-hit != 'true'
+        uses: actions/cache/save@v6
+        with:
+          path: ${{ env.CONDA }}/envs/${{ env.CONDA_ENV_NAME }}
+          key: numerical-stack-${{ env.NUMERICAL_STACK_CACHE_VERSION }}-${{ inputs.runner_label }}-${{ runner.arch }}-official-casadi-${{ env.CASADI_VERSION }}-${{ hashFiles('.github/cycling-benchmark-linux-environment.yml', '.github/scripts/install_biorbd_casadi_linux.sh') }}
+
+      - name: Install the pinned Bioptim integration
+        run: python -m pip install --no-deps -e .benchmark-deps/bioptim
 
       - name: Validate the seed on the assisted target
         env:
@@ -173,10 +203,118 @@ jobs:
           if-no-files-found: error
           retention-days: 30
 
+  prepare-madnlp-stack:
+    name: Prepare the cached MadNLP numerical stack
+    runs-on: ${{ inputs.runner_label }}
[…]
