### 2026-07-29 e979f67
Add cached ACADOS cycling smoke CI
