### 2026-07-29 56b5cf1
Sanitize ACADOS code generation tags
