### 2026-07-29 589bf64
Separate ACADOS stationarity tolerance
