+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -66,6 +66,20 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "pulse_width_active_margin",
     "acados_wheel_q_slack",
     "acados_terminal_wheel_q_slack",
+    "max_acados_iterations",
+    "acados_tolerance",
+    "acados_stationarity_tolerance",
+    "acados_qp_solver",
+    "acados_integrator_type",
+    "acados_collocation_type",
+    "acados_sim_stages",
+    "acados_sim_steps",
+    "acados_newton_iter",
+    "acados_hessian_approx",
+    "acados_nlp_solver_type",
+    "acados_control_homotopy_radii",
+    "acados_control_homotopy_tolerance",
+    "acados_control_homotopy_stage_iterations",
     "terminal_wheel_q_reference_mode",
     "cycles_per_window",
     "stimulations_per_cycle",
