### 2026-07-29 ff3b203
Release reduced ACADOS control homotopy
