+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -80,6 +80,12 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "acados_control_homotopy_radii",
     "acados_control_homotopy_tolerance",
     "acados_control_homotopy_stage_iterations",
+    "acados_control_homotopy_keep_final_radius",
+    "acados_transfer_irk_rollout",
+    "acados_transfer_phase_one",
+    "acados_transfer_mechanical_restoration",
+    "acados_transfer_sqp_restarts",
+    "acados_cyclical_transfer_mode",
     "terminal_wheel_q_reference_mode",
     "cycles_per_window",
     "stimulations_per_cycle",
