### 2026-07-29 56d655c
Expose ACADOS homotopy radius policy
