### 2026-07-29 fa2c223
Grow reduced ACADOS transfer radius
