+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -81,6 +81,7 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "acados_control_homotopy_tolerance",
     "acados_control_homotopy_stage_iterations",
     "acados_control_homotopy_keep_final_radius",
+    "acados_control_homotopy_window_growth",
     "acados_transfer_irk_rollout",
     "acados_transfer_phase_one",
     "acados_transfer_mechanical_restoration",
@@ -2224,6 +2225,7 @@ def main(
     optional_nlp_periodic_ipopt_hot_start: bool = True,
     acados_max_iter: int = 100,
     acados_control_homotopy_keep_final_radius: bool | None = None,
+    acados_control_homotopy_window_growth: float = 1.0,
     control_regularization_weight: float = 0.0,
     acados_control_regularization_weight: float | None = None,
     control_regularization_target: float | None = None,
@@ -2610,6 +2612,9 @@ def main(
     acados_args.acados_stationarity_tolerance = acados_stationarity_tolerance
     acados_args.acados_control_homotopy_keep_final_radius = (
         acados_control_homotopy_keep_final_radius
+    )
+    acados_args.acados_control_homotopy_window_growth = (
+        acados_control_homotopy_window_growth
     )
     acados_args.acados_dual_warm_start_mode = acados_dual_warm_start_mode
     acados_args.acados_proximal_control_weights = acados_proximal_control_weights
@@ -3396,6 +3401,12 @@ def build_cli() -> argparse.ArgumentParser:
         action="store_false",
     )
     parser.add_argument(
+        "--acados-control-homotopy-window-growth",
+        type=float,
+        default=1.0,
+        help="Factor applied to the retained control radius after each RHO window.",
+    )
+    parser.add_argument(
         "--acados-integrator-type", choices=("ERK", "IRK", "DISCRETE"), default="IRK"
     )
     parser.add_argument(
@@ -3791,6 +3802,9 @@ if __name__ == "__main__":
         acados_max_iter=args.acados_max_iter,
         acados_control_homotopy_keep_final_radius=(
             args.acados_control_homotopy_keep_final_radius
+        ),
+        acados_control_homotopy_window_growth=(
+            args.acados_control_homotopy_window_growth
         ),
         control_regularization_weight=args.control_regularization_weight,
         acados_control_regularization_weight=args.acados_control_regularization_weight,
