### 2026-07-29 939d32e
Cap ACADOS transfer radius growth
