### 2026-07-29 f2e4f06
Tighten reduced ACADOS transfer cap
