### 2026-07-29 d697fa0
Document ACADOS 30-RHO benchmark
