+++ docs/acados_environment.md
@@ -42,8 +42,16 @@ are reported separately from the ACADOS window solve times.
 The Linux benchmark workflow performs the equivalent installation from the
 ACADOS submodule pinned by Bioptim, installs it into `$CONDA_PREFIX`, and
 caches the complete native/Python stack. Its separate `acados-smoke` matrix
-tests five full-mechanics and five experimental reduced-mechanics RHO by
-default. Increase the workflow input `acados_smoke_rhos` to 30 only after the
-short SQP jobs are stable; RTI is deliberately excluded from this first
-validation. Set `cycles=acados` to run only seed preparation, the cached
-ACADOS stack preparation, and these two smoke jobs.
+tests 30 full-mechanics and 30 experimental reduced-mechanics RHO by default.
+The reduced transfer grows its retained pulse-width radius from `0.1` to
+`10 microseconds`, then keeps that cap while recentering it every cycle.
+RTI is deliberately excluded until the full-SQP reference is more repeatable.
+Set `cycles=acados` to run only seed preparation, the cached ACADOS stack
+preparation, and these two benchmark jobs.
+
+Linux run `30415853682` reached `ACADOS_SUCCESS` on 29/30 windows for both
+formulations. The full model had one stationarity-only maximum-iteration
+window at RHO 30; the reduced model had one maximum-iteration window at RHO 26
+and recovered for RHO 27--30. Both attempted all 30 windows and retained the
+absolute `0.002 rad` terminal-angle target. These are strong timing results,
+but not yet a 30/30 convergence certificate.
