### 2026-07-29 10c3ff7
Expand cycling endurance solver benchmark
