+++ .github/scripts/summarize_cycling_benchmark.py
@@ -36,12 +36,17 @@ COMPARABILITY_FIELDS = (
 DEFAULT_EXPECTED_SOLVERS = ("ipopt", "fatrop", "madnlp")
 DEFAULT_EXPECTED_CASES = (
     "ipopt/full",
-    "fatrop/full",
+    "fatrop-rk4-compiled/full",
+    "fatrop-collocation-compiled/full",
     "madnlp-pardiso/full",
     "madnlp-mumps/full",
+    "acados/full",
     "ipopt/reduced",
+    "fatrop-rk4-compiled/reduced",
+    "fatrop-collocation-compiled/reduced",
     "madnlp-pardiso/reduced",
     "madnlp-mumps/reduced",
+    "acados/reduced",
 )
 
 
@@ -75,6 +80,13 @@ def _entry_case(entry: dict) -> str:
 
 def _solver_variant(entry: dict) -> str:
     solver = entry["result"].get("solver", "unknown").lower()
+    if solver == "fatrop":
+        configuration = entry.get("configuration", {})
+        transcription = str(configuration.get("ode_solver") or "unknown").lower()
+        compilation = (
+            "-compiled" if configuration.get("fatrop_c_compile") is True else ""
+        )
+        return f"fatrop-{transcription}{compilation}"
     if solver != "madnlp":
         return solver
     linear_solver = str(
@@ -514,6 +526,64 @@ def write_stimulation_csv(path: Path, entries: list[dict]) -> None:
                         )
 
 
+def write_muscle_fatigue_csv(path: Path, entries: list[dict]) -> None:
+    fieldnames = (
+        "solver",
+        "mechanical_formulation",
+        "case",
+        "muscle",
+        "executed_fatigue_objective",
+        "cumulative_normalized_fatigue_cycles",
+        "final_capacity_ratio",
+    )
+    with path.open("w", newline="", encoding="utf-8") as stream:
+        writer = csv.DictWriter(stream, fieldnames=fieldnames)
+        writer.writeheader()
+        for entry in entries:
+            for row in entry["result"].get("muscle_fatigue") or []:
+                writer.writerow(
+                    {
+                        "solver": entry["result"]["solver"],
+                        "mechanical_formulation": _mechanical_formulation(entry),
+                        "case": _entry_case(entry),
+                        **{field: row.get(field) for field in fieldnames[3:]},
+                    }
+                )
+
+
+def write_pulse_width_variation_csv(path: Path, entries: list[dict]) -> None:
+    fieldnames = (
+        "solver",
+        "mechanical_formulation",
+        "case",
+        "muscle",
+        "from_cycle",
+        "to_cycle",
+        "mean_absolute_change_us",
+        "root_mean_square_change_us",
+        "maximum_absolute_change_us",
+    )
+    with path.open("w", newline="", encoding="utf-8") as stream:
+        writer = csv.DictWriter(stream, fieldnames=fieldnames)
+        writer.writeheader()
+        for entry in entries:
+            variation = entry["result"].get("pulse_width_cycle_variation") or {}
+            for muscle in variation.get("muscles") or []:
+                for transition in muscle.get("transitions") or []:
+                    writer.writerow(
+                        {
+                            "solver": entry["result"]["solver"],
+                            "mechanical_formulation": _mechanical_formulation(entry),
+                            "case": _entry_case(entry),
+                            "muscle": muscle.get("muscle"),
+                            **{
+                                field: transition.get(field)
+                                for field in fieldnames[4:]
+                            },
+                        }
+                    )
+
+
 def render_markdown(
     entries: list[dict],
     mismatches: list[dict],
[…]
