### 2026-07-29 7c0a413
Stabilize ACADOS and FATROP benchmark probes
