+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -28,6 +28,7 @@ try:
         ACADOS_STATUS_NAMES,
         DEFAULT_CRANK_TORQUE_NM,
         build_argument_parser,
+        parse_control_homotopy_radii,
         parse_crank_assistance,
         parse_proximal_control_weights,
         parse_terminal_wheel_q_slacks,
@@ -38,6 +39,7 @@ except ImportError:
         ACADOS_STATUS_NAMES,
         DEFAULT_CRANK_TORQUE_NM,
         build_argument_parser,
+        parse_control_homotopy_radii,
         parse_crank_assistance,
         parse_proximal_control_weights,
         parse_terminal_wheel_q_slacks,
@@ -77,9 +79,11 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "acados_newton_iter",
     "acados_hessian_approx",
     "acados_nlp_solver_type",
+    "acados_assisted_hot_start",
     "acados_control_homotopy_radii",
     "acados_control_homotopy_tolerance",
     "acados_control_homotopy_stage_iterations",
+    "acados_control_homotopy_max_restarts",
     "acados_control_homotopy_keep_final_radius",
     "acados_control_homotopy_window_growth",
     "acados_control_homotopy_window_max_radius",
@@ -1070,6 +1074,11 @@ def _solver_config(
     ipopt_linear_solver: str,
     ipopt_dual_warm_start_mode: str,
     acados_max_iter: int,
+    acados_assisted_hot_start: bool,
+    acados_control_homotopy_radii: tuple[float, ...] | None,
+    acados_control_homotopy_tolerance: float,
+    acados_control_homotopy_stage_iterations: int,
+    acados_control_homotopy_max_restarts: int,
     control_regularization_weight: float,
     control_regularization_target: float | None,
     control_regularization_target_source: str,
@@ -1280,6 +1289,15 @@ def _solver_config(
             objective_shape=objective_shape,
             constant_crank_torque=resistive_torque,
             max_acados_iterations=acados_max_iter,
+            acados_assisted_hot_start=acados_assisted_hot_start,
+            acados_control_homotopy_radii=acados_control_homotopy_radii,
+            acados_control_homotopy_tolerance=acados_control_homotopy_tolerance,
+            acados_control_homotopy_stage_iterations=(
+                acados_control_homotopy_stage_iterations
+            ),
+            acados_control_homotopy_max_restarts=(
+                acados_control_homotopy_max_restarts
+            ),
             ipopt_linear_solver=ipopt_linear_solver,
             ipopt_dual_warm_start_mode=ipopt_dual_warm_start_mode,
             n_windows=n_windows,
@@ -2340,6 +2358,11 @@ def main(
     alpaqa_max_no_progress: int | None = None,
     optional_nlp_periodic_ipopt_hot_start: bool = True,
     acados_max_iter: int = 100,
+    acados_assisted_hot_start: bool = True,
+    acados_control_homotopy_radii: tuple[float, ...] | None = None,
+    acados_control_homotopy_tolerance: float = 5e-4,
+    acados_control_homotopy_stage_iterations: int = 50,
+    acados_control_homotopy_max_restarts: int = 1,
     acados_control_homotopy_keep_final_radius: bool | None = None,
     acados_control_homotopy_window_growth: float = 1.0,
     acados_control_homotopy_window_max_radius: float | None = None,
@@ -2498,6 +2521,15 @@ def main(
         ipopt_linear_solver=ipopt_linear_solver,
         ipopt_dual_warm_start_mode=ipopt_dual_warm_start_mode,
         acados_max_iter=acados_max_iter,
+        acados_assisted_hot_start=acados_assisted_hot_start,
[…]
