### 2026-07-29 a512f8c
Release full ACADOS controls after homotopy
