### 2026-07-29 f7f4020
Audit ACADOS shooting residuals
