+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -568,9 +568,12 @@ jobs:
             else
               # The full model needs the homotopy to construct a nearby
               # feasible point, but retaining the final ±1 us radius prevents
-              # the SQP from following the physically required PW branch.
+              # the SQP from following the physically required PW branch. A
+              # full step also tests whether repeated relinearization can turn
+              # the useful first RTI direction into a converged SQP trajectory.
               mechanics_options+=(
                 --acados-control-homotopy-release-final-radius
+                --acados-globalization FIXED_STEP
               )
             fi
             if [[ "$nlp_solver" == "SQP_RTI" ]]; then
