### 2026-07-29 8c0f1c2
Apply the common feasibility threshold to ACADOS
