### 2026-07-29 e56e2ae
Enforce ACADOS physical feasibility in benchmarks
