+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -559,11 +559,16 @@ jobs:
             fi
             if [[ "$mechanics" == "reduced" ]]; then
               stationarity=5e-3
+              # The 10 us cap was hit by the triceps and coincided with
+              # isolated SQP stationarity failures at RHO 34 and 79. A
+              # 50 us trust region covers the observed triceps transitions;
+              # rare larger recruitment switches still require the physical
+              # [pd0, 600 us] fallback outside this benchmark screen.
               mechanics_options+=(
                 --mechanical-formulation reduced
                 --experimental-reduced-acados
                 --acados-control-homotopy-window-growth 10
-                --acados-control-homotopy-window-max-radius 1e-5
+                --acados-control-homotopy-window-max-radius 5e-5
               )
             else
               # The full model needs the homotopy to construct a nearby
