### 2026-07-29 4f777a0
Align full and reduced crank velocity bounds
