### 2026-07-29 9126639
Keep the stable ACADOS pulse-width trust region
