### 2026-07-29 fc4546b
Screen ACADOS feasibility restoration strategies
