+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -81,6 +81,7 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "acados_newton_iter",
     "acados_hessian_approx",
     "acados_nlp_solver_type",
+    "acados_search_direction_mode",
     "acados_assisted_hot_start",
     "acados_control_homotopy_radii",
     "acados_control_homotopy_tolerance",
@@ -91,6 +92,12 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "acados_control_homotopy_window_max_radius",
     "acados_transfer_irk_rollout",
     "acados_transfer_phase_one",
+    "acados_transfer_bound_homotopy",
+    "acados_transfer_bound_homotopy_fractions",
+    "acados_transfer_bound_homotopy_padding",
+    "acados_transfer_bound_homotopy_iterations",
+    "acados_transfer_bound_homotopy_tolerance",
+    "acados_transfer_bound_homotopy_solver_tolerance",
     "acados_transfer_mechanical_restoration",
     "acados_transfer_sqp_restarts",
     "acados_cyclical_transfer_mode",
