### 2026-07-29 9e8f615
Forward ACADOS transfer homotopy options
