### 2026-07-29 7eb3931
Extend feasible-QP ACADOS screen to 100 RHO
