### 2026-07-29 5d63b86
Preserve common seed warmup semantics
