### 2026-07-29 617d35d
Use SX for IPOPT and MadNLP cycling OCPs
