+++ tests/shard1/test_periodic_pulse_width.py
@@ -3440,10 +3440,10 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
         flags=re.DOTALL,
     )
     assert "timeout_minutes: 360" in workflow
-    assert "Screen IPOPT SX full over 30 RHO" in workflow
+    assert "Screen IPOPT MX full over 30 RHO" in workflow
     assert "Screen Fatrop RK4 SX reduced over 30 RHO" in workflow
     assert "Screen Fatrop collocation SX full over 30 RHO" in workflow
-    assert "Screen MadNLP MUMPS SX reduced over 30 RHO" in workflow
+    assert "Screen MadNLP MUMPS MX reduced over 30 RHO" in workflow
     assert "expected_graph_screens=4" in workflow
     assert "Download the SX graph-mode screens" in workflow
     assert "--ipopt-use-sx" in (
