### 2026-07-29 f59ef25
Document cycling benchmark gains and next steps
