+++ README.md
@@ -30,6 +30,7 @@ Supports predictive musculoskeletal simulation driven by FES, moving time horizo
   - [Available FES models](#available-fes-models)
   - [Musculoskeletal model driven by FES](#musculoskeletal-model-driven-by-fes)
   - [Moving time horizons](#moving-time-horizons)
+  - [Cycling fatigue solver benchmark](#cycling-fatigue-solver-benchmark)
   - [Identification](#identification)
   - [Initial value problem](#initial-value-problem)
   - [Summation truncation](#summation-truncation)
@@ -232,6 +233,151 @@ For longer time span simulation and apprehend muscle fatigue apparition, `Cocofe
 The cycling fatigue MHE can also benchmark the optional MadNLP and Alpaqa
 Bioptim integrations. See the
 [solver setup, warm-start policy, and benchmark results](docs/optional_nlp_solvers.md).
+
+<a id="cycling-fatigue-solver-benchmark"></a>
+## Cycling fatigue solver benchmark
+
+The endurance benchmark minimizes fatigue with four Ding fatigue muscle
+models. Its default is **one pedalling cycle per OCP** and 30 pulse-width
+controls per muscle and cycle. Therefore, `cycles=100` means 100 consecutive
+one-cycle receding-horizon optimizations (100 RHO), not a single OCP containing
+100 cycles. A two-cycle OCP remains available with `cycles_per_window=2`, but
+ACADOS is currently benchmarked only with one cycle.
+
+The benchmark uses zero external crank assistance, an absolute terminal crank
+angle target with a `0.002 rad` tolerance, and two allowed consecutive
+failures. Pulse widths are validated and clipped to each muscle's physical
+interval `[pd0, 600 us]`; `pd0` is the Ding model's zero-recruitment pulse
+duration, approximately `131.405 us`, rather than a numerical pulse width of
+zero.
+
+### Changes made during the solver study
+
+1. The sign of the external crank torque was made explicit and the current
+   endurance problem was changed to `0 N.m` assistance to expose fatigue
+   earlier.
+2. The terminal crank target was made absolute in cycle count. It no longer
+   follows the previous solution, which prevents accumulated angular drift.
+3. Historical seeds are validated at load time, including model,
+   transcription, torque, horizon and pulse-width metadata. Pulse widths are
+   clipped to `[pd0, 600 us]`, with warnings for repaired legacy values.
+4. A common, solver-independent primal seed is produced once with IPOPT and
+   consumed by IPOPT, Fatrop, MadNLP and ACADOS. Its recorded consumed warmup
+   cycle must be preserved by every consumer.
+5. NLP variables use time-major ordering for the sparse IPOPT, Fatrop and
+   MadNLP interfaces. ACADOS retains its native stage-major organization.
+6. A reduced mechanical formulation keeps only crank angle `theta` and speed
+   `omega`, while retaining all 20 Ding muscle/fatigue states. It does not
+   impose constant crank speed.
+7. IPOPT, MadNLP/MUMPS, MadNLP/PARDISO, Fatrop/RK4,
+   Fatrop/collocation and ACADOS are run from the same physical seed and
+   feasibility threshold. Alpaqa is documented but excluded from endurance
+   runs because its current sparse CasADi/Bioptim path is not operational.
+8. The Linux action uses independent IPOPT, Fatrop, MadNLP and ACADOS runners.
[…]
