### 2026-07-29 8d74efc
Reuse compiled NLP graphs across cycling RHO
