+++ tests/shard1/test_periodic_pulse_width.py
@@ -3006,6 +3006,12 @@ def test_benchmark_json_summary_contains_comparable_fatigue_metrics(tmp_path):
                 "t_wall_nlp_hess_l": 0.75,
             }
         ],
+        compiled_nlp_reuse={
+            "enabled": True,
+            "compiled_library_build_count": 1,
+            "compiled_library_reused": True,
+            "graph_rebuild_detected": False,
+        },
     )
     result["args"].objective = "fatigue"
     result["args"].objective_shape = "quadratic"
@@ -3058,6 +3064,7 @@ def test_benchmark_json_summary_contains_comparable_fatigue_metrics(tmp_path):
     assert [window["rho"] for window in row["windows"]] == [1, 2]
     assert row["windows"][0]["native_status"] == "Solve_Succeeded"
     assert row["nlp_solver_stats"][0]["t_wall_nlp_hess_l"] == 0.75
+    assert row["compiled_nlp_reuse"]["compiled_library_build_count"] == 1
 
 
 def test_independent_bound_violation_accepts_infinite_bounds():
@@ -3384,6 +3391,13 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
         workflow,
     )
     assert "acados_option_rhos:" in workflow
+    assert "compile_nlp_evaluators:" in workflow
+    assert (
+        '"$BENCHMARK_CYCLES" "${{ inputs.compile_nlp_evaluators }}"'
+        in workflow
+    )
+    assert ".compiled_nlp_reuse.compiled_library_build_count == 1" in workflow
+    assert ".compiled_nlp_reuse.graph_rebuild_detected == false" in workflow
     assert "inputs.cycles != 'screen' && inputs.cycles != 'acados'" in workflow
     assert "prepare-acados-stack:" in workflow
     assert "ACADOS_COMMIT: 48e223e85f0408ebfd1d8c6d6fb0589e9c41b3aa" in workflow
@@ -3430,6 +3444,14 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
         "benchmark-probes 1 true"
     ) in workflow
     assert "cycling-fatrop-compile-probe-full-${{ github.run_id }}" in workflow
+    benchmark_runner = (
+        Path(__file__).resolve().parents[2]
+        / ".github"
+        / "scripts"
+        / "run_cycling_benchmark_case.sh"
+    ).read_text(encoding="utf-8")
+    assert 'solver_options+=(--ipopt-c-compile)' in benchmark_runner
+    assert 'solver_options+=(--madnlp-c-compile)' in benchmark_runner
     assert "Checkpoint IPOPT full" in workflow
     assert "Checkpoint Fatrop RK4 reduced" in workflow
     assert "max-parallel: 3" in workflow
@@ -4927,6 +4949,128 @@ def test_window_cache_callback_runs_before_window_is_advanced():
     MyCyclicNMPC.advance_window_bounds_states(nmpc, solution)
 
     assert events == ["before_window_advance", "decision_states"]
+
+
+def test_compiled_nlp_tracker_accepts_runtime_bound_changes_without_graph_rebuild():
+    state_bounds = SimpleNamespace(
+        min=np.array([[0.0, -10.0, -2.0 * np.pi]]),
+        max=np.array([[0.0, 10.0, -2.0 * np.pi]]),
+    )
+    control_bounds = SimpleNamespace(
+        min=np.array([[131.405e-6, 131.405e-6]]),
+        max=np.array([[600e-6, 600e-6]]),
+    )
+    constraint_bounds = SimpleNamespace(
+        min=np.array([[-0.002]]),
+        max=np.array([[0.002]]),
+    )
+    compiled_solver = object()
[…]
