### 2026-07-30 c19a14f
Certify cycling mechanics and compile solver benchmarks
