### 2026-07-30 199b34d
Fix compiled cycling solver benchmark paths
