### 2026-07-30 d7d957b
Document pinned Bioptim benchmark revisions
