### 2026-07-30 5727942
Make cycling benchmark SX-only with MadNLP MUMPS
