+++ tests/shard1/test_periodic_pulse_width.py
@@ -2733,13 +2733,13 @@ def test_solver_clis_accept_explicit_thread_count():
     periodic_args = periodic_example.build_argument_parser().parse_args(
         ["--n-threads", "8"]
     )
-    pardiso_args = comparison_example.build_cli().parse_args(
-        ["--madnlp-linear-solver", "pardiso_mkl"]
+    mumps_args = comparison_example.build_cli().parse_args(
+        ["--madnlp-linear-solver", "mumps"]
     )
     comparison_args = comparison_example.build_cli().parse_args(["--n-threads", "8"])
 
     assert periodic_args.n_threads == 8
-    assert pardiso_args.madnlp_linear_solver == "pardiso_mkl"
+    assert mumps_args.madnlp_linear_solver == "mumps"
     assert comparison_args.n_threads == 8
 
 
@@ -3664,17 +3664,9 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
     assert "--acados-control-homotopy-window-max-radius 1e-5" in workflow
     assert "--max-consecutive-failing 2" in workflow
     assert "cycling-acados-smoke-${{ github.run_id }}" in workflow
-    assert (
-        "run_cycling_benchmark_case.sh fatrop-collocation fatrop full structured "
-        'collocation benchmark-results "$BENCHMARK_CYCLES" '
-        '"${{ inputs.compile_nlp_evaluators }}" mx full'
-    ) in workflow
+    assert "case_slug: fatrop" not in workflow
+    assert "madnlp-pardiso" not in workflow
     assert "fatrop-rk4" not in workflow
-    assert "Compare Fatrop interpreted and compiled evaluators over 5 RHO" in workflow
-    assert "cycling-compile-ablation-fatrop-${{ github.run_id }}" in workflow
-    assert "bioptim-fatrop-scaled-gaps.patch" in workflow
-    assert "Compare Fatrop state scaling over 5 RHO" in workflow
-    assert "cycling-scaling-ablation-fatrop-${{ github.run_id }}" in workflow
     benchmark_runner = (
         Path(__file__).resolve().parents[2]
         / ".github"
@@ -3685,54 +3677,55 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
     assert "MadNLP C compilation is not validated" in benchmark_runner
     assert 'solver_options+=(--madnlp-c-compile)' not in benchmark_runner
     assert '--fatrop-state-scaling "$fatrop_state_scaling"' in benchmark_runner
+    assert 'case "$graph_mode" in' in benchmark_runner
+    assert "The endurance benchmark is SX-only" in benchmark_runner
+    assert "libMAD WARNING: option linear_solver is of unknown type" in (
+        benchmark_runner
+    )
+    assert ".configurations[$solver].use_sx == true" in benchmark_runner
     assert "--ipopt-enforce-start-constraints" in benchmark_runner
     assert "--reduced-cycling-profile benchmark-seed/" in benchmark_runner
     assert "Compare IPOPT interpreted and compiled evaluators over 5 RHO" in workflow
     assert "Compare MadNLP MUMPS interpreted and compiled evaluators" not in workflow
-    assert (
-        "Compile IPOPT/Fatrop CasADi evaluators once"
-        in workflow
-    )
+    assert "Compile IPOPT CasADi evaluators once" in workflow
     assert "Checkpoint IPOPT full" in workflow
-    assert "Checkpoint Fatrop collocation reduced" in workflow
-    assert "max-parallel: 4" in workflow
-    assert "case_slug: fatrop-probes" not in workflow
[…]
