### 2026-07-30 89f644e
Reuse certified SX seed for ACADOS benchmarks
