### 2026-07-30 e6c4a46
Avoid redundant OCP builds in endurance CI
