### 2026-07-30 fdb948c
Harden compiled NLP and ACADOS CI audits
