+++ docs/message_kevin_benchmark.md
@@ -64,7 +64,10 @@ d’infrastructure, sans graphe non-SX et sans option libMad ignorée.
 La compilation IPOPT est validée directement dans les cas full et reduced :
 la bibliothèque doit être construite une seule fois, le source `nlp.c` doit
 garder le même hash et les bornes mobiles doivent changer sans reconstruction
-du graphe. L’ancienne ablation intégrée qui reconstruisait deux OCP full
+du graphe. Chaque cas possède un répertoire de codegen isolé et la CI exige
+une observation de ce source à chaque RHO, ce qui empêche un `nlp.c` full
+résiduel de certifier artificiellement le reduced. L’ancienne ablation
+intégrée qui reconstruisait deux OCP full
 supplémentaires a été retirée après un arrêt `143` du runner. Si tu veux
 remesurer compilé contre interprété, il faut lancer deux workflows identiques
 en changeant seulement `compile_nlp_evaluators`; les artefacts principaux
