### 2026-07-30 aac9ff5
Recenter absolute cycle target after seed loading
