### 2026-07-30 578dc3f
Document final SX-only endurance campaign
