+++ docs/cycling_solver_benchmark/README.md
@@ -1063,7 +1063,123 @@ La formulation réduite effectue davantage d’itérations, mais chaque itérati
 est environ 2.2 fois moins coûteuse. Le gain vient donc surtout du graphe de
 dérivées, pas d’une meilleure convergence itérative.
 
-## 11. Résultats Linux historiques de référence
+## 11. Résultats Linux
+
+### 11.1 Campagne SX-only finale, 100 RHO
+
+Le run [`30522170340`](https://github.com/mickaelbegon/cocofest/actions/runs/30522170340)
+est la référence courante. Il a été lancé après deux paliers verts à 5 et
+30 RHO, avec :
+
+- Cocofest `aac9ff5c2ccec2f16adb6fb1f46932d44e15b7f7`;
+- Bioptim `3523f1745e315f07761159d7e06bd2d876026704`;
+- libMad `5529f23a6bff33c566ad954da38d352f1f172356`;
+- ACADOS `48e223e85f0408ebfd1d8c6d6fb0589e9c41b3aa`;
+- un cycle et 30 stimulations par OCP;
+- 100 RHO successifs;
+- couple externe nul;
+- objectif de fatigue seul;
+- cible angulaire absolue recalée après chargement du seed, slack
+  `0.002 rad`;
+- graphes SX uniquement;
+- IPOPT compilé une fois par formulation;
+- MadNLP interprété avec `MumpsSolver`;
+- deux échecs consécutifs permis.
+
+| Solveur | Mécanique | Préfixe strict | Convergés et faisables isolément/tentés | Médiane chaude | P90 chaud | Mur-à-mur | Coût fatigue | AUC | Min. \(A/A_\mathrm{scale}\) |
+|---|---|---:|---:|---:|---:|---:|---:|---:|---:|
+| IPOPT compilé/MUMPS | full | 100/100 | 100/100 | **1.444 s** | 1.698 s | 612.4 s | 11406.60 | 16.489 | 0.86736 |
+| MadNLP/MUMPS | full | 100/100 | 100/100 | 2.600 s | 2.899 s | **350.1 s** | 11344.72 | 16.460 | 0.86795 |
+| ACADOS SQP-IRK | full | 0/100 | 0/2 | — | — | 19.7 s | — | — | — |
+| IPOPT compilé/MUMPS | reduced | 100/100 | 100/100 | **1.122 s** | 1.302 s | 255.6 s | 668.10 | 4.849 | 0.97649 |
+| MadNLP/MUMPS | reduced | 100/100 | 100/100 | 1.180 s | 1.425 s | **178.4 s** | 652.99 | 4.812 | 0.97707 |
+| ACADOS SQP-IRK | reduced | 13/100 | 18/22 | **0.102 s** | 0.104 s | 25.6 s | 46.33¹ | 0.400¹ | 0.98509¹ |
+
+¹ Les métriques ACADOS sont calculées sur le préfixe strict, pas sur
+100 cycles. Après le premier échec au RHO 14, les fenêtres 15–18 et 20
+convergent et sont faisables lorsqu'elles sont évaluées isolément. Elles ne
+prolongent toutefois pas une chaîne MHE valide : le préfixe d'endurance reste
+donc limité à 13 cycles. Les échecs consécutifs 21–22 mettent fin à la
+campagne.
+
+Les quatre cas NLP terminent donc 100/100 avec succès solveur et physique.
+L’erreur angulaire absolue maximale vaut environ `1.470 mrad` en full et
+`2.004–2.006 mrad` en reduced; le seuil physique inclut la petite tolérance
+numérique autour du slack de `2 mrad`. Entre la fin du premier cycle et celle
[…]
