### 2026-07-30 e7f133e
Integrate ACADOS v0.5.5 benchmark audit
