+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1374,6 +1374,52 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Initialize the first QP warm start from the current NLP iterate.",
     )
     parser.add_argument(
+        "--acados-reset-solver-before-solve",
+        action="store_true",
+        help=(
+            "Reset ACADOS and the QP-solver memory before each NLP solve. "
+            "This is an ACADOS >=0.5.5 diagnostic option; the projected "
+            "Cocofest primal is injected again after the reset."
+        ),
+    )
+    parser.add_argument(
+        "--acados-check-reuse-possible",
+        action="store_true",
+        help=(
+            "Ask ACADOS >=0.5.5 to validate whether its generated code can be "
+            "reused. Each benchmark case must use an isolated JSON/code folder."
+        ),
+    )
+    parser.add_argument(
+        "--acados-code-reuse-tolerance",
+        type=float,
+        default=1e-12,
+        help="Numerical comparison tolerance used by the ACADOS code-reuse audit.",
+    )
+    parser.add_argument(
+        "--acados-with-anderson-acceleration",
+        action="store_true",
+        help=(
+            "Enable ACADOS Anderson acceleration. It is accepted only with "
+            "FIXED_STEP globalization."
+        ),
+    )
+    parser.add_argument(
+        "--acados-anderson-activation-threshold",
+        type=float,
+        default=0.1,
+        help="Residual threshold below which ACADOS activates Anderson acceleration.",
+    )
+    parser.add_argument(
+        "--acados-byrd-omojokon-slack-relaxation-factor",
+        type=float,
+        default=1.00001,
+        help=(
+            "Slack relaxation factor for the Byrd-Omojokun feasible-QP direction; "
+            "must be finite and at least one."
+        ),
+    )
+    parser.add_argument(
         "--acados-dual-warm-start-mode",
         choices=("preserve", "reset", "shift"),
         default="reset",
@@ -3336,6 +3382,20 @@ def _codegen_signature(args: argparse.Namespace) -> str:
         "acados_warm_start_first_qp_from_nlp": (
             args.acados_warm_start_first_qp_from_nlp
         ),
+        "acados_reset_solver_before_solve": (
+            args.acados_reset_solver_before_solve
+        ),
+        "acados_check_reuse_possible": args.acados_check_reuse_possible,
+        "acados_code_reuse_tolerance": args.acados_code_reuse_tolerance,
+        "acados_with_anderson_acceleration": (
+            args.acados_with_anderson_acceleration
+        ),
+        "acados_anderson_activation_threshold": (
+            args.acados_anderson_activation_threshold
+        ),
+        "acados_byrd_omojokon_slack_relaxation_factor": (
+            args.acados_byrd_omojokon_slack_relaxation_factor
+        ),
         "acados_qpscaling_scale_objective": args.acados_qpscaling_scale_objective,
         "acados_qpscaling_scale_constraints": args.acados_qpscaling_scale_constraints,
         "acados_ext_qp_res": args.acados_ext_qp_res,
@@ -3389,6 +3449,45 @@ def set_acados_unsafe_option(solver: Solver.ACADOS, value, name: str) -> None:
     setattr(solver, f"_{name}", value)
 
 
+def validate_acados_v055_options(
+    *,
+    nlp_solver_type: str,
+    globalization: str,
+    ext_qp_res: bool,
+    code_reuse_tolerance: float,
+    with_anderson_acceleration: bool,
+    anderson_activation_threshold: float,
+    byrd_omojokon_slack_relaxation_factor: float,
+) -> None:
+    """Reject option combinations that ACADOS 0.5.5 cannot use safely."""
+
[…]
