+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -914,6 +914,8 @@ def prepare_nmpc(
     if mechanical_formulation not in ("full", "reduced"):
         raise ValueError("mechanical_formulation must be 'full' or 'reduced'.")
     reduced_dynamics = simulation_conditions.get("reduced_cycling_dynamics")
+    if isinstance(reduced_dynamics, (str, os.PathLike)):
+        reduced_dynamics = ReducedCyclingDynamics.load(reduced_dynamics)
     # --- Pickle file info --- #
     initial_guess_path = simulation_conditions["init_guess_file_path"]
 
@@ -959,8 +961,6 @@ def prepare_nmpc(
             raise ValueError(
                 "reduced_cycling_dynamics is required for the reduced formulation."
             )
-        if isinstance(reduced_dynamics, (str, os.PathLike)):
-            reduced_dynamics = ReducedCyclingDynamics.load(reduced_dynamics)
         theta_init, omega_init, projection_audit = (
             reduced_dynamics.kinematics.project_generalized_trajectory(
                 np.asarray(full_mechanical_init["q"].init, dtype=float),
@@ -1030,6 +1030,15 @@ def prepare_nmpc(
         )
     else:
         x_init = full_mechanical_init
+        coordinate_qdot_bounds = None
+        if reduced_dynamics is not None:
+            coordinate_qdot_bounds = (
+                full_coordinate_qdot_bounds_from_reduced_profile(
+                    reduced_dynamics,
+                    physical_crank_velocity_target=wheel_qdot_regularization_target,
+                    physical_crank_velocity_margin=wheel_qdot_bound_margin,
+                )
+            )
         x_bounds, x_init = set_x_bounds(
             model=model,
             x_init=x_init,
@@ -1037,6 +1046,7 @@ def prepare_nmpc(
             ode_solver=ode_solver,
             init_file_path=initial_guess_path,
             wheel_qdot_bound_margin=wheel_qdot_bound_margin,
+            coordinate_qdot_bounds=coordinate_qdot_bounds,
         )
 
     # --- Set states scaling --- #
@@ -1260,6 +1270,7 @@ def set_x_bounds(
     ode_solver: OdeSolver,
     init_file_path: str,
     wheel_qdot_bound_margin: float = 3.0,
+    coordinate_qdot_bounds: tuple[np.ndarray, np.ndarray] | None = None,
 ) -> tuple[BoundsList, InitialGuessList]:
     if wheel_qdot_bound_margin <= 0:
         raise ValueError("wheel_qdot_bound_margin must be strictly positive.")
@@ -1345,6 +1356,33 @@ def set_x_bounds(
     qdot_x_bounds.max[1] = [forearm_qdot[1], forearm_qdot[1], forearm_qdot[1]]
     qdot_x_bounds.min[2] = [wheel_qdot[0], wheel_qdot[0], wheel_qdot[0]]
     qdot_x_bounds.max[2] = [wheel_qdot[1], wheel_qdot[1], wheel_qdot[1]]
+    if coordinate_qdot_bounds is not None:
+        coordinate_lower, coordinate_upper = (
+            np.asarray(values, dtype=float).reshape(-1)
+            for values in coordinate_qdot_bounds
+        )
+        if (
+            coordinate_lower.size != model.nb_qdot
+            or coordinate_upper.size != model.nb_qdot
+            or not np.all(np.isfinite(coordinate_lower))
+            or not np.all(np.isfinite(coordinate_upper))
+            or np.any(coordinate_lower >= coordinate_upper)
+        ):
+            raise ValueError(
+                "coordinate_qdot_bounds must contain finite, ordered bounds "
+                f"for all {model.nb_qdot} generalized velocities."
+            )
+        for index in range(model.nb_qdot):
+            lower = min(
[…]
