### 2026-07-30 46cb42c
Keep ACADOS cadence validation external
