### 2026-07-30 23eda2e
Separate ACADOS contact stabilization
