+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe.py
@@ -866,6 +866,10 @@ def prepare_nmpc(
         "enforce_contact_constraints_all_nodes",
         False,
     )
+    enforce_contact_position_all_nodes = cycling_info.get(
+        "enforce_contact_position_all_nodes",
+        False,
+    )
     enforce_physical_crank_velocity_bounds = cycling_info.get(
         "enforce_physical_crank_velocity_bounds",
         False,
@@ -1076,6 +1080,7 @@ def prepare_nmpc(
         enforce_contact_constraints_all_nodes=(
             enforce_contact_constraints_all_nodes
         ),
+        enforce_contact_position_all_nodes=enforce_contact_position_all_nodes,
         x_init=x_init,
         cycle_len=cycle_len,
         n_cycles_simultaneous=n_cycles_simultaneous,
@@ -1896,6 +1901,7 @@ def set_constraints(
     enforce_start_constraints: bool = True,
     *,
     enforce_contact_constraints_all_nodes: bool = False,
+    enforce_contact_position_all_nodes: bool = False,
     x_init: InitialGuessList | None = None,
     cycle_len: int | None = None,
     n_cycles_simultaneous: int = 1,
@@ -1909,19 +1915,34 @@ def set_constraints(
 ):
     constraints = ConstraintList()
     is_reduced = isinstance(bio_model, ReducedFesCyclingModel)
-    if enforce_contact_constraints_all_nodes and is_reduced:
+    if (
+        enforce_contact_constraints_all_nodes
+        or enforce_contact_position_all_nodes
+    ) and is_reduced:
         raise ValueError(
             "All-node marker contact constraints apply only to the full "
             "mechanical formulation."
         )
-    if (enforce_start_constraints or enforce_contact_constraints_all_nodes) and not is_reduced:
-        contact_node = (
+    if (
+        enforce_start_constraints
+        or enforce_contact_constraints_all_nodes
+        or enforce_contact_position_all_nodes
+    ) and not is_reduced:
+        velocity_node = (
             Node.ALL if enforce_contact_constraints_all_nodes else Node.START
+        )
+        position_node = (
+            Node.ALL
+            if (
+                enforce_contact_constraints_all_nodes
+                or enforce_contact_position_all_nodes
+            )
+            else Node.START
         )
         # --- Constraining wheel center position to a fix position --- #
         constraints.add(
             ConstraintFcn.TRACK_MARKERS_VELOCITY,
-            node=contact_node,
+            node=velocity_node,
             marker_index=bio_model.marker_index("wheel_center"),
             axes=[Axis.X, Axis.Y],
         )
@@ -1929,7 +1950,7 @@ def set_constraints(
             ConstraintFcn.SUPERIMPOSE_MARKERS,
             first_marker="wheel_center",
             second_marker="global_wheel_center",
-            node=contact_node,
+            node=position_node,
             axes=[Axis.X, Axis.Y],
         )
     if enforce_physical_crank_velocity_bounds:
