### 2026-07-30 4b380c6
Keep ACADOS contact stabilization independent
