### 2026-07-30 dc7eb2d
Validate ACADOS scaled constraints in endurance CI
