+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -73,7 +73,7 @@ env:
   BIOPTIM_REPOSITORY: mickaelbegon/BiorbdOptim
   # Dedicated multi-solver integration branch:
   # codex/cocofest-acados-v055-exploration
-  BIOPTIM_PRODUCTION_COMMIT: 04f0e1487f5b2836f4f724664dc0313ee09e9773
+  BIOPTIM_PRODUCTION_COMMIT: 9ae08f35a95f8ca167b4f05dfdf69b8e643f667e
   BIOPTIM_ALPAQA_COMMIT: d84e7e43534360fc048e0be26a3bd69a2abc2d77
   ACADOS_COMMIT: 59d93e17d2985fdd73fc58b8a83ed8f83a024171
   ACADOS_INSTALL_SCRIPT_BLOB: 5ac8064ab613251e62560b5de8cbbb9550f5c5d0
@@ -564,6 +564,12 @@ jobs:
             fi
           done
           mkdir -p acados-smoke-results
+          : > acados-smoke-results/expected-cases.txt
+          acados_extended=false
+          if (( ACADOS_SMOKE_RHOS > 5 || ACADOS_OPTION_RHOS > 5 )); then
+            acados_extended=true
+          fi
+          echo "$acados_extended" > acados-smoke-results/extended-campaign.txt
           run_case() {
             local variant="$1"
             local mechanics="$2"
@@ -578,6 +584,7 @@ jobs:
             local solver_options=()
             local stationarity=1e-3
             mkdir -p "$case_dir"
+            echo "${variant}-${mechanics}" >> acados-smoke-results/expected-cases.txt
             if [[ "$integrator" == "ERK" ]]; then
               collocation_type=EXPLICIT_RUNGE_KUTTA
             fi
@@ -736,23 +743,35 @@ jobs:
             echo "$solver_exit" > "$case_dir/process-exit-code.txt"
           }
 
-          for mechanics in full reduced; do
-            run_case sqp-irk-reference "$mechanics" "$ACADOS_SMOKE_RHOS" SQP IRK 5 5
-            run_case sqp-irk-reset-memory "$mechanics" "$ACADOS_OPTION_RHOS" SQP IRK 5 5
-            run_case sqp-irk-qp-hot "$mechanics" "$ACADOS_OPTION_RHOS" SQP IRK 5 5
-            run_case sqp-irk-fixed "$mechanics" "$ACADOS_OPTION_RHOS" SQP IRK 5 5
-            run_case sqp-irk-anderson "$mechanics" "$ACADOS_OPTION_RHOS" SQP IRK 5 5
-            run_case sqp-irk-two-stage "$mechanics" "$ACADOS_SMOKE_RHOS" SQP IRK 5 5
-            run_case sqp-feasible-qp-irk "$mechanics" "$ACADOS_SMOKE_RHOS" SQP_WITH_FEASIBLE_QP IRK 5 5
-            run_case sqp-feasibility-qp-irk "$mechanics" "$ACADOS_SMOKE_RHOS" SQP_WITH_FEASIBLE_QP IRK 5 5
-            run_case sqp-rti-irk "$mechanics" "$ACADOS_OPTION_RHOS" SQP_RTI IRK 5 5
-            run_case sqp-erk "$mechanics" "$ACADOS_OPTION_RHOS" SQP ERK 20 5
-            run_case sqp-irk-light "$mechanics" "$ACADOS_OPTION_RHOS" SQP IRK 2 3
-          done
-          run_case sqp-irk-contact full "$ACADOS_OPTION_RHOS" SQP IRK 5 5
-          run_case sqp-rti-irk-contact full "$ACADOS_OPTION_RHOS" SQP_RTI IRK 5 5
[…]
