### 2026-07-30 d8ac9c3
Validate ACADOS initial contact anchoring
