+++ docs/cycling_solver_benchmark/README.md
@@ -57,15 +57,15 @@ provenance humaine.
 
 | Composant | Version Bioptim réellement utilisée |
 |---|---|
-| Construction et certification des seeds | `9ae08f35a95f8ca167b4f05dfdf69b8e643f667e` |
-| IPOPT full/reduced | `9ae08f35a95f8ca167b4f05dfdf69b8e643f667e` |
-| MadNLP/MUMPS full/reduced | `9ae08f35a95f8ca167b4f05dfdf69b8e643f667e` |
-| ACADOS full/reduced et variantes | `9ae08f35a95f8ca167b4f05dfdf69b8e643f667e` |
+| Construction et certification des seeds | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
+| IPOPT full/reduced | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
+| MadNLP/MUMPS full/reduced | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
+| ACADOS full/reduced et variantes | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
 
 Ce commit appartient à la branche dédiée
 `codex/cocofest-acados-v055-exploration`. Il part exactement de
 `3523f1745e315f07761159d7e06bd2d876026704`, utilisé par la campagne publiée,
-et ajoute sans modifier les interfaces MadNLP/Fatrop :
+et rassemble les adaptations communes aux différents solveurs :
 
 - ACADOS `v0.5.5`, sous-module
   `59d93e17d2985fdd73fc58b8a83ed8f83a024171`, contre `v0.5.1` auparavant;
@@ -81,12 +81,21 @@ et ajoute sans modifier les interfaces MadNLP/Fatrop :
   réutiliser la même bibliothèque avec de nouvelles données nodales;
 - l’export des contraintes non linéaires avec les variables décisionnelles
   scalées attendues par les fonctions de pénalité Bioptim. L’ancienne voie
-  appliquait le scaling une seconde fois et créait des résidus artificiels.
+  appliquait le scaling une seconde fois et créait des résidus artificiels;
+- les contraintes non linéaires ACADOS au `Node.START`, nécessaires pour
+  ancrer le contact full sans répéter des égalités redondantes à tous les
+  nœuds;
+- le scaling du guess terminal ACADOS, qui était auparavant injecté en unités
+  physiques alors que les nœuds `0..N-1` étaient correctement scalés;
+- la normalisation des contraintes de continuité FATROP en coordonnées
+  scalées, tout en conservant le resserrement des bornes physiques et
+  l’organisation temporelle des variables.
 
-La branche Bioptim a passé 17 tests ciblés et un test réel ACADOS `v0.5.5` :
-la seconde OCP a réutilisé la bibliothèque compilée tout en recevant de
-nouvelles valeurs aux six nœuds. Le workflow effectue toujours un checkout par
-SHA complet; le nom de branche documente seulement la provenance.
+Les tests FATROP ciblés passent localement (`7/7`). Les tests ACADOS ajoutés
+requièrent l’environnement Linux qui installe `acados_template`; ils sont donc
+validés par le palier CI 5 RHO avant toute promotion à 30 RHO. Le workflow
+effectue toujours un checkout par SHA complet; le nom de branche documente
+seulement la provenance.
 
 Les résultats 100 RHO de la section 11 ont été produits avec l’ancien SHA
 `3523f1745e315f07761159d7e06bd2d876026704`. Ils restent historiques jusqu’à
@@ -177,6 +186,14 @@ A_m^\mathrm{eff}(PW_m)
[…]

[…]
