### 2026-07-30 70870cb
Restore FATROP collocation to staged benchmarks
