### 2026-07-30 48ef359
Enforce and report strict RHO state continuity
