+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -74,7 +74,9 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "wheel_qdot_regularization_target",
     "wheel_qdot_bound_margin",
     "acados_wheel_q_slack",
+    "acados_wheel_qdot_slack",
     "acados_terminal_wheel_q_slack",
+    "rho_state_continuity_mode",
     "max_acados_iterations",
     "acados_tolerance",
     "acados_stationarity_tolerance",

[…]
