### 2026-07-30 e534ca2
Fix FATROP compiled plugin loading
