+++ docs/cycling_solver_benchmark/README.md
@@ -6,10 +6,11 @@ mathématiques, la discrétisation, les stratégies de warm-start, les critères
 de validation, l’organisation GitHub Actions et l’interprétation des résultats.
 
 La campagne courante est volontairement **SX-only** et compare IPOPT/MUMPS,
-MadNLP/MUMPS et ACADOS, en mécanique full et reduced. PARDISO est archivé :
-il n’a apporté aucun gain à MadNLP. Fatrop est également sorti de l’endurance
-courante, car sa formulation full échoue encore en SX lors de l’identification
-de la structure des gaps. Alpaqa reste documenté comme intégration non
+MadNLP/MUMPS et ACADOS, en mécanique full et reduced. FATROP/collocation est
+réintégré comme sonde graduelle sur une machine séparée, mais n’est pas encore
+certifié pour l’endurance : sa formulation full échouait encore en SX lors de
+l’identification de la structure des gaps. PARDISO est archivé, car il n’a
+apporté aucun gain à MadNLP. Alpaqa reste documenté comme intégration non
 fonctionnelle.
 
 Les résultats historiques d’endurance proviennent principalement du
@@ -33,8 +34,10 @@ Ils servent de référence avant la nouvelle campagne séquentielle 5, 30 puis
   runner refuse explicitement une demande MX.
 - MadNLP utilise exclusivement MUMPS, transmis à libMad sous le nom typé
   exact `MumpsSolver`. La CI échoue si libMad signale une option inconnue.
-- PARDISO, Fatrop et RK4 ne font pas partie de la campagne active. Leurs
-  anciens résultats sont conservés comme diagnostics historiques.
+- PARDISO et RK4 ne font pas partie de la campagne active. FATROP/collocation
+  est exécuté dans les paliers 5, 30 puis 100 RHO, mais ses résultats restent
+  diagnostiques tant que les cas full et reduced ne passent pas les mêmes
+  audits physiques que les autres solveurs.
 - MadNLP reste interprété : la compilation C est refusée par l’interface
   Bioptim/CasADi épinglée et n’est pas comptée comme une variante valide.
 - La convergence du solveur ne suffit pas : chaque RHO est soumis à un audit
@@ -60,6 +63,7 @@ provenance humaine.
 | Construction et certification des seeds | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
 | IPOPT full/reduced | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
 | MadNLP/MUMPS full/reduced | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
+| FATROP/collocation full/reduced | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
 | ACADOS full/reduced et variantes | `a3499cab16d7605b8efa7255cf89f1af6a7c59c9` |
 
 Ce commit appartient à la branche dédiée
@@ -899,7 +903,7 @@ Le benchmark garde le transfert des multiplicateurs désactivé. Les blocs duaux
 ne sont pas encore décalés structurellement avec le RHO; le primal décalé est
 le warm-start fiable.
 
-### 7.3 Fatrop — diagnostic historique, hors campagne SX-only
+### 7.3 FATROP — diagnostic actif, certification endurance en attente
 
 Fatrop exploite la structure d’OCP pour résoudre les systèmes linéaires par
 une factorisation de type Riccati plutôt que comme une matrice KKT générique.
@@ -941,11 +945,13 @@ récent; la branche complète n’est pas utilisée car elle est en retard de
 28 commits sur cette base.
 
 Les 7 tests Bioptim dédiés au patch passent localement, mais cela ne suffit
[…]
