### 2026-07-30 cc738fe
Fix shared phase-one benchmark option
