### 2026-07-30 cec91d9
Preserve the certified first RHO seed
