### 2026-07-30 6a57e4b
Keep the full IPOPT reference interpreted
