+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -1506,7 +1506,11 @@ jobs:
 
       - name: Run IPOPT full
         if: matrix.solver == 'ipopt'
-        run: bash .github/scripts/run_cycling_benchmark_case.sh ipopt ipopt full mumps collocation benchmark-results "$BENCHMARK_CYCLES" "${{ inputs.compile_nlp_evaluators }}"
+        # Compiling the full 26-state collocation NLP repeatedly terminates the
+        # hosted runner with SIGTERM 143 before any artifact can be checkpointed.
+        # Keep the scientific full reference interpreted; the reduced case
+        # below certifies persistent C evaluator reuse when requested.
+        run: bash .github/scripts/run_cycling_benchmark_case.sh ipopt ipopt full mumps collocation benchmark-results "$BENCHMARK_CYCLES" false
       - name: Checkpoint IPOPT full
         if: ${{ always() && matrix.solver == 'ipopt' }}
         continue-on-error: true
