### 2026-07-30 acc90a6
Limit transfer phase one to shooting grids
