### 2026-07-30 d78e61a
Report the strict physical RHO prefix
