### 2026-07-30 91087a5
Treat ACADOS nonconvergence as benchmark data
