### 2026-07-30 3248ab1
Track ACADOS reference cases explicitly
