### 2026-07-30 591aada
Keep the common ACADOS full reference
