+++ docs/cycling_solver_benchmark/README.md
@@ -1863,11 +1863,13 @@ cinq. Le gate vérifie désormais la présence de tous les cas, l’usage de SX,
 un `n_windows > 5` réellement transmis et l’absence d’erreur
 d’infrastructure sur les références full/reduced; une non-convergence ACADOS
 reste dans le rapport sans rendre le workflow rouge.
-Le run de vérification `30568985981` a ensuite révélé un second défaut purement
-CI : le cas full retenu après le screen porte le nom
-`sqp-irk-contact-position-full`, pas `sqp-irk-reference-full`. Les deux cas de
-référence sont désormais consignés dans `reference-cases.txt`; le gate lit ce
-manifeste au lieu de reconstruire leurs noms.
+Le run de vérification `30568985981` a ensuite révélé que la campagne longue
+remplaçait le cas full commun par `sqp-irk-contact-position-full`. Cette
+ablation n'a pas le même ensemble admissible et ne peut pas remplacer la
+référence dans une comparaison multi-solveurs. Les cas
+`sqp-irk-reference-full` et `sqp-irk-reference-reduced` sont donc toujours
+exécutés et consignés dans `reference-cases.txt`; les contraintes de contact à
+tous les nœuds restent une variante ACADOS séparée.
 
 Pour mesurer la compilation, relancer sur le même type de runner avec :
 
