### 2026-07-30 9d1073a
Report the physically certified RHO prefix
