### 2026-07-30 f74a796
Document the strict 30 RHO campaign
