### 2026-07-30 da3560a
Document the strict 100 RHO results
