+++ docs/cycling_solver_benchmark/README.md
@@ -6,19 +6,18 @@ mathématiques, la discrétisation, les stratégies de warm-start, les critères
 de validation, l’organisation GitHub Actions et l’interprétation des résultats.
 
 La campagne courante est volontairement **SX-only** et compare IPOPT/MUMPS,
-MadNLP/MUMPS et ACADOS, en mécanique full et reduced. FATROP/collocation est
-réintégré comme sonde graduelle sur une machine séparée, mais n’est pas encore
-certifié pour l’endurance : sa formulation full échouait encore en SX lors de
-l’identification de la structure des gaps. PARDISO est archivé, car il n’a
-apporté aucun gain à MadNLP. Alpaqa reste documenté comme intégration non
-fonctionnelle.
+MadNLP/MUMPS, FATROP/collocation et ACADOS, en mécanique full et reduced.
+FATROP reduced est certifié 100/100; sa formulation full reste une sonde
+négative, bloquée lors de l’identification de la structure des gaps. PARDISO
+est archivé, car il n’a apporté aucun gain à MadNLP. Alpaqa reste documenté
+comme intégration non fonctionnelle.
 
 Les résultats historiques d’endurance proviennent principalement du
 [run Linux 30487321536](https://github.com/mickaelbegon/cocofest/actions/runs/30487321536).
 La preuve comparative SX/MX vient du
 [run Linux 30475768127](https://github.com/mickaelbegon/cocofest/actions/runs/30475768127).
-Ils servent de référence avant la nouvelle campagne séquentielle 5, 30 puis
-100 RHO.
+Ils constituent l’historique antérieur à la campagne séquentielle stricte 5,
+30 puis 100 RHO maintenant terminée.
 
 ## Résumé opérationnel
 
@@ -35,16 +34,16 @@ Ils servent de référence avant la nouvelle campagne séquentielle 5, 30 puis
 - MadNLP utilise exclusivement MUMPS, transmis à libMad sous le nom typé
   exact `MumpsSolver`. La CI échoue si libMad signale une option inconnue.
 - PARDISO et RK4 ne font pas partie de la campagne active. FATROP/collocation
-  est exécuté dans les paliers 5, 30 puis 100 RHO, mais ses résultats restent
-  diagnostiques tant que les cas full et reduced ne passent pas les mêmes
-  audits physiques que les autres solveurs.
+  reduced est certifié sur les paliers 5, 30 puis 100 RHO. FATROP full reste
+  diagnostique tant que l’interface n’accepte pas sa structure de collocation.
 - MadNLP reste interprété : la compilation C est refusée par l’interface
   Bioptim/CasADi épinglée et n’est pas comptée comme une variante valide.
 - La convergence du solveur ne suffit pas : chaque RHO est soumis à un audit
   indépendant de faisabilité physique avec un seuil de `1e-5`.
 - La mécanique réduite est plus rapide, mais le grand écart de fatigue observé
-  n’est pas encore une réduction physiologique démontrée. Les problèmes full
-  et reduced ne sont actuellement pas mécaniquement équivalents.
+  historiquement n’est pas une réduction physiologique démontrée. L’équivalence
+  full/reduced sur l’endurance reste à établir, car aucun préfixe full strict
+  ne dépasse actuellement un cycle.
 
 ## 0. Versions Bioptim et reproductibilité
 
@@ -56,7 +55,7 @@ et effectue chaque checkout par SHA complet. Le SHA est la référence
[…]
