+++ docs/message_kevin_benchmark.md
@@ -2,84 +2,117 @@
 
 Salut Kevin,
 
-Le benchmark du RHO de pédalage est documenté ici :
+Nous avons repris le benchmark du RHO de pédalage en corrigeant plusieurs
+incohérences qui rendaient l’ancienne comparaison full/reduced difficile à
+interpréter. Le rapport détaillé est ici :
 
 [Benchmark des solveurs du RHO de pédalage FES](cycling_solver_benchmark/README.md)
 
-Le problème minimise uniquement la fatigue, sans couple externe. Chaque OCP
-contient un cycle de pédalage et 30 stimulations par muscle; une campagne de
-100 RHO enchaîne donc 100 OCP d’un cycle. Les PW sont bornées dans
-`[pd0, 600 µs]`, avec `pd0 ≈ 131.405 µs`, et la cible terminale d’angle est
-absolue avec un slack de `0.002 rad`.
-La référence absolue est recalée après le chargement du seed, y compris quand
-le seed et le solveur utilisent tous deux la formulation full ou tous deux la
-formulation reduced. Ce point corrige un décalage initial full de `3.94 mrad`
-observé avec cinq statuts IPOPT pourtant égaux à zéro; le slack n’a pas été
-relâché pour masquer cette incohérence.
+Le problème minimise uniquement la fatigue, avec un couple externe nul. Chaque
+OCP contient un cycle de pédalage et 30 stimulations par muscle. Les PW sont
+bornées dans `[pd0, 600 µs]`, avec `pd0 ≈ 131.405 µs`; les anciens seeds sont
+validés et tronqués dans cette plage au chargement. La cible terminale d’angle
+est absolue, avec un slack de `0.002 rad`, afin d’éviter toute dérive cumulative
+du pédalier.
 
-La nouvelle campagne active compare :
+Les changements importants sont :
 
-- IPOPT/MUMPS full et reduced;
-- MadNLP/MUMPS full et reduced;
-- ACADOS full et reduced, avec ses variantes de restauration et RTI.
+- continuité exacte entre deux RHO de la cadence et des 20 états Ding;
+- seed reduced commun, relevé sur la variété de contact pour initialiser le
+  full au même instant physiologique;
+- même loi de force passive en full et reduced;
+- bornes de cadence exprimées dans les coordonnées physiques
+  `theta/omega`, et non directement sur `qdot[2]`;
+- audit indépendant de la variété de contact, de la vitesse tangentielle et de
+  la progression angulaire;
+- séparation entre RHO convergés isolément et préfixe strict réellement
+  exécutable;
+- graphes SX seulement, car ils sont nettement plus rapides que MX sur ce
+  problème.
 
-Tous les graphes de la campagne sont maintenant SX, y compris le warm-up
-IPOPT. ACADOS consomme directement ce seed IPOPT/collocation SX certifié :
-son raffinement IPOPT auxiliaire n’est plus répété dans chaque variante,
-car deux runners ont été arrêtés après environ 150 s pendant sa construction
-full redondante. Ce choix SX est mesuré, pas
-seulement supposé : sur 30 RHO, SX a réduit la médiane chaude de 57.5 à
-60.5 % par rapport à MX, soit une accélération de 2.35× à 2.53×. Les quatre
-comparaisons IPOPT/MadNLP full/reduced ont convergé 30/30, avec un écart
-maximal d’objectif compris entre `8.6e-12` et `5.01e-11`. Le runner refuse
-désormais MX et le rapport vérifie `use_sx=true`.
+La branche Bioptim dédiée est
[…]
