### 2026-07-30 50f3bcd
fix: stabilize full cycling transfers and acados warm starts
