+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -18,6 +18,7 @@ from time import perf_counter
 import traceback
 
 import numpy as np
+from bioptim import SolutionMerge
 
 REPO_ROOT = Path(__file__).resolve().parents[3]
 if str(REPO_ROOT) not in sys.path:
@@ -58,8 +59,13 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "model_formulation",
     "mechanical_formulation",
     "mechanical_equivalence_audit",
+    "full_contact_constraints_terminal",
+    "full_contact_position_terminal",
+    "full_contact_position_tolerance",
     "full_contact_constraints_all_nodes",
     "full_contact_position_all_nodes",
+    "transfer_contact_manifold_projection",
+    "transfer_contact_manifold_projection_mode",
     "torque_application",
     "ode_solver",
     "collocation_degree",
@@ -117,6 +123,9 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "acados_transfer_bound_homotopy_solver_tolerance",
     "acados_transfer_mechanical_restoration",
     "acados_transfer_sqp_restarts",
+    "acados_transfer_active_set_guard_radius",
+    "acados_transfer_active_set_guard_margin",
+    "acados_transfer_active_set_threshold",
     "acados_cyclical_transfer_mode",
     "terminal_wheel_q_reference_mode",
     "cycles_per_window",
@@ -2069,6 +2078,101 @@ def stimulation_pattern_snapshots(
     }
 
 
+def isolated_window_checkpoint_snapshots(
+    result: dict,
+    cycles: tuple[int, ...] = BENCHMARK_STIMULATION_PATTERN_CYCLES,
+) -> dict[str, dict]:
+    """Preserve selected RHO decisions even after the strict prefix stops.
+
+    These checkpoints are deliberately separate from ``stimulation_patterns``:
+    they are numerical diagnostics, not an executed endurance trajectory.
+    Keeping their terminal A states and PW vectors makes the 100th isolated
+    full solve auditable without accidentally certifying all preceding seams.
+    """
+
+    cycles_per_window = int(getattr(result["args"], "cycles_per_window", 1))
+    windows = result.get("window_solutions") or []
+    objectives = result.get("window_objectives") or []
+    feasibility = result.get("window_feasibility") or []
+    strict_prefix = _physically_validated_cycle_count(result)
+    capacity_scales = result.get("fatigue_capacity_scales") or {}
+    snapshots = {}
+    for cycle in cycles:
+        snapshot = {
+            "cycle": int(cycle),
+            "rho": int(cycle) if cycles_per_window == 1 else None,
+            "available": False,
+            "diagnostic_only": True,
+            "belongs_to_strict_prefix": bool(cycle <= strict_prefix),
+            "reason": None,
+            "status": None,
+            "objective": None,
+            "primal_feasible": None,
+            "capacity_states": {},
+            "pulse_width_us": {},
+        }
+        snapshots[f"cycle_{cycle}"] = snapshot
+        if cycles_per_window != 1:
+            snapshot["reason"] = "requires_one_cycle_per_window"
+            continue
+        if cycle < 1:
+            snapshot["reason"] = "cycle_must_be_positive"
+            continue
+        if cycle > len(windows):
+            snapshot["reason"] = f"only_{len(windows)}_windows_were_attempted"
+            continue
+
+        window_index = cycle - 1
+        solution = windows[window_index]
+        snapshot["status"] = getattr(solution, "status", None)
+        if window_index < len(objectives):
+            snapshot["objective"] = _finite_float(objectives[window_index])
+        if window_index < len(feasibility):
+            snapshot["primal_feasible"] = bool(
[…]
