+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1072,6 +1072,34 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--acados-transfer-active-set-guard-radius",
+        type=float,
+        default=None,
+        help=(
+            "Optional larger pulse-width radius, in seconds, applied only near "
+            "circular active/inactive phase transitions after an MHE transfer. "
+            "The ordinary transfer trust radius is retained at every other node."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-active-set-guard-margin",
+        type=int,
+        default=1,
+        help=(
+            "Circular number of neighboring stimulation nodes released on each "
+            "side of an active/inactive pulse-width transition."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-active-set-threshold",
+        type=float,
+        default=1e-6,
+        help=(
+            "Pulse width above the model pd0, in seconds, used to classify the "
+            "phase-aligned transferred control as recruited."
+        ),
+    )
+    parser.add_argument(
         "--acados-fes-state-trust-radius",
         type=float,
         default=None,
@@ -1572,6 +1600,24 @@ def build_argument_parser() -> argparse.ArgumentParser:
             "Reintegrate the appended cycle with the complete dynamics after each "
             "window transfer. This solver-independent rollout is available to IPOPT "
             "and ACADOS."
+        ),
+    )
+    parser.add_argument(
+        "--transfer-contact-manifold-projection",
+        action="store_true",
+        help=(
+            "For full mechanics, project the free first-node q/qdot components "
+            "onto the reduced contact manifold after each RHO transfer while "
+            "preserving the exactly bound crank coordinate and velocity."
+        ),
+    )
+    parser.add_argument(
+        "--transfer-contact-manifold-projection-mode",
+        choices=("position", "position_velocity"),
+        default="position",
+        help=(
+            "Project only the dominant position seam by default, or also "
+            "project the free generalized velocities."
         ),
     )
     parser.add_argument(
@@ -2455,6 +2501,33 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Disable the start-of-window posture constraints.",
     )
     parser.add_argument(
+        "--full-contact-constraints-terminal",
+        action="store_true",
+        help=(
+            "For full mechanics, close wheel-centre position and velocity at "
+            "the terminal node so the next RHO starts on the contact manifold."
+        ),
+    )
+    parser.add_argument(
+        "--full-contact-position-terminal",
+        action="store_true",
+        help=(
+            "For full mechanics, close only wheel-centre position at the "
+            "terminal node; this targets the observed RHO seam without adding "
+            "a terminal velocity constraint."
+        ),
+    )
+    parser.add_argument(
+        "--full-contact-position-tolerance",
+        type=float,
+        default=0.0,
+        metavar="M",
+        help=(
+            "Symmetric spatial tolerance on full wheel-centre position "
+            "constraints. The default 0 keeps the historical equality."
+        ),
+    )
+    parser.add_argument(
         "--full-contact-constraints-all-nodes",
         action="store_true",
         help=(
@@ -3084,6 +3157,15 @@ def _common_initial_solution_metadata(args: argparse.Namespace) -> dict:
         "constant_crank_torque": float(args.constant_crank_torque),
         "torque_application": args.torque_application,
         "enforce_start_constraints": bool(args.enforce_start_constraints),
+        "full_contact_constraints_terminal": bool(
[…]

[…]
