### 2026-07-30 005b8e6
ci: isolate bioptim acados interface test
