### 2026-07-30 4438b42
ci: validate acados warm-start integration
