### 2026-07-30 7183c44
fix: relax reduced mechanics in acados transfer homotopy
