### 2026-07-30 edd12d8
fix: continue acados transfer bounds gradually
