+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -690,7 +690,7 @@ jobs:
               # then run the normal fatigue-optimal SQP solve.
               solver_options+=(
                 --acados-transfer-bound-homotopy
-                --acados-transfer-bound-homotopy-fractions 0,1
+                --acados-transfer-bound-homotopy-fractions 0,0.125,0.25,0.375,0.5,0.625,0.75,0.875,1
                 --acados-transfer-bound-homotopy-padding 0.05
                 --acados-transfer-bound-homotopy-iterations 20
                 --acados-transfer-bound-homotopy-tolerance 1e-4
