### 2026-07-30 8c0f8af
feat: adapt acados transfer continuation steps
