+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1705,6 +1705,24 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--acados-transfer-bound-homotopy-min-fraction-step",
+        type=float,
+        default=0.0,
+        help=(
+            "Minimum interpolation step allowed when a failed transfer-bound "
+            "homotopy stage is bisected. Zero disables adaptive refinement."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-bound-homotopy-max-refinements",
+        type=int,
+        default=0,
+        help=(
+            "Maximum number of adaptive bisections across one transfer-bound "
+            "homotopy. Zero disables adaptive refinement."
+        ),
+    )
+    parser.add_argument(
         "--acados-transfer-sqp-restarts",
         type=int,
         default=0,
@@ -4525,7 +4543,9 @@ def run_acados_control_homotopy(
                 )
                 residuals = diagnostics.get("residuals")
                 summary = {
+                    "sequence_index": len(summaries),
                     "stage": stage_index,
+                    "queue_index": stage_index,
                     "attempt": attempt,
                     "kind": kind,
                     "radius": radius,
@@ -8625,6 +8645,8 @@ def run_acados_transfer_bound_homotopy(
     stage_iterations: int,
     solver_tolerance: float | None = None,
     max_restarts: int = 1,
+    minimum_fraction_step: float = 0.0,
+    max_refinements: int = 0,
     echo: bool = True,
     solve_stage=None,
 ) -> dict:
@@ -8663,8 +8685,14 @@ def run_acados_transfer_bound_homotopy(
     accepted_controls = snapshot_container(periodic_nmpc.nlp[0].u_init)
     summaries = []
     completed = False
+    scheduled_fractions = [float(fraction) for fraction in fractions]
+    accepted_fraction = None
+    refinement_count = 0
+    stage_index = 0
+    termination_reason = "not_started"
     try:
-        for stage_index, fraction in enumerate(fractions):
+        while stage_index < len(scheduled_fractions):
+            fraction = scheduled_fractions[stage_index]
             apply_transfer_state_bound_fraction(
                 periodic_nmpc, original_bounds, relaxed_bounds, fraction
             )
@@ -8718,6 +8746,8 @@ def run_acados_transfer_bound_homotopy(
                     or (
                         intermediate_stage
                         and _status_is_success(solution.status)
+                        and residuals is None
+                        and final_history_residuals is None
                     )
                 )
                 retryable = bool(
@@ -8786,10 +8816,123 @@ def run_acados_transfer_bound_homotopy(
                 reset_acados_solver_memory(periodic_nmpc)
 
             if not stage_accepted:
-                break
+                fraction_step = (
+                    np.inf
+                    if accepted_fraction is None
+                    else fraction - accepted_fraction
+                )
+                can_refine = bool(
+                    accepted_fraction is not None
+                    and refinement_count < max_refinements
+                    and fraction_step > minimum_fraction_step
+                    and fraction_step > np.finfo(float).eps
+                )
+                if can_refine:
+                    refined_fraction = 0.5 * (accepted_fraction + fraction)
+                    if (
+                        refined_fraction - accepted_fraction
+                        < minimum_fraction_step
+                    ):
+                        can_refine = False
+                if not can_refine:
+                    if accepted_fraction is None:
+                        termination_reason = "initial_stage_failed"
+                    elif refinement_count >= max_refinements:
+                        termination_reason = "maximum_refinements_reached"
+                    else:
+                        termination_reason = "minimum_fraction_step_reached"
+                    break
+                apply_transfer_state_bound_fraction(
+                    periodic_nmpc,
+                    original_bounds,
+                    relaxed_bounds,
[…]
