### 2026-07-30 eb96617
fix: constrain full cadence at collocation stages
