+++ tests/shard1/test_periodic_pulse_width.py
@@ -148,6 +148,31 @@ def test_full_contact_position_stabilization_avoids_velocity_redundancy(monkeypa
     assert captured[0][1]["max_bound"] == 2e-5
 
 
+def test_full_cadence_constraint_covers_collocation_stages_and_terminal(monkeypatch):
+    captured = []
+
+    class FakeConstraintList:
+        def add(self, constraint, **kwargs):
+            captured.append((constraint, kwargs))
+
+    monkeypatch.setattr(mhe_example, "ConstraintList", FakeConstraintList)
+    model = SimpleNamespace(marker_index=lambda name: 7)
+
+    mhe_example.set_constraints(
+        model,
+        enforce_start_constraints=False,
+        enforce_physical_crank_velocity_bounds=True,
+    )
+
+    assert len(captured) == 2
+    assert captured[0][0] is (
+        mhe_example.physical_crank_velocity_all_collocation_points_constraint
+    )
+    assert captured[0][1]["node"] == Node.ALL_SHOOTING
+    assert captured[1][0] is mhe_example.physical_crank_velocity_constraint
+    assert captured[1][1]["node"] == Node.END
+
+
 def test_full_transfer_contact_projection_preserves_bound_crank_states():
     class Kinematics:
         @staticmethod
@@ -4284,7 +4309,7 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
         in workflow
     )
     assert "--acados-transfer-bound-homotopy-max-refinements 16" in workflow
-    assert "--acados-transfer-bound-homotopy-iterations 20" in workflow
+    assert "--acados-transfer-bound-homotopy-iterations 40" in workflow
     assert "--acados-transfer-bound-homotopy-solver-tolerance 1e-4" in workflow
     assert "run_case sqp-feasible-qp-irk" in workflow
     assert "run_case sqp-feasibility-qp-irk" in workflow
@@ -4851,6 +4876,81 @@ def test_physical_crank_velocity_uses_center_to_hand_kinematics():
     omega = mhe_example.physical_crank_velocity_constraint(controller)
 
     assert omega == pytest.approx(-5.0)
+
+
+def test_physical_crank_velocity_includes_all_collocation_points():
+    class FakeBioModel:
+        @staticmethod
+        def marker_index(name):
+            return {"hand": 0, "global_wheel_center": 1}[name]
+
+        @staticmethod
+        def marker(index):
+            values = (
+                np.array([1.0, 0.0, 0.0])
+                if index == 0
+                else np.array([0.0, 0.0, 0.0])
+            )
+            return lambda q, parameters: values
+
+        @staticmethod
+        def marker_velocity(index):
+            if index == 0:
+                return lambda q, qdot, parameters: np.array(
+                    [0.0, qdot[0], 0.0]
+                )
+            return lambda q, qdot, parameters: np.zeros(3)
+
+    controller = SimpleNamespace(
+        model=SimpleNamespace(bio_model=FakeBioModel()),
+        states={
+            "q": SimpleNamespace(
+                cx_start=np.zeros(3),
+                cx_intermediates_list=[np.zeros(3), np.zeros(3)],
+            ),
+            "qdot": SimpleNamespace(
+                cx_start=np.array([-5.0, 0.0, 0.0]),
+                cx_intermediates_list=[
+                    np.array([-7.0, 0.0, 0.0]),
+                    np.array([-9.0, 0.0, 0.0]),
+                ],
+            ),
+        },
+        parameters=SimpleNamespace(cx=np.array([])),
+    )
+
[…]
