### 2026-07-30 eb4fb2d
fix: exclude direct-shooting pseudo stages
