### 2026-07-30 90f47da
feat: retain acados iterates and compile madnlp
