+++ tests/shard1/test_periodic_pulse_width.py
@@ -4259,6 +4259,21 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
         'benchmark-results "$BENCHMARK_CYCLES" false'
         in workflow
     )
+    assert (
+        "run_cycling_benchmark_case.sh madnlp-mumps madnlp full mumps collocation "
+        'benchmark-results "$BENCHMARK_CYCLES" false'
+        in workflow
+    )
+    assert (
+        "run_cycling_benchmark_case.sh madnlp-mumps madnlp reduced mumps collocation "
+        'benchmark-results "$BENCHMARK_CYCLES" "${{ inputs.compile_nlp_evaluators }}"'
+        in workflow
+    )
+    assert (
+        'if [[ "$BENCHMARK_SOLVER" == "madnlp" && '
+        '"$result" == *"madnlp-mumps-full/"* ]]'
+        in workflow
+    )
     assert "specified structure of A does not correspond" in workflow
     assert 'case_requires_compile="$COMPILE_NLP_EVALUATORS"' in workflow
     assert 'case_requires_compile=false' in workflow
@@ -4274,7 +4289,7 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
     assert "prepare-acados-stack:" in workflow
     assert (
         "BIOPTIM_PRODUCTION_COMMIT: "
-        "733e442c7b429e20a67a7cf4c2b69694c54513b3"
+        "efd59c39777c83f97058f8d6c1ef472f78f9925d"
     ) in workflow
     assert "ACADOS_COMMIT: 59d93e17d2985fdd73fc58b8a83ed8f83a024171" in workflow
     assert "ACADOS_INSTALL_SCRIPT_BLOB: 5ac8064ab613251e62560b5de8cbbb9550f5c5d0" in workflow
@@ -4363,8 +4378,7 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
         / "run_cycling_benchmark_case.sh"
     ).read_text(encoding="utf-8")
     assert 'solver_options+=(--ipopt-c-compile)' in benchmark_runner
-    assert "MadNLP C compilation is not validated" in benchmark_runner
-    assert 'solver_options+=(--madnlp-c-compile)' not in benchmark_runner
+    assert 'solver_options+=(--madnlp-c-compile)' in benchmark_runner
     assert '--fatrop-state-scaling "$fatrop_state_scaling"' in benchmark_runner
     assert 'case "$graph_mode" in' in benchmark_runner
     assert "The endurance benchmark is SX-only" in benchmark_runner
@@ -4388,7 +4402,7 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
     assert "Compare IPOPT interpreted and compiled evaluators over 5 RHO" not in workflow
     assert "cycling-compile-ablation-" not in workflow
     assert "Compare MadNLP MUMPS interpreted and compiled evaluators" not in workflow
-    assert "Compile IPOPT CasADi evaluators once" in workflow
+    assert "Compile and reuse reduced IPOPT/MadNLP" in workflow
     assert "Checkpoint IPOPT full" in workflow
     assert "Checkpoint MadNLP MUMPS reduced" in workflow
     assert "max-parallel: 2" in workflow
@@ -6243,30 +6257,6 @@ def test_compiled_nlp_tracker_requires_source_at_every_solve(tmp_path, monkeypat
     assert summary["compiled_source_reused"] is False
 
 
-def test_compiled_nlp_plugin_name_patch_only_changes_compiled_interfaces():
-    calls = []
-
-    class FakeInterface:
-        _cocofest_compiled_solver_name_patch = False
-
-        def solve(self, marker):
[…]
