### 2026-07-30 728fbad
fix: align benchmark bioptim revisions
