### 2026-07-31 924acea
docs: report corrected full reduced 30-rho results
