### 2026-07-31 7aee4a0
fix: audit and warm start acados paths
