+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -6535,6 +6535,7 @@ def audit_mechanical_trajectory(
     crank_velocity_target_rad_s: float = DEFAULT_CRANK_QDOT_RAD_S,
     crank_velocity_margin_rad_s: float = 3.0,
     cadence_node_stride: int = 1,
+    shooting_interval_duration_s: float | None = None,
 ) -> tuple[np.ndarray, np.ndarray, dict]:
     """Express either formulation in the same physical crank coordinates."""
 
@@ -6544,6 +6545,16 @@ def audit_mechanical_trajectory(
         raise ValueError("The physical crank-velocity margin must be positive.")
     if cadence_node_stride < 1:
         raise ValueError("cadence_node_stride must be strictly positive.")
+    if (
+        shooting_interval_duration_s is not None
+        and (
+            not np.isfinite(shooting_interval_duration_s)
+            or shooting_interval_duration_s <= 0.0
+        )
+    ):
+        raise ValueError(
+            "shooting_interval_duration_s must be finite and positive."
+        )
     kinematics = reduced_cycling_dynamics.kinematics
     if "theta" in state_traces and "omega" in state_traces:
         theta = np.asarray(state_traces["theta"], dtype=float)
@@ -6594,6 +6605,27 @@ def audit_mechanical_trajectory(
             float(np.max(physical_omega)) - omega_upper,
         )
     )
+    interval_average_omega = np.array([], dtype=float)
+    interval_average_omega_violation = 0.0
+    if shooting_interval_duration_s is not None and projected_theta.shape[1] > 1:
+        # ACADOS enforces state bounds at shooting nodes, not at the internal
+        # Runge--Kutta stages.  The interval-average crank speed is therefore
+        # an independent necessary condition: if Δtheta / Δt is outside the
+        # cadence bounds, at least one internal point necessarily violates
+        # those bounds even when every exported omega node is admissible.
+        interval_average_omega = (
+            np.diff(projected_theta[0]) / float(shooting_interval_duration_s)
+        )
+        interval_average_omega_violation = float(
+            max(
+                0.0,
+                omega_lower - float(np.min(interval_average_omega)),
+                float(np.max(interval_average_omega)) - omega_upper,
+            )
+        )
+    combined_omega_violation = max(
+        all_node_omega_violation, interval_average_omega_violation
+    )
     audit.update(
         {
             "source_formulation": source_formulation,
@@ -6623,17 +6655,33 @@ def audit_mechanical_trajectory(
             "maximum_all_node_crank_velocity_bound_violation_rad_s": (
                 all_node_omega_violation
             ),
+            "interval_average_crank_velocity_available": bool(
+                interval_average_omega.size
+            ),
+            "minimum_interval_average_crank_velocity_rad_s": (
+                float(np.min(interval_average_omega))
+                if interval_average_omega.size
+                else None
+            ),
+            "maximum_interval_average_crank_velocity_rad_s": (
+                float(np.max(interval_average_omega))
+                if interval_average_omega.size
+                else None
+            ),
+            "maximum_interval_average_crank_velocity_bound_violation_rad_s": (
+                interval_average_omega_violation
+            ),
             "passes_configuration_tolerance": (
                 configuration_error <= configuration_tolerance_rad
             ),
             "passes_velocity_tolerance": velocity_error <= velocity_tolerance_rad_s,
             "passes_physical_crank_velocity_bounds": (
-                all_node_omega_violation <= velocity_tolerance_rad_s
+                combined_omega_violation <= velocity_tolerance_rad_s
             ),
             "passes_tolerance": (
                 configuration_error <= configuration_tolerance_rad
                 and velocity_error <= velocity_tolerance_rad_s
-                and all_node_omega_violation <= velocity_tolerance_rad_s
[…]
