### 2026-07-31 41f9a59
feat: audit stiff calcium with refined collocation
