### 2026-07-31 85984c7
fix: preserve partial acados certificates
