### 2026-07-31 bff552e
fix: refine collocation warm starts
