### 2026-07-31 e72ed09
ci: screen acados cadence regularization
