+++ tests/shard1/test_periodic_pulse_width.py
@@ -4416,6 +4416,10 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
     assert "--acados-transfer-bound-homotopy-fractions 0,1" in workflow
     assert "run_case sqp-irk-two-stage-cadence-guard reduced" in workflow
     assert "--wheel-qdot-bound-margin 2.5" in workflow
+    assert "run_case sqp-irk-two-stage-cadence-reg-0p1 reduced" in workflow
+    assert "--acados-wheel-qdot-regularization-weight 0.1" in workflow
+    assert "run_case sqp-irk-two-stage-cadence-reg-1 reduced" in workflow
+    assert "--acados-wheel-qdot-regularization-weight 1" in workflow
     assert "--shared-transfer-contact-projection" in workflow
     assert (
         "--shared-transfer-contact-projection-mode position_velocity"
