### 2026-07-31 1dfbb02
ci: replace infeasible acados cadence guard
