### 2026-07-31 746093c
docs: record acados continuous cadence audit
