### 2026-07-31 ab09f1d
fix: keep robust full acados transfer
