### 2026-07-31 73d662a
fix: preserve collocation stride in physical audit
