### 2026-07-31 cec1b52
ci: allow refined collocation convergence
