+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -1638,7 +1638,11 @@ jobs:
 
       - name: Run IPOPT reduced Radau degree 5
         if: ${{ matrix.solver == 'ipopt' && inputs.refined_collocation_validation }}
-        run: bash .github/scripts/run_cycling_benchmark_case.sh ipopt-radau5 ipopt reduced mumps collocation benchmark-results "$REFINED_COLLOCATION_RHOS" false sx none 5
+        # Degree 5 is a transcription audit, not a real-time candidate.  Its
+        # first two solves needed 1692 and >2000 iterations in run 30594633810,
+        # while remaining primal-feasible to 1e-7.  Give this offline
+        # validation enough iterations to reach optimality.
+        run: BENCHMARK_MAX_ITER=5000 bash .github/scripts/run_cycling_benchmark_case.sh ipopt-radau5 ipopt reduced mumps collocation benchmark-results "$REFINED_COLLOCATION_RHOS" false sx none 5
       - name: Checkpoint IPOPT reduced Radau degree 5
         if: ${{ always() && matrix.solver == 'ipopt' && inputs.refined_collocation_validation }}
         continue-on-error: true
@@ -1677,7 +1681,7 @@ jobs:
 
       - name: Run MadNLP MUMPS reduced Radau degree 5
         if: ${{ matrix.solver == 'madnlp' && inputs.refined_collocation_validation }}
-        run: bash .github/scripts/run_cycling_benchmark_case.sh madnlp-mumps-radau5 madnlp reduced mumps collocation benchmark-results "$REFINED_COLLOCATION_RHOS" false sx none 5
+        run: BENCHMARK_MAX_ITER=5000 bash .github/scripts/run_cycling_benchmark_case.sh madnlp-mumps-radau5 madnlp reduced mumps collocation benchmark-results "$REFINED_COLLOCATION_RHOS" false sx none 5
       - name: Checkpoint MadNLP MUMPS reduced Radau degree 5
         if: ${{ always() && matrix.solver == 'madnlp' && inputs.refined_collocation_validation }}
         continue-on-error: true
