### 2026-07-31 eff08e2
ci: focus the 100 RHO acados campaign
