### 2026-07-31 e86c1fb
ci: seed regularized ACADOS from feasible full RHO
