### 2026-07-31 12baa73
ci: retain bounded refined-collocation audit
