### 2026-07-31 5f80483
Refine MadNLP seed before compiled RHO loop
