### 2026-07-31 19bd26f
Complete full reduced endurance validation
