+++ tests/shard1/test_periodic_pulse_width.py
@@ -51,7 +51,9 @@ def _muscle_model():
     )
 
 
-def test_full_contact_stabilization_constrains_every_shooting_node(monkeypatch):
+def test_full_contact_stabilization_constrains_every_shooting_node(
+    monkeypatch,
+):
     captured = []
 
     class FakeConstraintList:
@@ -101,7 +103,9 @@ def test_full_contact_terminal_stabilization_closes_only_rho_seam(monkeypatch):
     assert captured[3][1]["second_marker"] == "global_wheel_center"
 
 
-def test_full_contact_terminal_position_avoids_velocity_redundancy(monkeypatch):
+def test_full_contact_terminal_position_avoids_velocity_redundancy(
+    monkeypatch,
+):
     captured = []
 
     class FakeConstraintList:
@@ -125,7 +129,9 @@ def test_full_contact_terminal_position_avoids_velocity_redundancy(monkeypatch):
     assert captured[-1][1]["first_marker"] == "wheel_center"
 
 
-def test_full_contact_position_stabilization_avoids_velocity_redundancy(monkeypatch):
+def test_full_contact_position_stabilization_avoids_velocity_redundancy(
+    monkeypatch,
+):
     captured = []
 
     class FakeConstraintList:
@@ -148,7 +154,9 @@ def test_full_contact_position_stabilization_avoids_velocity_redundancy(monkeypa
     assert captured[0][1]["max_bound"] == 2e-5
 
 
-def test_full_cadence_constraint_covers_collocation_stages_and_terminal(monkeypatch):
+def test_full_cadence_constraint_covers_collocation_stages_and_terminal(
+    monkeypatch,
+):
     captured = []
 
     class FakeConstraintList:
@@ -191,7 +199,14 @@ def test_full_transfer_contact_projection_preserves_bound_crank_states():
     q = np.array([[0.0, 1.0], [0.0, 1.0], [4.0, 5.0]])
     qdot = np.array([[0.0, 1.0], [0.0, 1.0], [-6.0, -7.0]])
     nmpc = SimpleNamespace(
-        nlp=[SimpleNamespace(x_init={"q": SimpleNamespace(init=q), "qdot": SimpleNamespace(init=qdot)})],
+        nlp=[
+            SimpleNamespace(
+                x_init={
+                    "q": SimpleNamespace(init=q),
+                    "qdot": SimpleNamespace(init=qdot),
+                }
+            )
+        ],
         wheel_state_index=2,
         _cocofest_mechanical_equivalence_dynamics=SimpleNamespace(
             kinematics=Kinematics()
@@ -211,17 +226,13 @@ def test_full_transfer_contact_projection_preserves_bound_crank_states():
     np.testing.assert_allclose(nmpc.nlp[0].x_init["q"].init[:, 1], [1.0, 1.0, 5.0])
     np.testing.assert_allclose(nmpc.nlp[0].x_init["qdot"].init[:, 1], [1.0, 1.0, -7.0])
 
-    nmpc.nlp[0].x_init["q"].init[:, :] = np.array(
-        [[0.0, 1.0], [0.0, 1.0], [4.0, 5.0]]
-    )
+    nmpc.nlp[0].x_init["q"].init[:, :] = np.array([[0.0, 1.0], [0.0, 1.0], [4.0, 5.0]])
     nmpc.nlp[0].x_init["qdot"].init[:, :] = np.array(
         [[0.0, 1.0], [0.0, 1.0], [-6.0, -7.0]]
     )
     summary = mhe_example.project_full_first_node_initial_guess_to_contact(nmpc)
     assert summary["mode"] == "position"
[…]
