+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -54,6 +54,7 @@ BENCHMARK_SOLVERS = ("ipopt", "acados", "fatrop", "madnlp")
 BENCHMARK_STIMULATION_PATTERN_CYCLES = (10, 30, 100)
 BENCHMARK_CONFIGURATION_FIELDS = (
     "solver",
+    "single_shot",
     "objective",
     "objective_shape",
     "model_formulation",
@@ -148,6 +149,7 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "legacy_standard_warmup_seed_signed_torque",
     "common_initial_solution",
     "common_initial_solution_output",
+    "allow_partial_receding_horizon_solution_output",
     "ipopt_hsl_library",
     "ipopt_c_compile",
     "ipopt_print_level",
@@ -488,9 +490,7 @@ def _shooting_node_physical_trace(
             f"and stride {stride}."
         )
     requested_cycles = min(int(cycle_count), available_cycles)
-    return values[
-        ..., : requested_cycles * shooting_per_cycle * stride + 1 : stride
-    ]
+    return values[..., : requested_cycles * shooting_per_cycle * stride + 1 : stride]
 
 
 def _truncate_result_to_cycles(result: dict, cycle_count: int) -> dict:
@@ -694,9 +694,7 @@ def _executed_fatigue_objective(
 ) -> float | None:
     """Return the sum of the independently re-evaluated muscle costs."""
 
-    rows = _executed_fatigue_objective_by_muscle(
-        result, cycle_count, weight=weight
-    )
+    rows = _executed_fatigue_objective_by_muscle(result, cycle_count, weight=weight)
     if not rows:
         return None
     return float(sum(row["executed_fatigue_objective"] for row in rows))
@@ -751,7 +749,9 @@ def _external_crank_power_metrics(result: dict, cycle_count: int) -> dict:
     role = (
         "driving"
         if mean_power > power_tolerance
-        else "resistive" if mean_power < -power_tolerance else "neutral"
+        else "resistive"
+        if mean_power < -power_tolerance
+        else "neutral"
     )
     metrics.update(
         mean_power_w=mean_power,
@@ -1161,7 +1161,8 @@ def _shared_initial_guess_comparison(ipopt_result: dict, acados_result: dict) ->
             common_values += ipopt_array.size
             if ipopt_array.size:
                 max_error = max(
-                    max_error, float(np.max(np.abs(ipopt_array - acados_array)))
+                    max_error,
+                    float(np.max(np.abs(ipopt_array - acados_array))),
                 )
 
     ipopt_audits = ipopt_result.get("initial_guess_audits") or []
@@ -1190,6 +1191,7 @@ def _pick(value, fallback):
 
 def _solver_config(
     solver_name: str,
+    single_shot: bool,
     objective: str,
     objective_shape: str,
     cycles_per_window: int,
@@ -1289,7 +1291,7 @@ def _solver_config(
 
         return _namespace_from_cli(
             solver="ipopt",
-            single_shot=False,
+            single_shot=single_shot,
             model_formulation=_pick(
                 ipopt_model_formulation, defaults["model_formulation"]
             ),
@@ -1407,7 +1409,7 @@ def _solver_config(
     if solver_name == "acados":
         return _namespace_from_cli(
             solver="acados",
-            single_shot=False,
+            single_shot=single_shot,
             model_formulation="periodic_node",
             torque_application="constant",
             cycles_per_window=cycles_per_window,
@@ -1422,9 +1424,7 @@ def _solver_config(
             acados_control_homotopy_stage_iterations=(
                 acados_control_homotopy_stage_iterations
             ),
-            acados_control_homotopy_max_restarts=(
-                acados_control_homotopy_max_restarts
-            ),
[…]
