### 2026-07-31 de5b4a5
Fix full horizon workflow dispatch
