### 2026-07-31 8fb127c
Fix full horizon log path handling
