### 2026-07-31 1563f08
Retry ACADOS from the best feasible iterate
