+++ tests/shard1/test_periodic_pulse_width.py
@@ -2,6 +2,7 @@ import importlib.util
 import numpy as np
 import pytest
 import re
+import warnings
 from casadi import Function, SX
 from pathlib import Path
 from types import SimpleNamespace
@@ -890,6 +891,323 @@ def test_acados_iterate_storage_is_opt_in():
     assert periodic_example._codegen_signature(
         parser.parse_args([])
     ) != periodic_example._codegen_signature(stored_args)
+
+
+def test_acados_conditional_maxiter_retry_options_are_parsed():
+    parser = periodic_example.build_argument_parser()
+    args = parser.parse_args(
+        [
+            "--acados-store-iterates",
+            "--acados-maxiter-retries",
+            "1",
+            "--acados-maxiter-retry-iterations",
+            "20",
+            "--acados-maxiter-retry-feasibility-tolerance",
+            "0.0025",
+        ]
+    )
+
+    assert args.acados_store_iterates is True
+    assert args.acados_maxiter_retries == 1
+    assert args.acados_maxiter_retry_iterations == 20
+    assert args.acados_maxiter_retry_feasibility_tolerance == pytest.approx(0.0025)
+
+
+def test_acados_maxiter_retry_candidate_requires_nearly_feasible_history():
+    diagnostics = {
+        "res_stat_all": np.array([20.0, 0.2, 0.3]),
+        "res_eq_all": np.array([0.4, 0.0019, 0.003]),
+        "res_ineq_all": np.array([0.01, 1e-8, 1e-7]),
+        "res_comp_all": np.array([0.0, 1e-9, 1e-8]),
+    }
+
+    candidate = periodic_example._acados_maxiter_retry_candidate(
+        2, diagnostics, feasibility_tolerance=0.0025
+    )
+    rejected = periodic_example._acados_maxiter_retry_candidate(
+        2, diagnostics, feasibility_tolerance=0.001
+    )
+
+    assert candidate["eligible"] is True
+    assert candidate["best_index"] == 1
+    assert candidate["best_feasibility"] == pytest.approx(0.0019)
+    assert rejected["eligible"] is False
+    assert rejected["reason"] == "best_iterate_not_nearly_feasible"
+    assert periodic_example._acados_maxiter_retry_candidate(
+        0, diagnostics, feasibility_tolerance=0.0025
+    ) == {"eligible": False, "reason": "status_not_maxiter"}
+
+
+def test_conditional_maxiter_retry_ignores_successful_main_solve():
+    class FakeAcadosInterface:
+        status = -1
+        real_time_to_optimize = 0.1
+
+        def solve(self):
+            self.status = 0
+            return self.get_optimized_value()
+
+        def get_optimized_value(self):
+            return {
+                "status": self.status,
+                "solver_time_to_optimize": 0.1,
+                "real_time_to_optimize": self.real_time_to_optimize,
+                "iter": 3,
+            }
+
+    interface = FakeAcadosInterface()
+    nmpc = SimpleNamespace(
+        ocp_solver=interface,
+        total_optimization_run=4,
+        _cocofest_acados_main_window_retry_armed=True,
+    )
+    summaries = []
+
+    periodic_example.install_acados_conditional_maxiter_retry(
+        nmpc,
+        max_retries=1,
+        retry_iterations=20,
+        feasibility_tolerance=0.0025,
+        nominal_iterations=100,
+        summaries=summaries,
+        echo=False,
+        residual_diagnostics_function=lambda current_nmpc: pytest.fail(
+            "A successful solve must not inspect the residual history."
+        ),
+    )
+
+    output = interface.solve()
+
+    assert output["status"] == 0
+    assert summaries == []
+
+
+def test_conditional_maxiter_retry_replaces_solution_and_accounts_full_budget():
[…]
