+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -15,6 +15,7 @@ import re
 import sys
 from sys import platform as sys_platform
 from time import perf_counter
+from types import MethodType
 import warnings
 
 import numpy as np
@@ -1779,6 +1780,31 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help=(
             "Retain every SQP iterate so a restart can select an intermediate "
             "primal. This substantially increases memory use on long MHE runs."
+        ),
+    )
+    parser.add_argument(
+        "--acados-maxiter-retries",
+        type=int,
+        default=0,
+        help=(
+            "Maximum number of conditional retries after ACADOS_MAXITER. Each "
+            "retry starts from the stored iterate with the lowest feasibility "
+            "residual and leaves the physical OCP bounds unchanged."
+        ),
+    )
+    parser.add_argument(
+        "--acados-maxiter-retry-iterations",
+        type=int,
+        default=20,
+        help="Additional SQP iteration budget for each conditional MAXITER retry.",
+    )
+    parser.add_argument(
+        "--acados-maxiter-retry-feasibility-tolerance",
+        type=float,
+        default=2.5e-3,
+        help=(
+            "Largest dynamics, inequality, or complementarity residual allowed "
+            "for the stored iterate used by a conditional MAXITER retry."
         ),
     )
     parser.add_argument(
@@ -4480,7 +4506,10 @@ def reset_acados_solver_memory(periodic_nmpc) -> bool:
 
 
 def apply_acados_capsule_primal_to_initial_guess(
-    periodic_nmpc, iterate_index: int | None = None
+    periodic_nmpc,
+    iterate_index: int | None = None,
+    *,
+    require_stored_iterate: bool = False,
 ) -> dict:
     """Copy the current scaled ACADOS primal, including failed iterates, to Bioptim."""
 
@@ -4498,6 +4527,13 @@ def apply_acados_capsule_primal_to_initial_guess(
             stored_iterate = get_iterate(-1 if iterate_index is None else iterate_index)
         except (IndexError, RuntimeError, TypeError, ValueError) as exc:
             stored_iterate_error = str(exc)
+    if require_stored_iterate and stored_iterate is None:
+        return {
+            "applied": False,
+            "reason": "stored_iterate_unavailable",
+            "iterate_index": iterate_index,
+            "stored_iterate_error": stored_iterate_error,
+        }
 
     nlp = periodic_nmpc.nlp[0]
     first_state_key = next(iter(nlp.x_init.keys()))
@@ -8798,6 +8834,311 @@ def _acados_residual_history_summary(diagnostics: dict) -> dict:
     }
 
 
+def _acados_interface_residual_diagnostics(periodic_nmpc) -> dict:
+    """Snapshot only the native residual data needed by a post-MAXITER retry."""
+
+    interface = getattr(periodic_nmpc, "ocp_solver", None)
+    acados_solver = getattr(interface, "ocp_solver", None)
+    if acados_solver is None:
+        return {}
+    return {
+        field: deepcopy(_safe_acados_stat(acados_solver, field))
+        for field in (
+            "sqp_iter",
+            "time_tot",
+            "residuals",
+            "res_stat_all",
+            "res_eq_all",
+            "res_ineq_all",
+            "res_comp_all",
+        )
+    }
+
+
+def _acados_stat_scalar(value, default: float = 0.0) -> float:
+    """Return a native scalar statistic without relying on ndarray truthiness."""
+
+    if value is None or isinstance(value, dict):
+        return float(default)
+    values = np.asarray(value, dtype=float).reshape(-1)
+    if values.size != 1 or not np.isfinite(values[0]):
+        return float(default)
[…]

[…]
