### 2026-07-31 8bd6728
Record active sets around failed RHO windows
