### 2026-07-31 356ad30
Strengthen full horizon warm start diagnostics
