+++ .github/scripts/run_full_horizon_benchmark.py
@@ -499,7 +499,11 @@ def _full_horizon_command(
         "madnlp",
         *_common_solver_options(args),
         "--ipopt-no-use-sx",
+        "--ipopt-disable-standard-warmup",
         "--optional-nlp-periodic-ipopt-hot-start",
+        "--periodic-ipopt-refinement-use-sx",
+        "--periodic-ipopt-refinement-iterations",
+        str(args.max_iterations),
         "--single-shot",
         "--cycles-per-window",
         str(cycles),
