### 2026-07-31 3ea4a9c
Preserve full horizon seed chronology
