### 2026-07-31 f718bd1
Restore complete ACADOS retry iterates
