+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -15,7 +15,7 @@ import re
 import sys
 from sys import platform as sys_platform
 from time import perf_counter
-from types import MethodType
+from types import MethodType, SimpleNamespace
 import warnings
 
 import numpy as np
@@ -8825,7 +8825,10 @@ def apply_transfer_state_bound_fraction(
     periodic_nmpc._sync_acados_state_bounds()
 
 
-def _acados_residual_history_summary(diagnostics: dict) -> dict:
+def _acados_residual_history_summary(
+    diagnostics: dict,
+    feasibility_tolerance: float | None = None,
+) -> dict:
     rows = []
     for key in ("res_stat_all", "res_eq_all", "res_ineq_all", "res_comp_all"):
         values = diagnostics.get(key)
@@ -8843,20 +8846,36 @@ def _acados_residual_history_summary(diagnostics: dict) -> dict:
     feasibility = np.max(np.abs(history[1:]), axis=0)
     stationarity = np.abs(history[0])
     candidate_indices = np.flatnonzero(finite_columns)
+    selection = "minimum_feasibility"
+    if feasibility_tolerance is not None:
+        feasible_candidates = np.flatnonzero(
+            finite_columns & (feasibility <= feasibility_tolerance)
+        )
+        if feasible_candidates.size:
+            candidate_indices = feasible_candidates
+            selection = "minimum_stationarity_within_feasibility"
+            sort_keys = (
+                feasibility[candidate_indices],
+                stationarity[candidate_indices],
+            )
+        else:
+            sort_keys = (
+                stationarity[candidate_indices],
+                feasibility[candidate_indices],
+            )
+    else:
+        sort_keys = (
+            stationarity[candidate_indices],
+            feasibility[candidate_indices],
+        )
     best_index = int(
-        candidate_indices[
-            np.lexsort(
-                (
-                    stationarity[candidate_indices],
-                    feasibility[candidate_indices],
-                )
-            )[0]
-        ]
+        candidate_indices[np.lexsort(sort_keys)[0]]
     )
     return {
         "initial": history[:, 0],
         "best": history[:, best_index],
         "best_index": best_index,
+        "selection": selection,
         "componentwise_best": np.min(np.abs(history), axis=1),
         "final": history[:, -1],
     }
@@ -8903,7 +8922,9 @@ def _acados_maxiter_retry_candidate(
 
     if status != 2:
         return {"eligible": False, "reason": "status_not_maxiter"}
-    history = _acados_residual_history_summary(diagnostics)
+    history = _acados_residual_history_summary(
+        diagnostics, feasibility_tolerance=feasibility_tolerance
+    )
     if not history:
         return {"eligible": False, "reason": "residual_history_unavailable"}
     best = np.asarray(history["best"], dtype=float).reshape(-1)
@@ -8924,6 +8945,7 @@ def _acados_maxiter_retry_candidate(
             "best_residuals": best,
             "best_feasibility": best_feasibility,
             "final_feasibility": final_feasibility,
+            "selection": history["selection"],
         }
     return {
         "eligible": True,
@@ -8932,7 +8954,152 @@ def _acados_maxiter_retry_candidate(
         "best_residuals": best,
         "best_feasibility": best_feasibility,
         "final_feasibility": final_feasibility,
+        "selection": history["selection"],
     }
+
+
+def capture_acados_stored_primal_dual_iterate(
+    periodic_nmpc,
+    iterate_index: int,
+) -> tuple[dict, dict | None]:
+    """Detach a complete stored ACADOS iterate for restoration after a QP reset."""
+
+    interface = getattr(periodic_nmpc, "ocp_solver", None)
[…]
