+++ tests/shard1/test_periodic_pulse_width.py
@@ -938,6 +938,102 @@ def test_acados_maxiter_retry_candidate_requires_nearly_feasible_history():
     ) == {"eligible": False, "reason": "status_not_maxiter"}
 
 
+def test_acados_maxiter_retry_prefers_stationarity_within_feasible_candidates():
+    diagnostics = {
+        "res_stat_all": np.array([20.0, 0.4, 0.01]),
+        "res_eq_all": np.array([0.4, 1e-5, 0.002]),
+        "res_ineq_all": np.array([0.01, 1e-8, 1e-7]),
+        "res_comp_all": np.array([0.0, 1e-9, 1e-8]),
+    }
+
+    candidate = periodic_example._acados_maxiter_retry_candidate(
+        2, diagnostics, feasibility_tolerance=0.0025
+    )
+
+    assert candidate["eligible"] is True
+    assert candidate["best_index"] == 2
+    assert candidate["selection"] == "minimum_stationarity_within_feasibility"
+
+
+def test_capture_and_queue_acados_stored_primal_dual_iterate():
+    fields = ("x", "u", "pi", "lam", "sl", "su")
+
+    class StoredIterate:
+        def __init__(self):
+            self.arrays = {
+                field: np.full(2, fields.index(field) + 1.0) for field in fields
+            }
+
+        def flatten(self):
+            return SimpleNamespace(**self.arrays)
+
+    class Capsule:
+        stored = StoredIterate()
+
+        def get_iterate(self, index):
+            assert index == 8
+            return self.stored
+
+        def get_dim_flat(self, field):
+            if field == "z":
+                return 0
+            assert field in fields
+            return 2
+
+    class Interface:
+        def __init__(self):
+            self.ocp_solver = Capsule()
+            self.queued = None
+
+        def get_solver_state(self):
+            return {
+                "solver": "ACADOS",
+                "format_version": 1,
+                "n_horizon": 30,
+                "iterate": {field: np.zeros(2) for field in fields},
+            }
+
+        def set_lagrange_multiplier(self, solution):
+            self.queued = solution.solver_state
+
+    interface = Interface()
+    nmpc = SimpleNamespace(ocp_solver=interface)
+
+    summary, state = periodic_example.capture_acados_stored_primal_dual_iterate(
+        nmpc, 8
+    )
+    queued = periodic_example.queue_acados_primal_dual_solver_state(nmpc, state)
+    interface.ocp_solver.stored.arrays["x"][:] = -99.0
+
+    assert summary["captured"] is True
+    assert summary["field_sizes"] == {field: 2 for field in fields}
+    assert queued == {"queued": True, "reason": None}
+    for index, field in enumerate(fields):
+        np.testing.assert_allclose(interface.queued["iterate"][field], index + 1.0)
+
+
+def test_capture_rejects_acados_algebraic_states():
+    class Capsule:
+        def get_iterate(self, index):
+            return SimpleNamespace(flatten=lambda: SimpleNamespace())
+
+        def get_dim_flat(self, field):
+            return 1 if field == "z" else 0
+
+    interface = SimpleNamespace(
+        ocp_solver=Capsule(),
+        get_solver_state=lambda: {"iterate": {}},
+    )
+
+    summary, state = periodic_example.capture_acados_stored_primal_dual_iterate(
+        SimpleNamespace(ocp_solver=interface), 0
+    )
+
+    assert summary["captured"] is False
+    assert summary["reason"] == "algebraic_states_not_supported"
+    assert state is None
+
+
 def test_conditional_maxiter_retry_ignores_successful_main_solve():
     class FakeAcadosInterface:
         status = -1
[…]
