### 2026-07-31 2370aa1
Match full horizon warmup to seed size
