### 2026-07-31 fdc38c5
Test feasibility-first ACADOS warm starts
