+++ .github/scripts/run_full_horizon_benchmark.py
@@ -311,6 +311,7 @@ def _benchmark_success(
     expected_mode: str,
     expected_cycles: int,
     expected_solver: str,
+    expected_mechanics: str | None = None,
 ) -> bool:
     """Require a solver and physical certificate for the complete horizon."""
 
@@ -318,7 +319,10 @@ def _benchmark_success(
         payload = json.loads(result_path.read_text(encoding="utf-8"))
         result = payload["results"][0]
         configuration = payload["configurations"][expected_solver]
-        expected_mechanics = "full" if expected_mode == "single_shot" else "reduced"
+        if expected_mechanics is None:
+            expected_mechanics = (
+                "full" if expected_mode == "single_shot" else "reduced"
+            )
         expected_use_sx = expected_solver == "ipopt"
         return bool(
             result["success"]
@@ -488,7 +492,11 @@ def _full_horizon_command(
     seed_path: Path,
     result_path: Path,
     solution_path: Path,
+    *,
+    mechanical_formulation: str = "full",
 ) -> list[str]:
+    if mechanical_formulation not in {"full", "reduced"}:
+        raise ValueError("mechanical_formulation must be 'full' or 'reduced'.")
     command = [
         args.python,
         str(
@@ -513,6 +521,11 @@ def _full_horizon_command(
     command.extend(
         [
         "--optional-nlp-periodic-ipopt-hot-start",
+        # Despite its historical name, this flag also prints solver-neutral
+        # initial-guess defects.  Keeping the block/FES/RK4 diagnostics in the
+        # full-horizon artifact is essential when the reduced-to-full bridge
+        # is geometrically accurate but the NLP refinement still diverges.
+        "--acados-diagnostics",
         "--periodic-ipopt-refinement-use-sx",
         "--periodic-ipopt-refinement-iterations",
         str(args.max_iterations),
@@ -522,7 +535,7 @@ def _full_horizon_command(
         "--n-windows",
         str(cycles),
         "--mechanical-formulation",
-        "full",
+        mechanical_formulation,
         "--full-contact-position-tolerance",
         "2e-5",
         "--common-initial-solution",
@@ -541,6 +554,8 @@ def _attempt_record(
     phase: str,
     monitored: MonitoredRun,
     result_path: Path,
+    *,
+    mechanical_formulation: str = "full",
 ) -> dict:
     unknown_mumps_warning = _log_has_unknown_mumps_warning(
         Path(monitored.log_path)
@@ -550,6 +565,7 @@ def _attempt_record(
         expected_mode="single_shot",
         expected_cycles=cycles,
         expected_solver="madnlp",
+        expected_mechanics=mechanical_formulation,
     )
     infrastructure_error = bool(
         not monitored.memory_limit_exceeded
@@ -579,6 +595,7 @@ def _attempt_record(
     )
     return {
         "cycles": cycles,
+        "mechanical_formulation": mechanical_formulation,
         "phase": phase,
         "success": success,
         "failure_kind": failure_kind,
@@ -622,9 +639,34 @@ def _write_markdown(path: Path, report: dict) -> None:
         ),
         f"- Arrêt : `{report['stop_reason']}`",
         "",
+        "## Contrôle apparié reduced/MX",
+        "",
+    ]
+    paired_attempts = report.get("paired_reduced_control_attempts", [])
+    if paired_attempts:
+        paired_last = paired_attempts[-1]
+        lines.extend(
+            [
+                (
+                    f"- Horizon : `{paired_last['cycles']} cycles`; "
+                    f"succès : `{'oui' if paired_last['success'] else 'non'}`; "
+                    f"échec : `{paired_last.get('failure_kind') or '—'}`; "
[…]
