### 2026-07-31 76ca1b6
Use implemented ACADOS feasibility search
