### 2026-07-31 0605e5e
Test certified ACADOS dual preservation
