+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -100,6 +100,9 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "acados_hessian_approx",
     "acados_nlp_solver_type",
     "acados_search_direction_mode",
+    "acados_dual_warm_start_mode",
+    "acados_disable_direction_mode_switch_to_nominal",
+    "acados_use_constraint_hessian_in_feas_qp",
     "acados_store_iterates",
     "acados_maxiter_retries",
     "acados_maxiter_retry_iterations",
