+++ docs/cycling_solver_benchmark/README.md
@@ -3309,6 +3309,53 @@ le mode effectivement appliqué ainsi que les options
 d'un point non admissible contamine le RHO suivant et rend l'ablation
 reproductible.
 
+Le run
+[`30608917200`](https://github.com/mickaelbegon/cocofest/actions/runs/30608917200),
+au SHA `0605e5eca1d804185f7e6fa90df47f302d605239`, ne montre pas de gain
+robuste :
+
+| Byrd–Omojokun full | Dual reset | Dual preserve certifié |
+|---|---:|---:|
+| Préfixe strict | `13/100` | `13/20` |
+| Échecs dans les 20 premiers RHO | `14, 16, 18, 20` | `14, 16, 18, 20` |
+| Médiane chaude mur-à-mur | `0.592 s` | `0.623 s` |
+| P90 chaud mur-à-mur | `1.003 s` | `1.188 s` |
+| Fatigue exécutée du préfixe | `132.646468` | `132.647337` |
+
+Les résidus finals montrent pourquoi une moyenne seule serait trompeuse. Au
+RHO 16, `preserve` réduit la dynamique de `4.70e-3` à `1.43e-4` et la
+stationnarité de `0.458` à `0.056`; au RHO 18, il dégrade au contraire la
+dynamique de `5.68e-3` à `7.53e-3`; au RHO 20, elle passe de `9.90e-7` à
+`3.18e-6`. Aucun de ces changements ne déplace le premier échec. De plus, le
+premier RHO, pourtant explicitement en `reset` dans les deux capsules et
+chargé depuis la même signature de seed, prend respectivement `50` et `87`
+itérations. Sur les 20 RHO communs, `preserve` ajoute `49` itérations et
+`2.426 s` de solveur, soit `+8.8 %`.
+
+Cette différence au premier RHO précède cependant l'activation de
+`preserve` : l'homotopie contrôle initiale échoue dans les deux cas avec des
+résidus différents, puis la capsule est réutilisée. Le mode `reset` annule
+`pi/lam`, mais ne recrée pas la capsule et ne garantit pas la remise à zéro
+des slacks et de toute la mémoire QP interne. L'exécution OpenMP peut ensuite
+amplifier de faibles différences d'ordre numérique sur ce NLP sensible. Une
+ablation causale plus stricte devra recréer ou réinitialiser complètement la
+capsule après cette homotopie, imposer un thread, puis répéter chaque cas au
+moins trois fois. Les petits écarts actuels ne doivent pas être
+surinterprétés.
+
+Conclusion : la conservation duale peut améliorer spectaculairement un RHO
+isolé, mais elle n'est ni stable ni plus rapide dans cette campagne. Elle ne
+devient pas le réglage par défaut. Le prochain essai utile doit agir sur le
+primal transféré avant le solve, ou tester séparément l'interdiction du retour
+automatique vers `NOMINAL_QP`, plutôt que cumuler ces facteurs.
+
+Le premier artefact de cette ablation a aussi révélé une lacune de
+sérialisation : les modes étaient correctement visibles dans le log, mais le
+rapport agrégé ne recopiait que les duals des solveurs NLP. Le sérialiseur
+conserve désormais également `acados_dual_warm_start_summaries` dans
+`warm_start.dual_summaries` et joint le mode effectivement appliqué à chaque
+ligne `windows`.
+
 Cette garantie dépend de l'interface actuelle : aucun `solver_state` ACADOS
 n'est mis en attente après l'avance de fenêtre. Si Bioptim active plus tard
 une restauration native des multiplicateurs ACADOS, il faudra appliquer ce
[…]
