### 2026-07-31 bdd0254
Audit collocation warm starts across NLP solvers
