+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -1523,6 +1523,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         help="Print ACADOS residuals, SQP/QP stats and finite-value checks after the solve.",
     )
     parser.add_argument(
+        "--initial-guess-diagnostics",
+        action="store_true",
+        help=(
+            "Audit the physical primal seed before solving, independently of the "
+            "selected backend (bounds, kinematics, periodic Ding rollout, and "
+            "complete-dynamics RK4 defects)."
+        ),
+    )
+    parser.add_argument(
         "--acados-print-level",
         type=int,
         default=0,
@@ -5327,7 +5336,66 @@ def _trajectory_bounds_for_guess(bounds, n_nodes: int) -> tuple[np.ndarray, np.n
     return lower, upper
 
 
-def print_initial_guess_diagnostics(nmpc) -> None:
+def _initial_guess_shooting_node_indices(nmpc) -> tuple[np.ndarray, int]:
+    """Locate shooting endpoints in RK or collocation initial-guess layouts."""
+
+    nlp = nmpc.nlp[0]
+    if not nlp.x_init.keys() or not nlp.u_init.keys():
+        return np.array([], dtype=int), 0
+    first_state_key = next(iter(nlp.x_init.keys()))
+    first_control_key = next(iter(nlp.u_init.keys()))
+    n_state_nodes = int(nlp.x_init[first_state_key].init.shape[1])
+    n_control_nodes = int(nlp.u_init[first_control_key].init.shape[1])
+    if n_control_nodes < 1 or n_state_nodes < 2:
+        return np.array([], dtype=int), 0
+    interval_columns, remainder = divmod(n_state_nodes - 1, n_control_nodes)
+    if remainder or interval_columns < 1:
+        raise ValueError(
+            "Cannot identify shooting nodes in the initial guess: "
+            f"{n_state_nodes} state columns for {n_control_nodes} controls."
+        )
+    indices = np.arange(0, n_state_nodes, interval_columns, dtype=int)
+    if indices.size != n_control_nodes + 1 or indices[-1] != n_state_nodes - 1:
+        raise ValueError(
+            "The inferred initial-guess shooting-node layout does not end at "
+            f"the terminal state (stride={interval_columns})."
+        )
+    return indices, interval_columns
+
+
+def _lift_shooting_endpoint_update_to_state_columns(
+    original: np.ndarray,
+    projected_endpoints: np.ndarray,
+    shooting_indices: np.ndarray,
+) -> np.ndarray:
+    """Preserve collocation-stage structure while moving shooting endpoints."""
+
+    if projected_endpoints.shape != (original.shape[0], shooting_indices.size):
+        raise ValueError("Projected endpoint values do not match the shooting grid.")
+    updated = np.array(original, dtype=float, copy=True)
+    corrections = projected_endpoints - original[:, shooting_indices]
+    for interval, (start, stop) in enumerate(
+        zip(shooting_indices[:-1], shooting_indices[1:], strict=True)
+    ):
+        width = int(stop - start)
+        updated[:, start] = projected_endpoints[:, interval]
+        for offset in range(1, width):
+            # Radau's last internal stage represents the interval endpoint;
+            # interpolating the correction also remains benign for other
+            # collocation layouts while preserving each original stage shape.
+            fraction = offset / max(width - 1, 1)
+            correction = (
+                (1.0 - fraction) * corrections[:, interval]
+                + fraction * corrections[:, interval + 1]
+            )
+            updated[:, start + offset] += correction
[…]
