### 2026-07-31 88da5e3
Test ACADOS terminal bound homotopy
