### 2026-07-31 67e09d3
Add focused ACADOS homotopy campaign
