### 2026-07-31 49f1915
Fix ACADOS terminal homotopy seed contract
