+++ tests/shard1/test_periodic_pulse_width.py
@@ -1847,6 +1847,101 @@ def test_common_initial_solution_metadata_rejects_a_looser_start_seed(
         )
 
 
+def test_common_initial_solution_metadata_allows_a_stricter_terminal_seed(
+    tmp_path,
+):
+    args = SimpleNamespace(
+        model_formulation="periodic_node",
+        mechanical_formulation="full",
+        cycles_per_window=1,
+        stimulations_per_cycle=30,
+        objective="fatigue",
+        objective_shape="quadratic",
+        constant_crank_torque=0.0,
+        torque_application="constant",
+        enforce_start_constraints=False,
+        acados_wheel_q_slack=0.0,
+        acados_terminal_wheel_q_slack=0.01,
+        terminal_wheel_q_reference_mode="absolute_initial",
+        pulse_width_scaling=0.0025,
+        pulse_width_active_set="none",
+        ode_solver="collocation",
+        nlp_ordering_strategy="time_major",
+        solver="acados",
+        warmup_cycles_consumed=1,
+    )
+    metadata = periodic_example._common_initial_solution_metadata(args)
+    metadata["terminal_wheel_q_slack"] = 0.002
+    seed = periodic_example._WarmupSolutionAdapter({}, {}, metadata=metadata)
+
+    periodic_example._validate_common_initial_solution_metadata(
+        seed, args, tmp_path / "strict-terminal-common.npz"
+    )
+
+
+def test_common_initial_solution_metadata_records_the_terminal_homotopy_target():
+    args = SimpleNamespace(
+        model_formulation="periodic_node",
+        mechanical_formulation="full",
+        cycles_per_window=1,
+        stimulations_per_cycle=30,
+        objective="fatigue",
+        objective_shape="quadratic",
+        constant_crank_torque=0.0,
+        torque_application="constant",
+        enforce_start_constraints=False,
+        acados_wheel_q_slack=0.0,
+        acados_terminal_wheel_q_slack=0.01,
+        acados_terminal_wheel_q_homotopy_slacks=(0.01, 0.005, 0.002),
+        terminal_wheel_q_reference_mode="absolute_initial",
+        pulse_width_scaling=0.0025,
+        pulse_width_active_set="none",
+        ode_solver="collocation",
+        nlp_ordering_strategy="time_major",
+        solver="acados",
+        warmup_cycles_consumed=1,
+    )
+
+    metadata = periodic_example._common_initial_solution_metadata(args)
+
+    assert metadata["terminal_wheel_q_slack"] == 0.002
+    assert metadata["terminal_wheel_q_initial_slack"] == 0.01
+    assert metadata["terminal_wheel_q_homotopy_slacks"] == [0.01, 0.005, 0.002]
+
+
+def test_common_initial_solution_metadata_rejects_a_looser_terminal_seed(
+    tmp_path,
+):
+    args = SimpleNamespace(
+        model_formulation="periodic_node",
+        mechanical_formulation="full",
+        cycles_per_window=1,
+        stimulations_per_cycle=30,
+        objective="fatigue",
+        objective_shape="quadratic",
+        constant_crank_torque=0.0,
+        torque_application="constant",
+        enforce_start_constraints=False,
+        acados_wheel_q_slack=0.0,
+        acados_terminal_wheel_q_slack=0.002,
+        terminal_wheel_q_reference_mode="absolute_initial",
+        pulse_width_scaling=0.0025,
+        pulse_width_active_set="none",
+        ode_solver="collocation",
+        nlp_ordering_strategy="time_major",
+        solver="acados",
+        warmup_cycles_consumed=1,
+    )
[…]
