+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -83,6 +83,9 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "acados_wheel_q_slack",
     "acados_wheel_qdot_slack",
     "acados_terminal_wheel_q_slack",
+    "acados_terminal_wheel_q_target_slack",
+    "acados_terminal_wheel_q_homotopy_slacks",
+    "acados_terminal_wheel_q_homotopy_each_window",
     "rho_state_continuity_mode",
     "max_acados_iterations",
     "acados_tolerance",

[…]
