### 2026-07-31 0793f9e
Separate NLP and physical crank audits
