### 2026-07-31 8fcc647
Fix multi-cycle terminal crank target
