### 2026-07-31 cb855d4
Fix multi-cycle terminal warm-start bounds
