### 2026-07-31 9d6fc9c
Finalize terminal seeds across ACADOS continuations
