+++ tests/shard1/test_periodic_pulse_width.py
@@ -3233,6 +3233,7 @@ def test_terminal_wheel_bound_continuation_tightens_accepted_stages(
     applied_slacks = []
     applied_solutions = []
     nmpc = SimpleNamespace(
+        _cocofest_terminal_wheel_q_slack_scale=0.5,
         nlp=[
             SimpleNamespace(
                 x_bounds={
@@ -3288,7 +3289,13 @@ def test_terminal_wheel_bound_continuation_tightens_accepted_stages(
     )
 
     assert [item["accepted"] for item in summaries] == [True, True, True]
-    assert applied_slacks == [0.2, 0.1, 0.02, 0.02]
+    assert applied_slacks == [0.1, 0.05, 0.01, 0.01]
+    assert [item["slack"] for item in summaries] == [0.2, 0.1, 0.02]
+    assert [item["applied_wheel_q_slack"] for item in summaries] == [
+        0.1,
+        0.05,
+        0.01,
+    ]
     assert applied_solutions == [0, 0, 0]
     assert nmpc._cocofest_dual_warm_start_mode == "preserve"
     assert nmpc._cocofest_terminal_wheel_q_center == -1.0
@@ -6400,7 +6407,18 @@ def test_mechanical_audit_rejects_hidden_acados_interval_cadence_violation():
     assert audit["passes_tolerance"] is False
 
 
-def test_acados_mechanical_audit_excludes_failed_tail_from_validated_prefix():
+@pytest.mark.parametrize(
+    ("absolute_offset", "absolute_reference", "expected_nlp_issues"),
+    (
+        (0.0, None, []),
+        (0.01, 0.0, ["wheel_absolute_progress_out_of_bounds"]),
+    ),
+)
+def test_acados_mechanical_audit_excludes_failed_tail_from_validated_prefix(
+    absolute_offset,
+    absolute_reference,
+    expected_nlp_issues,
+):
     class FakeKinematics:
         @staticmethod
         def lift_generalized_trajectory(theta, omega):
@@ -6430,7 +6448,16 @@ def test_acados_mechanical_audit_excludes_failed_tail_from_validated_prefix():
         ),
         "mode": "rho",
         "state_traces": {
-            "theta": np.array([[0.0, -np.pi, -2.0 * np.pi, -20.0]]),
+            "theta": np.array(
+                [
+                    [
+                        absolute_offset,
+                        absolute_offset - np.pi,
+                        absolute_offset - 2.0 * np.pi,
+                        absolute_offset - 20.0,
+                    ]
+                ]
+            ),
             "omega": np.full((1, 4), -2.0 * np.pi),
         },
         "window_statuses": [0, 4],
@@ -6439,7 +6466,13 @@ def test_acados_mechanical_audit_excludes_failed_tail_from_validated_prefix():
             {"passes_tolerance": False},
         ],
         "covered_cycles": 2,
-        "diagnostics": {"is_physical": True, "issues": []},
+        "diagnostics": {
+            "is_physical": True,
+            "issues": [],
+            "absolute_cycle_reference": absolute_reference,
+            "cycle_progress_tolerance": 0.00402,
+            "absolute_cycle_tolerance": 0.00201,
+        },
         "physical_success": True,
         "success": True,
     }
@@ -6453,6 +6486,10 @@ def test_acados_mechanical_audit_excludes_failed_tail_from_validated_prefix():
     assert audit["audited_validated_cycles"] == 1
     assert audit["passes_tolerance"] is True
     assert summary["physical_crank_angle_trace"].shape == (3,)
[…]

[…]
