### 2026-07-31 85b20e6
Add exact NLP and projected ACADOS warm-start audits
