+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -45,6 +45,7 @@ from cocofest.optimization.receding_horizon_initial_guess import (
     copy_container_values,
     project_initial_guess_to_bounds,
     snapshot_container,
+    snapshot_initial_guess,
 )
 from cocofest.optimization.solver_backends import (
     NLP_SOLVER_NAMES,
@@ -1534,6 +1535,15 @@ def build_argument_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
+        "--exact-initial-nlp-audit",
+        action="store_true",
+        help=(
+            "Evaluate the canonical shaked g(x0) and submitted variable bounds "
+            "immediately before every NLP solve. Diagnostic only; its cost is "
+            "reported separately and excluded from solver timing."
+        ),
+    )
+    parser.add_argument(
         "--acados-print-level",
         type=int,
         default=0,
@@ -1889,6 +1899,58 @@ def build_argument_parser() -> argparse.ArgumentParser:
         type=float,
         default=1.0,
         help="Reject the inter-window rollout when a state exceeds its bounds by more than this value.",
+    )
+    parser.add_argument(
+        "--acados-transfer-select-projected-candidate",
+        action="store_true",
+        help=(
+            "Compare the box-projected shifted and ACADOS-IRK rollout seeds, "
+            "then keep the rollout only when its normalized mechanical defects "
+            "pass the separate q/qdot guards. The historical raw mixed-unit "
+            "rollout threshold remains unchanged when this option is absent."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-selector-max-q-bound-violation-rad",
+        type=float,
+        default=1.0,
+        help="Maximum raw angular bound violation accepted before projecting the rollout [rad].",
+    )
+    parser.add_argument(
+        "--acados-transfer-selector-max-qdot-bound-violation-rad-s",
+        type=float,
+        default=12.0,
+        help="Maximum raw angular-velocity bound violation accepted before projection [rad/s].",
+    )
+    parser.add_argument(
+        "--acados-transfer-selector-max-other-scaled-bound-violation",
+        type=float,
+        default=1.0,
+        help=(
+            "Maximum non-mechanical raw bound violation divided by the corresponding "
+            "ACADOS state scaling (dimensionless)."
+        ),
+    )
+    parser.add_argument(
+        "--acados-transfer-selector-max-scaled-q-defect",
+        type=float,
+        default=0.1,
+        help="Maximum projected rollout q dynamics defect divided by the q state scale.",
+    )
+    parser.add_argument(
+        "--acados-transfer-selector-max-scaled-qdot-defect",
+        type=float,
+        default=0.1,
+        help="Maximum projected rollout qdot dynamics defect divided by the qdot state scale.",
+    )
+    parser.add_argument(
+        "--acados-transfer-selector-improvement-ratio",
+        type=float,
+        default=0.95,
+        help=(
+            "Require the rollout mechanical score to be no larger than this fraction "
+            "of the shifted-seed score; must lie in (0, 1]."
+        ),
     )
     parser.add_argument(
         "--warmup-state-comparison-limit",
@@ -6551,6 +6613,28 @@ def snapshot_nlp_solver_stats(nmpc) -> dict:
     }
 
 
+def enable_exact_initial_nlp_audit(nmpc, solver) -> None:
+    """Enable Bioptim's opt-in canonical pre-solve feasibility audit."""
+
+    nmpc.set_ocp_solver(solver)
+    interface = getattr(nmpc, "ocp_solver", None)
+    enable = getattr(interface, "enable_initial_nlp_audit", None)
[…]
