+++ tests/shard1/test_periodic_pulse_width.py
@@ -5506,10 +5506,10 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
     assert "inputs.cycles != 'screen' && inputs.cycles != 'acados'" in workflow
     assert "prepare-acados-stack:" in workflow
     assert (
-        "BIOPTIM_PRODUCTION_COMMIT: " "efd59c39777c83f97058f8d6c1ef472f78f9925d"
+        "BIOPTIM_PRODUCTION_COMMIT: " "dad96b90d47c36126c1e97ec35f27c499abf4b12"
     ) in workflow
     assert (
-        workflow.count("bioptim_commit: efd59c39777c83f97058f8d6c1ef472f78f9925d") == 3
+        workflow.count("bioptim_commit: dad96b90d47c36126c1e97ec35f27c499abf4b12") == 3
     )
     assert "a3499cab16d7605b8efa7255cf89f1af6a7c59c9" not in workflow
     assert "ACADOS_COMMIT: 59d93e17d2985fdd73fc58b8a83ed8f83a024171" in workflow
@@ -5538,6 +5538,8 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
     assert "--acados-anderson-activation-threshold 0.1" in workflow
     assert "--acados-check-reuse-possible" in workflow
     assert "--acados-byrd-omojokon-slack-relaxation-factor 1.00001" in workflow
+    assert "run_case sqp-byrd-projected-selector-cadence-reg-1-irk" in workflow
+    assert "--acados-transfer-select-projected-candidate" in workflow
     assert "run_case sqp-irk-two-stage" in workflow
     assert (
         "--acados-transfer-bound-homotopy-fractions "
@@ -5632,7 +5634,7 @@ def test_github_acados_runner_uses_reference_and_option_profiles_sequentially():
     focused_campaign = workflow.split(
         'if [[ "$ACADOS_HOMOTOPY_ONLY" == "true" ]]; then', maxsplit=1
     )[1].split('elif [[ "$acados_long" == "true" ]]; then', maxsplit=1)[0]
-    assert focused_campaign.count("run_case ") == 4
+    assert focused_campaign.count("run_case ") == 5
     assert "run_case sqp-irk-reference full 1" in focused_campaign
     assert "run_case sqp-byrd-omojokun-cadence-reg-1-irk full" in focused_campaign
     assert (
@@ -6975,6 +6977,39 @@ def test_control_bounds_summary_preserves_physical_units():
     assert periodic_example._control_bounds_summary(nmpc) == {
         "last_pulse_width_Biceps": {"lower": 0.0001, "upper": 0.0007}
     }
+
+
+def test_exact_initial_nlp_audit_is_enabled_and_attached_without_mutation():
+    audits = [
+        {
+            "solver": "madnlp",
+            "constraints": {"maximum_bound_violation": 0.25},
+            "variables": {"maximum_bound_violation": 0.0},
+            "evaluation_time_s": 0.01,
+        }
+    ]
+
+    class FakeInterface:
+        initial_nlp_audits = audits
+
+        def enable_initial_nlp_audit(self):
+            self.enabled = True
+
+    nmpc = SimpleNamespace(ocp_solver=None)
+
+    def set_ocp_solver(solver):
+        nmpc.ocp_solver = solver
+
+    nmpc.set_ocp_solver = set_ocp_solver
+    interface = FakeInterface()
+
+    periodic_example.enable_exact_initial_nlp_audit(nmpc, interface)
+    summary = {}
+    periodic_example.attach_exact_initial_nlp_audits(summary, nmpc)
+
+    assert interface.enabled is True
[…]
