### 2026-07-31 2aa3633
Fix GitHub rendering of benchmark equations
