+++ docs/cycling_solver_benchmark/README.md
@@ -171,24 +171,24 @@ une modification ultérieure du README.
 
 Pour chaque muscle $m$, le modèle de Ding avec fatigue utilise cinq états :
 
-$$
+```math
 x_m =
 \begin{bmatrix}
 C_{N,m} & F_m & A_m & \tau_{1,m} & K_{\mathrm M,m}
 \end{bmatrix}^{\mathsf T}.
-$$
+```
 
 Ils représentent respectivement le complexe calcium-troponine, la force, la
 capacité de production de force et deux paramètres dynamiques qui évoluent avec
 la fatigue. Avec quatre muscles,
 
-$$
+```math
 n_{\mathrm{Ding}} = 4 \times 5 = 20.
-$$
+```
 
 La dynamique de force peut être écrite sous la forme :
 
-$$
+```math
 \dot F_m =
 \left[
 A_m^\mathrm{eff}(PW_m)
@@ -198,7 +198,7 @@ A_m^\mathrm{eff}(PW_m)
 {\tau_{1,m}+\tau_2\frac{C_{N,m}}{K_{\mathrm M,m}+C_{N,m}}}
 \right]
 \left(f_{\ell,m} f_{v,m}+f_{\mathrm{passif},m}\right),
-$$
+```
 
 Dans cette implémentation, le terme dit « passif » multiplie donc la dynamique
 de force; il ne correspond pas à une tension passive positive simplement
@@ -210,36 +210,36 @@ effet comme physiologique.
 
 où l’effet de la largeur d’impulsion est
 
-$$
+```math
 A_m^\mathrm{eff}(PW_m)
 =
 A_m\left[
 1-\exp\left(-\frac{PW_m-pd0_m}{pdt_m}\right)
 \right].
-$$
+```
 
 Les trois états lents de fatigue suivent :
 
-$$
+```math
 \dot A_m
 =
 -\frac{A_m-A_{\mathrm{scale},m}}{\tau_{\mathrm{fat},m}}
 +\alpha_{A,m}F_m,
-$$
+```
 
-$$
+```math
 \dot \tau_{1,m}
 =
 -\frac{\tau_{1,m}-\tau_{1,\mathrm{rest},m}}{\tau_{\mathrm{fat},m}}
 +\alpha_{\tau_1,m}F_m,
-$$
+```
 
-$$
+```math
 \dot K_{\mathrm M,m}
 =
 -\frac{K_{\mathrm M,m}-K_{\mathrm M,\mathrm{rest},m}}{\tau_{\mathrm{fat},m}}
 +\alpha_{K_m,m}F_m.
-$$
+```
 
 Dans les paramètres courants, $\alpha_A<0$, tandis que
 $\alpha_{\tau_1}>0$ et $\alpha_{K_m}>0$.
@@ -269,25 +269,25 @@ encore le faire.
 
 Chaque muscle possède 30 commandes de largeur d’impulsion par cycle :
 
-$$
+```math
 u_k =
 \begin{bmatrix}
 PW_{1,k} & PW_{2,k} & PW_{3,k} & PW_{4,k}
 \end{bmatrix}^{\mathsf T},
 \qquad k=0,\ldots,29.
-$$
+```
 
 Les bornes physiques sont
 
-$$
+```math
 pd0_m \le PW_{m,k} \le 600\ \mu\mathrm{s}.
-$$
+```
 
 Pour les paramètres Ding courants,
 
-$$
+```math
 pd0 \simeq 131.405\ \mu\mathrm{s}.
-$$
+```
 
 `pd0` est le vrai zéro de recrutement du modèle. Une commande inactive doit
[…]
