### 2026-08-01 46fa20a
Restructure RHO benchmark documentation
