+++ docs/cycling_solver_benchmark/development_history.md
@@ -1,0 +1,4024 @@
+# Historique du benchmark des solveurs du RHO de pédalage FES
+
+> [!IMPORTANT]
+> Ce document est une archive chronologique des développements, campagnes CI,
+> erreurs découvertes et options réfutées. Il ne définit plus la méthode à
+> utiliser. La configuration recommandée et ses critères scientifiques sont
+> décrits dans le [README du benchmark](README.md). Une ancienne « référence »
+> reste documentée ici pour assurer la traçabilité, mais elle ne doit pas être
+> reproduite lorsqu'une erreur de modèle ou de discrétisation a été démontrée.
+> Le [point de reprise](resume_and_todo.md) contient le bilan condensé et les
+> prochaines tâches; il doit être lu avant de reprendre une campagne.
+
+Ce document est la source de référence pour le benchmark de fatigue du
+pédalage FES. Il regroupe le problème scientifique, les formulations
+mathématiques, la discrétisation, les stratégies de warm-start, les critères
+de validation, l’organisation GitHub Actions et l’interprétation des résultats.
+
+La campagne courante est volontairement **SX-only** et compare IPOPT/MUMPS,
+MadNLP/MUMPS, FATROP/collocation et ACADOS, en mécanique full et reduced.
+FATROP reduced est certifié 100/100; sa formulation full reste une sonde
+négative, bloquée lors de l’identification de la structure des gaps. PARDISO
+est archivé, car il n’a apporté aucun gain à MadNLP. Alpaqa reste documenté
+comme intégration non fonctionnelle.
+
+Les résultats historiques d’endurance proviennent principalement du
+[run Linux 30487321536](https://github.com/mickaelbegon/cocofest/actions/runs/30487321536).
+La preuve comparative SX/MX vient du
+[run Linux 30475768127](https://github.com/mickaelbegon/cocofest/actions/runs/30475768127).
+Ils constituent l’historique antérieur à la campagne séquentielle stricte 5,
+30 puis 100 RHO maintenant terminée.
+
+## Résumé opérationnel
+
+- L’objectif est uniquement de minimiser la fatigue musculaire.
+- Le problème courant utilise un couple externe nul, `0 N.m`.
+- Un OCP contient un cycle de pédalage et 30 stimulations par muscle.
+- Un benchmark de 100 RHO résout donc 100 OCP successifs, et non un OCP de
+  100 cycles.
+- Les quatre muscles fournissent 20 états de Ding; la mécanique complète porte
+  le total à 26 états et la mécanique réduite à 22.
+- La référence NLP utilise une collocation Radau de degré 3.
+- Tous les OCP et warm-starts IPOPT de la campagne active utilisent SX. Le
+  runner refuse explicitement une demande MX.
+- MadNLP utilise exclusivement MUMPS, transmis à libMad sous le nom typé
+  exact `MumpsSolver`. La CI échoue si libMad signale une option inconnue.
+- PARDISO et RK4 ne font pas partie de la campagne active. FATROP/collocation
+  reduced est certifié sur les paliers 5, 30 puis 100 RHO. FATROP full reste
+  diagnostique tant que l’interface n’accepte pas sa structure de collocation.
+- MadNLP full reste interprété. MadNLP reduced dispose maintenant d’un chemin
+  C expérimental; la CI doit certifier un seul build, sa réutilisation et
[…]

[…]
