+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -2887,6 +2887,10 @@ def _warmup_cache_signature(
             args.wheel_qdot_regularization_target,
         )["role"],
         "torque_application": args.torque_application,
+        "activate_passive_force_relationship": getattr(
+            args, "activate_passive_force_relationship", True
+        ),
+        "ding_sum_stim_truncation": getattr(args, "ding_sum_stim_truncation", 6),
         "state_scaling": args.state_scaling,
         "pulse_width_scaling": args.pulse_width_scaling,
         "warmup_ipopt_linear_solver": _warmup_ipopt_linear_solver(args),
@@ -2947,6 +2951,16 @@ def _standard_warmup_metadata(args: argparse.Namespace) -> dict:
         "signed_crank_torque_nm": float(args.constant_crank_torque),
         "crank_torque_role": torque["role"],
         "torque_application": str(args.torque_application),
+        "producer_ode_solver": str(args.ode_solver),
+        "producer_collocation_degree": (
+            None
+            if getattr(args, "collocation_degree", None) is None
+            else int(args.collocation_degree)
+        ),
+        "producer_collocation_method": getattr(args, "collocation_method", None),
+        "activate_passive_force_relationship": bool(
+            getattr(args, "activate_passive_force_relationship", True)
+        ),
     }
 
 
@@ -3159,6 +3173,10 @@ def _periodic_ipopt_refinement_cache_path(
         "kind": "periodic_ipopt_refinement",
         "cache_version": int(cache_version),
         "model_formulation": args.model_formulation,
+        "activate_passive_force_relationship": getattr(
+            args, "activate_passive_force_relationship", True
+        ),
+        "ding_sum_stim_truncation": getattr(args, "ding_sum_stim_truncation", 6),
         "mechanical_formulation": args.mechanical_formulation,
         "cycles_per_window": args.cycles_per_window,
         "stimulations_per_cycle": args.stimulations_per_cycle,
@@ -3369,6 +3387,30 @@ def _common_initial_solution_metadata(args: argparse.Namespace) -> dict:
         "pulse_width_maximum_s": 0.0006,
         "warmup_cycles_consumed": int(getattr(args, "warmup_cycles_consumed", 0)),
         "ode_solver": args.ode_solver,
+        "producer_transcription_profile": getattr(
+            args, "transcription_profile", None
+        ),
+        "producer_collocation_degree": (
+            None
+            if getattr(args, "collocation_degree", None) is None
+            else int(args.collocation_degree)
+        ),
+        "producer_collocation_method": getattr(args, "collocation_method", None),
+        "calcium_forcing_formulation": getattr(
+            args, "calcium_forcing_formulation", None
+        ),
+        "ding_sum_stim_truncation": int(
+            getattr(args, "ding_sum_stim_truncation", 6)
+        ),
+        "activate_force_length_relationship": bool(
+            getattr(args, "activate_force_length_relationship", True)
+        ),
+        "activate_force_velocity_relationship": bool(
+            getattr(args, "activate_force_velocity_relationship", True)
+        ),
+        "activate_passive_force_relationship": bool(
+            getattr(args, "activate_passive_force_relationship", True)
+        ),
         "nlp_ordering_strategy": getattr(args, "nlp_ordering_strategy", None),
         "producer_solver": args.solver,
     }
@@ -3659,6 +3701,10 @@ def _horizon_seed_cache_signature(args: argparse.Namespace) -> str:
         "cache_version": 5,
         "nmpc_builder_version": 2,
         "model_formulation": args.model_formulation,
+        "activate_passive_force_relationship": getattr(
+            args, "activate_passive_force_relationship", True
+        ),
[…]
