+++ examples/fes_multibody/cycling/cycling_fes_solver_comparison.py
@@ -8,6 +8,7 @@ periodic formulation. The default objective is fatigue only.
 from __future__ import annotations
 
 import argparse
+import hashlib
 import json
 import os
 import platform
@@ -56,6 +57,13 @@ BENCHMARK_SOLVERS = ("ipopt", "acados", "fatrop", "madnlp")
 BENCHMARK_STIMULATION_PATTERN_CYCLES = (10, 30, 100)
 BENCHMARK_CONFIGURATION_FIELDS = (
     "solver",
+    "benchmark_profile",
+    "transcription_profile",
+    "ipopt_profile",
+    "profile_version",
+    "profile_hash",
+    "profile_integrity",
+    "scientific_status",
     "single_shot",
     "objective",
     "objective_shape",
@@ -73,6 +81,19 @@ BENCHMARK_CONFIGURATION_FIELDS = (
     "ode_solver",
     "collocation_degree",
     "collocation_method",
+    "calcium_forcing_formulation",
+    "calcium_initialization_regime",
+    "calcium_tau_s",
+    "calcium_stimulation_interval_s",
+    "calcium_post_stimulation_amplitude",
+    "calcium_analytical_periodic_value",
+    "ding_sum_stim_truncation",
+    "ding_states_per_muscle",
+    "muscle_names",
+    "control_decisions_per_cycle",
+    "activate_force_length_relationship",
+    "activate_force_velocity_relationship",
+    "activate_passive_force_relationship",
     "use_sx",
     "nlp_ordering_strategy",
     "state_scaling",
@@ -246,7 +267,31 @@ IPOPT_PROFILE_DEFAULTS = {
         "disable_periodic_fes_warmup_projection": False,
         "fatigue_warmstart_mode": None,
     },
+    "scientific_radau5": {
+        "model_formulation": "periodic_node",
+        "torque_application": "constant",
+        "ode_solver": "collocation",
+        "rk_steps": 1,
+        "collocation_degree": 5,
+        "collocation_method": "radau",
+        "use_sx": True,
+        "enforce_start_constraints": True,
+        "disable_standard_ipopt_warmup": False,
+        "disable_periodic_fes_warmup_projection": False,
+        "fatigue_warmstart_mode": None,
+    },
 }
+IPOPT_PROFILE_VERSION = 1
+
+
+def _profile_hash(profile: str) -> str:
+    contract = {
+        "name": profile,
+        "version": IPOPT_PROFILE_VERSION,
+        "defaults": IPOPT_PROFILE_DEFAULTS[profile],
+    }
+    encoded = json.dumps(contract, sort_keys=True, separators=(",", ":")).encode()
+    return hashlib.sha256(encoded).hexdigest()
 
 
 def _namespace_from_cli(**overrides) -> argparse.Namespace:
@@ -1290,7 +1335,7 @@ def _solver_config(
         if normalized_profile not in IPOPT_PROFILE_DEFAULTS:
             raise ValueError(
                 "--ipopt-profile must be 'historical', 'periodic_collocation', "
-                "or 'acados_like'."
+                "'scientific_radau5', or 'acados_like'."
             )
 
         defaults = IPOPT_PROFILE_DEFAULTS[normalized_profile]
@@ -1302,8 +1347,18 @@ def _solver_config(
         if fatigue_warmstart_default is None:
             fatigue_warmstart_default = acados_fatigue_warmstart_mode
 
-        return _namespace_from_cli(
+        solver_args = _namespace_from_cli(
             solver="ipopt",
+            benchmark_profile=normalized_profile.replace("_", "-"),
+            ipopt_profile=normalized_profile,
+            transcription_profile=normalized_profile.replace("_", "-"),
+            profile_version=IPOPT_PROFILE_VERSION,
[…]
