+++ .github/scripts/summarize_cycling_benchmark.py
@@ -16,6 +16,14 @@ COMPARABILITY_FIELDS = (
     "model_formulation",
     "mechanical_formulation",
     "torque_application",
+    "ode_solver",
+    "collocation_method",
+    "calcium_forcing_formulation",
+    "ding_sum_stim_truncation",
+    "activate_force_length_relationship",
+    "activate_force_velocity_relationship",
+    "activate_passive_force_relationship",
+    "control_decisions_per_cycle",
     "wheel_qdot_regularization_target",
     "wheel_qdot_bound_margin",
     "collocation_degree",

[…]
