### 2026-08-02 44093c2
Harden scientific Radau benchmark gate
