+++ .github/scripts/run_cycling_benchmark_case.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-if [[ "$#" -lt 5 || "$#" -gt 12 ]]; then
-  echo "usage: $0 CASE SOLVER MECHANICS BACKEND ODE [ROOT] [WINDOWS] [COMPILE] [GRAPH] [FATROP_SCALING] [COLLOCATION_DEGREE] [IPOPT_PROFILE]" >&2
+if [[ "$#" -lt 5 || "$#" -gt 14 ]]; then
+  echo "usage: $0 CASE SOLVER MECHANICS BACKEND ODE [ROOT] [WINDOWS] [COMPILE] [GRAPH] [FATROP_SCALING] [COLLOCATION_DEGREE] [IPOPT_PROFILE] [DUAL_WARM_START] [TARGET_REFINEMENT]" >&2
   exit 2
 fi
 
@@ -18,6 +18,8 @@ graph_mode="${9:-sx}"
 fatrop_state_scaling="${10:-none}"
 collocation_degree="${11:-3}"
 ipopt_profile="${12:-periodic_collocation}"
+dual_warm_start="${13:-auto}"
+target_refinement="${14:-auto}"
 workspace="${GITHUB_WORKSPACE:?GITHUB_WORKSPACE is required}"
 case_dir="${workspace}/${case_root}/${case_slug}-${mechanics}"
 result="$case_dir/result.json"
@@ -30,9 +32,21 @@ if ! [[ "$collocation_degree" =~ ^[2-9]$ ]]; then
   exit 2
 fi
 case "$ipopt_profile" in
-  historical|periodic_collocation|periodic-collocation|scientific_radau5|scientific-radau5|acados_like|acados-like) ;;
+  historical|periodic_collocation|periodic-collocation|scientific_radau4|scientific-radau4|scientific_radau5|scientific-radau5|scientific_radau6|scientific-radau6|acados_like|acados-like) ;;
   *) echo "IPOPT_PROFILE is not supported: '$ipopt_profile'." >&2; exit 2 ;;
 esac
+case "$dual_warm_start" in
+  auto|off|constraints|bounds|all) ;;
+  *) echo "DUAL_WARM_START must be auto, off, constraints, bounds, or all; got '$dual_warm_start'." >&2; exit 2 ;;
+esac
+case "$target_refinement" in
+  auto|true|false) ;;
+  *) echo "TARGET_REFINEMENT must be auto, true, or false; got '$target_refinement'." >&2; exit 2 ;;
+esac
+
+if [[ "$target_refinement" == "true" ]]; then
+  initialization_options=(--optional-nlp-periodic-ipopt-hot-start)
+fi
 
 mkdir -p "$case_dir"
 # CasADi emits fixed filenames such as nlp.c/nlp.so in the current directory.
@@ -49,14 +63,20 @@ case "$graph_mode" in
 esac
 
 if [[ "$solver" == "ipopt" ]]; then
+  if [[ "$dual_warm_start" == "auto" ]]; then
+    dual_warm_start="bounds"
+  fi
   solver_options+=(
     --ipopt-max-iter "$BENCHMARK_MAX_ITER"
-    --ipopt-dual-warm-start-mode bounds
+    --ipopt-dual-warm-start-mode "$dual_warm_start"
   )
   if [[ "$compile_mode" == "true" ]]; then
     solver_options+=(--ipopt-c-compile)
   fi
 elif [[ "$solver" == "fatrop" ]]; then
+  if [[ "$dual_warm_start" == "auto" ]]; then
+    dual_warm_start="off"
+  fi
   case "$fatrop_state_scaling" in
     none|full) ;;
[…]
