### 2026-08-02 4729cfa
Add dense Radau trajectory audit
