+++ examples/fes_multibody/cycling/cycling_pulse_width_mhe_acados_periodic.py
@@ -8164,6 +8164,182 @@ def high_accuracy_integrator_map_diagnostics(
     return rows
 
 
+def high_accuracy_trace_rollout_diagnostics(
+    nmpc,
+    state_traces: dict[str, np.ndarray],
+    control_traces: dict[str, np.ndarray],
+    *,
+    cycle_count: int,
+    capacity_scales: dict[str, float],
+    objective_weight: float = 10_000.0,
+) -> dict:
+    """Reintegrate an exported RHO prefix with one common DOP853 reference.
+
+    Unlike a collocation-defect check, this rollout never resets the reference
+    state at an optimized shooting node.  It therefore scores the exact same
+    PW sequence independently of the Radau degree that produced it.  Time is
+    reset at every RHO because the periodic-node OCP represents one crank cycle
+    per window while the Ding states remain continuous across windows.
+    """
+
+    from scipy.integrate import solve_ivp
+
+    if cycle_count < 1:
+        raise ValueError("cycle_count must be strictly positive.")
+    nlp = nmpc.nlp[0]
+    intervals_per_cycle = int(nmpc.cycle_len)
+    expected_intervals = cycle_count * intervals_per_cycle
+
+    def stack_traces(traces, variables, expected_nodes, label):
+        values = np.zeros((variables.shape, expected_nodes))
+        for key in variables.keys():
+            if key not in traces:
+                raise ValueError(f"Missing {label} trace '{key}'.")
+            trace = np.atleast_2d(np.asarray(traces[key], dtype=float))
+            if trace.shape[1] != expected_nodes:
+                raise ValueError(
+                    f"{label.capitalize()} trace '{key}' has {trace.shape[1]} "
+                    f"nodes, expected {expected_nodes}."
+                )
+            values[variables[key].index, :] = trace
+        if not np.all(np.isfinite(values)):
+            raise ValueError(f"The exported {label} traces are not finite.")
+        return values
+
+    states = stack_traces(
+        state_traces, nlp.states, expected_intervals + 1, "state"
+    )
+    controls = stack_traces(
+        control_traces, nlp.controls, expected_intervals, "control"
+    )
+    dt = float(nmpc.cycle_duration) / intervals_per_cycle
+
+    capacity_rows = []
+    for key in sorted(k for k in nlp.states.keys() if k.startswith("A_")):
+        scale = capacity_scales.get(key)
+        if scale is None or not np.isfinite(float(scale)) or float(scale) <= 0:
+            raise ValueError(f"Missing positive fatigue capacity scale for '{key}'.")
+        indexes = np.asarray(nlp.states[key].index).reshape(-1)
+        if indexes.size != 1:
+            raise ValueError(f"Fatigue capacity state '{key}' must be scalar.")
+        capacity_rows.append((key, int(indexes[0]), float(scale)))
+
+    n_states = states.shape[0]
+    n_capacities = len(capacity_rows)
+    augmented = np.concatenate(
+        (states[:, 0], np.zeros(2 * n_capacities, dtype=float))
+    )
+    maximum_absolute_endpoint_error = 0.0
+    maximum_scaled_endpoint_error = 0.0
+    maximum_absolute_by_state = {key: 0.0 for key in nlp.states.keys()}
+    reference_evaluations = 0
+    state_scales = np.maximum(
+        np.maximum(np.ptp(states, axis=1), np.max(np.abs(states), axis=1)), 1.0
+    )
+    # An absolute crank angle grows by 2*pi every RHO; scaling it by its final
+    # magnitude would make the same local drift look artificially smaller on
+    # longer prefixes.
+    for position_key in _phase_one_state_keys(nlp)["q"]:
+        state_scales[
[…]
