### 2026-08-02 61e60ea
Forward dense audit through benchmark CLI
