### 2026-08-02 96da425
Handle collocation nodes in trajectory audit
