### 2026-08-02 3f0e8c1
Enable FATROP full benchmark after stage fix
