+++ docs/cycling_solver_benchmark/README.md
@@ -23,7 +23,7 @@ La CI ne dépend pas d'un nom de branche flottant. Elle clone le fork
 effectue les checkouts par SHA complet :
 
 - intégration multi-solveurs active :
-  `dad96b90d47c36126c1e97ec35f27c499abf4b12`;
+  `4179bf076b724fe6c4702739b3462e29ae4adef4`;
 - intégration Alpaqa archivée :
   `d84e7e43534360fc048e0be26a3bd69a2abc2d77`;
 - écran MadNLP/MUMPS historique, conservé uniquement pour diagnostic :
@@ -315,20 +315,22 @@ le **même** RHO.
 | Besoin | Solveur recommandé | État actuel |
 |---|---|---|
 | Chaîne robuste de 100 RHO | IPOPT/MUMPS reduced, SX, compilation persistante | `100/100`, environ `1.0 s` chaud sur Radau 3 |
-| Collocation du calcium raffinée | MadNLP/MUMPS reduced, Radau 5 | `5/5`, environ `1.99 s` chaud; endurance à recertifier |
-| Contrôle indépendant de l'optimum reduced | FATROP/collocation reduced | `100/100`, mais plus lent qu'IPOPT reduced |
+| Collocation du calcium raffinée | MadNLP/MUMPS reduced, Radau 6 | `5/5`; meilleure fidélité DOP853, endurance à recertifier |
+| Contrôle indépendant de l'optimum | FATROP/collocation full et reduced | full corrigé localement `1/1`; gate Linux requis |
 | Cible sous-seconde | ACADOS IRK | Environ `0.4–0.6 s`, mais préfixe physique limité à 13 RHO |
 
 MUMPS est le backend portable retenu pour IPOPT et MadNLP. PARDISO n'a pas
-apporté de gain à MadNLP et reste archivé. FATROP full reste bloqué par
-l'identification de structure des gaps. Alpaqa n'est pas fonctionnel sur cette
+apporté de gain à MadNLP et reste archivé. L'échec FATROP full venait du
+rangement global des contraintes multi-thread de collocation dans Bioptim,
+pas de la formulation : le correctif `4179bf07` les redistribue par stage et
+le vrai OCP full passe localement `1/1`. Alpaqa n'est pas fonctionnel sur cette
 formulation.
 
 Le meilleur résultat chaud brut n'est pas nécessairement la meilleure
 méthode. ACADOS est actuellement le plus rapide par fenêtre convergée, mais
-IPOPT reduced est le choix robuste pour une endurance complète. MadNLP devient
-le meilleur candidat pour la cible scientifique raffinée, car Radau 5 y a
-convergé là où IPOPT a stagné sur un RHO pourtant primalement faisable.
+IPOPT reduced est le choix robuste pour une endurance complète.
+MadNLP/MUMPS reduced Radau 6 devient le meilleur candidat provisoire pour la
+cible scientifique raffinée; le palier 30 doit encore confirmer sa robustesse.
 
 ## 4. Protocole de certification
 
@@ -398,9 +400,10 @@ devraient décrire la méthode, par exemple `legacy-radau3`,
 Les profils CLI `scientific-radau4`, `scientific-radau5` et
 `scientific-radau6` sont des contrats verrouillés : `periodic_node`, couple
 constant, SX, collocation Radau au degré annoncé et contraintes initiales
-actives. Une surcharge contradictoire est refusée. Seul Radau 5 porte le statut
-`candidate`; Radau 4 et 6 sont des diagnostics de raffinement. Le nom du profil
-ne remplace pas la certification de la fatigue, de l'AUC et des bornes internes.
+actives. Une surcharge contradictoire est refusée. Les noms historiques des
[…]
