+++ docs/cycling_solver_benchmark/development_history.md
@@ -44,8 +44,9 @@ Ils constituent l’historique antérieur à la campagne séquentielle stricte 5
 - MadNLP utilise exclusivement MUMPS, transmis à libMad sous le nom typé
   exact `MumpsSolver`. La CI échoue si libMad signale une option inconnue.
 - PARDISO et RK4 ne font pas partie de la campagne active. FATROP/collocation
-  reduced est certifié sur les paliers 5, 30 puis 100 RHO. FATROP full reste
-  diagnostique tant que l’interface n’accepte pas sa structure de collocation.
+  reduced est certifié sur les paliers 5, 30 puis 100 RHO. Le rangement
+  stage-wise de FATROP full est corrigé dans Bioptim `4179bf07`; le smoke
+  local full passe `1/1` et attend son gate Linux.
 - MadNLP full reste interprété. MadNLP reduced dispose maintenant d’un chemin
   C expérimental; la CI doit certifier un seul build, sa réutilisation et
   l’équivalence numérique avant qu’un gain de performance soit revendiqué.
@@ -70,11 +71,11 @@ provenance humaine.
 
 | Composant | Version Bioptim réellement utilisée |
 |---|---|
-| Construction et certification des seeds | `dad96b90d47c36126c1e97ec35f27c499abf4b12` |
-| IPOPT full/reduced | `dad96b90d47c36126c1e97ec35f27c499abf4b12` |
-| MadNLP/MUMPS full/reduced | `dad96b90d47c36126c1e97ec35f27c499abf4b12` |
-| FATROP/collocation full/reduced | `dad96b90d47c36126c1e97ec35f27c499abf4b12` |
-| ACADOS full/reduced et variantes | `dad96b90d47c36126c1e97ec35f27c499abf4b12` |
+| Construction et certification des seeds | `4179bf076b724fe6c4702739b3462e29ae4adef4` |
+| IPOPT full/reduced | `4179bf076b724fe6c4702739b3462e29ae4adef4` |
+| MadNLP/MUMPS full/reduced | `4179bf076b724fe6c4702739b3462e29ae4adef4` |
+| FATROP/collocation full/reduced | `4179bf076b724fe6c4702739b3462e29ae4adef4` |
+| ACADOS full/reduced et variantes | `4179bf076b724fe6c4702739b3462e29ae4adef4` |
 
 Ce commit appartient à la branche dédiée
 `codex/cocofest-acados-v055-exploration`. Il part exactement de
@@ -4222,3 +4223,57 @@ où `n` est le nombre de cycles. Cette réévaluation emploie les tolérances
 conserve aussi l'erreur maximale, absolue et normalisée, entre ce rollout
 continu et les nœuds optimisés. Le test unitaire analytique vérifie à la fois
 la propagation d'état et les deux quadratures.
+
+## 22. Rollout DOP853 et correction FATROP full (2 août 2026)
+
+Le [run 30754413003](https://github.com/mickaelbegon/cocofest/actions/runs/30754413003)
+est le premier gate qui réintègre réellement les trajectoires scientifiques
+exportées. Une erreur d'extraction des états de collocation, détectée d'abord
+en local, a été corrigée : pour `N=30`, Radau 5 fournit `181` colonnes, soit un
+nœud initial puis six colonnes par intervalle. L'audit infère maintenant ce
+stride (`5`, `6` ou `7` pour R4, R5 ou R6) et compare DOP853 uniquement aux
+nœuds de tir.
+
+| Solveur | Formulation | Degré | Préfixe | Temps solveur | Coût DOP853 | Drift normalisé maximal |
+|---|---|---:|---:|---:|---:|---:|
[…]

[…]
