### 2026-08-02 638e295
Add ACADOS inter-node cadence guard
