+++ docs/cycling_solver_benchmark/README.md
@@ -310,6 +310,18 @@ par défaut. Une solution non convergée ne doit jamais alimenter le RHO suivant
 Le retry doit repartir du dernier checkpoint certifié et résoudre à nouveau
 le **même** RHO.
 
+L'audit du run `30754413003` a aussi séparé une réussite aux nœuds d'une
+violation cachée entre nœuds : la vitesse moyenne d'intervalle dépasse la
+borne rapide de `0.376–0.403 rad/s`, alors que la vitesse nodale ne la dépasse
+presque pas. La nouvelle expérience ACADOS conserve la borne physique et
+l'audit à `-2*pi +/- 3 rad/s`, mais resserre uniquement la borne interne du
+côté rapide à `-2*pi-2.60`, puis `-2*pi-2.55 rad/s`. Le côté lent reste à
+`-2*pi+3 rad/s`; une garde symétrique réduirait inutilement la faisabilité.
+Ces marges sont des candidats de CI, pas encore des réglages certifiés. Si
+elles échouent, l'étape rigoureuse suivante est un maillage à 60 nœuds avec
+30 PW maintenues par paires, plutôt qu'une augmentation des sous-pas IRK qui
+n'ajoute aucune contrainte aux stages internes.
+
 ### 3.5 Choisir le solveur selon le niveau de garantie
 
 | Besoin | Solveur recommandé | État actuel |

[…]
