### 2026-08-02 506ed61
Cover full ACADOS cadence guard bounds
