### 2026-08-02 27d6aff
Keep FATROP full benchmark interpreted
