+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -1952,7 +1952,12 @@ jobs:
 
       - name: Run FATROP collocation full
         if: matrix.solver == 'fatrop'
-        run: bash .github/scripts/run_cycling_benchmark_case.sh fatrop-collocation fatrop full fatrop collocation benchmark-results "$BENCHMARK_CYCLES" "${{ inputs.compile_nlp_evaluators }}" sx none
+        # The full SX graph generates a roughly 200 MB monolithic C file.
+        # Clang remained active for more than 40 minutes on macOS and two
+        # hosted Linux runners were reclaimed during GCC compilation. The
+        # interpreted full solve is therefore the reproducible benchmark;
+        # reduced still measures persistent C-library reuse.
+        run: bash .github/scripts/run_cycling_benchmark_case.sh fatrop-collocation fatrop full fatrop collocation benchmark-results "$BENCHMARK_CYCLES" false sx none
       - name: Checkpoint FATROP collocation full
         if: ${{ always() && matrix.solver == 'fatrop' }}
         continue-on-error: true
@@ -2055,6 +2060,9 @@ jobs:
               case_requires_compile=false
             fi
             if [[ "$BENCHMARK_SOLVER" == "madnlp" && "$result" == *"madnlp-mumps-full/"* ]]; then
+              case_requires_compile=false
+            fi
+            if [[ "$BENCHMARK_SOLVER" == "fatrop" && "$result" == *"fatrop-collocation-full/"* ]]; then
               case_requires_compile=false
             fi
             # The scientific Radau audit times interpreted evaluators in both
