### 2026-08-02 1db1609
Fix ACADOS cadence guard warm starts
