+++ tests/shard1/test_periodic_pulse_width.py
@@ -180,6 +180,10 @@ def test_full_cadence_constraint_covers_collocation_stages_and_terminal(
         model,
         enforce_start_constraints=False,
         enforce_physical_crank_velocity_bounds=True,
+        physical_crank_velocity_target=-2.0 * np.pi,
+        physical_crank_velocity_margin=3.0,
+        physical_crank_velocity_fast_margin=2.55,
+        physical_crank_velocity_slow_margin=3.0,
     )
 
     assert len(captured) == 2
@@ -187,8 +191,12 @@ def test_full_cadence_constraint_covers_collocation_stages_and_terminal(
         mhe_example.physical_crank_velocity_all_collocation_points_constraint
     )
     assert captured[0][1]["node"] == Node.ALL_SHOOTING
+    assert captured[0][1]["min_bound"] == pytest.approx(-2.0 * np.pi - 2.55)
+    assert captured[0][1]["max_bound"] == pytest.approx(-2.0 * np.pi + 3.0)
     assert captured[1][0] is mhe_example.physical_crank_velocity_constraint
     assert captured[1][1]["node"] == Node.END
+    assert captured[1][1]["min_bound"] == pytest.approx(-2.0 * np.pi - 2.55)
+    assert captured[1][1]["max_bound"] == pytest.approx(-2.0 * np.pi + 3.0)
 
 
 def test_full_transfer_contact_projection_preserves_bound_crank_states():
@@ -3627,6 +3635,81 @@ def test_control_homotopy_does_not_restart_a_linesearch_failure(monkeypatch):
     assert reset_calls == [1]
 
 
+def test_control_homotopy_tries_a_finite_radius_after_fixed_qp_failure(
+    monkeypatch,
+):
+    class FakeAcadosSolver:
+        def reset(self, reset_qp_solver_mem):
+            assert reset_qp_solver_mem == 1
+
+    class FakeSolver:
+        nlp_solver_max_iter = 50
+
+        def set_convergence_tolerance(self, value):
+            self.tolerance = value
+
+    bounds = SimpleNamespace(min=np.array([[0.1, 0.1]]), max=np.array([[0.6, 0.6]]))
+    nmpc = SimpleNamespace(
+        ocp_solver=SimpleNamespace(ocp_solver=FakeAcadosSolver()),
+        nlp=[
+            SimpleNamespace(
+                u_init={
+                    "last_pulse_width_Biceps": SimpleNamespace(
+                        init=np.array([[0.2, 0.4]])
+                    )
+                },
+                u_bounds={"last_pulse_width_Biceps": bounds},
+            )
+        ],
+    )
+    solutions = iter(
+        [
+            SimpleNamespace(
+                status=4,
+                residuals=np.array([150.0, 0.05, 0.0, 0.0]),
+                solver_time_to_optimize=1.0,
+                real_time_to_optimize=1.1,
+            ),
+            SimpleNamespace(
+                status=0,
+                residuals=np.zeros(4),
+                solver_time_to_optimize=2.0,
+                real_time_to_optimize=2.1,
+            ),
+        ]
+    )
+    monkeypatch.setattr(
+        periodic_example,
+        "snapshot_acados_diagnostics",
+        lambda result: {"residuals": result.residuals},
+    )
+    applied = []
+    monkeypatch.setattr(
+        periodic_example,
+        "apply_solution_directly_to_periodic_nmpc_initial_guess",
+        lambda _nmpc, result: applied.append(result.status),
+    )
+
+    summaries = periodic_example.run_acados_control_homotopy(
+        nmpc,
+        FakeSolver(),
[…]
