### 2026-08-02 c89ad41
Retry ACADOS cadence guard on the same RHO
