### 2026-08-02 3857ddf
Give ACADOS two bounded same-RHO retries
