### 2026-08-02 d0e8ff6
Document ACADOS 80-RHO convergence limit
