### 2026-08-02 15a3609
Add paired ACADOS RHO recovery screen
