+++ .github/workflows/cycling_solver_benchmark_linux.yml
@@ -10,7 +10,7 @@ on:
         default: ubuntu-24.04
         type: string
       cycles:
-        description: Number of RHO solves, "screen", "acados", "acados_guard", "acados_homotopy", or "full_horizon"
+        description: Number of RHO solves, "screen", "acados", "acados_guard", "acados_recovery", "acados_homotopy", or "full_horizon"
         required: true
         default: "100"
         type: string
@@ -583,6 +583,7 @@ jobs:
           ACADOS_OPTION_RHOS: ${{ inputs.acados_option_rhos }}
           ACADOS_HOMOTOPY_ONLY: ${{ inputs.cycles == 'acados_homotopy' }}
           ACADOS_CADENCE_GUARD_ONLY: ${{ inputs.cycles == 'acados_guard' }}
+          ACADOS_RECOVERY_ONLY: ${{ inputs.cycles == 'acados_recovery' }}
           BENCHMARK_ASSISTANCE: ${{ inputs.crank_assistance_nm }}
           BENCHMARK_Q_SLACK: ${{ inputs.terminal_wheel_q_slack }}
           COCOFEST_BENCHMARK_COMMIT: ${{ github.sha }}
@@ -646,7 +647,7 @@ jobs:
                 --common-initial-solution-output
                 "$GITHUB_WORKSPACE/acados-smoke-results/reference-reduced-feasible-seed.npz"
               )
-            elif [[ "$variant" == sqp-irk-two-stage-fast-guard-* ]]; then
+            elif [[ "$variant" == *fast-guard* ]]; then
               common_seed="$GITHUB_WORKSPACE/acados-smoke-results/reference-${mechanics}-feasible-seed.npz"
               if [[ ! -f "$common_seed" ]]; then
                 echo "The cadence-guard case requires its preceding ACADOS reference seed." >&2
@@ -781,7 +782,7 @@ jobs:
                 --acados-wheel-qdot-regularization-weight 1
               )
             fi
-            if [[ "$variant" == "sqp-irk-two-stage-fast-guard-2p6" ]]; then
+            if [[ "$variant" == *fast-guard-2p6* ]]; then
               # Preserve the physical audit at +/-3 rad/s, but keep the
               # optimized nodes 0.4 rad/s inside the fast-side bound. The
               # slow-side bound remains unchanged because the observed hidden
@@ -797,7 +798,7 @@ jobs:
                 --acados-wheel-qdot-fast-bound-margin 2.55
               )
             fi
-            if [[ "$mechanics" == "full" && "$variant" == sqp-irk-two-stage-fast-guard-* ]]; then
+            if [[ "$mechanics" == "full" && "$variant" == *fast-guard* ]]; then
               # Give MAXITER one genuine second chance on the same RHO from
               # the best stored SQP iterate. If that retry fails, stop before
               # Bioptim can advance the window with an uncertified solution.
@@ -809,6 +810,9 @@ jobs:
                 --acados-maxiter-retry-iterations 20
                 --acados-maxiter-retry-feasibility-tolerance 0.0025
                 --max-consecutive-failing 1
+                --receding-horizon-solution-output
+                "$case_dir/validated-rho-trajectory.npz"
+                --allow-partial-receding-horizon-solution-output
               )
             fi
             if [[ "$variant" == "sqp-byrd-terminal-homotopy-cadence-reg-1-irk" ]]; then
@@ -904,7 +908,15 @@ jobs:
                 --acados-transfer-active-set-threshold 1e-6
               )
             fi
-            if [[ "$variant" == *"-phase-one" ]]; then
[…]
