### 2026-08-02 99aa54e
Audit ACADOS phase-one recovery trajectories
