### 2026-08-02 a43acb2
Truncate failed ACADOS windows before dense audit
