### 2026-08-02 c7fc9b5
Document ACADOS mechanical phase-one recovery
