### 2026-08-03 856bffd
Add endurance solver benchmark campaigns
