### 2026-08-18 #10669 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] Qu et al. reference, be sure to ensure {Capitalization} of Kalman
- [ ] both Michaud et al. references: similar but for programming language names
- [ ] Summary: hyphenate "long-duration"
- [ ] Statement of need: (optional) "how to choose nonlinear dynamics" feels awkward to me, perhaps "how to choose nonlinear dynamics parameters" would be better? But if this phrasing is common usage in the subfield, feel free to leave it.
- [ ] whenever used as an adjective (e.g. modifying "parameters"), "pulse-train" should be hyphenated ([compound adjective](https://www.grammarly.com/blog/parts-of-speech/compound-adjectives/))
- [ ] similarly for "optimal-control" (e.g. optimal-control computations, optimal-control-driven)
- [ ] under State of the Field: it seems like this was intended to be a bullet-point list, check the formatting of that
- [ ] second paragraph of Software design: add "a" before "conventional"
- [ ] same paragraph: any subscripts that aren't themselves variables (e.g. "max") should be in `\mathrm{}` mode rather than italicized
