### 2026-08-19 #10669 @Kev1CO
Hi @rkurchin,

I went throught the remaning points.
- _Qu et al. reference, be sure to ensure {Capitalization} of Kalman_ (Done)
- _both Michaud et al. references: similar but for programming language names_ (Done)
- _Summary: hyphenate "long-duration"_ (Done)
- _Statement of need: (optional) "how to choose nonlinear dynamics" feels awkward to me, perhaps "how to choose nonlinear dynamics parameters" would be better? But if this phrasing is common usage in the subfield, feel free to leave it._ (Common usage)
- _whenever used as an adjective (e.g. modifying "parameters"), "pulse-train" should be hyphenated_ (Done)
- _similarly for "optimal-control" (e.g. optimal-control computations, optimal-control-driven)_ (Done)
- _under State of the Field: it seems like this was intended to be a bullet-point list, check the formatting of that_ (Done)
- _second paragraph of Software design: add "a" before "conventional"_ (Done)
- _same paragraph: any subscripts that aren't themselves variables (e.g. "max") should be in \mathrm{} mode rather than italicized_ (Done)

Tell me if I need to do anything else.
