### 2026-08-20 #10669 @Kev1CO
Thank you, everyone, for your help throughout the review process.

@Aravind-Sundararajan and @siboles, thank you for the time and effort you put into this review. Your comments have significantly improved the accessibility of the package and made it easier for everyone to get started with it.
@borgesaugusto, thank you for editing the submission and for keeping the process moving smoothly from start to finish.
Thank you, @rkurchin, for your help with the final step.

Best of luck with your respective work, and please feel free to reach out if I can be of any help.

Sincerely,

Kevin CO and all the co-authors
