### 2026-08-20 #10669 @rkurchin
Congrats to @Kev1CO and thanks to handling editor @borgesaugusto and reviewers @Aravind-Sundararajan and @siboles. JOSS wouldn't exist without your volunteer efforts!
